Get BOM Agriculture Bulletin

Fetch the BOM agricultural bulletin information and return a tidy data frame

get_bulletin(state = NULL)

Arguments

state
Australian state or territory as postal code, see details for instruction.

Value

Data frame of a Australia BOM agricultural bulletin information

obs-time-utc

Observation time (Time in UTC)

time-zone

Time zone for observation

site

Unique BOM identifier for each station

name

BOM station name

r

Rain to 9am (millimetres). Trace will be reported as 0.01

tn

Minimum temperature (Celsius)

tx

Maximum temperature (Celsius)

twd

Wetbulb depression (Celsius)

ev

Evaporation (millimetres)

tg

Terrestrial minimum temperature (Celsius)

sn

Sunshine (Hours)

t5

5cm soil temperature (Celsius)

t10

10cm soil temperature (Celsius)

t20

20cm soil temperature (Celsius)

t50

50cm soil temperature (Celsius)

t1m

1m soil temperature (Celsius)

wr

Wind run (kilometres)

state

State name (postal code abbreviation)

lat

Latitude (decimal degrees)

lon

Longitude (decimal degrees)

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns bulletin for all states

References

Australian Bureau of Meteorology (BOM) Weather Data Services Agriculture Bulletins http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml

Australian Bureau of Meteorology (BOM) Weather Data Services Observation of Rainfall http://www.bom.gov.au/climate/how/observations/rain-measure.shtml

Examples

not_run({
  ag_bulletin <- get_bulletin(state = "QLD")
})