Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-07-11 15:47:46 2021-07-11 15:47:46      10:00 2021-07-11 15:47:46
#>  2: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#>  3: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#>  4: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#>  5: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#>  6: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#>  7: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#>  8: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#>  9: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 10: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 11: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 12: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 13: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 14: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 15: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 16: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 17: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 18: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 19: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 20: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 21: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 22: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 23: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 24: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 25: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 26: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 27: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 28: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 29: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#> 30: 2021-07-11 17:00:00 2021-07-11 17:00:00      10:00 2021-07-11 17:00:00
#> 31: 2021-07-12 00:00:00 2021-07-12 00:00:00      10:00 2021-07-12 00:00:00
#> 32: 2021-07-13 00:00:00 2021-07-13 00:00:00      10:00 2021-07-13 00:00:00
#> 33: 2021-07-14 00:00:00 2021-07-14 00:00:00      10:00 2021-07-14 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-07-11 15:47:46
#>  2: 2021-07-11 17:00:00
#>  3: 2021-07-12 00:00:00
#>  4: 2021-07-13 00:00:00
#>  5: 2021-07-14 00:00:00
#>  6: 2021-07-11 17:00:00
#>  7: 2021-07-12 00:00:00
#>  8: 2021-07-13 00:00:00
#>  9: 2021-07-14 00:00:00
#> 10: 2021-07-11 17:00:00
#> 11: 2021-07-12 00:00:00
#> 12: 2021-07-13 00:00:00
#> 13: 2021-07-14 00:00:00
#> 14: 2021-07-11 17:00:00
#> 15: 2021-07-12 00:00:00
#> 16: 2021-07-13 00:00:00
#> 17: 2021-07-14 00:00:00
#> 18: 2021-07-11 17:00:00
#> 19: 2021-07-12 00:00:00
#> 20: 2021-07-13 00:00:00
#> 21: 2021-07-14 00:00:00
#> 22: 2021-07-11 17:00:00
#> 23: 2021-07-12 00:00:00
#> 24: 2021-07-13 00:00:00
#> 25: 2021-07-14 00:00:00
#> 26: 2021-07-11 17:00:00
#> 27: 2021-07-12 00:00:00
#> 28: 2021-07-13 00:00:00
#> 29: 2021-07-14 00:00:00
#> 30: 2021-07-11 17:00:00
#> 31: 2021-07-12 00:00:00
#> 32: 2021-07-13 00:00:00
#> 33: 2021-07-14 00:00:00
#>            end_time_utc
#>                                                                                                          forecast_seas
#>  1:                                                                                                               <NA>
#>  2:                                                                                                     Below 1 metre.
#>  3:                                                    Around 1 metre, increasing to 1 to 2 metres during the morning.
#>  4:                                                      1 to 1.5 metres, decreasing below 1 metre during the morning.
#>  5:                                               Below 1 metre, increasing to 1.5 to 2.5 metres during the afternoon.
#>  6:                                                                                                  Below 0.5 metres.
#>  7:                                                       Around 1 metre, increasing to 1 to 1.5 metres around midday.
#>  8:                                                      1 to 1.5 metres, decreasing below 1 metre during the morning.
#>  9:                                             1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the afternoon.
#> 10:                                                                                                     Below 1 metre.
#> 11:                                                 Below 1 metre, increasing to 1 to 1.5 metres during the afternoon.
#> 12:                                                         1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 13:                                                Around 1 metre, increasing to 1.5 to 2.5 metres during the morning.
#> 14:                                                                                                     Below 1 metre.
#> 15:                                       Around 1 metre, increasing to 1 to 1.5 metres offshore later in the evening.
#> 16:                                                                                                   1 to 1.5 metres.
#> 17:                                                    Around 1 metre, increasing to 1 to 2 metres during the morning.
#> 18:                                                                                                    Around 1 metre.
#> 19: Below 1 metre, decreasing below 0.5 metres during the morning, then increasing to around 1 metre by early evening.
#> 20:                                                                                                   1 to 1.5 metres.
#> 21:                                                             Below 1 metre, increasing to 1 to 1.5 metres offshore.
#> 22:                                                                                                     1 to 2 metres.
#> 23:                                                                                                     Below 1 metre.
#> 24:                                                                                                     Below 1 metre.
#> 25:                  Around 1 metre, increasing to 1 to 1.5 metres offshore north of Crowdy Head during the afternoon.
#> 26:                                                       1 to 1.5 metres, decreasing to 1 metre later in the evening.
#> 27:                                                      Around 1 metre, decreasing below 0.5 metres by early evening.
#> 28:                                                                                                     Below 1 metre.
#> 29:                                       Around 1 metre, increasing to 1 to 1.5 metres offshore during the afternoon.
#> 30:                                                               2 metres, decreasing to 1.5 metres by early evening.
#> 31:                                                           1 to 1.5 metres, decreasing below 1 metre around midday.
#> 32:                                                                                                     Below 1 metre.
#> 33:                                       Around 1 metre, increasing to 1 to 1.5 metres offshore during the afternoon.
#>                                                                                                          forecast_seas
#>                                                                                         forecast_weather
#>  1:                                                                                                 <NA>
#>  2:                                                                                        Mostly clear.
#>  3:                                                                                       Partly cloudy.
#>  4:                                                                                        Mostly sunny.
#>  5:                                                                Partly cloudy. 50% chance of showers.
#>  6:                                                                                        Mostly clear.
#>  7:                                                                                       Partly cloudy.
#>  8:  The chance of a thunderstorm offshore north of Moruya Heads in the morning. Mostly sunny afternoon.
#>  9:                                                                Partly cloudy. 60% chance of showers.
#> 10:                                                                                       Partly cloudy.
#> 11:                                                                Partly cloudy. 50% chance of showers.
#> 12:                        The chance of a thunderstorm offshore in the morning. Mostly sunny afternoon.
#> 13:                                                                Partly cloudy. 60% chance of showers.
#> 14:                                                                                       Partly cloudy.
#> 15:                                                                Partly cloudy. 60% chance of showers.
#> 16:                        The chance of a thunderstorm offshore in the morning. Mostly sunny afternoon.
#> 17:                                                                Partly cloudy. 60% chance of showers.
#> 18:                                                                                       Partly cloudy.
#> 19:                                                                Partly cloudy. 60% chance of showers.
#> 20: Partly cloudy. 50% chance of showers offshore. The chance of a thunderstorm offshore in the morning.
#> 21:                                                                Partly cloudy. 50% chance of showers.
#> 22:                                                                                       Partly cloudy.
#> 23:                                                                Partly cloudy. 70% chance of showers.
#> 24: Partly cloudy. 50% chance of showers offshore. The chance of a thunderstorm offshore in the morning.
#> 25:                                                                                       Partly cloudy.
#> 26:                                                                Partly cloudy. 50% chance of showers.
#> 27:                                                                Partly cloudy. 70% chance of showers.
#> 28:                        The chance of a thunderstorm offshore in the morning. Mostly sunny afternoon.
#> 29:                                                                                       Partly cloudy.
#> 30:                                                                                       Partly cloudy.
#> 31:                                                                Partly cloudy. 70% chance of showers.
#> 32:                                                                                       Partly cloudy.
#> 33:                                                                                       Partly cloudy.
#>                                                                                         forecast_weather
#>                                                                                                                                                        forecast_winds
#>  1:                                                                                                                                                              <NA>
#>  2:                                                                                    Northeasterly about 10 knots increasing to 10 to 15 knots in the late evening.
#>  3:                                                 Northerly 15 to 20 knots, reaching up to 25 knots offshore south of Green Cape in the late morning and afternoon.
#>  4:                                                                                     Northwesterly 10 to 15 knots becoming variable about 10 knots in the morning.
#>  5:                   Northwesterly about 10 knots increasing to 10 to 15 knots during the morning then tending north to northwesterly 20 to 30 knots during the day.
#>  6:                                                                                                                                          Variable about 10 knots.
#>  7:                                                               North to northeasterly 10 to 15 knots tending north to northwesterly 15 to 20 knots during the day.
#>  8:                                                                                                                                     Northwesterly 10 to 15 knots.
#>  9:                                                                                   Northwesterly 15 to 20 knots increasing to 20 to 25 knots during the afternoon.
#> 10:                                                                                                                                          Variable below 10 knots.
#> 11:                                Variable about 10 knots becoming northerly 15 to 20 knots in the middle of the day then tending northwesterly in the late evening.
#> 12:                                                                                  Northwesterly 15 to 20 knots turning westerly 10 to 15 knots during the morning.
#> 13:                                                                                   Northwesterly 15 to 20 knots increasing to 20 to 25 knots during the afternoon.
#> 14:                                                                                                                            South to southeasterly below 10 knots.
#> 15:                                                                                  Variable about 10 knots becoming northerly 15 to 20 knots in the late afternoon.
#> 16:                                                                                                                                     Northwesterly 15 to 20 knots.
#> 17:                                                                                                                                     Northwesterly 20 to 25 knots.
#> 18:                                                                                                                                     Southeasterly about 10 knots.
#> 19: Variable about 10 knots becoming northerly 10 to 15 knots in the early afternoon. Winds reaching up to 20 knots offshore south of Norah Head in the late evening.
#> 20:                                                            North to northwesterly 10 to 15 knots tending west to northwesterly 15 to 20 knots during the morning.
#> 21:                                                                                                                                     Northwesterly 15 to 20 knots.
#> 22:                                                                                                                            South to southeasterly 10 to 15 knots.
#> 23:                  Southeasterly 10 to 15 knots becoming variable about 10 knots in the morning then becoming north to northeasterly 10 to 15 knots in the evening.
#> 24:                                                                                Northerly 10 to 15 knots turning west to northwesterly below 10 knots before dawn.
#> 25:                                                                                                                                     Northwesterly 15 to 20 knots.
#> 26:                                                                 Southerly 10 to 15 knots, reaching up to 20 knots offshore north of Coffs Harbour in the evening.
#> 27:                                                                Southeast to southwesterly 10 to 15 knots becoming variable about 10 knots in the early afternoon.
#> 28:                                                                                                                                     Northwesterly 10 to 15 knots.
#> 29:                                                                                     Northwesterly about 10 knots increasing to 15 to 20 knots during the morning.
#> 30:                                                                                                                                         Southerly 15 to 20 knots.
#> 31:        South to southeasterly 10 to 15 knots, reaching up to 20 knots north of Yamba early in the morning. Winds becoming variable below 10 knots in the evening.
#> 32:                                                                                                                                          Variable about 10 knots.
#> 33:                                                                                                                                     Northwesterly 15 to 20 knots.
#>                                                                                                                                                        forecast_winds
#>                                                                     forecast_swell1
#>  1:                                                                            <NA>
#>  2:                                                         Easterly 2 to 3 metres.
#>  3:        Easterly 1 to 1.5 metres, decreasing to around 1 metre by early evening.
#>  4:                                                   Southeasterly around 1 metre.
#>  5:                                                   Southeasterly around 1 metre.
#>  6:                                                         Easterly 2 to 3 metres.
#>  7:            Easterly 2 metres, decreasing to 1 to 1.5 metres during the morning.
#>  8:                                                        Easterly around 1 metre.
#>  9:                                                        Easterly around 1 metre.
#> 10:                                                     Easterly 2.5 to 3.5 metres.
#> 11:                                                     Easterly 1.5 to 2.5 metres.
#> 12:                                                   Southeasterly around 1 metre.
#> 13:                                                        Easterly around 1 metre.
#> 14:                                                         Easterly 2 to 3 metres.
#> 15:                 Easterly 2 metres, decreasing to 1.5 metres during the morning.
#> 16:                                                            Easterly 1.5 metres.
#> 17:                                                        Easterly around 1 metre.
#> 18:                                                    Southeasterly 3 to 4 metres.
#> 19:             Easterly 1.5 to 2.5 metres, decreasing to 1.5 metres around midday.
#> 20:                                                       Easterly 1.5 to 2 metres.
#> 21:     Easterly around 1 metre, increasing to 1 to 1.5 metres north of Norah Head.
#> 22:                                                    Southeasterly 3 to 4 metres.
#> 23:                                                    Southeasterly 2 to 3 metres.
#> 24:                                                       Easterly 1 to 1.5 metres.
#> 25:      Easterly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 26:                                                         Southeasterly 3 metres.
#> 27:  Southeasterly 2 to 3 metres, decreasing to 1.5 to 2 metres during the morning.
#> 28:           Easterly 1 to 1.5 metres, decreasing to around 1 metre around midday.
#> 29:                                                       Easterly 1 to 1.5 metres.
#> 30:    Southeasterly 2.5 to 4 metres, decreasing to 2 to 3 metres by early evening.
#> 31:       Southeasterly 2 to 3 metres, decreasing to 1.5 to 2 metres around midday.
#> 32:      Southeasterly 1 to 1.5 metres, decreasing to around 1 metre around midday.
#> 33: Southeasterly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#>                                                                     forecast_swell1
#>                                                                                                   forecast_swell2
#>  1:                                                                                                          <NA>
#>  2:                                                                                 Southeasterly around 1 metre.
#>  3:                                                                                 Southeasterly around 1 metre.
#>  4:                                                                                      Easterly around 1 metre.
#>  5:                                                                                      Easterly around 1 metre.
#>  6:                                                                                     Southerly around 1 metre.
#>  7:                                                                                 Southeasterly around 1 metre.
#>  8:                                                                                                          <NA>
#>  9:                                                                                                          <NA>
#> 10:                                                                                    Southerly 1 to 1.5 metres.
#> 11:                                                                        South to southeasterly around 1 metre.
#> 12:                                                                                      Easterly around 1 metre.
#> 13:                                                                                                          <NA>
#> 14:                                          Southerly 1.5 metres, decreasing to around 1 metre by early evening.
#> 15:                                                                                     Southerly around 1 metre.
#> 16:                                                                                                          <NA>
#> 17:                                                                                                          <NA>
#> 18:                                                                                       Easterly 2 to 3 metres.
#> 19:                                                                                     Southerly around 1 metre.
#> 20:                                                                                 Southeasterly around 1 metre.
#> 21:                                                                                                          <NA>
#> 22:                                                                                                          <NA>
#> 23:                                                                                                          <NA>
#> 24:                                                                        South to southeasterly around 1 metre.
#> 25:                                                                                                          <NA>
#> 26: South to southeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Coffs Harbour.
#> 27:              Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Coffs Harbour.
#> 28:                                                                                     Southerly around 1 metre.
#> 29:                                                                                                          <NA>
#> 30:                                                                                  Southeasterly 2 to 3 metres.
#> 31:     South to southeasterly 1 to 1.5 metres inshore, increasing to 1 to 2 metres offshore north of Cape Byron.
#> 32:                 Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Cape Byron.
#> 33:                                                                                                          <NA>
#>                                                                                                   forecast_swell2
#>                                                                                                                                                        forecast_caution
#>  1:                                                                                                                                                                <NA>
#>  2: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  3:                                                                                                                                                                <NA>
#>  4:                                                                                                                                                                <NA>
#>  5:                                                                                                                                                                <NA>
#>  6: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  7:                                                                                                                                                                <NA>
#>  8:                                                                                                                                                                <NA>
#>  9:                                                                                                                                                                <NA>
#> 10: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 11:                                                                                                                                                                <NA>
#> 12:                                                                                                                                                                <NA>
#> 13:                                                                                                                                                                <NA>
#> 14: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 15:                                                                                                                                                                <NA>
#> 16:                                                                                                                                                                <NA>
#> 17:                                                                                                                                                                <NA>
#> 18: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 19:                                                                                                                                                                <NA>
#> 20:                                                                                                                                                                <NA>
#> 21:                                                                                                                                                                <NA>
#> 22: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 23:                                                                                                                                                                <NA>
#> 24:                                                                                                                                                                <NA>
#> 25:                                                                                                                                                                <NA>
#> 26: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 27:               Large and powerful surf conditions in the morning are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 28:                                                                                                                                                                <NA>
#> 29:                                                                                                                                                                <NA>
#> 30: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 31:               Large and powerful surf conditions in the morning are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 32:                                                                                                                                                                <NA>
#> 33:                                                                                                                                                                <NA>
#>                                                                                                                                                        forecast_caution
#>     marine_forecast tropical_system_location forecast_waves
#>  1:            <NA>                       NA             NA
#>  2:            <NA>                       NA             NA
#>  3:            <NA>                       NA             NA
#>  4:            <NA>                       NA             NA
#>  5:            <NA>                       NA             NA
#>  6:            <NA>                       NA             NA
#>  7:            <NA>                       NA             NA
#>  8:            <NA>                       NA             NA
#>  9:            <NA>                       NA             NA
#> 10:            <NA>                       NA             NA
#> 11:            <NA>                       NA             NA
#> 12:            <NA>                       NA             NA
#> 13:            <NA>                       NA             NA
#> 14:            <NA>                       NA             NA
#> 15:            <NA>                       NA             NA
#> 16:            <NA>                       NA             NA
#> 17:            <NA>                       NA             NA
#> 18:            <NA>                       NA             NA
#> 19:            <NA>                       NA             NA
#> 20:            <NA>                       NA             NA
#> 21:            <NA>                       NA             NA
#> 22:            <NA>                       NA             NA
#> 23:            <NA>                       NA             NA
#> 24:            <NA>                       NA             NA
#> 25:            <NA>                       NA             NA
#> 26:            <NA>                       NA             NA
#> 27:            <NA>                       NA             NA
#> 28:            <NA>                       NA             NA
#> 29:            <NA>                       NA             NA
#> 30:            <NA>                       NA             NA
#> 31:            <NA>                       NA             NA
#> 32:            <NA>                       NA             NA
#> 33:            <NA>                       NA             NA
#>     marine_forecast tropical_system_location forecast_waves

# }