Obtain Historical BOM Data

Retrieves daily observations for a given station.

get_historical(stationid = NULL, latlon = NULL, type = c("rain", "min",
  "max", "solar"))

Arguments

stationid
BOM station ID. See Details.

latlon
Length-2 numeric vector of Latitude/Longitude. See Details.

type
Measurement type, either daily "rain", "min" (temp), "max" (temp), or "solar" (exposure). Partial matching is performed.

Value

a complete data.frame of historical observations for the chosen station, with some subset of the following columns

Product_code:	BOM internal code.

Station_number:	BOM station ID.

Year:	Year of observation (YYYY).

Month:	Month of observation (1-12).

Day:	Day of observation (1-31).

Min_temperature:	Minimum daily recorded temperature (degrees C).

Max_temperature:	Maximum daily recorded temperature (degrees C).

Accum_days_min:	Accumulated number of days of minimum temperature.

Accum_days_max:	Accumulated number of days of maximum temperature.

Rainfall:	Daily recorded rainfall in mm.

Period:	Period over which rainfall was measured.

Solar_exposure:	Daily global solar exposure in MJ/m^2.

Temperature data prior to 1910 should be used with extreme caution as many stations, prior to that date, were exposed in non-standard shelters, some of which give readings which are several degrees warmer or cooler than those measured according to post-1910 standards.

Daily maximum temperatures usually occur in the afternoon and daily minimum temperatures overnight or near dawn. Occasionally, however, the lowest temperature in the 24 hours to prior to 9 am can occur around 9 am the previous day if the night was particularly warm.

Either stationid or latlon must be provided, but if both are, then stationid will be used as it is more reliable.

In some cases data is available back to the 1800s, so tens of thousands of daily records will be returned. Other stations will be newer and will return fewer observations.

Examples

# NOT RUN {
get_historical(stationid = "023000", type = "max") ## ~48,000+ daily records
# }
get_historical(latlon = c(-35.2809, 149.1300), type = "min") ## 3,500+ daily records

#> Closest station: 070351 (CANBERRA AIRPORT)

#> Data saved as /var/folders/hk/1xg925vs6fsgfv2q9f9k6fp00000gn/T//Rtmpfb4D5f/IDCJAC0011_070351_1800_Data.csv

#>      Product_code Station_number Year Month Day Min_temperature Accum_days_min
#> 1      IDCJAC0011          70351 2008     1   1              NA             NA
#> 2      IDCJAC0011          70351 2008     1   2              NA             NA
#> 3      IDCJAC0011          70351 2008     1   3              NA             NA
#> 4      IDCJAC0011          70351 2008     1   4              NA             NA
#> 5      IDCJAC0011          70351 2008     1   5              NA             NA
#> 6      IDCJAC0011          70351 2008     1   6              NA             NA
#> 7      IDCJAC0011          70351 2008     1   7              NA             NA
#> 8      IDCJAC0011          70351 2008     1   8              NA             NA
#> 9      IDCJAC0011          70351 2008     1   9              NA             NA
#> 10     IDCJAC0011          70351 2008     1  10              NA             NA
#> 11     IDCJAC0011          70351 2008     1  11              NA             NA
#> 12     IDCJAC0011          70351 2008     1  12              NA             NA
#> 13     IDCJAC0011          70351 2008     1  13              NA             NA
#> 14     IDCJAC0011          70351 2008     1  14              NA             NA
#> 15     IDCJAC0011          70351 2008     1  15              NA             NA
#> 16     IDCJAC0011          70351 2008     1  16              NA             NA
#> 17     IDCJAC0011          70351 2008     1  17              NA             NA
#> 18     IDCJAC0011          70351 2008     1  18              NA             NA
#> 19     IDCJAC0011          70351 2008     1  19              NA             NA
#> 20     IDCJAC0011          70351 2008     1  20              NA             NA
#> 21     IDCJAC0011          70351 2008     1  21              NA             NA
#> 22     IDCJAC0011          70351 2008     1  22              NA             NA
#> 23     IDCJAC0011          70351 2008     1  23              NA             NA
#> 24     IDCJAC0011          70351 2008     1  24              NA             NA
#> 25     IDCJAC0011          70351 2008     1  25              NA             NA
#> 26     IDCJAC0011          70351 2008     1  26              NA             NA
#> 27     IDCJAC0011          70351 2008     1  27              NA             NA
#> 28     IDCJAC0011          70351 2008     1  28              NA             NA
#> 29     IDCJAC0011          70351 2008     1  29              NA             NA
#> 30     IDCJAC0011          70351 2008     1  30              NA             NA
#> 31     IDCJAC0011          70351 2008     1  31              NA             NA
#> 32     IDCJAC0011          70351 2008     2   1              NA             NA
#> 33     IDCJAC0011          70351 2008     2   2              NA             NA
#> 34     IDCJAC0011          70351 2008     2   3              NA             NA
#> 35     IDCJAC0011          70351 2008     2   4              NA             NA
#> 36     IDCJAC0011          70351 2008     2   5              NA             NA
#> 37     IDCJAC0011          70351 2008     2   6              NA             NA
#> 38     IDCJAC0011          70351 2008     2   7              NA             NA
#> 39     IDCJAC0011          70351 2008     2   8              NA             NA
#> 40     IDCJAC0011          70351 2008     2   9              NA             NA
#> 41     IDCJAC0011          70351 2008     2  10              NA             NA
#> 42     IDCJAC0011          70351 2008     2  11              NA             NA
#> 43     IDCJAC0011          70351 2008     2  12              NA             NA
#> 44     IDCJAC0011          70351 2008     2  13              NA             NA
#> 45     IDCJAC0011          70351 2008     2  14              NA             NA
#> 46     IDCJAC0011          70351 2008     2  15              NA             NA
#> 47     IDCJAC0011          70351 2008     2  16              NA             NA
#> 48     IDCJAC0011          70351 2008     2  17              NA             NA
#> 49     IDCJAC0011          70351 2008     2  18              NA             NA
#> 50     IDCJAC0011          70351 2008     2  19              NA             NA
#> 51     IDCJAC0011          70351 2008     2  20              NA             NA
#> 52     IDCJAC0011          70351 2008     2  21              NA             NA
#> 53     IDCJAC0011          70351 2008     2  22              NA             NA
#> 54     IDCJAC0011          70351 2008     2  23              NA             NA
#> 55     IDCJAC0011          70351 2008     2  24              NA             NA
#> 56     IDCJAC0011          70351 2008     2  25              NA             NA
#> 57     IDCJAC0011          70351 2008     2  26              NA             NA
#> 58     IDCJAC0011          70351 2008     2  27              NA             NA
#> 59     IDCJAC0011          70351 2008     2  28              NA             NA
#> 60     IDCJAC0011          70351 2008     2  29              NA             NA
#> 61     IDCJAC0011          70351 2008     3   1              NA             NA
#> 62     IDCJAC0011          70351 2008     3   2              NA             NA
#> 63     IDCJAC0011          70351 2008     3   3              NA             NA
#> 64     IDCJAC0011          70351 2008     3   4              NA             NA
#> 65     IDCJAC0011          70351 2008     3   5              NA             NA
#> 66     IDCJAC0011          70351 2008     3   6              NA             NA
#> 67     IDCJAC0011          70351 2008     3   7              NA             NA
#> 68     IDCJAC0011          70351 2008     3   8              NA             NA
#> 69     IDCJAC0011          70351 2008     3   9              NA             NA
#> 70     IDCJAC0011          70351 2008     3  10              NA             NA
#> 71     IDCJAC0011          70351 2008     3  11              NA             NA
#> 72     IDCJAC0011          70351 2008     3  12              NA             NA
#> 73     IDCJAC0011          70351 2008     3  13              NA             NA
#> 74     IDCJAC0011          70351 2008     3  14              NA             NA
#> 75     IDCJAC0011          70351 2008     3  15              NA             NA
#> 76     IDCJAC0011          70351 2008     3  16              NA             NA
#> 77     IDCJAC0011          70351 2008     3  17              NA             NA
#> 78     IDCJAC0011          70351 2008     3  18              NA             NA
#> 79     IDCJAC0011          70351 2008     3  19              NA             NA
#> 80     IDCJAC0011          70351 2008     3  20              NA             NA
#> 81     IDCJAC0011          70351 2008     3  21              NA             NA
#> 82     IDCJAC0011          70351 2008     3  22              NA             NA
#> 83     IDCJAC0011          70351 2008     3  23              NA             NA
#> 84     IDCJAC0011          70351 2008     3  24              NA             NA
#> 85     IDCJAC0011          70351 2008     3  25              NA             NA
#> 86     IDCJAC0011          70351 2008     3  26              NA             NA
#> 87     IDCJAC0011          70351 2008     3  27              NA             NA
#> 88     IDCJAC0011          70351 2008     3  28              NA             NA
#> 89     IDCJAC0011          70351 2008     3  29              NA             NA
#> 90     IDCJAC0011          70351 2008     3  30              NA             NA
#> 91     IDCJAC0011          70351 2008     3  31              NA             NA
#> 92     IDCJAC0011          70351 2008     4   1              NA             NA
#> 93     IDCJAC0011          70351 2008     4   2              NA             NA
#> 94     IDCJAC0011          70351 2008     4   3              NA             NA
#> 95     IDCJAC0011          70351 2008     4   4              NA             NA
#> 96     IDCJAC0011          70351 2008     4   5              NA             NA
#> 97     IDCJAC0011          70351 2008     4   6              NA             NA
#> 98     IDCJAC0011          70351 2008     4   7              NA             NA
#> 99     IDCJAC0011          70351 2008     4   8              NA             NA
#> 100    IDCJAC0011          70351 2008     4   9              NA             NA
#> 101    IDCJAC0011          70351 2008     4  10              NA             NA
#> 102    IDCJAC0011          70351 2008     4  11              NA             NA
#> 103    IDCJAC0011          70351 2008     4  12              NA             NA
#> 104    IDCJAC0011          70351 2008     4  13              NA             NA
#> 105    IDCJAC0011          70351 2008     4  14              NA             NA
#> 106    IDCJAC0011          70351 2008     4  15              NA             NA
#> 107    IDCJAC0011          70351 2008     4  16              NA             NA
#> 108    IDCJAC0011          70351 2008     4  17              NA             NA
#> 109    IDCJAC0011          70351 2008     4  18              NA             NA
#> 110    IDCJAC0011          70351 2008     4  19              NA             NA
#> 111    IDCJAC0011          70351 2008     4  20              NA             NA
#> 112    IDCJAC0011          70351 2008     4  21              NA             NA
#> 113    IDCJAC0011          70351 2008     4  22              NA             NA
#> 114    IDCJAC0011          70351 2008     4  23              NA             NA
#> 115    IDCJAC0011          70351 2008     4  24              NA             NA
#> 116    IDCJAC0011          70351 2008     4  25              NA             NA
#> 117    IDCJAC0011          70351 2008     4  26              NA             NA
#> 118    IDCJAC0011          70351 2008     4  27              NA             NA
#> 119    IDCJAC0011          70351 2008     4  28              NA             NA
#> 120    IDCJAC0011          70351 2008     4  29              NA             NA
#> 121    IDCJAC0011          70351 2008     4  30              NA             NA
#> 122    IDCJAC0011          70351 2008     5   1              NA             NA
#> 123    IDCJAC0011          70351 2008     5   2              NA             NA
#> 124    IDCJAC0011          70351 2008     5   3              NA             NA
#> 125    IDCJAC0011          70351 2008     5   4              NA             NA
#>      Quality
#> 1           
#> 2           
#> 3           
#> 4           
#> 5           
#> 6           
#> 7           
#> 8           
#> 9           
#> 10          
#> 11          
#> 12          
#> 13          
#> 14          
#> 15          
#> 16          
#> 17          
#> 18          
#> 19          
#> 20          
#> 21          
#> 22          
#> 23          
#> 24          
#> 25          
#> 26          
#> 27          
#> 28          
#> 29          
#> 30          
#> 31          
#> 32          
#> 33          
#> 34          
#> 35          
#> 36          
#> 37          
#> 38          
#> 39          
#> 40          
#> 41          
#> 42          
#> 43          
#> 44          
#> 45          
#> 46          
#> 47          
#> 48          
#> 49          
#> 50          
#> 51          
#> 52          
#> 53          
#> 54          
#> 55          
#> 56          
#> 57          
#> 58          
#> 59          
#> 60          
#> 61          
#> 62          
#> 63          
#> 64          
#> 65          
#> 66          
#> 67          
#> 68          
#> 69          
#> 70          
#> 71          
#> 72          
#> 73          
#> 74          
#> 75          
#> 76          
#> 77          
#> 78          
#> 79          
#> 80          
#> 81          
#> 82          
#> 83          
#> 84          
#> 85          
#> 86          
#> 87          
#> 88          
#> 89          
#> 90          
#> 91          
#> 92          
#> 93          
#> 94          
#> 95          
#> 96          
#> 97          
#> 98          
#> 99          
#> 100         
#> 101         
#> 102         
#> 103         
#> 104         
#> 105         
#> 106         
#> 107         
#> 108         
#> 109         
#> 110         
#> 111         
#> 112         
#> 113         
#> 114         
#> 115         
#> 116         
#> 117         
#> 118         
#> 119         
#> 120         
#> 121         
#> 122         
#> 123         
#> 124         
#> 125         
#>  [ reached getOption("max.print") -- omitted 3657 rows ]