Get BOM 0900 or 1500 weather bulletin

Fetch the daily BOM 0900 or 1500 weather bulletins and return a data frame for a specified state or territory.

get_weather_bulletin(state = "qld", morning = TRUE)

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

morning
If TRUE, return the 9am bulletin for the nominated state; otherwise return the 3pm bulletin.

Value

Data frame as a data.table object of Australian 9am or 3pm weather observations for a state. For full details of fields and units returned see Appendix 4, "Appendix 4 - Output from get_weather_bulletin()" in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes:

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

It is not possible to return weather bulletins for the entire country in a single call. Rainfall figures for the 9am bulletin are generally for the preceding 24 hours, while those for the 3pm bulletin are for the preceding 6 hours since 9am. Note that values are manually entered into the bulletins and sometimes contain typographical errors which may lead to warnings about "NAs introduced by coercion".

References

Daily observation data come from Australian Bureau of Meteorology (BOM) website. The 3pm bulletin for Queensland is, for example,
http://www.bom.gov.au/qld/observations/3pm_bulletin.shtml

Author

Mark Padgham, mark.padgham@email.com

Examples

# \donttest{
qld_weather <- get_weather_bulletin(state = "QLD", morning = FALSE)
qld_weather

#>            stations cld8ths wind_dir wind_speed_kmh temp_c_dry temp_c_dew
#>   1:     Coconut Is      NA     CALM             NA         27         NA
#>   2:        Coen Ap      NA        E             17         30         22
#>   3:        Horn Is       7      ESE             17         25         23
#>   4: Lockhart River       8        S             17         26         25
#>   5:    Palmerville      NA       NE              7         31         18
#>  ---                                                                     
#> 128:         Hobart      NA        N             20         21         10
#> 129:       Adelaide      NA      WSW             13         26         10
#> 130:       Perth Ap      NA      WNW             19         30         19
#> 131:         Darwin       7        E             26         32         20
#> 132:  Alice Springs      NA      ESE             30         28          7
#>      temp_c_max temp_c_min temp_c_gr barhpa rain_mm_6hr weather
#>   1:         NA         NA        NA   1004          NA    <NA>
#>   2:         32         NA        NA   1006          NA    <NA>
#>   3:         26         NA        NA   1004          40    Rain
#>   4:         31         NA        NA   1006           5    <NA>
#>   5:         32         NA        NA   1007          NA    <NA>
#>  ---                                                           
#> 128:         NA         NA        NA   1021          NA    <NA>
#> 129:         NA         NA        NA   1024          NA    <NA>
#> 130:         NA         NA        NA   1015          NA    <NA>
#> 131:         33         NA        NA   1005          NA    <NA>
#> 132:         29         NA        NA   1014          NA    <NA>
#>      seastate_sea_swell_dir
#>   1:                   <NA>
#>   2:                   <NA>
#>   3:                   <NA>
#>   4:                   <NA>
#>   5:                   <NA>
#>  ---                       
#> 128:                   <NA>
#> 129:                   <NA>
#> 130:                   <NA>
#> 131:                   <NA>
#> 132:                   <NA>

# }