Manage locally cached bomrang files

The user is given an option when downloading the bomrang data to cache or not to cache the data for later use. If cache == TRUE, then the bomrang data files are saved in a directory in the users' home file space. These functions provide facilities for interacting and managing these files.

bomrang_cache_list()

bomrang_cache_delete(files, force = TRUE)

bomrang_cache_delete_all(force = TRUE)

bomrang_cache_details(files = NULL)

Arguments

files
Character. One or more complete file names with no file path

force
Logical. Should files be force deleted? Defaults to : TRUE

Details

bomrang_cache_delete only accepts one file name, while bomrang_cache_delete_all does not accept any names, but deletes all files. For deleting many specific files, use cache_delete in a lapply type call.

We files cache using user_cache_dir, find your cache folder by executing rappdirs::user_cache_dir("bomrang")

Note

These functions were adapted from rOpenSci's cc_cache.

Functions

bomrang_cache_list() returns a character vector of full path file names

bomrang_cache_delete() deletes one or more files, returns nothing

bomrang_cache_delete_all() delete all files, returns nothing

bomrang_cache_details() prints file name and file size for each file, supply with one or more files, or no files (and get details for all available)

Examples

# NOT RUN {
# List files in cache
bomrang_cache_list()

# List info for single files
bomrang_cache_details(files = bomrang_cache_list()[1])
bomrang_cache_details(files = bomrang_cache_list()[2])

# List info for all files
bomrang_cache_details()

# Delete files by name in cache
bomrang_cache_delete(files = bomrang_cache_list()[1])

# Delete all files in cache
bomrang_cache_delete_all()
# }