Manage locally cached bomrang files

Manage cached bomrang satellite imagery files with hoardr.

Details

The default cache directory is
file.path(rappdirs::user_cache_dir(), "R/bomrang"), but you can set your own path using manage_cache$cache_path_set()

manage_cache$cache_delete only accepts one file name, while manage_cache$cache_delete_all does not accept any names, but deletes all files. For deleting many specific files, use
manage_cache$cache_delete in an lapply type call.

Useful user functions

manage_cache$cache_path_get() - get cache path

manage_cache$cache_path_set() - set cache path

manage_cache$list() - returns a character vector of full path file names

manage_cache$files() - returns file objects with metadata

manage_cache$details() - returns files with details

manage_cache$delete() - delete specific files

manage_cache$delete_all() - delete all files, returns nothing

Examples

if (FALSE) {

# list files in cache

imagery <- get_satellite_imagery(product_id = "IDE00425",
                                 scans = 1,
                                 cache = TRUE)

manage_cache$list()

# delete certain database files
manage_cache$delete("file path")
manage_cache$list()

# delete all files in cache
manage_cache$delete_all()
manage_cache$list()

# set a different cache path from the default
manage_cache$cache_path_set("~/tmp")
}