Parse local BOM coastal waters forecast XML files

Parse local BOM daily coastal waters forecast XML file(s) and return a data frame for a specified state or territory or all Australia.

parse_coastal_forecast(state = "AUS", filepath)

Arguments

state
Required value of an Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

filepath
A string providing the directory location of the coastal forecast file(s) to parse. See Details for more.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

The filepath argument will only accept a directory where files are located for parsing. DO NOT supply the full path including the file name. This function will only parse the requested state or all of Australia in the same fashion as get_coastal_forecast, provided that the files are all present in the directory.

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
# parse the coastal forecast for Queensland

#download to tempfile() using basename() to keep original name
download.file(url = "ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml",
              destfile = file.path(tempdir(),
              basename("ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml")),
              mode = "wb")

coastal_forecast <- parse_coastal_forecast(state = "QLD",
                                           filepath = tempdir())
coastal_forecast

#>     index product_id    type state_code
#>  1:  <NA>   IDQ11290    <NA>        QLD
#>  2:  <NA>   IDQ11290    <NA>        QLD
#>  3:  <NA>   IDQ11290    <NA>        QLD
#>  4:  <NA>   IDQ11290    <NA>        QLD
#>  5:     0   IDQ11290    <NA>        QLD
#>  6:     1   IDQ11290    <NA>        QLD
#>  7:     2   IDQ11290    <NA>        QLD
#>  8:     0   IDQ11290    <NA>        QLD
#>  9:     1   IDQ11290    <NA>        QLD
#> 10:     2   IDQ11290    <NA>        QLD
#> 11:     0   IDQ11290 Coastal        QLD
#> 12:     1   IDQ11290 Coastal        QLD
#> 13:     2   IDQ11290 Coastal        QLD
#> 14:     0   IDQ11290    <NA>        QLD
#> 15:     1   IDQ11290    <NA>        QLD
#> 16:     2   IDQ11290    <NA>        QLD
#> 17:     0   IDQ11290    <NA>        QLD
#> 18:     1   IDQ11290    <NA>        QLD
#> 19:     2   IDQ11290    <NA>        QLD
#> 20:     0   IDQ11290    <NA>        QLD
#> 21:     1   IDQ11290    <NA>        QLD
#> 22:     2   IDQ11290    <NA>        QLD
#> 23:     0   IDQ11290    <NA>        QLD
#> 24:     1   IDQ11290    <NA>        QLD
#> 25:     2   IDQ11290    <NA>        QLD
#> 26:     0   IDQ11290    <NA>        QLD
#> 27:     1   IDQ11290    <NA>        QLD
#> 28:     2   IDQ11290    <NA>        QLD
#> 29:     0   IDQ11290    <NA>        QLD
#> 30:     1   IDQ11290    <NA>        QLD
#> 31:     2   IDQ11290    <NA>        QLD
#> 32:     0   IDQ11290   Local        QLD
#> 33:     1   IDQ11290   Local        QLD
#> 34:     2   IDQ11290   Local        QLD
#> 35:     0   IDQ11290    <NA>        QLD
#> 36:     1   IDQ11290    <NA>        QLD
#> 37:     2   IDQ11290    <NA>        QLD
#> 38:     0   IDQ11290    <NA>        QLD
#> 39:     1   IDQ11290    <NA>        QLD
#> 40:     2   IDQ11290    <NA>        QLD
#> 41:     0   IDQ11290   Local        QLD
#> 42:     1   IDQ11290   Local        QLD
#> 43:     2   IDQ11290   Local        QLD
#> 44:     0   IDQ11290    <NA>        QLD
#> 45:     1   IDQ11290    <NA>        QLD
#> 46:     2   IDQ11290    <NA>        QLD
#> 47:     0   IDQ11290    <NA>        QLD
#> 48:     1   IDQ11290    <NA>        QLD
#> 49:     2   IDQ11290    <NA>        QLD
#>     index product_id    type state_code
#>                                                          dist_name pt_1_name
#>  1:                                                     Queensland      <NA>
#>  2:                                 Queensland Gulf of Carpentaria      <NA>
#>  3:                                               North Queensland      <NA>
#>  4:                                               South Queensland      <NA>
#>  5: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  6: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  7: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  8:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#>  9:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 10:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 11:                                                  Torres Strait      <NA>
#> 12:                                                  Torres Strait      <NA>
#> 13:                                                  Torres Strait      <NA>
#> 14:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 15:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 16:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 17:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 18:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 19:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 20:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 21:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 22:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 23:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 24:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 25:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 26:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 27:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 28:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 29:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 30:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 31:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 32:                                                     Hervey Bay      <NA>
#> 33:                                                     Hervey Bay      <NA>
#> 34:                                                     Hervey Bay      <NA>
#> 35:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 36:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 37:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 38:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 39:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 40:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 41:                                                    Moreton Bay      <NA>
#> 42:                                                    Moreton Bay      <NA>
#> 43:                                                    Moreton Bay      <NA>
#> 44:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 45:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 46:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 47:                             Great Barrier Reef Offshore Waters      <NA>
#> 48:                             Great Barrier Reef Offshore Waters      <NA>
#> 49:                             Great Barrier Reef Offshore Waters      <NA>
#>                                                          dist_name pt_1_name
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>  1:      <NA> QLD_FA001 2020-12-11 10:00:12 2020-12-11 10:00:12      10:00
#>  2:      <NA> QLD_FA002 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#>  3:      <NA> QLD_FA003 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#>  4:      <NA> QLD_FA004 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#>  5:      <NA> QLD_MW001 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#>  6:      <NA> QLD_MW001 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#>  7:      <NA> QLD_MW001 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#>  8:      <NA> QLD_MW002 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#>  9:      <NA> QLD_MW002 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 10:      <NA> QLD_MW002 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 11:      <NA> QLD_MW003 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 12:      <NA> QLD_MW003 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 13:      <NA> QLD_MW003 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 14:      <NA> QLD_MW004 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 15:      <NA> QLD_MW004 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 16:      <NA> QLD_MW004 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 17:      <NA> QLD_MW005 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 18:      <NA> QLD_MW005 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 19:      <NA> QLD_MW005 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 20:      <NA> QLD_MW006 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 21:      <NA> QLD_MW006 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 22:      <NA> QLD_MW006 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 23:      <NA> QLD_MW007 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 24:      <NA> QLD_MW007 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 25:      <NA> QLD_MW007 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 26:      <NA> QLD_MW008 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 27:      <NA> QLD_MW008 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 28:      <NA> QLD_MW008 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 29:      <NA> QLD_MW009 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 30:      <NA> QLD_MW009 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 31:      <NA> QLD_MW009 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 32:      <NA> QLD_MW010 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 33:      <NA> QLD_MW010 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 34:      <NA> QLD_MW010 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 35:      <NA> QLD_MW011 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 36:      <NA> QLD_MW011 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 37:      <NA> QLD_MW011 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 38:      <NA> QLD_MW012 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 39:      <NA> QLD_MW012 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 40:      <NA> QLD_MW012 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 41:      <NA> QLD_MW013 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 42:      <NA> QLD_MW013 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 43:      <NA> QLD_MW013 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 44:      <NA> QLD_MW014 2020-12-11 10:00:00 2020-12-11 10:00:00      10:00
#> 45:      <NA> QLD_MW014 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 46:      <NA> QLD_MW014 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#> 47:      <NA> QLD_MW015 2020-12-11 05:00:00 2020-12-11 05:00:00      10:00
#> 48:      <NA> QLD_MW015 2020-12-12 00:00:00 2020-12-12 00:00:00      10:00
#> 49:      <NA> QLD_MW015 2020-12-13 00:00:00 2020-12-13 00:00:00      10:00
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>          start_time_utc        end_time_utc
#>  1: 2020-12-11 10:00:12 2020-12-11 10:00:12
#>  2: 2020-12-11 05:00:00 2020-12-11 05:00:00
#>  3: 2020-12-11 05:00:00 2020-12-11 05:00:00
#>  4: 2020-12-11 05:00:00 2020-12-11 05:00:00
#>  5: 2020-12-11 05:00:00 2020-12-11 05:00:00
#>  6: 2020-12-12 00:00:00 2020-12-12 00:00:00
#>  7: 2020-12-13 00:00:00 2020-12-13 00:00:00
#>  8: 2020-12-11 05:00:00 2020-12-11 05:00:00
#>  9: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 10: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 11: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 12: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 13: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 14: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 15: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 16: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 17: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 18: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 19: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 20: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 21: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 22: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 23: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 24: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 25: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 26: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 27: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 28: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 29: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 30: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 31: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 32: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 33: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 34: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 35: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 36: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 37: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 38: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 39: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 40: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 41: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 42: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 43: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 44: 2020-12-11 10:00:00 2020-12-11 10:00:00
#> 45: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 46: 2020-12-13 00:00:00 2020-12-13 00:00:00
#> 47: 2020-12-11 05:00:00 2020-12-11 05:00:00
#> 48: 2020-12-12 00:00:00 2020-12-12 00:00:00
#> 49: 2020-12-13 00:00:00 2020-12-13 00:00:00
#>          start_time_utc        end_time_utc
#>                                                                             forecast_seas
#>  1:                                                                                  <NA>
#>  2:                                                                                  <NA>
#>  3:                                                                                  <NA>
#>  4:                                                                                  <NA>
#>  5:                                                                       Around 1 metre.
#>  6:                                                                        Below 1 metre.
#>  7:                                                                        Below 1 metre.
#>  8:                                                                        Below 1 metre.
#>  9:                                                                       Around 1 metre.
#> 10:                                                                        Below 1 metre.
#> 11:                                                                        Below 1 metre.
#> 12:                                                                        Below 1 metre.
#> 13:                                                                        Below 1 metre.
#> 14:                        Below 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 15:                        Below 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 16:                                                                       Around 1 metre.
#> 17:                                                                       Around 1 metre.
#> 18:    Around 1 metre, increasing to 1 to 1.5 metres outside the reef during the morning.
#> 19:                                                                       Around 1 metre.
#> 20:       Below 1 metre, increasing to 1 to 1.5 metres outside the reef by early evening.
#> 21:                       Around 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 22:                                                                       Around 1 metre.
#> 23:                        1 to 1.5 metres, increasing to 1 to 2 metres by early evening.
#> 24:                                                                        1 to 2 metres.
#> 25:                                                                        1 to 2 metres.
#> 26:                        1 to 1.5 metres, increasing to 1 to 2 metres by early evening.
#> 27:                            1 to 1.5 metres, increasing to 1.5 to 2.5 metres offshore.
#> 28:                                      Around 1 metre, increasing to 2 metres offshore.
#> 29:                  1 to 1.5 metres, increasing to 1.5 to 2 metres during the afternoon.
#> 30:                                                                    1.5 to 2.5 metres.
#> 31:                            2 to 3 metres, reaching up to 4 metres south of Gladstone.
#> 32:                     Around 1 metre, increasing to 1 to 2 metres during the afternoon.
#> 33:                  1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the morning.
#> 34:                                                                        2 to 3 metres.
#> 35:                            Around 1 metre, increasing to 1 to 2 metres around midday.
#> 36:                                                                    1.5 to 2.5 metres.
#> 37:                           3 to 4 metres, increasing to 4 to 4.5 metres around midday.
#> 38:                   Around 1 metre, increasing to 1.5 to 2 metres during the afternoon.
#> 39:                                    2 to 2.5 metres, reaching 3 metres in the evening.
#> 40:                             3 to 4 metres, increasing to 4 to 5 metres around midday.
#> 41:                                                                       Around 1 metre.
#> 42: Around 1 metre, increasing to 1 to 1.5 metres in the northern bay during the morning.
#> 43:                                                                      2 to 2.5 metres.
#> 44:                        Around 1 metre, increasing to 1.5 metres during the afternoon.
#> 45:                      1 to 1.5 metres, increasing to 2 to 3 metres during the morning.
#> 46:                          3 to 4 metres, increasing to 4 to 5 metres in the afternoon.
#> 47:                    1 to 1.5 metres, increasing to 1 to 2 metres during the afternoon.
#> 48:                                                                    1.5 to 2.5 metres.
#> 49:                                                                    1.5 to 2.5 metres.
#>                                                                             forecast_seas
#>                                                                                                     forecast_weather
#>  1:                                                                                                             <NA>
#>  2:                                                                                                             <NA>
#>  3:                                                                                                             <NA>
#>  4:                                                                                                             <NA>
#>  5:                                              Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#>  6:                                              Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#>  7:                                              Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#>  8:                                              Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#>  9:                                              Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 10:                                              Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#> 11:                   Partly cloudy. 50% chance of showers. The chance of a thunderstorm offshore west of Cape York.
#> 12:                                              Partly cloudy. 50% chance of showers. The chance of a thunderstorm.
#> 13:                                                                     Partly cloudy. The chance of a thunderstorm.
#> 14:                                                                            Partly cloudy. 60% chance of showers.
#> 15:                The chance of a thunderstorm south of Cape Sidmouth in the early morning. Mostly sunny afternoon.
#> 16:                                                      Partly cloudy. The chance of a thunderstorm in the evening.
#> 17:                                                                                                    Mostly sunny.
#> 18:                                                                                                    Mostly sunny.
#> 19:                                                                                                   Partly cloudy.
#> 20:                                                                            Partly cloudy. 60% chance of showers.
#> 21:                                                                            Partly cloudy. 50% chance of showers.
#> 22:                                                                                                   Partly cloudy.
#> 23:                                                                                  Partly cloudy. Possible shower.
#> 24:                                                                                                   Partly cloudy.
#> 25:                                                                                                   Partly cloudy.
#> 26:                                                                                                   Partly cloudy.
#> 27:                                                                                                   Partly cloudy.
#> 28:                                                                            Partly cloudy. 50% chance of showers.
#> 29:                               Partly cloudy. The chance of a thunderstorm south of Cape Capricorn later tonight.
#> 30: Cloudy. 80% chance of showers offshore south of Agnes Water, 50% chance elsewhere. The chance of a thunderstorm.
#> 31:                                                                            Partly cloudy. 50% chance of showers.
#> 32:                                                       Partly cloudy. The chance of a thunderstorm later tonight.
#> 33:                                                     Cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 34:                                                                                   Cloudy. 80% chance of showers.
#> 35:                                                       Partly cloudy. The chance of a thunderstorm later tonight.
#> 36:                                              Partly cloudy. 90% chance of showers. The chance of a thunderstorm.
#> 37:                                                        Rain and thunderstorms likely, with heavy falls possible.
#> 38:                                              Partly cloudy. The chance of a thunderstorm offshore later tonight.
#> 39:                                                     Cloudy. 95% chance of showers. The chance of a thunderstorm.
#> 40:                                                        Rain and thunderstorms likely, with heavy falls possible.
#> 41:                                            Partly cloudy. 50% chance of showers late this afternoon and evening.
#> 42:                                      Cloudy. 95% chance of showers. The chance of a thunderstorm in the morning.
#> 43:                                                        Rain and thunderstorms likely, with heavy falls possible.
#> 44:                            Partly cloudy. 50% chance of showers. The chance of a thunderstorm offshore at night.
#> 45:                                                     Cloudy. 95% chance of showers. The chance of a thunderstorm.
#> 46:                                                        Rain and thunderstorms likely, with heavy falls possible.
#> 47:                                                                                                   Partly cloudy.
#> 48:               Partly cloudy. 50% chance of showers from the late morning. The chance of a thunderstorm at night.
#> 49:                                                                                           50% chance of showers.
#>                                                                                                     forecast_weather
#>                                                                                                                                                                           forecast_winds
#>  1:                                                                                                                                                                                 <NA>
#>  2:                                                                                                                                                                                 <NA>
#>  3:                                                                                                                                                                                 <NA>
#>  4:                                                                                                                                                                                 <NA>
#>  5:                                          Northwest to northeasterly 10 to 15 knots. Winds becoming variable about 10 knots in the afternoon with inshore sea breezes up to 15 knots.
#>  6: Variable about 10 knots becoming northeasterly 10 to 15 knots early in the morning. Winds becoming variable about 10 knots in the afternoon with inshore sea breezes up to 15 knots.
#>  7:                                                                                                           Variable about 10 knots with inshore afternoon sea breezes up to 15 knots.
#>  8:                                                                                  East to northeasterly 10 to 15 knots. Inshore winds tending west to southwesterly in the afternoon.
#>  9:                            East to northeasterly 10 to 15 knots becoming variable about 10 knots during the afternoon. Inshore winds tending west to southwesterly in the afternoon.
#> 10:                                                                                                                          Variable about 10 knots with afternoon inshore sea breezes.
#> 11:                                                                                            East to southeasterly about 10 knots increasing to 10 to 15 knots in the early afternoon.
#> 12:                                                                                                                                                             Easterly 10 to 15 knots.
#> 13:                                                                                                                                                East to southeasterly 10 to 15 knots.
#> 14:                      East to southeasterly 10 to 15 knots. Inshore winds south of Port Stewart tending northeasterly and reaching up to 20 knots in the afternoon and early evening.
#> 15:                      East to southeasterly 10 to 15 knots. Inshore winds south of Port Stewart tending northeasterly and reaching up to 20 knots in the afternoon and early evening.
#> 16:                                                                                                                                                East to southeasterly 10 to 15 knots.
#> 17:                                                                                                                      Southeasterly 10 to 15 knots, reaching up to 20 knots at times.
#> 18:                                                                                                                                                        Southeasterly 15 to 20 knots.
#> 19:                                                                                                                                                        Southeasterly 10 to 15 knots.
#> 20:                                                                                                    Southeasterly 10 to 15 knots, reaching up to 20 knots at times on exposed waters.
#> 21:                                                                                                                                                        Southeasterly 15 to 20 knots.
#> 22:                                                                                                                                                        Southeasterly 10 to 15 knots.
#> 23:                                                                                       East to southeasterly 15 to 20 knots with east to northeasterly afternoon sea breezes inshore.
#> 24:          Southeasterly 15 to 20 knots, reaching up to 25 knots offshore south of Cape Bowling Green. Inshore winds around Townsville tending east to northeasterly in the afternoon.
#> 25:                                                                        Southeasterly 15 to 20 knots. Inshore winds around Townsville tending east to northeasterly in the afternoon.
#> 26:                                                     Easterly 10 to 15 knots turning southeasterly 15 to 20 knots in the late morning, increasing to 20 to 25 knots in the afternoon.
#> 27:                                                                               Southeasterly 20 to 25 knots, reaching up to 30 knots offshore south of Mackay in the early afternoon.
#> 28:                                                                                                                      Southeasterly 15 to 20 knots, reaching up to 25 knots at times.
#> 29:                                              Southeasterly 10 to 15 knots increasing to 15 to 20 knots during the late morning. Winds increasing to 20 to 25 knots in the afternoon.
#> 30:                                                                                                             Southeasterly 20 to 25 knots, reaching 25 to 30 knots from late morning.
#> 31:                                                                                                   South to southeasterly 20 to 25 knots, reaching up to 33 knots south of Gladstone.
#> 32:                                   Southeasterly 10 to 15 knots increasing to 15 to 20 knots in the middle of the day, reaching up to 25 knots at times in the afternoon and evening.
#> 33:                                    Southeasterly 15 to 20 knots, increasing to 20 to 25 knots during the morning then reaching up to 30 knots at times in the afternoon and evening.
#> 34:                                                                                                         Southeasterly 20 to 25 knots turning southerly 25 to 33 knots around midday.
#> 35:                                                                                    Southeasterly 15 to 20 knots in the morning. Winds increasing to 20 to 25 knots in the afternoon.
#> 36:                                                                                                            Southeasterly 20 to 25 knots increasing to 20 to 30 knots in the morning.
#> 37:                                                       Southeasterly 25 to 30 knots, increasing to 30 to 35 knots in the morning. Squalls to 45 knots with showers and thunderstorms.
#> 38:                                                    Southeasterly 10 to 15 knots increasing to 15 to 20 knots in the middle of the day. Winds reaching up to 25 knots in the evening.
#> 39:                                                                                                          Southeasterly 20 to 25 knots increasing to 25 to 30 knots in the afternoon.
#> 40:                                                       Southeasterly 25 to 30 knots, increasing to 30 to 40 knots in the morning. Squalls to 50 knots with showers and thunderstorms.
#> 41:                                   East to southeasterly 10 to 15 knots turning southeasterly 15 to 20 knots in the afternoon. Winds reaching up to 25 knots at times in the evening.
#> 42:                                                                                                       Southeasterly 20 to 25 knots increasing to 20 to 30 knots in the late evening.
#> 43:                                                            Southeasterly 25 to 30 knots, increasing to 30 to 35 knots by midday. Squalls to 50 knots with showers and thunderstorms.
#> 44:                                                                          East to southeasterly 15 to 20 knots during the morning. Winds increasing to 20 to 25 knots in the evening.
#> 45:                                                                                                          Southeasterly 15 to 25 knots increasing to 25 to 30 knots in the afternoon.
#> 46:                                                            Southeasterly 25 to 30 knots, increasing to 30 to 40 knots by midday. Squalls to 50 knots with showers and thunderstorms.
#> 47:                                                                                                                                                        Southeasterly 15 to 20 knots.
#> 48:                                                                                                    Southeasterly 20 to 25 knots, reaching up to 30 knots at times from late morning.
#> 49:                                                                                                                                               South to southeasterly 15 to 25 knots.
#>                                                                                                                                                                           forecast_winds
#>                                                                                       forecast_swell1
#>  1:                                                                                              <NA>
#>  2:                                                                                              <NA>
#>  3:                                                                                              <NA>
#>  4:                                                                                              <NA>
#>  5:                                                                                              <NA>
#>  6:                                                                                              <NA>
#>  7:                                                                                              <NA>
#>  8:                                                                                              <NA>
#>  9:                                                                                              <NA>
#> 10:                                                                                              <NA>
#> 11:                                                                                 Below 0.5 metres.
#> 12:                                                              East to southeasterly below 1 metre.
#> 13:                                                              East to southeasterly below 1 metre.
#> 14:                                                                      Southeasterly below 1 metre.
#> 15:                                                    Southeasterly around 1 metre outside the reef.
#> 16:                                                    Southeasterly around 1 metre outside the reef.
#> 17:                                                              East to southeasterly below 1 metre.
#> 18:                                            East to southeasterly around 1 metre outside the reef.
#> 19:  Southeasterly around 1 metre outside the reef, increasing to 1 to 1.5 metres during the morning.
#> 20:                                            East to southeasterly around 1 metre outside the reef.
#> 21:  Southeasterly around 1 metre outside the reef, increasing to 1 to 1.5 metres during the morning.
#> 22:                        Southeasterly below 1 metre, increasing to 1 to 2 metres outside the reef.
#> 23:                                                         Easterly around 1 metre outside the reef.
#> 24:                           Easterly below 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 25:                             Easterly below 1 metre, increasing to 1 to 2 metres outside the reef.
#> 26:                                                            Southeasterly around 1 metre offshore.
#> 27:              East to southeasterly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 28:                        Southeasterly below 1 metre inshore, increasing to 1 to 2 metres offshore.
#> 29:                            Easterly below 1 metre inshore, increasing to around 1 metre offshore.
#> 30:                     Easterly 1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the morning.
#> 31:           East to southeasterly around 1 metre inshore, increasing to 2.5 to 3.5 metres offshore.
#> 32:                                         North to northeasterly below 1 metre in the northern bay.
#> 33:                                                 Northeasterly around 1 metre in the northern bay.
#> 34: Northeasterly around 1 metre, increasing to 1 to 2 metres in the northern bay during the evening.
#> 35:                                                                          Easterly around 1 metre.
#> 36:                  East to southeasterly 1.5 to 2.5 metres, reaching up to 3 metres in the evening.
#> 37:              East to southeasterly 2.5 to 3 metres, increasing to 3 to 4 metres in the afternoon.
#> 38:                                                              East to southeasterly below 1 metre.
#> 39:                          Southeasterly 1 to 2 metres, increasing to 2 to 3 metres in the evening.
#> 40:              East to southeasterly 2.5 to 3 metres, increasing to 3 to 4 metres in the afternoon.
#> 41:                                                                                              <NA>
#> 42:                                                                                              <NA>
#> 43:                                                                                              <NA>
#> 44:                                                              East to southeasterly below 1 metre.
#> 45:                       Southeasterly around 1 metre, increasing to 2 to 3 metres in the afternoon.
#> 46:              East to southeasterly 2.5 to 3 metres, increasing to 3 to 4 metres in the afternoon.
#> 47:                                                                          Easterly around 1 metre.
#> 48:                                                              East to southeasterly 1 to 2 metres.
#> 49:         Southeasterly 1 to 1.5 metres, increasing to 2 to 3 metres through the Capricorn Channel.
#>                                                                                       forecast_swell1
#>     forecast_swell2 forecast_caution
#>  1:            <NA>             <NA>
#>  2:            <NA>             <NA>
#>  3:            <NA>             <NA>
#>  4:            <NA>             <NA>
#>  5:            <NA>             <NA>
#>  6:            <NA>             <NA>
#>  7:            <NA>             <NA>
#>  8:            <NA>             <NA>
#>  9:            <NA>             <NA>
#> 10:            <NA>             <NA>
#> 11:            <NA>             <NA>
#> 12:            <NA>             <NA>
#> 13:            <NA>             <NA>
#> 14:            <NA>             <NA>
#> 15:            <NA>             <NA>
#> 16:            <NA>             <NA>
#> 17:            <NA>             <NA>
#> 18:            <NA>             <NA>
#> 19:            <NA>             <NA>
#> 20:            <NA>             <NA>
#> 21:            <NA>             <NA>
#> 22:            <NA>             <NA>
#> 23:            <NA>             <NA>
#> 24:            <NA>             <NA>
#> 25:            <NA>             <NA>
#> 26:            <NA>             <NA>
#> 27:            <NA>             <NA>
#> 28:            <NA>             <NA>
#> 29:            <NA>             <NA>
#> 30:            <NA>             <NA>
#> 31:            <NA>             <NA>
#> 32:            <NA>             <NA>
#> 33:            <NA>             <NA>
#> 34:            <NA>             <NA>
#> 35:            <NA>             <NA>
#> 36:            <NA>             <NA>
#> 37:            <NA>             <NA>
#> 38:            <NA>             <NA>
#> 39:            <NA>             <NA>
#> 40:            <NA>             <NA>
#> 41:            <NA>             <NA>
#> 42:            <NA>             <NA>
#> 43:            <NA>             <NA>
#> 44:            <NA>             <NA>
#> 45:            <NA>             <NA>
#> 46:            <NA>             <NA>
#> 47:            <NA>             <NA>
#> 48:            <NA>             <NA>
#> 49:            <NA>             <NA>
#>     forecast_swell2 forecast_caution
#>                                                marine_forecast
#>  1:                                                       <NA>
#>  2:                                                       <NA>
#>  3:                                                       <NA>
#>  4:                                                       <NA>
#>  5:                                                       <NA>
#>  6:                                                       <NA>
#>  7:                                                       <NA>
#>  8:                                                       <NA>
#>  9:                                                       <NA>
#> 10:                                                       <NA>
#> 11:                                                       <NA>
#> 12:                                                       <NA>
#> 13:                                                       <NA>
#> 14:                                                       <NA>
#> 15:                                                       <NA>
#> 16:                                                       <NA>
#> 17:                                                       <NA>
#> 18:                                                       <NA>
#> 19:                                                       <NA>
#> 20:                                                       <NA>
#> 21:                                                       <NA>
#> 22:                                                       <NA>
#> 23:                                                       <NA>
#> 24:                                                       <NA>
#> 25:                                                       <NA>
#> 26:                                                       <NA>
#> 27:          Strong Wind Warning for Saturday for Mackay Coast
#> 28:                                                       <NA>
#> 29:                                                       <NA>
#> 30:     Strong Wind Warning for Saturday for Capricornia Coast
#> 31:                                                       <NA>
#> 32:                                                       <NA>
#> 33:            Strong Wind Warning for Saturday for Hervey Bay
#> 34:                                                       <NA>
#> 35:                                                       <NA>
#> 36:   Strong Wind Warning for Saturday for Fraser Island Coast
#> 37:                                                       <NA>
#> 38:                                                       <NA>
#> 39: Strong Wind Warning for Saturday for Sunshine Coast Waters
#> 40:                                                       <NA>
#> 41:                                                       <NA>
#> 42:           Strong Wind Warning for Saturday for Moreton Bay
#> 43:                                                       <NA>
#> 44:                                                       <NA>
#> 45:     Strong Wind Warning for Saturday for Gold Coast Waters
#> 46:                                                       <NA>
#> 47:                                                       <NA>
#> 48:                                                       <NA>
#> 49:                                                       <NA>
#>                                                marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#> 34:                       NA             NA
#> 35:                       NA             NA
#> 36:                       NA             NA
#> 37:                       NA             NA
#> 38:                       NA             NA
#> 39:                       NA             NA
#> 40:                       NA             NA
#> 41:                       NA             NA
#> 42:                       NA             NA
#> 43:                       NA             NA
#> 44:                       NA             NA
#> 45:                       NA             NA
#> 46:                       NA             NA
#> 47:                       NA             NA
#> 48:                       NA             NA
#> 49:                       NA             NA
#>     tropical_system_location forecast_waves

# }