Update internal databases with latest BOM station locations and metadata

Download the latest station locations and metadata and update bomrang's internal databases that support the use of link{get_current_weather} and get_ag_bulletin. There is no need to use this unless you know that a station exists in BOM's database that is not available in the databases distributed with bomrang.

update_station_locations()

Value

Updated internal databases of BOM station locations and JSON URLs

References

Australian Bureau of Meteorology (BOM) Site Numbers http://www.bom.gov.au/climate/cdo/about/site-num.shtml

Examples

not_run({
  update_station_locations()
})