BoM Agriculture Bulletin Information for Select Stations

Fetch the BoM agricultural bulletin information and return a tidy data frame

get_ag_bulletin(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via base::agrep is done. Defaults to "AUS" returning all state bulletins, see details for more.

Value

Tidy data frame of a Australia BoM agricultural bulletin information. For full details of fields and units returned see Appendix 3 in the bomrang vignette, use vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns bulletin for all states and NT

References

Australian Bureau of Meteorology (BoM) Weather Data Services Agriculture Bulletins http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml

Australian Bureau of Meteorology (BoM) Weather Data Services Observation of Rainfall http://www.bom.gov.au/climate/how/observations/rain-measure.shtml

Examples

# NOT RUN {
ag_bulletin <- get_ag_bulletin(state = "QLD")
# }