Get a listing of available BOM satellite GeoTIFF imagery

Fetch a listing of BOM 'GeoTIFF' satellite imagery from ftp://ftp.bom.gov.au/anon/gen/gms/ to determine which files are currently available for download. Files are available at ten minute update frequency with a 24 hour delete time. Useful to know the most recent files available and then specify in the get_satellite_imagery function.

get_available_imagery(product_id = "all")

Arguments

product_id
Character. BOM product ID of interest for which a list of available images will be returned. Defaults to all images currently available.

Value

A vector of all available files for the requested Product ID(s).

Details

Valid BOM satellite Product IDs for 'GeoTIFF' files include:

IDE00420

AHI cloud cover only 2km FD GEOS GIS

IDE00421

AHI IR (Ch13) greyscale 2km FD GEOS GIS

IDE00422

AHI VIS (Ch3) greyscale 2km FD GEOS GIS

IDE00423

AHI IR (Ch13) Zehr 2km FD GEOS GIS

IDE00425

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD GEOS GIS

IDE00426

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km FD GEOS GIS

IDE00427

AHI WV (Ch8) 2km FD GEOS GIS

IDE00430

AHI cloud cover only 2km AUS equirect. GIS

IDE00431

AHI IR (Ch13) greyscale 2km AUS equirect. GIS

IDE00432

AHI VIS (Ch3) greyscale 2km AUS equirect. GIS

IDE00433

AHI IR (Ch13) Zehr 2km AUS equirect. GIS

IDE00435

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km AUS equirect. GIS

IDE00436

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km AUS equirect. GIS

IDE00437

AHI WV (Ch8) 2km AUS equirect. GIS

IDE00439

AHI VIS (Ch3) greyscale 0.5km AUS equirect. GIS

References

Australian Bureau of Meteorology (BOM) high-definition satellite images http://www.bom.gov.au/australia/satellite/index.shtml

Author

Adam H. Sparks, adamhsparks@gmail.com

Examples

# \donttest{
# Check availability of AHI VIS (true colour) / IR (Ch13 greyscale) composite
# 1km FD GEOS GIS images
imagery <- get_available_imagery(product_id = "IDE00425")

#> 
#> The following files are currently available for download:

#>   [1] "IDE00425.202103300010.tif" "IDE00425.202103300020.tif"
#>   [3] "IDE00425.202103300030.tif" "IDE00425.202103300040.tif"
#>   [5] "IDE00425.202103300050.tif" "IDE00425.202103300100.tif"
#>   [7] "IDE00425.202103300110.tif" "IDE00425.202103300120.tif"
#>   [9] "IDE00425.202103300130.tif" "IDE00425.202103300140.tif"
#>  [11] "IDE00425.202103300150.tif" "IDE00425.202103300200.tif"
#>  [13] "IDE00425.202103300210.tif" "IDE00425.202103300220.tif"
#>  [15] "IDE00425.202103300230.tif" "IDE00425.202103300250.tif"
#>  [17] "IDE00425.202103300300.tif" "IDE00425.202103300310.tif"
#>  [19] "IDE00425.202103300320.tif" "IDE00425.202103300330.tif"
#>  [21] "IDE00425.202103300340.tif" "IDE00425.202103300350.tif"
#>  [23] "IDE00425.202103300400.tif" "IDE00425.202103300410.tif"
#>  [25] "IDE00425.202103300420.tif" "IDE00425.202103300430.tif"
#>  [27] "IDE00425.202103300440.tif" "IDE00425.202103300450.tif"
#>  [29] "IDE00425.202103300500.tif" "IDE00425.202103300510.tif"
#>  [31] "IDE00425.202103300520.tif" "IDE00425.202103300530.tif"
#>  [33] "IDE00425.202103300540.tif" "IDE00425.202103300550.tif"
#>  [35] "IDE00425.202103300600.tif" "IDE00425.202103300610.tif"
#>  [37] "IDE00425.202103300620.tif" "IDE00425.202103300630.tif"
#>  [39] "IDE00425.202103300640.tif" "IDE00425.202103300650.tif"
#>  [41] "IDE00425.202103300700.tif" "IDE00425.202103300710.tif"
#>  [43] "IDE00425.202103300720.tif" "IDE00425.202103300730.tif"
#>  [45] "IDE00425.202103300740.tif" "IDE00425.202103300750.tif"
#>  [47] "IDE00425.202103300800.tif" "IDE00425.202103300810.tif"
#>  [49] "IDE00425.202103300820.tif" "IDE00425.202103300830.tif"
#>  [51] "IDE00425.202103300840.tif" "IDE00425.202103300850.tif"
#>  [53] "IDE00425.202103300900.tif" "IDE00425.202103300910.tif"
#>  [55] "IDE00425.202103300920.tif" "IDE00425.202103300930.tif"
#>  [57] "IDE00425.202103300940.tif" "IDE00425.202103300950.tif"
#>  [59] "IDE00425.202103301000.tif" "IDE00425.202103301010.tif"
#>  [61] "IDE00425.202103301020.tif" "IDE00425.202103301030.tif"
#>  [63] "IDE00425.202103301040.tif" "IDE00425.202103301050.tif"
#>  [65] "IDE00425.202103301100.tif" "IDE00425.202103301110.tif"
#>  [67] "IDE00425.202103301120.tif" "IDE00425.202103301130.tif"
#>  [69] "IDE00425.202103301140.tif" "IDE00425.202103301150.tif"
#>  [71] "IDE00425.202103301200.tif" "IDE00425.202103301210.tif"
#>  [73] "IDE00425.202103301220.tif" "IDE00425.202103301230.tif"
#>  [75] "IDE00425.202103301240.tif" "IDE00425.202103301250.tif"
#>  [77] "IDE00425.202103301300.tif" "IDE00425.202103301310.tif"
#>  [79] "IDE00425.202103301320.tif" "IDE00425.202103301330.tif"
#>  [81] "IDE00425.202103301340.tif" "IDE00425.202103301350.tif"
#>  [83] "IDE00425.202103301400.tif" "IDE00425.202103301410.tif"
#>  [85] "IDE00425.202103301420.tif" "IDE00425.202103301430.tif"
#>  [87] "IDE00425.202103301450.tif" "IDE00425.202103301500.tif"
#>  [89] "IDE00425.202103301510.tif" "IDE00425.202103301520.tif"
#>  [91] "IDE00425.202103301530.tif" "IDE00425.202103301540.tif"
#>  [93] "IDE00425.202103301550.tif" "IDE00425.202103301600.tif"
#>  [95] "IDE00425.202103301610.tif" "IDE00425.202103301620.tif"
#>  [97] "IDE00425.202103301630.tif" "IDE00425.202103301640.tif"
#>  [99] "IDE00425.202103301650.tif" "IDE00425.202103301700.tif"
#> [101] "IDE00425.202103301710.tif" "IDE00425.202103301720.tif"
#> [103] "IDE00425.202103301730.tif" "IDE00425.202103301740.tif"
#> [105] "IDE00425.202103301750.tif" "IDE00425.202103301800.tif"
#> [107] "IDE00425.202103301810.tif" "IDE00425.202103301820.tif"
#> [109] "IDE00425.202103301830.tif" "IDE00425.202103301840.tif"
#> [111] "IDE00425.202103301850.tif" "IDE00425.202103301900.tif"
#> [113] "IDE00425.202103301910.tif" "IDE00425.202103301920.tif"
#> [115] "IDE00425.202103301930.tif" "IDE00425.202103301940.tif"
#> [117] "IDE00425.202103301950.tif" "IDE00425.202103302000.tif"
#> [119] "IDE00425.202103302010.tif" "IDE00425.202103302020.tif"
#> [121] "IDE00425.202103302030.tif" "IDE00425.202103302040.tif"
#> [123] "IDE00425.202103302050.tif" "IDE00425.202103302100.tif"
#> [125] "IDE00425.202103302110.tif" "IDE00425.202103302120.tif"
#> [127] "IDE00425.202103302130.tif" "IDE00425.202103302140.tif"
#> [129] "IDE00425.202103302150.tif" "IDE00425.202103302200.tif"
#> [131] "IDE00425.202103302210.tif" "IDE00425.202103302220.tif"
#> [133] "IDE00425.202103302230.tif" "IDE00425.202103302240.tif"
#> [135] "IDE00425.202103302250.tif" "IDE00425.202103302300.tif"
#> [137] "IDE00425.202103302310.tif" "IDE00425.202103302320.tif"
#> [139] "IDE00425.202103302330.tif" "IDE00425.202103302340.tif"
#> [141] "IDE00425.202103302350.tif" "IDE00425.202103310000.tif"
#> [143] "IDE00425.202103310010.tif" "IDE00425.202103310020.tif"
#> [145] "IDE00425.202103310030.tif" "IDE00425.202103310040.tif"

# }