Get BOM Coastal Waters Forecast

Fetch the BOM daily Coastal Waters Forecast and return a tidy data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to "AUS" returning all state forecasts, see details for further information.

Value

Tidy data.frame of a Australia BOM Coastal Waters Forecast.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

Examples

coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>    index product_id    type state_code dist_name       pt_1_name
#> 1      0   IDN11001 Coastal        NSW      Eden Montague Island
#> 2      1   IDN11001 Coastal        NSW      Eden Montague Island
#> 3      2   IDN11001 Coastal        NSW      Eden Montague Island
#> 4      0   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 5      1   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 6      2   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 7      0   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 8      1   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 9      2   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 10     0   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 11     1   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 12     2   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 13     0   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 14     1   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 15     2   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 16     0   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 17     1   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 18     2   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 19     0   IDN11001 Coastal        NSW     Coffs           Wooli
#> 20     1   IDN11001 Coastal        NSW     Coffs           Wooli
#> 21     2   IDN11001 Coastal        NSW     Coffs           Wooli
#> 22     0   IDN11001 Coastal        NSW     Byron    Point Danger
#> 23     1   IDN11001 Coastal        NSW     Byron    Point Danger
#> 24     2   IDN11001 Coastal        NSW     Byron    Point Danger
#>          pt_2_name       aac    start_time_local end_time_local utc_offset
#> 1      Gabo Island NSW_MW001 2019-03-22 05:00:00     2019-03-23      11:00
#> 2      Gabo Island NSW_MW001 2019-03-23 00:00:00     2019-03-24      11:00
#> 3      Gabo Island NSW_MW001 2019-03-24 00:00:00     2019-03-25      11:00
#> 4  Montague Island NSW_MW002 2019-03-22 05:00:00     2019-03-23      11:00
#> 5  Montague Island NSW_MW002 2019-03-23 00:00:00     2019-03-24      11:00
#> 6  Montague Island NSW_MW002 2019-03-24 00:00:00     2019-03-25      11:00
#> 7        Ulladulla NSW_MW003 2019-03-22 05:00:00     2019-03-23      11:00
#> 8        Ulladulla NSW_MW003 2019-03-23 00:00:00     2019-03-24      11:00
#> 9        Ulladulla NSW_MW003 2019-03-24 00:00:00     2019-03-25      11:00
#> 10    Port Hacking NSW_MW004 2019-03-22 05:00:00     2019-03-23      11:00
#> 11    Port Hacking NSW_MW004 2019-03-23 00:00:00     2019-03-24      11:00
#> 12    Port Hacking NSW_MW004 2019-03-24 00:00:00     2019-03-25      11:00
#> 13      Broken Bay NSW_MW005 2019-03-22 05:00:00     2019-03-23      11:00
#> 14      Broken Bay NSW_MW005 2019-03-23 00:00:00     2019-03-24      11:00
#> 15      Broken Bay NSW_MW005 2019-03-24 00:00:00     2019-03-25      11:00
#> 16      Seal Rocks NSW_MW006 2019-03-22 05:00:00     2019-03-23      11:00
#> 17      Seal Rocks NSW_MW006 2019-03-23 00:00:00     2019-03-24      11:00
#> 18      Seal Rocks NSW_MW006 2019-03-24 00:00:00     2019-03-25      11:00
#> 19      Smoky Cape NSW_MW007 2019-03-22 05:00:00     2019-03-23      11:00
#> 20      Smoky Cape NSW_MW007 2019-03-23 00:00:00     2019-03-24      11:00
#> 21      Smoky Cape NSW_MW007 2019-03-24 00:00:00     2019-03-25      11:00
#> 22           Wooli NSW_MW008 2019-03-22 05:00:00     2019-03-23      11:00
#> 23           Wooli NSW_MW008 2019-03-23 00:00:00     2019-03-24      11:00
#> 24           Wooli NSW_MW008 2019-03-24 00:00:00     2019-03-25      11:00
#>         start_time_utc        end_time_utc
#> 1  2019-03-21 18:00:00 2019-03-22 13:00:00
#> 2  2019-03-22 13:00:00 2019-03-23 13:00:00
#> 3  2019-03-23 13:00:00 2019-03-24 13:00:00
#> 4  2019-03-21 18:00:00 2019-03-22 13:00:00
#> 5  2019-03-22 13:00:00 2019-03-23 13:00:00
#> 6  2019-03-23 13:00:00 2019-03-24 13:00:00
#> 7  2019-03-21 18:00:00 2019-03-22 13:00:00
#> 8  2019-03-22 13:00:00 2019-03-23 13:00:00
#> 9  2019-03-23 13:00:00 2019-03-24 13:00:00
#> 10 2019-03-21 18:00:00 2019-03-22 13:00:00
#> 11 2019-03-22 13:00:00 2019-03-23 13:00:00
#> 12 2019-03-23 13:00:00 2019-03-24 13:00:00
#> 13 2019-03-21 18:00:00 2019-03-22 13:00:00
#> 14 2019-03-22 13:00:00 2019-03-23 13:00:00
#> 15 2019-03-23 13:00:00 2019-03-24 13:00:00
#> 16 2019-03-21 18:00:00 2019-03-22 13:00:00
#> 17 2019-03-22 13:00:00 2019-03-23 13:00:00
#> 18 2019-03-23 13:00:00 2019-03-24 13:00:00
#> 19 2019-03-21 18:00:00 2019-03-22 13:00:00
#> 20 2019-03-22 13:00:00 2019-03-23 13:00:00
#> 21 2019-03-23 13:00:00 2019-03-24 13:00:00
#> 22 2019-03-21 18:00:00 2019-03-22 13:00:00
#> 23 2019-03-22 13:00:00 2019-03-23 13:00:00
#> 24 2019-03-23 13:00:00 2019-03-24 13:00:00
#>                                                                 forecast_seas
#> 1                                                              Below 1 metre.
#> 2            Below 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 3                                                              1 to 2 metres.
#> 4                                                              Below 1 metre.
#> 5            Below 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 6  1 to 1.5 metres, decreasing below 1 metre during the afternoon or evening.
#> 7                                                              Below 1 metre.
#> 8                 Below 1 metre, increasing to 1 to 1.5 metres around midday.
#> 9                  1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 10                                                          Below 0.5 metres.
#> 11         Below 0.5 metres, increasing to around 1 metre during the morning.
#> 12                 1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 13                                                          Below 0.5 metres.
#> 14         Below 0.5 metres, increasing to around 1 metre during the morning.
#> 15              1 to 1.5 metres, decreasing below 1 metre during the morning.
#> 16                                                             Below 1 metre.
#> 17                                                             Below 1 metre.
#> 18              1 to 1.5 metres, decreasing below 1 metre during the morning.
#> 19                                                             Below 1 metre.
#> 20         Below 0.5 metres, increasing to around 1 metre during the morning.
#> 21                                                           1 to 1.5 metres.
#> 22                                                          Below 0.5 metres.
#> 23              Below 0.5 metres, increasing to around 1 metre around midday.
#> 24                                                           1 to 1.5 metres.
#>                                                                                               forecast_weather
#> 1                         Partly cloudy. The chance of a thunderstorm during this afternoon and early evening.
#> 2                                    Partly cloudy. The chance of a thunderstorm in the morning and afternoon.
#> 3                                                                                               Partly cloudy.
#> 4                                                  Partly cloudy. The chance of a thunderstorm this afternoon.
#> 5                              Partly cloudy. The chance of a thunderstorm in the afternoon and early evening.
#> 6                                                                                               Partly cloudy.
#> 7                                                  Partly cloudy. The chance of a thunderstorm this afternoon.
#> 8                                    Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 9                                                                                               Partly cloudy.
#> 10                                                 Partly cloudy. The chance of a thunderstorm this afternoon.
#> 11                                   Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 12                                                                                              Partly cloudy.
#> 13                                                 Partly cloudy. The chance of a thunderstorm this afternoon.
#> 14                           Mostly sunny day. The chance of a thunderstorm in the late afternoon and evening.
#> 15                                                                                              Partly cloudy.
#> 16 Partly cloudy. 50% chance of showers. The chance of a thunderstorm during this afternoon and early evening.
#> 17                                                                 Mostly sunny. The chance of a thunderstorm.
#> 18                                                                                               Mostly sunny.
#> 19 Partly cloudy. 50% chance of showers. The chance of a thunderstorm during this afternoon and early evening.
#> 20                                   Partly cloudy. The chance of a thunderstorm in the morning and afternoon.
#> 21                                                                                               Mostly sunny.
#> 22                              Partly cloudy. The chance of a thunderstorm during this afternoon and evening.
#> 23                                   Partly cloudy. The chance of a thunderstorm in the morning and afternoon.
#> 24                                                                                               Mostly sunny.
#>                                                                                                                                                           forecast_winds
#> 1                                                                                                                                               Variable about 10 knots.
#> 2                                                                                     East to northeasterly below 10 knots tending northerly 15 to 20 knots before dawn.
#> 3        North to northwesterly 15 to 20 knots shifting southerly 10 to 15 knots in the morning, then turning northeasterly 15 to 20 knots in the afternoon and evening.
#> 4                                                                                                                                               Variable about 10 knots.
#> 5                                                                                                  Northerly about 10 knots increasing to 15 to 20 knots in the morning.
#> 6       North to northwesterly 15 to 20 knots turning south to southeasterly during the morning then tending north to northeasterly 10 to 15 knots during the afternoon.
#> 7                                                                                   South to southwesterly about 10 knots becoming east to southeasterly in the evening.
#> 8                                                                                              Variable about 10 knots becoming northerly 15 to 20 knots in the morning.
#> 9                                                               North to northwesterly 15 to 20 knots shifting south to southeasterly 10 to 15 knots during the morning.
#> 10                                                                                  South to southeasterly below 10 knots becoming east to northeasterly in the evening.
#> 11                   North to northeasterly about 10 knots increasing to 10 to 15 knots in the middle of the day. Winds reaching up to 20 knots offshore in the evening.
#> 12                              North to northwesterly 10 to 15 knots shifting south to southeasterly during the day, then tending east to northeasterly in the evening.
#> 13                                                                             South to southeasterly below 10 knots becoming east to northeasterly in the late evening.
#> 14 North to northeasterly below 10 knots increasing to 10 to 15 knots in the morning. Winds reaching up to 20 knots offshore south of Port Stephens in the late evening.
#> 15                      North to northwesterly 10 to 15 knots turning south to southeasterly during the afternoon then tending east to northeasterly during the evening.
#> 16                                                                                  South to southwesterly about 10 knots becoming east to southeasterly in the evening.
#> 17            Variable below 10 knots becoming northerly 10 to 15 knots in the morning. Winds reaching up to 20 knots offshore north of Crowdy Head in the late evening.
#> 18                                                                            North to northwesterly 15 to 20 knots tending northerly 10 to 15 knots during the evening.
#> 19                                                                        South to southwesterly below 10 knots becoming east to southeasterly in the middle of the day.
#> 20                           Variable below 10 knots becoming northerly 10 to 15 knots early in the morning. Winds reaching up to 20 knots offshore in the late evening.
#> 21                                                                                                                                             Northerly 15 to 20 knots.
#> 22                                                                                               South to southeasterly below 10 knots becoming easterly in the evening.
#> 23                                                                                             Variable below 10 knots becoming northerly 10 to 15 knots in the morning.
#> 24                                                                                                                                North to northwesterly 15 to 20 knots.
#>             forecast_swell1 forecast_swell2 forecast_caution marine_forecast
#> 1  Easterly around 1 metre.              NA               NA              NA
#> 2  Easterly around 1 metre.              NA               NA              NA
#> 3  Easterly around 1 metre.              NA               NA              NA
#> 4  Easterly around 1 metre.              NA               NA              NA
#> 5  Easterly around 1 metre.              NA               NA              NA
#> 6  Easterly around 1 metre.              NA               NA              NA
#> 7  Easterly around 1 metre.              NA               NA              NA
#> 8  Easterly around 1 metre.              NA               NA              NA
#> 9  Easterly around 1 metre.              NA               NA              NA
#> 10 Easterly around 1 metre.              NA               NA              NA
#> 11 Easterly around 1 metre.              NA               NA              NA
#> 12 Easterly around 1 metre.              NA               NA              NA
#> 13 Easterly around 1 metre.              NA               NA              NA
#> 14 Easterly around 1 metre.              NA               NA              NA
#> 15 Easterly around 1 metre.              NA               NA              NA
#> 16 Easterly around 1 metre.              NA               NA              NA
#> 17 Easterly around 1 metre.              NA               NA              NA
#> 18 Easterly around 1 metre.              NA               NA              NA
#> 19 Easterly around 1 metre.              NA               NA              NA
#> 20 Easterly around 1 metre.              NA               NA              NA
#> 21 Easterly around 1 metre.              NA               NA              NA
#> 22 Easterly around 1 metre.              NA               NA              NA
#> 23 Easterly around 1 metre.              NA               NA              NA
#> 24 Easterly around 1 metre.              NA               NA              NA