Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     0   IDN11001 <NA>        NSW
#>  6:     1   IDN11001 <NA>        NSW
#>  7:     2   IDN11001 <NA>        NSW
#>  8:     0   IDN11001 <NA>        NSW
#>  9:     1   IDN11001 <NA>        NSW
#> 10:     2   IDN11001 <NA>        NSW
#> 11:     0   IDN11001 <NA>        NSW
#> 12:     1   IDN11001 <NA>        NSW
#> 13:     2   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     0   IDN11001 <NA>        NSW
#> 18:     1   IDN11001 <NA>        NSW
#> 19:     2   IDN11001 <NA>        NSW
#> 20:     0   IDN11001 <NA>        NSW
#> 21:     1   IDN11001 <NA>        NSW
#> 22:     2   IDN11001 <NA>        NSW
#> 23:     0   IDN11001 <NA>        NSW
#> 24:     1   IDN11001 <NA>        NSW
#> 25:     2   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#>  9:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 12:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 13:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 14:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 15:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 16:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 17:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 18:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 19:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 20:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 21:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 22:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 23:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 24:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 25:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-03-31 04:10:08 2021-03-31 04:10:08      11:00 2021-03-31 04:10:08
#>  2: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#>  3: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#>  4: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#>  5: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#>  6: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#>  7: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#>  8: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#>  9: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 10: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#> 11: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#> 12: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 13: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#> 14: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#> 15: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 16: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#> 17: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#> 18: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 19: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#> 20: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#> 21: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 22: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#> 23: 2021-03-31 05:00:00 2021-03-31 05:00:00      11:00 2021-03-31 05:00:00
#> 24: 2021-04-01 00:00:00 2021-04-01 00:00:00      11:00 2021-04-01 00:00:00
#> 25: 2021-04-02 00:00:00 2021-04-02 00:00:00      11:00 2021-04-02 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-03-31 04:10:08
#>  2: 2021-03-31 05:00:00
#>  3: 2021-04-01 00:00:00
#>  4: 2021-04-02 00:00:00
#>  5: 2021-03-31 05:00:00
#>  6: 2021-04-01 00:00:00
#>  7: 2021-04-02 00:00:00
#>  8: 2021-03-31 05:00:00
#>  9: 2021-04-01 00:00:00
#> 10: 2021-04-02 00:00:00
#> 11: 2021-03-31 05:00:00
#> 12: 2021-04-01 00:00:00
#> 13: 2021-04-02 00:00:00
#> 14: 2021-03-31 05:00:00
#> 15: 2021-04-01 00:00:00
#> 16: 2021-04-02 00:00:00
#> 17: 2021-03-31 05:00:00
#> 18: 2021-04-01 00:00:00
#> 19: 2021-04-02 00:00:00
#> 20: 2021-03-31 05:00:00
#> 21: 2021-04-01 00:00:00
#> 22: 2021-04-02 00:00:00
#> 23: 2021-03-31 05:00:00
#> 24: 2021-04-01 00:00:00
#> 25: 2021-04-02 00:00:00
#>            end_time_utc
#>                                                   forecast_seas
#>  1:                                                        <NA>
#>  2: 1 to 1.5 metres, decreasing below 1 metre by early evening.
#>  3:                                              Below 1 metre.
#>  4:                                              Below 1 metre.
#>  5:       Around 1 metre, increasing to 1 to 2 metres offshore.
#>  6:                                             Around 1 metre.
#>  7:                                              Below 1 metre.
#>  8:      1 to 1.5 metres, increasing to 1 to 2 metres offshore.
#>  9:                                            1 to 1.5 metres.
#> 10:                                              Below 1 metre.
#> 11:                                            1 to 1.5 metres.
#> 12:                                            1 to 1.5 metres.
#> 13:                                              Below 1 metre.
#> 14:                                            1 to 1.5 metres.
#> 15:                                            1 to 1.5 metres.
#> 16:                                              Below 1 metre.
#> 17:                                            1.5 to 2 metres.
#> 18:                                            1 to 1.5 metres.
#> 19:     Around 1 metre, increasing to 1 to 1.5 metres offshore.
#> 20:                                                   2 metres.
#> 21:                                            1.5 to 2 metres.
#> 22:                                            1 to 1.5 metres.
#> 23:                                            2 to 2.5 metres.
#> 24:                                            1.5 to 2 metres.
#> 25:                                            1.5 to 2 metres.
#>                                                   forecast_seas
#>                          forecast_weather
#>  1:                                  <NA>
#>  2:                        Partly cloudy.
#>  3:                         Mostly sunny.
#>  4:                         Mostly sunny.
#>  5:                        Partly cloudy.
#>  6:                         Mostly sunny.
#>  7:                         Mostly sunny.
#>  8: Partly cloudy. 50% chance of showers.
#>  9:                        Partly cloudy.
#> 10:                         Mostly sunny.
#> 11: Partly cloudy. 60% chance of showers.
#> 12:                        Partly cloudy.
#> 13:                        Partly cloudy.
#> 14: Partly cloudy. 60% chance of showers.
#> 15:                        Partly cloudy.
#> 16:                        Partly cloudy.
#> 17: Partly cloudy. 70% chance of showers.
#> 18: Partly cloudy. 50% chance of showers.
#> 19:                        Partly cloudy.
#> 20: Partly cloudy. 70% chance of showers.
#> 21: Partly cloudy. 70% chance of showers.
#> 22:                        Partly cloudy.
#> 23: Partly cloudy. 80% chance of showers.
#> 24: Partly cloudy. 70% chance of showers.
#> 25:                        Partly cloudy.
#>                          forecast_weather
#>                                                                                                                                               forecast_winds
#>  1:                                                                                                                                                     <NA>
#>  2:                                                             Southwesterly 15 to 20 knots turning south to southeasterly below 10 knots in the afternoon.
#>  3:                            Variable about 10 knots becoming northwesterly 10 to 15 knots in the morning then tending northerly in the middle of the day.
#>  4:                                                             Variable about 10 knots becoming north to northeasterly 10 to 15 knots during the afternoon.
#>  5:        South to southwesterly 10 to 15 knots reaching 20 knots offshore, tending south to southeasterly 10 to 15 knots during the afternoon and evening.
#>  6: South to southeasterly 10 to 15 knots becoming variable about 10 knots before dawn then becoming north to northeasterly 10 to 15 knots in the afternoon.
#>  7:                                                                                                                                 Variable about 10 knots.
#>  8:                                                                                   Southerly 15 to 20 knots, reaching 25 knots offshore in the afternoon.
#>  9:                                                                   South to southeasterly 10 to 15 knots becoming variable about 10 knots in the morning.
#> 10:                                                                                                                                 Variable about 10 knots.
#> 11:                                                                                Southerly 15 to 20 knots, reaching up to 25 knots inshore in the evening.
#> 12:                                                                Southerly 10 to 15 knots turning east to southeasterly below 10 knots during the morning.
#> 13:                                                                                                                                 Variable about 10 knots.
#> 14:                                                                                                                                Southerly 15 to 20 knots.
#> 15:                                                                      Southerly 10 to 15 knots turning southeasterly 15 to 20 knots early in the morning.
#> 16:                                                                    East to southeasterly about 10 knots increasing to 10 to 15 knots during the evening.
#> 17:                                                                                                                                Southerly 15 to 20 knots.
#> 18:                                                                          Southerly 15 to 20 knots turning southeasterly 10 to 15 knots in the afternoon.
#> 19:                                                                                        Southeasterly 10 to 15 knots turning easterly during the evening.
#> 20:                                                                                                                                Southerly 15 to 25 knots.
#> 21:                                                                                         Southerly 15 to 20 knots turning southeasterly in the afternoon.
#> 22:                                                                                                                            Southeasterly 15 to 20 knots.
#> 23:                                                                         Southerly 15 to 25 knots turning southeasterly during the afternoon and evening.
#> 24:                                                                                                                            Southeasterly 15 to 20 knots.
#> 25:                                                                                                                            Southeasterly 15 to 25 knots.
#>                                                                                                                                               forecast_winds
#>                                                                                                                              forecast_swell1
#>  1:                                                                                                                                     <NA>
#>  2:                                                                                                               Southerly 1 to 1.5 metres.
#>  3:                                                                                                                Southerly around 1 metre.
#>  4:                                                                                                             Southeasterly below 1 metre.
#>  5:                                                                                                               Southerly 1 to 1.5 metres.
#>  6:                                                                                                               Southerly 1 to 1.5 metres.
#>  7:                                                                                                             Southeasterly below 1 metre.
#>  8:                                                                                                               Southerly 1.5 to 2 metres.
#>  9: Southerly 1 to 1.5 metres, tending southeasterly 1 to 1.5 metres during the morning, then tending southerly below 1 metre around midday.
#> 10:                                                                                                             Southeasterly below 1 metre.
#> 11:                                                                    Southerly 1.5 to 2 metres, decreasing to 1.5 metres by early evening.
#> 12:                                                                Southerly 1 to 1.5 metres, decreasing below 1 metre during the afternoon.
#> 13:                                                                                                             Southeasterly below 1 metre.
#> 14:                                                                                                             Southerly 1.5 to 2.5 metres.
#> 15:                                                 Southerly 1.5 to 2 metres, tending south to southeasterly 1 to 1.5 metres around midday.
#> 16:                                                                                                            Southeasterly around 1 metre.
#> 17:                                                                                                               Southerly 2 to 2.5 metres.
#> 18:                                                                                                               Southerly 1.5 to 2 metres.
#> 19:                                                            Southeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 20:                                                                                                               Southerly 2 to 2.5 metres.
#> 21:                                                                                                               Southerly 1.5 to 2 metres.
#> 22:                                                          Southeasterly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 23:                                                                                                               Southerly 2 to 2.5 metres.
#> 24:                                                   Southerly 1.5 to 2 metres, tending southeasterly 1 to 1.5 metres during the afternoon.
#> 25:                                                                                                           Southeasterly 1 to 1.5 metres.
#>                                                                                                                              forecast_swell1
#>                    forecast_swell2 forecast_caution marine_forecast
#>  1:                           <NA>             <NA>            <NA>
#>  2:                           <NA>             <NA>            <NA>
#>  3:                           <NA>             <NA>            <NA>
#>  4:                           <NA>             <NA>            <NA>
#>  5:                           <NA>             <NA>            <NA>
#>  6:                           <NA>             <NA>            <NA>
#>  7:                           <NA>             <NA>            <NA>
#>  8:                           <NA>             <NA>            <NA>
#>  9:                           <NA>             <NA>            <NA>
#> 10:                           <NA>             <NA>            <NA>
#> 11:                           <NA>             <NA>            <NA>
#> 12:                           <NA>             <NA>            <NA>
#> 13:                           <NA>             <NA>            <NA>
#> 14:                           <NA>             <NA>            <NA>
#> 15:                           <NA>             <NA>            <NA>
#> 16:                           <NA>             <NA>            <NA>
#> 17:                           <NA>             <NA>            <NA>
#> 18:                           <NA>             <NA>            <NA>
#> 19:                           <NA>             <NA>            <NA>
#> 20:                           <NA>             <NA>            <NA>
#> 21:                           <NA>             <NA>            <NA>
#> 22:                           <NA>             <NA>            <NA>
#> 23:                           <NA>             <NA>            <NA>
#> 24:                           <NA>             <NA>            <NA>
#> 25: Southeasterly 1 to 1.5 metres.             <NA>            <NA>
#>                    forecast_swell2 forecast_caution marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#>     tropical_system_location forecast_waves

# }