Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     0   IDN11001 <NA>        NSW
#>  6:     1   IDN11001 <NA>        NSW
#>  7:     2   IDN11001 <NA>        NSW
#>  8:     0   IDN11001 <NA>        NSW
#>  9:     1   IDN11001 <NA>        NSW
#> 10:     2   IDN11001 <NA>        NSW
#> 11:     0   IDN11001 <NA>        NSW
#> 12:     1   IDN11001 <NA>        NSW
#> 13:     2   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     0   IDN11001 <NA>        NSW
#> 18:     1   IDN11001 <NA>        NSW
#> 19:     2   IDN11001 <NA>        NSW
#> 20:     0   IDN11001 <NA>        NSW
#> 21:     1   IDN11001 <NA>        NSW
#> 22:     2   IDN11001 <NA>        NSW
#> 23:     0   IDN11001 <NA>        NSW
#> 24:     1   IDN11001 <NA>        NSW
#> 25:     2   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#>  9:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 12:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 13:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 14:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 15:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 16:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 17:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 18:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 19:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 20:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 21:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 22:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 23:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 24:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 25:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-04-08 10:01:07 2021-04-08 10:01:07      10:00 2021-04-08 10:01:07
#>  2: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#>  3: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#>  4: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#>  5: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#>  6: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#>  7: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#>  8: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#>  9: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 10: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#> 11: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#> 12: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 13: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#> 14: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#> 15: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 16: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#> 17: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#> 18: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 19: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#> 20: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#> 21: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 22: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#> 23: 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00 2021-04-08 05:00:00
#> 24: 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00 2021-04-09 00:00:00
#> 25: 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00 2021-04-10 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-04-08 10:01:07
#>  2: 2021-04-08 05:00:00
#>  3: 2021-04-09 00:00:00
#>  4: 2021-04-10 00:00:00
#>  5: 2021-04-08 05:00:00
#>  6: 2021-04-09 00:00:00
#>  7: 2021-04-10 00:00:00
#>  8: 2021-04-08 05:00:00
#>  9: 2021-04-09 00:00:00
#> 10: 2021-04-10 00:00:00
#> 11: 2021-04-08 05:00:00
#> 12: 2021-04-09 00:00:00
#> 13: 2021-04-10 00:00:00
#> 14: 2021-04-08 05:00:00
#> 15: 2021-04-09 00:00:00
#> 16: 2021-04-10 00:00:00
#> 17: 2021-04-08 05:00:00
#> 18: 2021-04-09 00:00:00
#> 19: 2021-04-10 00:00:00
#> 20: 2021-04-08 05:00:00
#> 21: 2021-04-09 00:00:00
#> 22: 2021-04-10 00:00:00
#> 23: 2021-04-08 05:00:00
#> 24: 2021-04-09 00:00:00
#> 25: 2021-04-10 00:00:00
#>            end_time_utc
#>                                                                                                      forecast_seas
#>  1:                                                                                                           <NA>
#>  2:                                                1 to 1.5 metres increasing to 1.5 to 2 metres by early evening.
#>  3:                                                      1 to 2 metres, increasing to 2 to 4 metres around midday.
#>  4:                                                                                                 2 to 3 metres.
#>  5:                                               Below 1 metre, increasing to 1 to 2 metres during the afternoon.
#>  6:                                               1 to 2 metres, increasing to 2 to 3 metres during the afternoon.
#>  7:                                              1.5 to 2.5 metres, decreasing below 1 metre during the afternoon.
#>  8:                                                                                                 Below 1 metre.
#>  9:                                         1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the afternoon.
#> 10:                                                   1.5 to 2.5 metres, decreasing to 1 metre during the morning.
#> 11:                                                                                                 Below 1 metre.
#> 12: Around 1 metre, increasing to 1 to 1.5 metres around midday, then increasing to 2 metres later in the evening.
#> 13:                                                    1 to 2 metres, decreasing below 1 metre during the morning.
#> 14:                                                                                                Around 1 metre.
#> 15:                                                Around 1 metre, increasing to 1 to 2 metres during the morning.
#> 16:                                                                                                 1 to 2 metres.
#> 17:                                                                                               1 to 1.5 metres.
#> 18:                                                                                               1 to 1.5 metres.
#> 19:                                                     1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 20:                                                                                               1.5 to 2 metres.
#> 21:                                                            1 to 2 metres, decreasing to 1 metre around midday.
#> 22:                                                     1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 23:                                                                                               1.5 to 2 metres.
#> 24:                                                                                                 1 to 2 metres.
#> 25:                                                                                                Around 1 metre.
#>                                                                                                      forecast_seas
#>                                                                                         forecast_weather
#>  1:                                                                                                 <NA>
#>  2: The chance of a thunderstorm offshore north of Merimbula early this morning. Mostly sunny afternoon.
#>  3:                                                                Partly cloudy. 50% chance of showers.
#>  4:                                                                                       Partly cloudy.
#>  5:                                   The chance of a thunderstorm this morning. Mostly sunny afternoon.
#>  6:                                          Partly cloudy. 50% chance of showers from the late morning.
#>  7:                                                                                       Partly cloudy.
#>  8: Partly cloudy. 60% chance of showers. The chance of a thunderstorm during the morning and afternoon.
#>  9:                                   Partly cloudy. 50% chance of showers in the afternoon and evening.
#> 10:                                                                                       Partly cloudy.
#> 11: Partly cloudy. 80% chance of showers. The chance of a thunderstorm during the morning and afternoon.
#> 12:                                                                                       Partly cloudy.
#> 13:                                                                                       Partly cloudy.
#> 14:        Cloudy. 90% chance of showers. The chance of a thunderstorm during the morning and afternoon.
#> 15:                                                                                        Mostly sunny.
#> 16:                                                                                       Partly cloudy.
#> 17:                                   Cloudy. Near 100% chance of showers. The chance of a thunderstorm.
#> 18:                                                                                               Sunny.
#> 19:                                                                                       Partly cloudy.
#> 20: Cloudy. Near 100% chance of showers. The chance of a thunderstorm during this afternoon and evening.
#> 21:                                                                                               Sunny.
#> 22:                                                                                       Partly cloudy.
#> 23:       Cloudy. 90% chance of showers. The chance of a thunderstorm during this afternoon and evening.
#> 24:                                                                                        Mostly sunny.
#> 25:                                                                                        Mostly sunny.
#>                                                                                         forecast_weather
#>                                                                                                                                                                         forecast_winds
#>  1:                                                                                                                                                                               <NA>
#>  2:                                                                                                              Northerly about 10 knots increasing to 15 to 25 knots in the morning.
#>  3:                    North to northwesterly 15 to 20 knots shifting southwesterly 20 to 30 knots during the morning. Winds reaching up to 35 knots during the afternoon and evening.
#>  4:                                                                                                   Southwesterly 20 to 30 knots tending westerly 15 to 25 knots during the evening.
#>  5:                     North to northeasterly about 10 knots increasing to 15 to 20 knots in the early afternoon. Winds reaching up to 25 knots south of Moruya Heads in the evening.
#>  6:                  North to northwesterly 15 to 20 knots shifting southerly 20 to 25 knots in the middle of the day. Winds reaching up to 30 knots during the afternoon and evening.
#>  7:                                                 Southwesterly 20 to 30 knots decreasing to 15 to 20 knots during the morning then becoming variable about 10 knots during the day.
#>  8:                                   Variable about 10 knots becoming east to northeasterly 10 to 15 knots in the early afternoon then tending north to northeasterly in the evening.
#>  9:                               North to northwesterly 15 to 20 knots shifting southerly 15 to 25 knots in the early afternoon. Winds reaching up to 30 knots in the late afternoon.
#> 10:                                                                                              South to southwesterly 15 to 25 knots turning westerly 10 to 15 knots during the day.
#> 11:                                                                                              East to southeasterly 10 to 15 knots becoming variable below 10 knots in the evening.
#> 12:                                                                                             West to northwesterly 10 to 15 knots shifting southerly 20 to 25 knots in the evening.
#> 13:                                                                                                         Southerly 15 to 20 knots turning west to southwesterly during the morning.
#> 14:                                                                                                   Southeasterly 10 to 15 knots turning south to southwesterly in the late evening.
#> 15:                                                     Variable about 10 knots becoming westerly 15 to 20 knots before dawn then turning southerly20 to 25 knots in the late evening.
#> 16:                                                                                                             Southerly 15 to 20 knots turning west to southwesterly during the day.
#> 17:                              Southeasterly 10 to 15 knots, reaching up to 20 knots offshore during the day. Winds turning south to southwesterly during the afternoon and evening.
#> 18:                                                 Southwesterly 15 to 20 knots tending west to northwesterly in the morning then tending south to southwesterly in the late evening.
#> 19:               Southerly 15 to 20 knots tending northeast to southeasterly 10 to 15 knots during the day then tending northwest to southwesterly 15 to 20 knots during the evening.
#> 20:                                   South to southeasterly 15 to 20 knots tending south to southwesterly during the day. Winds reaching up to 25 knots offshore in the late evening.
#> 21:                  Southwesterly 10 to 15 knots, reaching up to 20 knots offshore during the morning. Winds shifting north to northwesterly in the late morning and early afternoon.
#> 22:           Northwest to southwesterly 15 to 20 knots tending south to southeasterly 10 to 15 knots during the morning then tending northeast to southeasterly during the afternoon.
#> 23:                                                                                                                                             South to southwesterly 15 to 25 knots.
#> 24: Southwesterly 10 to 15 knots, reaching up to 20 knots offshore early in the morning. Winds shifting north to northwesterly 10 to 15 knots in the late morning and early afternoon.
#> 25:                                                                                   Northwest to southwesterly 10 to 15 knots tending northeast to southeasterly during the morning.
#>                                                                                                                                                                         forecast_winds
#>                                                                                                                               forecast_swell1
#>  1:                                                                                                                                      <NA>
#>  2:                                                                                                             Northeasterly around 1 metre.
#>  3:                                                                                                            Northeasterly 1 to 1.5 metres.
#>  4:                                                                Southerly 1 to 1.5 metres, increasing to 1 to 2 metres during the morning.
#>  5:                                                             Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#>  6:                                                                                                              Southerly 1.5 to 2.5 metres.
#>  7:                                                                                                              Southerly 1.5 to 2.5 metres.
#>  8:                                                                                                                Southerly 1 to 1.5 metres.
#>  9:                                                                                                                  Southerly 1 to 2 metres.
#> 10:                                                                                                              Southerly 1.5 to 2.5 metres.
#> 11:                                                            Northeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 12:                     Northeasterly 2 metres, tending easterly 2 metres during the morning, then decreasing to 1.5 metres by early evening.
#> 13:                                                                Southerly 1 to 1.5 metres, increasing to 2 to 3 metres during the morning.
#> 14:                                                            Southerly 1 to 1.5 metres, increasing to 1.5 to 2 metres later in the evening.
#> 15:                                                                      Easterly 2 to 2.5 metres, decreasing to 1.5 metres by early evening.
#> 16:                                                                  Southerly below 1 metre, increasing to 2 to 3 metres during the morning.
#> 17:                                                                                                                Southerly 1.5 to 2 metres.
#> 18: East to southeasterly 2 to 3 metres, decreasing to 2 metres around midday, then tending southeasterly 1 to 2 metres during the afternoon.
#> 19:                                                                                                              Southerly 1.5 to 2.5 metres.
#> 20:                                                                          Southerly 1.5 metres, increasing to 2 metres during the morning.
#> 21:                                                            Southeasterly 2 to 3 metres, decreasing to 1 to 2 metres during the afternoon.
#> 22:                                                               Southerly around 1 metre, increasing to 1 to 2 metres during the afternoon.
#> 23:                                                                                                              Southerly 1.5 to 2.5 metres.
#> 24:                                                              Southerly 2 to 3 metres, decreasing to 1.5 to 2 metres during the afternoon.
#> 25:                    South to southeasterly below 1 metre, increasing to 1 to 2 metres south of Cape Byron during the afternoon or evening.
#>                                                                                                                               forecast_swell1
#>                                                                                           forecast_swell2
#>  1:                                                                                                  <NA>
#>  2:                                                                                                  <NA>
#>  3:                                                                             Southerly around 1 metre.
#>  4:                    Northeasterly 1 to 1.5 metres, tending easterly around 1 metre during the morning.
#>  5:                                                                                                  <NA>
#>  6:                                                                        Northeasterly 1 to 1.5 metres.
#>  7:                                 Easterly 1.5 metres, decreasing to around 1 metre during the morning.
#>  8:                                                                             Northeasterly 1.5 metres.
#>  9:                                                                East to northeasterly 1.5 to 2 metres.
#> 10:                                                                East to southeasterly 1 to 1.5 metres.
#> 11:            Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore by early evening.
#> 12:                                                                                                  <NA>
#> 13:                                                                East to southeasterly 1 to 1.5 metres.
#> 14: Northeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore south of Port Stephens.
#> 15:                                                                                                  <NA>
#> 16:                                                                             Easterly 1 to 1.5 metres.
#> 17:                                                                East to northeasterly 1.5 to 2 metres.
#> 18:                                                                                                  <NA>
#> 19:                                                                             Easterly 1 to 1.5 metres.
#> 20:                                                                East to northeasterly 1.5 to 2 metres.
#> 21:                                                                                                  <NA>
#> 22:      Southeasterly 1 to 1.5 metres, tending east to northeasterly below 1 metre during the afternoon.
#> 23:                                                                             Easterly 1.5 to 2 metres.
#> 24:                                                                                                  <NA>
#> 25:                                                               South to southeasterly 1 to 1.5 metres.
#>                                                                                           forecast_swell2
#>     forecast_caution                                    marine_forecast
#>  1:             <NA>                                               <NA>
#>  2:             <NA>                                               <NA>
#>  3:             <NA>             Gale Warning for Friday for Eden Coast
#>  4:             <NA>                                               <NA>
#>  5:             <NA>                                               <NA>
#>  6:             <NA>  Strong Wind Warning for Friday for Batemans Coast
#>  7:             <NA>                                               <NA>
#>  8:             <NA>                                               <NA>
#>  9:             <NA> Strong Wind Warning for Friday for Illawarra Coast
#> 10:             <NA>                                               <NA>
#> 11:             <NA>                                               <NA>
#> 12:             <NA>                                               <NA>
#> 13:             <NA>                                               <NA>
#> 14:             <NA>                                               <NA>
#> 15:             <NA>                                               <NA>
#> 16:             <NA>                                               <NA>
#> 17:             <NA>                                               <NA>
#> 18:             <NA>                                               <NA>
#> 19:             <NA>                                               <NA>
#> 20:             <NA>                                               <NA>
#> 21:             <NA>                                               <NA>
#> 22:             <NA>                                               <NA>
#> 23:             <NA>                                               <NA>
#> 24:             <NA>                                               <NA>
#> 25:             <NA>                                               <NA>
#>     forecast_caution                                    marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#>     tropical_system_location forecast_waves

# }