Get BOM Coastal Waters Forecast

Fetch the BOM daily Coastal Waters Forecast and return a tidy data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to "AUS" returning all state forecasts, see details for further information.

Value

Tidy data.table of a Australia BOM Coastal Waters Forecast.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

Examples

coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2019-04-03 16:05:04 2019-04-03 16:05:04      11:00 2019-04-03 16:05:04
#>  2: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#>  3: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#>  4: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#>  5: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#>  6: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#>  7: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#>  8: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#>  9: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 10: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 11: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 12: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 13: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 14: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 15: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 16: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 17: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 18: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 19: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 20: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 21: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 22: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 23: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 24: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 25: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 26: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 27: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 28: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 29: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#> 30: 2019-04-03 17:00:00 2019-04-03 17:00:00      11:00 2019-04-03 17:00:00
#> 31: 2019-04-04 00:00:00 2019-04-04 00:00:00      11:00 2019-04-04 00:00:00
#> 32: 2019-04-05 00:00:00 2019-04-05 00:00:00      11:00 2019-04-05 00:00:00
#> 33: 2019-04-06 00:00:00 2019-04-06 00:00:00      11:00 2019-04-06 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2019-04-03 16:05:04
#>  2: 2019-04-03 17:00:00
#>  3: 2019-04-04 00:00:00
#>  4: 2019-04-05 00:00:00
#>  5: 2019-04-06 00:00:00
#>  6: 2019-04-03 17:00:00
#>  7: 2019-04-04 00:00:00
#>  8: 2019-04-05 00:00:00
#>  9: 2019-04-06 00:00:00
#> 10: 2019-04-03 17:00:00
#> 11: 2019-04-04 00:00:00
#> 12: 2019-04-05 00:00:00
#> 13: 2019-04-06 00:00:00
#> 14: 2019-04-03 17:00:00
#> 15: 2019-04-04 00:00:00
#> 16: 2019-04-05 00:00:00
#> 17: 2019-04-06 00:00:00
#> 18: 2019-04-03 17:00:00
#> 19: 2019-04-04 00:00:00
#> 20: 2019-04-05 00:00:00
#> 21: 2019-04-06 00:00:00
#> 22: 2019-04-03 17:00:00
#> 23: 2019-04-04 00:00:00
#> 24: 2019-04-05 00:00:00
#> 25: 2019-04-06 00:00:00
#> 26: 2019-04-03 17:00:00
#> 27: 2019-04-04 00:00:00
#> 28: 2019-04-05 00:00:00
#> 29: 2019-04-06 00:00:00
#> 30: 2019-04-03 17:00:00
#> 31: 2019-04-04 00:00:00
#> 32: 2019-04-05 00:00:00
#> 33: 2019-04-06 00:00:00
#>            end_time_utc
#>                                                                                                                   forecast_seas
#>  1:                                                                                                                        <NA>
#>  2:                                                                                                              Below 1 metre.
#>  3:                                                                 1 to 1.5 metres, increasing to 1 to 2 metres around midday.
#>  4: 1 to 1.5 metres, decreasing to 1 metre during the morning, increasing to 1 to 1.5 metres south of Green Cape around midday.
#>  5:                                                       1 to 2 metres, decreasing to 1 metre during the afternoon or evening.
#>  6:                                                                                                              Below 1 metre.
#>  7:                                                            Below 1 metre, increasing to 1 to 1.5 metres during the morning.
#>  8:                                                                  1 to 1.5 metres, decreasing to 1 metre during the morning.
#>  9:                                                                     Around 1 metre, increasing to 1 to 1.5 metres offshore.
#> 10:                                                                                                              Below 1 metre.
#> 11:                                     Below 1 metre, increasing to 1 to 1.5 metres south of Port Kembla during the afternoon.
#> 12:                                                                     Around 1 metre, increasing to 1 to 1.5 metres offshore.
#> 13:        Around 1 metre, increasing to 1 to 1.5 metres offshore south of Point Perpendicular during the afternoon or evening.
#> 14:                                                                                                           Below 0.5 metres.
#> 15:                                                                                                              Below 1 metre.
#> 16:                                                                                                             Around 1 metre.
#> 17:                                                                                                             Around 1 metre.
#> 18:                                                                                                           Below 0.5 metres.
#> 19:                                                                                                              Below 1 metre.
#> 20:                                                                                                             Around 1 metre.
#> 21:                                                                                                             Around 1 metre.
#> 22:                                                                                                              Below 1 metre.
#> 23:                                                                                                              Below 1 metre.
#> 24:                                                                                                              Below 1 metre.
#> 25:                                                                                                              Below 1 metre.
#> 26:                                                                                                             Around 1 metre.
#> 27:                                              Around 1 metre, increasing to 1 to 1.5 metres offshore north of Coffs Harbour.
#> 28:                           Around 1 metre, increasing to 1 to 1.5 metres offshore north of Coffs Harbour during the morning.
#> 29:                                                                                                              Below 1 metre.
#> 30:                                                                                                            1 to 1.5 metres.
#> 31:                                                                      1 to 1.5 metres, increasing to 1 to 2 metres offshore.
#> 32:                                                                                                            1 to 1.5 metres.
#> 33:                                                      Around 1 metre, increasing to 1 to 1.5 metres offshore north of Yamba.
#>                                                                                                                   forecast_seas
#>                                                                                                    forecast_weather
#>  1:                                                                                                            <NA>
#>  2:                                                                                                  Partly cloudy.
#>  3:   Cloudy. 90% chance of showers. The chance of a thunderstorm north of Green Cape in the afternoon and evening.
#>  4:                                                                                                  Partly cloudy.
#>  5:                                       Partly cloudy. The chance of a thunderstorm in the morning and afternoon.
#>  6:                                                                                                  Partly cloudy.
#>  7:                          Cloudy. 90% chance of rain. The chance of a thunderstorm in the afternoon and evening.
#>  8:                                                                           Partly cloudy. 50% chance of showers.
#>  9:                                                                    Partly cloudy. The chance of a thunderstorm.
#> 10:                                                                                                   Mostly clear.
#> 11:           Partly cloudy. 80% chance of showers. The chance of a thunderstorm in the late afternoon and evening.
#> 12:               Partly cloudy. 60% chance of showers. The chance of a thunderstorm offshore in the early morning.
#> 13:                                       Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 14:                                                                                                   Mostly clear.
#> 15:                              Partly cloudy. 70% chance of showers. The chance of a thunderstorm in the evening.
#> 16:               Partly cloudy. 70% chance of showers. The chance of a thunderstorm offshore in the early morning.
#> 17:                                                     Partly cloudy. The chance of a thunderstorm in the evening.
#> 18:                                                                                                   Mostly clear.
#> 19: Partly cloudy. 50% chance of showers. The chance of a thunderstorm offshore south of Norah Head in the evening.
#> 20:                                                                           Partly cloudy. 60% chance of showers.
#> 21:                                                                                                  Partly cloudy.
#> 22:                                                                                                   Mostly clear.
#> 23:                                                                                                  Partly cloudy.
#> 24:                                                                           Partly cloudy. 60% chance of showers.
#> 25:                                                                                                  Partly cloudy.
#> 26:                                                                                                  Partly cloudy.
#> 27:                                                                                                  Partly cloudy.
#> 28:                                                                           Partly cloudy. 50% chance of showers.
#> 29:                                                                                                  Partly cloudy.
#> 30:                                  Partly cloudy. The chance of a thunderstorm north of Yamba early this evening.
#> 31:                                                                           Partly cloudy. 50% chance of showers.
#> 32:        Partly cloudy. 50% chance of showers. The chance of a thunderstorm north of Cape Byron in the afternoon.
#> 33:                                                                                                  Partly cloudy.
#>                                                                                                    forecast_weather
#>                                                                                                                                                                                                forecast_winds
#>  1:                                                                                                                                                                                                      <NA>
#>  2:                                                      Variable about 10 knots becoming southerly 10 to 15 knots in the evening. Winds reaching up to 20 knots offshore south of Green Cape in the evening.
#>  3:                                                Southerly 15 to 20 knots turning east to southeasterly during the morning. Winds reaching up to 25 knots offshore south of Green Cape in the late morning.
#>  4:                                                                                                                                  Easterly 10 to 15 knots turning northerly 15 to 20 knots in the evening.
#>  5:                                                                                                                            Northerly 15 to 25 knots shifting southerly 15 to 20 knots during the evening.
#>  6:                                                                                                                                                                                  Variable about 10 knots.
#>  7: Variable about 10 knots becoming southerly 10 to 15 knots before dawn then turning east to southeasterly in the early afternoon. Winds reaching up to 20 knots offshore during the afternoon and evening.
#>  8:                                                                                                                           East to southeasterly 10 to 15 knots turning northeasterly in the late evening.
#>  9:                                                                                                                                        Northeasterly 10 to 15 knots turning northerly during the morning.
#> 10:                                                                                                                                                                    North to northeasterly 10 to 15 knots.
#> 11:                                  Variable about 10 knots becoming south to southeasterly 10 to 15 knots early in the morning. Winds reaching up to 20 knots offshore south of Port Kembla in the evening.
#> 12:                                                                                                                           Southeasterly 10 to 15 knots turning east to northeasterly in the late evening.
#> 13:                                                                                                                                        Northeasterly 10 to 15 knots turning northerly during the morning.
#> 14:                                                                                                                                                                             Northeasterly about 10 knots.
#> 15:                                                                               Variable about 10 knots becoming southerly 10 to 15 knots in the early afternoon then tending southeasterly in the evening.
#> 16:                                                                     Southeasterly 10 to 15 knots tending southerly in the morning then shifting east to northeasterly below 10 knots in the late evening.
#> 17:                                                                                                              Variable about 10 knots becoming north to northeasterly 15 to 20 knots during the afternoon.
#> 18:                                                                                                                                                                     East to northeasterly about 10 knots.
#> 19:                                                                                                                      East to northeasterly about 10 knots becoming south to southeasterly in the morning.
#> 20:                                       South to southeasterly 10 to 15 knots tending south to southwesterly in the early afternoon then shifting east to northeasterly below 10 knots in the late evening.
#> 21:                                                                Variable about 10 knots becoming northwest to northeasterly 10 to 15 knots during the day then tending northeasterly during the afternoon.
#> 22:                                                                                                                                                                     East to southeasterly about 10 knots.
#> 23:                                                                                                                          Southeasterly about 10 knots increasing to 10 to 15 knots in the late afternoon.
#> 24:                                                                 Southeasterly 10 to 15 knots turning south to southwesterly during the morning then tending south to southeasterly in the late afternoon.
#> 25:                                                        Variable about 10 knots becoming east to northeasterly 10 to 15 knots during the afternoon then tending north to northeasterly during the evening.
#> 26:                                                                                                                                                                             Southeasterly 10 to 15 knots.
#> 27:                                                                                                                                                                             Southeasterly 10 to 15 knots.
#> 28:                                                                                                                                                                             Southeasterly 15 to 20 knots.
#> 29:                                                                                                                                                                                  Variable about 10 knots.
#> 30:                                                                                                                                                                             Southeasterly 15 to 20 knots.
#> 31:                                                                                                                                                                             Southeasterly 15 to 20 knots.
#> 32:                                                                                                                                                                             Southeasterly 15 to 20 knots.
#> 33:                                                                                                                                                                             Southeasterly 10 to 15 knots.
#>                                                                                                                                                                                                forecast_winds
#>                                                                                                                      forecast_swell1
#>  1:                                                                                                                             <NA>
#>  2:                                                                                           South to southeasterly around 1 metre.
#>  3:                                                Southerly around 1 metre, tending southeasterly 1 to 2 metres during the morning.
#>  4:                                                                                                        Southeasterly 1.5 metres.
#>  5:                                          East to southeasterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#>  6:                                                                                                        Southerly around 1 metre.
#>  7:                                              Southerly around 1 metre, tending southeasterly 1 to 2 metres during the afternoon.
#>  8:                                                      Southeasterly 1.5 to 2 metres, decreasing to 1.5 metres during the morning.
#>  9:                                                     Southeasterly 1 to 1.5 metres inshore, increasing to 1 to 2 metres offshore.
#> 10:                                                                                                        Southerly around 1 metre.
#> 11:          Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore south of Point Perpendicular by early evening.
#> 12:                                               Southerly 1 to 2 metres, tending southeasterly 1.5 to 2 metres during the morning.
#> 13:                                                                                                   Southeasterly 1 to 1.5 metres.
#> 14:                                                                                                        Southerly around 1 metre.
#> 15:                                                                                                        Southerly around 1 metre.
#> 16: Southerly around 1 metre, increasing to 1.5 to 2 metres during the morning, then tending southeasterly 1.5 metres around midday.
#> 17:                                                                                                        Southeasterly 1.5 metres.
#> 18:                                    Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Norah Head.
#> 19:                                                       Easterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 20:                                                         Southerly below 1 metre, increasing to 1 to 2 metres during the morning.
#> 21:                                                                                                        Southeasterly 1.5 metres.
#> 22:                                                        Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 23:                                                                                                        Easterly 1 to 1.5 metres.
#> 24:                                                      Southerly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 25:                                                     Southerly 1.5 metres, tending southeasterly 1.5 metres during the afternoon.
#> 26:                                 Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Coffs Harbour.
#> 27:                                      Easterly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore north of Coffs Harbour.
#> 28:            Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore south of Coffs Harbour during the afternoon.
#> 29:                                                        Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 30:                                                                                                        Easterly 1.5 to 2 metres.
#> 31:                                                                                                        Easterly 1.5 to 2 metres.
#> 32:                                         Easterly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore north of Cape Byron.
#> 33:                                                                                                             Easterly 1.5 metres.
#>                                                                                                                      forecast_swell1
#>                                                                                     forecast_swell2
#>  1:                                                                                            <NA>
#>  2:                                                                                            <NA>
#>  3:                                                                                            <NA>
#>  4:                                                                                            <NA>
#>  5:                      Easterly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#>  6:                                                                                            <NA>
#>  7:                                                                        Easterly around 1 metre.
#>  8:                                                                        Easterly around 1 metre.
#>  9:                      Easterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 10:                                                                        Easterly around 1 metre.
#> 11:                                                                        Easterly around 1 metre.
#> 12:                                                                        Easterly around 1 metre.
#> 13:                                                                        Easterly around 1 metre.
#> 14:                        Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 15:                                                                        Easterly around 1 metre.
#> 16:                                                                        Easterly around 1 metre.
#> 17:                                                                        Easterly around 1 metre.
#> 18:                                                                       Easterly 1 to 1.5 metres.
#> 19:                                                                       Southerly around 1 metre.
#> 20: Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Port Stephens.
#> 21:                                                                        Easterly around 1 metre.
#> 22:                                                                            Easterly 1.5 metres.
#> 23:                                                                       Southerly around 1 metre.
#> 24:                                                                       Easterly 1 to 1.5 metres.
#> 25:                                                                        Easterly around 1 metre.
#> 26:     Easterly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore north of Coffs Harbour.
#> 27:                                                                       Southerly around 1 metre.
#> 28:                                                                            Easterly 1.5 metres.
#> 29:                        Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 30:                        Southerly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 31:                                                                       Southerly around 1 metre.
#> 32:                                                                                            <NA>
#> 33:        Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore south of Yamba.
#>                                                                                     forecast_swell2
#>     forecast_caution marine_forecast tropical_system_location forecast_waves
#>  1:             <NA>            <NA>                       NA             NA
#>  2:             <NA>            <NA>                       NA             NA
#>  3:             <NA>            <NA>                       NA             NA
#>  4:             <NA>            <NA>                       NA             NA
#>  5:             <NA>            <NA>                       NA             NA
#>  6:             <NA>            <NA>                       NA             NA
#>  7:             <NA>            <NA>                       NA             NA
#>  8:             <NA>            <NA>                       NA             NA
#>  9:             <NA>            <NA>                       NA             NA
#> 10:             <NA>            <NA>                       NA             NA
#> 11:             <NA>            <NA>                       NA             NA
#> 12:             <NA>            <NA>                       NA             NA
#> 13:             <NA>            <NA>                       NA             NA
#> 14:             <NA>            <NA>                       NA             NA
#> 15:             <NA>            <NA>                       NA             NA
#> 16:             <NA>            <NA>                       NA             NA
#> 17:             <NA>            <NA>                       NA             NA
#> 18:             <NA>            <NA>                       NA             NA
#> 19:             <NA>            <NA>                       NA             NA
#> 20:             <NA>            <NA>                       NA             NA
#> 21:             <NA>            <NA>                       NA             NA
#> 22:             <NA>            <NA>                       NA             NA
#> 23:             <NA>            <NA>                       NA             NA
#> 24:             <NA>            <NA>                       NA             NA
#> 25:             <NA>            <NA>                       NA             NA
#> 26:             <NA>            <NA>                       NA             NA
#> 27:             <NA>            <NA>                       NA             NA
#> 28:             <NA>            <NA>                       NA             NA
#> 29:             <NA>            <NA>                       NA             NA
#> 30:             <NA>            <NA>                       NA             NA
#> 31:             <NA>            <NA>                       NA             NA
#> 32:             <NA>            <NA>                       NA             NA
#> 33:             <NA>            <NA>                       NA             NA
#>     forecast_caution marine_forecast tropical_system_location forecast_waves