Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     0   IDN11001 <NA>        NSW
#>  6:     1   IDN11001 <NA>        NSW
#>  7:     2   IDN11001 <NA>        NSW
#>  8:     0   IDN11001 <NA>        NSW
#>  9:     1   IDN11001 <NA>        NSW
#> 10:     2   IDN11001 <NA>        NSW
#> 11:     0   IDN11001 <NA>        NSW
#> 12:     1   IDN11001 <NA>        NSW
#> 13:     2   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     0   IDN11001 <NA>        NSW
#> 18:     1   IDN11001 <NA>        NSW
#> 19:     2   IDN11001 <NA>        NSW
#> 20:     0   IDN11001 <NA>        NSW
#> 21:     1   IDN11001 <NA>        NSW
#> 22:     2   IDN11001 <NA>        NSW
#> 23:     0   IDN11001 <NA>        NSW
#> 24:     1   IDN11001 <NA>        NSW
#> 25:     2   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#>  9:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 12:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 13:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 14:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 15:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 16:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 17:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 18:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 19:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 20:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 21:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 22:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 23:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 24:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 25:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-04-11 10:00:11 2021-04-11 10:00:11      10:00 2021-04-11 10:00:11
#>  2: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#>  3: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#>  4: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#>  5: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#>  6: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#>  7: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#>  8: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#>  9: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 10: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#> 11: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#> 12: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 13: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#> 14: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#> 15: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 16: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#> 17: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#> 18: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 19: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#> 20: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#> 21: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 22: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#> 23: 2021-04-11 10:00:00 2021-04-11 10:00:00      10:00 2021-04-11 10:00:00
#> 24: 2021-04-12 00:00:00 2021-04-12 00:00:00      10:00 2021-04-12 00:00:00
#> 25: 2021-04-13 00:00:00 2021-04-13 00:00:00      10:00 2021-04-13 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-04-11 10:00:11
#>  2: 2021-04-11 10:00:00
#>  3: 2021-04-12 00:00:00
#>  4: 2021-04-13 00:00:00
#>  5: 2021-04-11 10:00:00
#>  6: 2021-04-12 00:00:00
#>  7: 2021-04-13 00:00:00
#>  8: 2021-04-11 10:00:00
#>  9: 2021-04-12 00:00:00
#> 10: 2021-04-13 00:00:00
#> 11: 2021-04-11 10:00:00
#> 12: 2021-04-12 00:00:00
#> 13: 2021-04-13 00:00:00
#> 14: 2021-04-11 10:00:00
#> 15: 2021-04-12 00:00:00
#> 16: 2021-04-13 00:00:00
#> 17: 2021-04-11 10:00:00
#> 18: 2021-04-12 00:00:00
#> 19: 2021-04-13 00:00:00
#> 20: 2021-04-11 10:00:00
#> 21: 2021-04-12 00:00:00
#> 22: 2021-04-13 00:00:00
#> 23: 2021-04-11 10:00:00
#> 24: 2021-04-12 00:00:00
#> 25: 2021-04-13 00:00:00
#>            end_time_utc
#>                                                                                   forecast_seas
#>  1:                                                                                        <NA>
#>  2:                                                                              3 to 5 metres.
#>  3:                              2.5 to 4 metres, decreasing to 2 to 3 metres by early evening.
#>  4:                                     Around 1 metre, increasing to 1 to 1.5 metres offshore.
#>  5: 2 to 3 metres, increasing to 3 to 4 metres offshore south of Moruya Heads by early evening.
#>  6:                            2 to 4 metres, decreasing to 1.5 to 2.5 metres by early evening.
#>  7:             1 to 1.5 metres, increasing to 1.5 to 2 metres during the afternoon or evening.
#>  8:                                                                              2 to 3 metres.
#>  9:                                                                              2 to 3 metres.
#> 10:                                                                            1 to 1.5 metres.
#> 11:                                      1 to 2 metres, increasing to 2.5 to 3 metres offshore.
#> 12:                                                                          1.5 to 2.5 metres.
#> 13:                                                                            1 to 1.5 metres.
#> 14:                                                                              2 to 3 metres.
#> 15:                                                                          1.5 to 2.5 metres.
#> 16:                                     Around 1 metre, increasing to 1 to 1.5 metres offshore.
#> 17:                                                                              2 to 3 metres.
#> 18:                              1.5 to 2.5 metres, decreasing to 1 metre later in the evening.
#> 19:                  Around 1 metre, increasing to 1 to 1.5 metres offshore during the evening.
#> 20:                            1 to 2 metres, increasing to 2 to 3 metres later in the evening.
#> 21:                              1.5 to 2.5 metres, decreasing to 1 metre during the afternoon.
#> 22:      Around 1 metre, increasing to 1 to 1.5 metres inshore during the afternoon or evening.
#> 23:                                                                            1 to 1.5 metres.
#> 24:                                      1 to 2 metres, decreasing to 1 metre by early evening.
#> 25:                                                                             Around 1 metre.
#>                                                                                   forecast_seas
#>                                                                                                                            forecast_weather
#>  1:                                                                                                                                    <NA>
#>  2:                  Partly cloudy. 95% chance of showers offshore south of Green Cape, 60% chance elsewhere. The chance of a thunderstorm.
#>  3:                                                                                                          Cloudy. 50% chance of showers.
#>  4:                                                                                                                           Mostly sunny.
#>  5: Partly cloudy. 50% chance of showers offshore, near zero chance elsewhere. The chance of a thunderstorm offshore from the late morning.
#>  6:                                                                                                                          Partly cloudy.
#>  7:                                                                                                                           Mostly sunny.
#>  8:                          Mostly sunny morning. The chance of a thunderstorm offshore south of Port Kembla in the afternoon and evening.
#>  9:                                                                                                                          Partly cloudy.
#> 10:                                                                                                                           Mostly sunny.
#> 11:                                                                                                                           Mostly sunny.
#> 12:                                                                                                                          Partly cloudy.
#> 13:                                                                                                                         Cloud clearing.
#> 14:                                                                                                                           Mostly sunny.
#> 15:                                                                                                                          Partly cloudy.
#> 16:                                                                                                                         Cloud clearing.
#> 17:                                                                                                                           Mostly sunny.
#> 18:                                                                                                                           Mostly sunny.
#> 19:                                                                                                                           Mostly sunny.
#> 20:                                                                                                                                  Sunny.
#> 21:                                                                                                                                  Sunny.
#> 22:                                                                                                                                  Sunny.
#> 23:                                                                                                                                  Sunny.
#> 24:                                                                                                                                  Sunny.
#> 25:                                                                                                                           Mostly sunny.
#>                                                                                                                            forecast_weather
#>                                                                                                                                                               forecast_winds
#>  1:                                                                                                                                                                     <NA>
#>  2:                                                                                                                                            Southwesterly 30 to 40 knots.
#>  3:                                      Southwesterly 20 to 30 knots, reaching up to 40 knots early in the morning. Winds decreasing to 15 to 20 knots in the late evening.
#>  4:                                                                                                      Southwesterly 15 to 20 knots shifting northerly during the morning.
#>  5:                                                                     Southwesterly 20 to 30 knots, reaching up to 35 knots offshore south of Moruya Heads in the evening.
#>  6:                                                                                     Southwesterly 15 to 25 knots, reaching up to 35 knots offshore early in the morning.
#>  7:                         Southwesterly 15 to 20 knots shifting north to northwesterly 10 to 15 knots during the morning then increasing to 15 to 25 knots during the day.
#>  8:                                                                                                                                            Southwesterly 20 to 30 knots.
#>  9:                                                                                                Southwesterly 20 to 30 knots decreasing to 15 to 20 knots in the evening.
#> 10:                     Southwesterly 15 to 20 knots becoming variable about 10 knots during the morning then becoming north to northwesterly 15 to 20 knots during the day.
#> 11:                                                                                                  Westerly 25 to 30 knots turning southwesterly in the middle of the day.
#> 12:                       Southwesterly 20 to 25 knots, reaching up to 30 knots offshore early in the morning. Winds turning southerly 15 to 20 knots in the late afternoon.
#> 13: South to southwesterly 10 to 15 knots turning west to northwesterly below 10 knots during the morning then tending north to northwesterly 15 to 20 knots during the day.
#> 14:                                                                                                                                    West to southwesterly 25 to 30 knots.
#> 15:                              Southwesterly 15 to 25 knots, reaching up to 30 knots offshore early in the morning. Winds turning southerly 10 to 15 knots in the evening.
#> 16:              South to southwesterly 10 to 15 knots turning west to northwesterly below 10 knots during the morning then tending northerly 15 to 20 knots during the day.
#> 17:                                                                                                                                            Southwesterly 20 to 30 knots.
#> 18:                                                                Southwesterly 15 to 25 knots, reaching up to 30 knots offshore south of Crowdy Head early in the morning.
#> 19:                                                                                           Southwesterly 10 to 15 knots shifting northerly 15 to 20 knots during the day.
#> 20:                                                                                                    Southwesterly 15 to 25 knots, reaching up to 30 knots in the evening.
#> 21:                      Southwesterly 15 to 20 knots, reaching up to 25 knots offshore early in the morning. Winds turning southerly 10 to 15 knots in the early afternoon.
#> 22:                                                                                           South to southwesterly 10 to 15 knots shifting northerly during the afternoon.
#> 23:                   Southwesterly 15 to 25 knots tending south to southwesterly in the late evening. Winds reaching up to 30 knots offshore south of Yamba in the evening.
#> 24:                  Southerly 15 to 20 knots, reaching up to 25 knots offshore south of Yamba early in the morning. Winds decreasing to about 10 knots in the late evening.
#> 25:                                                                                           South to southeasterly 10 to 15 knots shifting northerly during the afternoon.
#>                                                                                                                                                               forecast_winds
#>                                                                                                       forecast_swell1
#>  1:                                                                                                              <NA>
#>  2: Southerly 2 to 3 metres inshore, increasing to 2.5 to 4 metres offshore south of Green Cape during the afternoon.
#>  3:                                            Southerly 2 to 3 metres inshore, increasing to 2 to 4 metres offshore.
#>  4:                                    Southerly 2 to 2.5 metres, decreasing to 1 to 1.5 metres during the afternoon.
#>  5:                                      Southerly 1.5 to 2 metres, increasing to 2 to 3 metres during the afternoon.
#>  6:   Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore north of Moruya Heads during the morning.
#>  7:                                          Southerly 2 to 3 metres, decreasing to 1 to 2 metres during the morning.
#>  8:                                          Southerly 1.5 to 2 metres, increasing to 2 to 3 metres by early evening.
#>  9:                                               Southerly 2 to 3 metres, increasing to 3 to 4 metres around midday.
#> 10:                                        Southerly 2 to 3 metres, decreasing to 1 to 2 metres during the afternoon.
#> 11:                                                                                      Southerly 1.5 to 2.5 metres.
#> 12:                              Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore around midday.
#> 13:                           Southerly 2 to 3 metres, decreasing to 1.5 to 2 metres during the afternoon or evening.
#> 14:                                                                                      Southerly 1.5 to 2.5 metres.
#> 15:   Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore south of Norah Head during the afternoon.
#> 16:                                          Southerly 3 to 4 metres, decreasing to 2 to 3 metres during the morning.
#> 17:                                  Southerly 1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the afternoon.
#> 18:                                        Southerly 1.5 to 2 metres, increasing to 2 to 3 metres during the morning.
#> 19:                                        Southerly 2 to 3 metres, decreasing to 1.5 to 2 metres during the evening.
#> 20:                                      Southerly 1 to 1.5 metres, increasing to 1.5 to 2.5 metres by early evening.
#> 21:                                                                                      Southerly 1.5 to 2.5 metres.
#> 22:                                            Southerly 1 to 2 metres inshore, increasing to 2 to 3 metres offshore.
#> 23:                             Southerly around 1 metre inshore, increasing to 1 to 2 metres offshore around midday.
#> 24:                                      Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2.5 metres offshore.
#> 25:                                                                                      Southerly 1.5 to 2.5 metres.
#>                                                                                                       forecast_swell1
#>     forecast_swell2
#>  1:            <NA>
#>  2:            <NA>
#>  3:            <NA>
#>  4:            <NA>
#>  5:            <NA>
#>  6:            <NA>
#>  7:            <NA>
#>  8:            <NA>
#>  9:            <NA>
#> 10:            <NA>
#> 11:            <NA>
#> 12:            <NA>
#> 13:            <NA>
#> 14:            <NA>
#> 15:            <NA>
#> 16:            <NA>
#> 17:            <NA>
#> 18:            <NA>
#> 19:            <NA>
#> 20:            <NA>
#> 21:            <NA>
#> 22:            <NA>
#> 23:            <NA>
#> 24:            <NA>
#> 25:            <NA>
#>     forecast_swell2
#>                                                                                                                           forecast_caution
#>  1:                                                                                                                                   <NA>
#>  2:                                                                                                                                   <NA>
#>  3: Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  4:                                                                                                                                   <NA>
#>  5:                                                                                                                                   <NA>
#>  6: Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  7:                                                                                                                                   <NA>
#>  8:                                                                                                                                   <NA>
#>  9: Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 10:                                                                                                                                   <NA>
#> 11:                                                                                                                                   <NA>
#> 12: Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 13:                                                                                                                                   <NA>
#> 14:                                                                                                                                   <NA>
#> 15: Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 16:                                                                                                                                   <NA>
#> 17:                                                                                                                                   <NA>
#> 18:                                                                                                                                   <NA>
#> 19:                                                                                                                                   <NA>
#> 20:                                                                                                                                   <NA>
#> 21:                                                                                                                                   <NA>
#> 22:                                                                                                                                   <NA>
#> 23:                                                                                                                                   <NA>
#> 24:                                                                                                                                   <NA>
#> 25:                                                                                                                                   <NA>
#>                                                                                                                           forecast_caution
#>                                        marine_forecast tropical_system_location
#>  1:                                               <NA>                       NA
#>  2:             Gale Warning for Sunday for Eden Coast                       NA
#>  3:             Gale Warning for Monday for Eden Coast                       NA
#>  4:                                               <NA>                       NA
#>  5:         Gale Warning for Sunday for Batemans Coast                       NA
#>  6:         Gale Warning for Monday for Batemans Coast                       NA
#>  7:                                               <NA>                       NA
#>  8: Strong Wind Warning for Sunday for Illawarra Coast                       NA
#>  9: Strong Wind Warning for Monday for Illawarra Coast                       NA
#> 10:                                               <NA>                       NA
#> 11:    Strong Wind Warning for Sunday for Sydney Coast                       NA
#> 12:    Strong Wind Warning for Monday for Sydney Coast                       NA
#> 13:                                               <NA>                       NA
#> 14:    Strong Wind Warning for Sunday for Hunter Coast                       NA
#> 15:    Strong Wind Warning for Monday for Hunter Coast                       NA
#> 16:                                               <NA>                       NA
#> 17: Strong Wind Warning for Sunday for Macquarie Coast                       NA
#> 18: Strong Wind Warning for Monday for Macquarie Coast                       NA
#> 19:                                               <NA>                       NA
#> 20:     Strong Wind Warning for Sunday for Coffs Coast                       NA
#> 21:                                               <NA>                       NA
#> 22:                                               <NA>                       NA
#> 23:     Strong Wind Warning for Sunday for Byron Coast                       NA
#> 24:                                               <NA>                       NA
#> 25:                                               <NA>                       NA
#>                                        marine_forecast tropical_system_location
#>     forecast_waves
#>  1:             NA
#>  2:             NA
#>  3:             NA
#>  4:             NA
#>  5:             NA
#>  6:             NA
#>  7:             NA
#>  8:             NA
#>  9:             NA
#> 10:             NA
#> 11:             NA
#> 12:             NA
#> 13:             NA
#> 14:             NA
#> 15:             NA
#> 16:             NA
#> 17:             NA
#> 18:             NA
#> 19:             NA
#> 20:             NA
#> 21:             NA
#> 22:             NA
#> 23:             NA
#> 24:             NA
#> 25:             NA
#>     forecast_waves

# }