Get BOM Coastal Waters Forecast

Fetch the BOM daily Coastal Waters Forecast and return a tidy data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to "AUS" returning all state forecasts, see details for further information.

Value

Tidy data.frame of a Australia BOM Coastal Waters Forecast.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

Examples

coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>    index product_id    type state_code dist_name       pt_1_name
#> 1      0   IDN11001 Coastal        NSW      Eden Montague Island
#> 2      1   IDN11001 Coastal        NSW      Eden Montague Island
#> 3      2   IDN11001 Coastal        NSW      Eden Montague Island
#> 4      3   IDN11001 Coastal        NSW      Eden Montague Island
#> 5      0   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 6      1   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 7      2   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 8      3   IDN11001 Coastal        NSW  Batemans       Ulladulla
#> 9      0   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 10     1   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 11     2   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 12     3   IDN11001 Coastal        NSW Illawarra    Port Hacking
#> 13     0   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 14     1   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 15     2   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 16     3   IDN11001 Coastal        NSW    Sydney      Broken Bay
#> 17     0   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 18     1   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 19     2   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 20     3   IDN11001 Coastal        NSW    Hunter      Seal Rocks
#> 21     0   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 22     1   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 23     2   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 24     3   IDN11001 Coastal        NSW Macquarie      Smoky Cape
#> 25     0   IDN11001 Coastal        NSW     Coffs           Wooli
#> 26     1   IDN11001 Coastal        NSW     Coffs           Wooli
#> 27     2   IDN11001 Coastal        NSW     Coffs           Wooli
#> 28     3   IDN11001 Coastal        NSW     Coffs           Wooli
#> 29     0   IDN11001 Coastal        NSW     Byron    Point Danger
#> 30     1   IDN11001 Coastal        NSW     Byron    Point Danger
#> 31     2   IDN11001 Coastal        NSW     Byron    Point Danger
#> 32     3   IDN11001 Coastal        NSW     Byron    Point Danger
#>          pt_2_name       aac    start_time_local end_time_local utc_offset
#> 1      Gabo Island NSW_MW001 2019-02-14 17:00:00     2019-02-15      11:00
#> 2      Gabo Island NSW_MW001 2019-02-15 00:00:00     2019-02-16      11:00
#> 3      Gabo Island NSW_MW001 2019-02-16 00:00:00     2019-02-17      11:00
#> 4      Gabo Island NSW_MW001 2019-02-17 00:00:00     2019-02-18      11:00
#> 5  Montague Island NSW_MW002 2019-02-14 17:00:00     2019-02-15      11:00
#> 6  Montague Island NSW_MW002 2019-02-15 00:00:00     2019-02-16      11:00
#> 7  Montague Island NSW_MW002 2019-02-16 00:00:00     2019-02-17      11:00
#> 8  Montague Island NSW_MW002 2019-02-17 00:00:00     2019-02-18      11:00
#> 9        Ulladulla NSW_MW003 2019-02-14 17:00:00     2019-02-15      11:00
#> 10       Ulladulla NSW_MW003 2019-02-15 00:00:00     2019-02-16      11:00
#> 11       Ulladulla NSW_MW003 2019-02-16 00:00:00     2019-02-17      11:00
#> 12       Ulladulla NSW_MW003 2019-02-17 00:00:00     2019-02-18      11:00
#> 13    Port Hacking NSW_MW004 2019-02-14 17:00:00     2019-02-15      11:00
#> 14    Port Hacking NSW_MW004 2019-02-15 00:00:00     2019-02-16      11:00
#> 15    Port Hacking NSW_MW004 2019-02-16 00:00:00     2019-02-17      11:00
#> 16    Port Hacking NSW_MW004 2019-02-17 00:00:00     2019-02-18      11:00
#> 17      Broken Bay NSW_MW005 2019-02-14 17:00:00     2019-02-15      11:00
#> 18      Broken Bay NSW_MW005 2019-02-15 00:00:00     2019-02-16      11:00
#> 19      Broken Bay NSW_MW005 2019-02-16 00:00:00     2019-02-17      11:00
#> 20      Broken Bay NSW_MW005 2019-02-17 00:00:00     2019-02-18      11:00
#> 21      Seal Rocks NSW_MW006 2019-02-14 17:00:00     2019-02-15      11:00
#> 22      Seal Rocks NSW_MW006 2019-02-15 00:00:00     2019-02-16      11:00
#> 23      Seal Rocks NSW_MW006 2019-02-16 00:00:00     2019-02-17      11:00
#> 24      Seal Rocks NSW_MW006 2019-02-17 00:00:00     2019-02-18      11:00
#> 25      Smoky Cape NSW_MW007 2019-02-14 17:00:00     2019-02-15      11:00
#> 26      Smoky Cape NSW_MW007 2019-02-15 00:00:00     2019-02-16      11:00
#> 27      Smoky Cape NSW_MW007 2019-02-16 00:00:00     2019-02-17      11:00
#> 28      Smoky Cape NSW_MW007 2019-02-17 00:00:00     2019-02-18      11:00
#> 29           Wooli NSW_MW008 2019-02-14 17:00:00     2019-02-15      11:00
#> 30           Wooli NSW_MW008 2019-02-15 00:00:00     2019-02-16      11:00
#> 31           Wooli NSW_MW008 2019-02-16 00:00:00     2019-02-17      11:00
#> 32           Wooli NSW_MW008 2019-02-17 00:00:00     2019-02-18      11:00
#>         start_time_utc        end_time_utc
#> 1  2019-02-14 06:00:00 2019-02-14 13:00:00
#> 2  2019-02-14 13:00:00 2019-02-15 13:00:00
#> 3  2019-02-15 13:00:00 2019-02-16 13:00:00
#> 4  2019-02-16 13:00:00 2019-02-17 13:00:00
#> 5  2019-02-14 06:00:00 2019-02-14 13:00:00
#> 6  2019-02-14 13:00:00 2019-02-15 13:00:00
#> 7  2019-02-15 13:00:00 2019-02-16 13:00:00
#> 8  2019-02-16 13:00:00 2019-02-17 13:00:00
#> 9  2019-02-14 06:00:00 2019-02-14 13:00:00
#> 10 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 11 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 12 2019-02-16 13:00:00 2019-02-17 13:00:00
#> 13 2019-02-14 06:00:00 2019-02-14 13:00:00
#> 14 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 15 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 16 2019-02-16 13:00:00 2019-02-17 13:00:00
#> 17 2019-02-14 06:00:00 2019-02-14 13:00:00
#> 18 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 19 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 20 2019-02-16 13:00:00 2019-02-17 13:00:00
#> 21 2019-02-14 06:00:00 2019-02-14 13:00:00
#> 22 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 23 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 24 2019-02-16 13:00:00 2019-02-17 13:00:00
#> 25 2019-02-14 06:00:00 2019-02-14 13:00:00
#> 26 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 27 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 28 2019-02-16 13:00:00 2019-02-17 13:00:00
#> 29 2019-02-14 06:00:00 2019-02-14 13:00:00
#> 30 2019-02-14 13:00:00 2019-02-15 13:00:00
#> 31 2019-02-15 13:00:00 2019-02-16 13:00:00
#> 32 2019-02-16 13:00:00 2019-02-17 13:00:00
#>                                                                                                      forecast_seas
#> 1                                                                                                   Below 1 metre.
#> 2                                                                                                   Below 1 metre.
#> 3                                  Around 1 metre, increasing to 1 to 1.5 metres south of Merimbula around midday.
#> 4                                                                                                 1 to 1.5 metres.
#> 5                                                                                                   Below 1 metre.
#> 6                                                                                                   Below 1 metre.
#> 7  Below 1 metre, decreasing below 0.5 metres during the morning, then increasing to around 1 metre around midday.
#> 8                                  1 to 1.5 metres, increasing to 1.5 to 2 metres during the afternoon or evening.
#> 9                                                                                                  Around 1 metre.
#> 10                                                                                                  Below 1 metre.
#> 11                                                                                                  Below 1 metre.
#> 12                                 1 to 1.5 metres, increasing to 1.5 to 2 metres during the afternoon or evening.
#> 13                                                                                                  Below 1 metre.
#> 14                                                                                                  Below 1 metre.
#> 15                                                                                                  Below 1 metre.
#> 16                                     Below 1 metre, increasing to 1 to 2 metres during the afternoon or evening.
#> 17                                                                                                  Below 1 metre.
#> 18                                                                                                 Around 1 metre.
#> 19                                                                                                 Around 1 metre.
#> 20                                                Below 1 metre, increasing to 1 to 2 metres during the afternoon.
#> 21                                 Around 1 metre, increasing to 1 to 1.5 metres offshore north of Port Macquarie.
#> 22                Around 1 metre, increasing to 1 to 1.5 metres offshore north of Port Macquarie by early evening.
#> 23                                                                                                 Around 1 metre.
#> 24          Around 1 metre, increasing to 1 to 1.5 metres south of Port Macquarie during the afternoon or evening.
#> 25                                                                                                1.5 to 2 metres.
#> 26                                                                                                1 to 1.5 metres.
#> 27                                                      1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 28                                                                                                 Around 1 metre.
#> 29                                                                                                2 to 2.5 metres.
#> 30                                                                                              1.5 to 2.5 metres.
#> 31                                              1.5 to 2 metres, decreasing to 1 to 1.5 metres during the morning.
#> 32                                                      1 to 1.5 metres, decreasing to 1 metre during the morning.
#>                                                                           forecast_weather
#> 1                                                                                   Clear.
#> 2                                                                           Partly cloudy.
#> 3                                                                           Partly cloudy.
#> 4                                                                                   Sunny.
#> 5                                                                            Mostly clear.
#> 6                                                                           Partly cloudy.
#> 7                                                                           Partly cloudy.
#> 8                                                                                   Sunny.
#> 9                                                                                   Clear.
#> 10                                                                          Partly cloudy.
#> 11                                                                          Partly cloudy.
#> 12                                                                           Mostly sunny.
#> 13                                                                                  Clear.
#> 14                                                                          Partly cloudy.
#> 15                                                                          Partly cloudy.
#> 16                                                                           Mostly sunny.
#> 17                                                                           Mostly clear.
#> 18                                                                          Partly cloudy.
#> 19                                                                          Partly cloudy.
#> 20                                                                         Cloud clearing.
#> 21                                                                          Partly cloudy.
#> 22                                                                                 Cloudy.
#> 23                                                                                 Cloudy.
#> 24                                                                          Partly cloudy.
#> 25                                                                          Partly cloudy.
#> 26                                                                                 Cloudy.
#> 27                                                                                 Cloudy.
#> 28                                                                          Partly cloudy.
#> 29 Cloudy. 50% chance of showers offshore north of Cape Byron, near zero chance elsewhere.
#> 30                                                                                 Cloudy.
#> 31                                                                                 Cloudy.
#> 32                                                                          Partly cloudy.
#>                                                                                                                                                        forecast_winds
#> 1                                                                                                                                            Variable about 10 knots.
#> 2                                                                                                                                            Variable about 10 knots.
#> 3                                                            East to northeasterly about 10 knots tending north to northeasterly 10 to 15 knots early in the morning.
#> 4                                                                                                                                           Northerly 15 to 20 knots.
#> 5                                                                              East to northeasterly 10 to 15 knots decreasing to below 10 knots in the late evening.
#> 6                                                                                                                                            Variable about 10 knots.
#> 7                                  Variable about 10 knots becoming northeasterly 10 to 15 knots in the middle of the day then tending northerly in the late evening.
#> 8                                                                                                                                           Northerly 15 to 25 knots.
#> 9  East to northeasterly 10 to 15 knots, reaching up to 20 knots south of Point Perpendicular in the evening. Winds decreasing to about 10 knots in the late evening.
#> 10                                                                                                                              East to northeasterly about 10 knots.
#> 11                                                                                East to northeasterly 10 to 15 knots tending north to northeasterly in the evening.
#> 12                                                                                                                             North to northeasterly 15 to 25 knots.
#> 13                                                                        Northeast to southeasterly about 10 knots increasing to 10 to 15 knots in the late evening.
#> 14                                                                                        Easterly about 10 knots increasing to 10 to 15 knots in the late afternoon.
#> 15                                                                                                                              East to northeasterly about 10 knots.
#> 16                                                                                 North to northeasterly about 10 knots increasing to 15 to 20 knots during the day.
#> 17                                                                                                                              East to southeasterly about 10 knots.
#> 18                                                                        East to southeasterly about 10 knots increasing to 10 to 15 knots in the middle of the day.
#> 19                                                                                         East to southeasterly 10 to 15 knots turning northeasterly in the evening.
#> 20                                                                                                                                      Northeasterly 15 to 20 knots.
#> 21                                                             Southeasterly 10 to 15 knots, reaching up to 20 knots offshore north of Port Macquarie in the evening.
#> 22                                                             Southeasterly 10 to 15 knots, reaching up to 20 knots offshore north of Port Macquarie in the evening.
#> 23                                                                                   Southeasterly 10 to 15 knots turning easterly in the late morning and afternoon.
#> 24          East to northeasterly 10 to 15 knots becoming variable about 10 knots during the morning then becoming northeasterly 15 to 20 knots during the afternoon.
#> 25                                                              Southeasterly 15 to 20 knots, reaching up to 25 knots offshore north of Coffs Harbour in the evening.
#> 26                                                                                                                                      Southeasterly 15 to 20 knots.
#> 27                                                                                                                                      Southeasterly 15 to 20 knots.
#> 28                                                              East to southeasterly 10 to 15 knots tending east to northeasterly below 10 knots during the evening.
#> 29                                                                               Southeasterly 20 to 25 knots, reaching up to 30 knots north of Yamba in the evening.
#> 30                                                  Southeasterly 15 to 20 knots, reaching up to 25 knots north of Yamba during the morning and again in the evening.
#> 31                                                                                                                                      Southeasterly 15 to 20 knots.
#> 32                                                                                                                                      Southeasterly 10 to 15 knots.
#>                                                                                                                     forecast_swell1
#> 1                                                                                                        Southerly 1 to 1.5 metres.
#> 2                                                       Southerly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 3                                                                                                          Southerly below 1 metre.
#> 4                  Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore south of Green Cape during the morning.
#> 5                                                          Southerly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 6                                                                                                         Southerly around 1 metre.
#> 7                                                                                                          Southerly below 1 metre.
#> 8                                                                                                         Southerly around 1 metre.
#> 9                                                         Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 10                                                                                                        Southerly around 1 metre.
#> 11                                                                                                          Easterly below 1 metre.
#> 12                                                                                                        Southerly around 1 metre.
#> 13                                                        Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 14                                                         Southerly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 15                                                                                                        Southerly around 1 metre.
#> 16                                                                                                         Easterly around 1 metre.
#> 17                                     Southerly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore north of Port Stephens.
#> 18                                                        Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 19               Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Port Stephens during the morning.
#> 20                                     Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Norah Head.
#> 21                                                       Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 22                                                                                                       Southerly 1 to 1.5 metres.
#> 23                                                       Easterly around 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 24                                                                                                             Easterly 1.5 metres.
#> 25                                                                                                            Southerly 1.5 metres.
#> 26                                                     Southerly 1.5 to 2 metres, decreasing to 1 to 1.5 metres during the morning.
#> 27                                     Easterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore during the morning.
#> 28                                                             Easterly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 29                                   Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore north of Cape Byron.
#> 30 Southerly 1 to 1.5 metres, decreasing to around 1 metre during the morning, then increasing to 1 to 1.5 metres by early evening.
#> 31                                                      Easterly 1 to 1.5 metres, increasing to 1.5 to 2 metres during the morning.
#> 32                                                                                                               Easterly 2 metres.
#>                                                                                forecast_swell2
#> 1                                                                                         <NA>
#> 2                                                                                         <NA>
#> 3                                                                                         <NA>
#> 4                                                                                         <NA>
#> 5                                                                                         <NA>
#> 6                                                                                         <NA>
#> 7                                                                                         <NA>
#> 8                                                                                         <NA>
#> 9                                                                                         <NA>
#> 10                                                                                        <NA>
#> 11                                                                                        <NA>
#> 12                                                                    Easterly around 1 metre.
#> 13                                                                                        <NA>
#> 14                                                                                        <NA>
#> 15                                                                    Easterly around 1 metre.
#> 16                                                                                        <NA>
#> 17                                                                                        <NA>
#> 18                                                                                        <NA>
#> 19                                                                   Southerly around 1 metre.
#> 20                                                                                        <NA>
#> 21                                                                                        <NA>
#> 22                                                                    Easterly around 1 metre.
#> 23                                                                   Southerly around 1 metre.
#> 24                                                                                        <NA>
#> 25                                                                   Easterly 1 to 1.5 metres.
#> 26 Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 27                                                                   Southerly around 1 metre.
#> 28                                                                                        <NA>
#> 29                                                                        Easterly 1.5 metres.
#> 30                    Easterly 1 to 1.5 metres, decreasing to around 1 metre by early evening.
#> 31                                                                   Southerly around 1 metre.
#> 32                                                                                        <NA>
#>    forecast_caution                                  marine_forecast
#> 1                NA                                             <NA>
#> 2                NA                                             <NA>
#> 3                NA                                             <NA>
#> 4                NA                                             <NA>
#> 5                NA                                             <NA>
#> 6                NA                                             <NA>
#> 7                NA                                             <NA>
#> 8                NA                                             <NA>
#> 9                NA                                             <NA>
#> 10               NA                                             <NA>
#> 11               NA                                             <NA>
#> 12               NA                                             <NA>
#> 13               NA                                             <NA>
#> 14               NA                                             <NA>
#> 15               NA                                             <NA>
#> 16               NA                                             <NA>
#> 17               NA                                             <NA>
#> 18               NA                                             <NA>
#> 19               NA                                             <NA>
#> 20               NA                                             <NA>
#> 21               NA                                             <NA>
#> 22               NA                                             <NA>
#> 23               NA                                             <NA>
#> 24               NA                                             <NA>
#> 25               NA                                             <NA>
#> 26               NA                                             <NA>
#> 27               NA                                             <NA>
#> 28               NA                                             <NA>
#> 29               NA Strong Wind Warning for Thursday for Byron Coast
#> 30               NA                                             <NA>
#> 31               NA                                             <NA>
#> 32               NA                                             <NA>