Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     0   IDN11001 <NA>        NSW
#>  6:     1   IDN11001 <NA>        NSW
#>  7:     2   IDN11001 <NA>        NSW
#>  8:     0   IDN11001 <NA>        NSW
#>  9:     1   IDN11001 <NA>        NSW
#> 10:     2   IDN11001 <NA>        NSW
#> 11:     0   IDN11001 <NA>        NSW
#> 12:     1   IDN11001 <NA>        NSW
#> 13:     2   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     0   IDN11001 <NA>        NSW
#> 18:     1   IDN11001 <NA>        NSW
#> 19:     2   IDN11001 <NA>        NSW
#> 20:     0   IDN11001 <NA>        NSW
#> 21:     1   IDN11001 <NA>        NSW
#> 22:     2   IDN11001 <NA>        NSW
#> 23:     0   IDN11001 <NA>        NSW
#> 24:     1   IDN11001 <NA>        NSW
#> 25:     2   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#>  9:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 12:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 13:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 14:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 15:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 16:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 17:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 18:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 19:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 20:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 21:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 22:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 23:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 24:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 25:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-04-03 04:10:08 2021-04-03 04:10:08      11:00 2021-04-03 04:10:08
#>  2: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#>  3: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#>  4: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#>  5: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#>  6: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#>  7: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#>  8: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#>  9: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 10: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#> 11: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#> 12: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 13: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#> 14: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#> 15: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 16: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#> 17: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#> 18: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 19: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#> 20: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#> 21: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 22: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#> 23: 2021-04-03 05:00:00 2021-04-03 05:00:00      11:00 2021-04-03 05:00:00
#> 24: 2021-04-04 00:00:00 2021-04-04 00:00:00      11:00 2021-04-04 00:00:00
#> 25: 2021-04-05 00:00:00 2021-04-05 00:00:00      10:00 2021-04-05 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-04-03 04:10:08
#>  2: 2021-04-03 05:00:00
#>  3: 2021-04-04 00:00:00
#>  4: 2021-04-05 00:00:00
#>  5: 2021-04-03 05:00:00
#>  6: 2021-04-04 00:00:00
#>  7: 2021-04-05 00:00:00
#>  8: 2021-04-03 05:00:00
#>  9: 2021-04-04 00:00:00
#> 10: 2021-04-05 00:00:00
#> 11: 2021-04-03 05:00:00
#> 12: 2021-04-04 00:00:00
#> 13: 2021-04-05 00:00:00
#> 14: 2021-04-03 05:00:00
#> 15: 2021-04-04 00:00:00
#> 16: 2021-04-05 00:00:00
#> 17: 2021-04-03 05:00:00
#> 18: 2021-04-04 00:00:00
#> 19: 2021-04-05 00:00:00
#> 20: 2021-04-03 05:00:00
#> 21: 2021-04-04 00:00:00
#> 22: 2021-04-05 00:00:00
#> 23: 2021-04-03 05:00:00
#> 24: 2021-04-04 00:00:00
#> 25: 2021-04-05 00:00:00
#>            end_time_utc
#>                                                                                forecast_seas
#>  1:                                                                                     <NA>
#>  2:                               Around 1 metre, increasing to 1 to 2 metres around midday.
#>  3:                                 1 to 2 metres, decreasing to 1 metre during the morning.
#>  4:                                                                          Around 1 metre.
#>  5:                               Around 1 metre, increasing to 1 to 2 metres around midday.
#>  6:                          1.5 to 2.5 metres, decreasing to 1 to 1.5 metres around midday.
#>  7:               Around 1 metre, increasing to 1 to 1.5 metres offshore during the morning.
#>  8:                        Around 1 metre, increasing to 1 to 2 metres during the afternoon.
#>  9:                               1 to 2 metres, decreasing to 1 metre later in the evening.
#> 10:                                                                          Around 1 metre.
#> 11:                              Below 1 metre, increasing to around 1 metre in the evening.
#> 12:                                                                          Around 1 metre.
#> 13:                                                                          Around 1 metre.
#> 14:                                                                           Below 1 metre.
#> 15:                                                                         1 to 1.5 metres.
#> 16:                               1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 17:                                                                           Below 1 metre.
#> 18: Around 1 metre, increasing to 1 to 1.5 metres south of Crowdy Head during the afternoon.
#> 19:                                                                         1 to 1.5 metres.
#> 20:                          1 to 1.5 metres, decreasing below 1 metre during the afternoon.
#> 21:                                                                          Around 1 metre.
#> 22:                        Around 1 metre, increasing to 1 to 1.5 metres during the morning.
#> 23:             1 to 1.5 metres, increasing to 1.5 to 2 metres offshore north of Cape Byron.
#> 24:             1 to 1.5 metres, increasing to 1.5 to 2 metres offshore north of Cape Byron.
#> 25:                     1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the morning.
#>                                                                                forecast_seas
#>                                              forecast_weather
#>  1:                                                      <NA>
#>  2:                                                    Sunny.
#>  3:                                             Mostly sunny.
#>  4:                                                   Cloudy.
#>  5:                                                    Sunny.
#>  6:                                             Mostly sunny.
#>  7:                            Cloudy. 60% chance of showers.
#>  8:                                                    Sunny.
#>  9:                                             Mostly sunny.
#> 10:                     Partly cloudy. 60% chance of showers.
#> 11:                                             Mostly sunny.
#> 12:                                             Mostly sunny.
#> 13:                     Partly cloudy. 60% chance of showers.
#> 14:                                            Partly cloudy.
#> 15:                                            Partly cloudy.
#> 16:                     Partly cloudy. 60% chance of showers.
#> 17:                                                   Cloudy.
#> 18:                                            Partly cloudy.
#> 19:                            Cloudy. 80% chance of showers.
#> 20:                                            Partly cloudy.
#> 21:                                            Partly cloudy.
#> 22:                               Cloudy. 95% chance of rain.
#> 23:                                            Partly cloudy.
#> 24:                            Cloudy. 70% chance of showers.
#> 25: Cloudy. 95% chance of rain. The chance of a thunderstorm.
#>                                              forecast_weather
#>                                                                                                                                                                                                                                  forecast_winds
#>  1:                                                                                                                                                                                                                                        <NA>
#>  2:                                                                                                                                          Northerly 15 to 20 knots, reaching up to 25 knots offshore north of Merimbula in the late evening.
#>  3: Northerly 10 to 15 knots, reaching up to 25 knots offshore north of Merimbula early in the morning. Winds becoming variable about 10 knots in the morning then becoming southeast to southwesterly 10 to 15 knots in the middle of the day.
#>  4:                                                                                                                                                                   Southerly 10 to 15 knots decreasing to about 10 knots during the evening.
#>  5:                                                                                                                                                                       Northerly about 10 knots increasing to 15 to 25 knots in the morning.
#>  6:                                                                                                                                                                   Northerly 15 to 25 knots becoming variable about 10 knots in the evening.
#>  7:                                                                                                                                                         Southerly 10 to 15 knots turning southeasterly 15 to 20 knots during the afternoon.
#>  8:                                                                    Northerly about 10 knots increasing to 15 to 20 knots in the middle of the day. Winds reaching up to 25 knots offshore south of Point Perpendicular in the late evening.
#>  9:                                                                                                                                                                       Northerly 15 to 25 knots decreasing to 10 to 15 knots in the evening.
#> 10:                                                                                                                                                     South to southwesterly 10 to 15 knots turning east to southeasterly during the morning.
#> 11:                                                                                                                                              Variable below 10 knots becoming north to northeasterly 10 to 15 knots in the early afternoon.
#> 12:                                                                                                                                                                                                                   Northerly 10 to 15 knots.
#> 13:                                                                                                               Northerly 10 to 15 knots shifting southerly in the morning, then tending east to southeasterly below 10 knots during the day.
#> 14:                                                                                                                                       East to northeasterly 10 to 15 knots tending north to northeasterly during the afternoon and evening.
#> 15:                                                                                                                                                                                                                   Northerly 15 to 20 knots.
#> 16:                                                                                                                                                                    Northerly 15 to 20 knots turning easterly 10 to 15 knots during the day.
#> 17:                                                                                                                                                             Easterly 10 to 15 knots turning northeasterly during the afternoon and evening.
#> 18:                                                                                                                                                                                                      North to northeasterly 15 to 20 knots.
#> 19:                                                                                                                                      North to northeasterly 10 to 15 knots tending east to northeasterly 15 to 20 knots during the morning.
#> 20:                                                                                                              Southeasterly 10 to 15 knots, reaching up to 20 knots offshore during the morning. Winds turning easterly in the late morning.
#> 21:                                                                                                                                                                                                       East to northeasterly 10 to 15 knots.
#> 22:                                                                                                                                                                                                                    Easterly 15 to 20 knots.
#> 23:                                                                                                 Southeasterly 15 to 20 knots turning easterly in the evening. Winds offshore north of Cape Byron possibly reaching 25 knots in the morning.
#> 24:                                                                                                                                                                                                                    Easterly 15 to 20 knots.
#> 25:                                                                                                                                                       Easterly 15 to 25 knots increasing to 20 to 30 knots during the afternoon or evening.
#>                                                                                                                                                                                                                                  forecast_winds
#>                                                                                                 forecast_swell1
#>  1:                                                                                                        <NA>
#>  2:                                                                                    Southerly below 1 metre.
#>  3:                  Southerly below 1 metre, increasing to around 1 metre south of Merimbula by early evening.
#>  4:                                Southerly around 1 metre, increasing to 1 to 1.5 metres south of Green Cape.
#>  5:                                                                                Southeasterly below 1 metre.
#>  6:                                                                   Northeast to southeasterly below 1 metre.
#>  7:                                                                                   Southerly around 1 metre.
#>  8:                                                                                Southeasterly below 1 metre.
#>  9:                                                                        East to southeasterly below 1 metre.
#> 10:                                                                                   Southerly around 1 metre.
#> 11:                                                                                Southeasterly below 1 metre.
#> 12:                                                                        East to southeasterly below 1 metre.
#> 13:                                                                                   Southerly around 1 metre.
#> 14:                                                                                Southeasterly below 1 metre.
#> 15:                                   Southeasterly below 1 metre tending east to northeasterly during the day.
#> 16:            Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 17:                                                                               Southeasterly around 1 metre.
#> 18:             East to southeasterly around 1 metre, tending northeasterly 1 to 1.5 metres during the morning.
#> 19:                                                                              Northeasterly 1 to 1.5 metres.
#> 20:                                                                              Southeasterly 1 to 1.5 metres.
#> 21: East to northeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore during the afternoon.
#> 22:                                   Northeasterly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 23:                            Southeasterly 1 to 1.5 metres, tending easterly during the afternoon or evening.
#> 24:                        Easterly 1.5 metres, tending east to northeasterly 1.5 to 2 metres in the afternoon.
#> 25:                                                                      East to northeasterly 2 to 2.5 metres.
#>                                                                                                 forecast_swell1
#>                                                                                      forecast_swell2
#>  1:                                                                                             <NA>
#>  2:                                                                                             <NA>
#>  3:                                                                    Northeasterly around 1 metre.
#>  4:                                                                                             <NA>
#>  5:                                                                                             <NA>
#>  6:                                                                                             <NA>
#>  7:                                                                                             <NA>
#>  8:                                                                                             <NA>
#>  9:                                                                                             <NA>
#> 10:                                                                    Northeasterly around 1 metre.
#> 11:                                                                                             <NA>
#> 12:                                                                                             <NA>
#> 13: Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 14:                                                                                             <NA>
#> 15:                                                                                             <NA>
#> 16:                                                                        Southerly around 1 metre.
#> 17:                                                                                             <NA>
#> 18:                                                                                             <NA>
#> 19:                                                                        Southerly around 1 metre.
#> 20:                                                                                             <NA>
#> 21:                                                                                             <NA>
#> 22:                                                                                             <NA>
#> 23:                                                                                             <NA>
#> 24:                                                                                             <NA>
#> 25:                                                                                             <NA>
#>                                                                                      forecast_swell2
#>     forecast_caution marine_forecast tropical_system_location forecast_waves
#>  1:             <NA>            <NA>                       NA             NA
#>  2:             <NA>            <NA>                       NA             NA
#>  3:             <NA>            <NA>                       NA             NA
#>  4:             <NA>            <NA>                       NA             NA
#>  5:             <NA>            <NA>                       NA             NA
#>  6:             <NA>            <NA>                       NA             NA
#>  7:             <NA>            <NA>                       NA             NA
#>  8:             <NA>            <NA>                       NA             NA
#>  9:             <NA>            <NA>                       NA             NA
#> 10:             <NA>            <NA>                       NA             NA
#> 11:             <NA>            <NA>                       NA             NA
#> 12:             <NA>            <NA>                       NA             NA
#> 13:             <NA>            <NA>                       NA             NA
#> 14:             <NA>            <NA>                       NA             NA
#> 15:             <NA>            <NA>                       NA             NA
#> 16:             <NA>            <NA>                       NA             NA
#> 17:             <NA>            <NA>                       NA             NA
#> 18:             <NA>            <NA>                       NA             NA
#> 19:             <NA>            <NA>                       NA             NA
#> 20:             <NA>            <NA>                       NA             NA
#> 21:             <NA>            <NA>                       NA             NA
#> 22:             <NA>            <NA>                       NA             NA
#> 23:             <NA>            <NA>                       NA             NA
#> 24:             <NA>            <NA>                       NA             NA
#> 25:             <NA>            <NA>                       NA             NA
#>     forecast_caution marine_forecast tropical_system_location forecast_waves

# }