Get BOM Coastal Waters Forecast

Fetch the BOM daily Coastal Waters Forecast and return a tidy data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to "AUS" returning all state forecasts, see details for further information.

Value

Tidy data.table of a Australia BOM Coastal Waters Forecast.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

Examples

coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2019-03-21 16:05:05 2019-03-21 16:05:05      11:00 2019-03-21 16:05:05
#>  2: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#>  3: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#>  4: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#>  5: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#>  6: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#>  7: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#>  8: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#>  9: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 10: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 11: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 12: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 13: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 14: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 15: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 16: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 17: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 18: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 19: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 20: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 21: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 22: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 23: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 24: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 25: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 26: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 27: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 28: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 29: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#> 30: 2019-03-21 17:00:00 2019-03-21 17:00:00      11:00 2019-03-21 17:00:00
#> 31: 2019-03-22 00:00:00 2019-03-22 00:00:00      11:00 2019-03-22 00:00:00
#> 32: 2019-03-23 00:00:00 2019-03-23 00:00:00      11:00 2019-03-23 00:00:00
#> 33: 2019-03-24 00:00:00 2019-03-24 00:00:00      11:00 2019-03-24 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2019-03-21 16:05:05
#>  2: 2019-03-21 17:00:00
#>  3: 2019-03-22 00:00:00
#>  4: 2019-03-23 00:00:00
#>  5: 2019-03-24 00:00:00
#>  6: 2019-03-21 17:00:00
#>  7: 2019-03-22 00:00:00
#>  8: 2019-03-23 00:00:00
#>  9: 2019-03-24 00:00:00
#> 10: 2019-03-21 17:00:00
#> 11: 2019-03-22 00:00:00
#> 12: 2019-03-23 00:00:00
#> 13: 2019-03-24 00:00:00
#> 14: 2019-03-21 17:00:00
#> 15: 2019-03-22 00:00:00
#> 16: 2019-03-23 00:00:00
#> 17: 2019-03-24 00:00:00
#> 18: 2019-03-21 17:00:00
#> 19: 2019-03-22 00:00:00
#> 20: 2019-03-23 00:00:00
#> 21: 2019-03-24 00:00:00
#> 22: 2019-03-21 17:00:00
#> 23: 2019-03-22 00:00:00
#> 24: 2019-03-23 00:00:00
#> 25: 2019-03-24 00:00:00
#> 26: 2019-03-21 17:00:00
#> 27: 2019-03-22 00:00:00
#> 28: 2019-03-23 00:00:00
#> 29: 2019-03-24 00:00:00
#> 30: 2019-03-21 17:00:00
#> 31: 2019-03-22 00:00:00
#> 32: 2019-03-23 00:00:00
#> 33: 2019-03-24 00:00:00
#>            end_time_utc
#>                                                                  forecast_seas
#>  1:                                                                       <NA>
#>  2:                                                             Below 1 metre.
#>  3:                                                             Below 1 metre.
#>  4:           Below 1 metre, increasing to 1 to 1.5 metres during the morning.
#>  5:                                                             1 to 2 metres.
#>  6:                                                             Below 1 metre.
#>  7:                                                             Below 1 metre.
#>  8:           Below 1 metre, increasing to 1 to 1.5 metres during the morning.
#>  9: 1 to 1.5 metres, decreasing below 1 metre during the afternoon or evening.
#> 10:                                                             Below 1 metre.
#> 11:                                                             Below 1 metre.
#> 12:                Below 1 metre, increasing to 1 to 1.5 metres around midday.
#> 13:                 1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 14:                                                            Around 1 metre.
#> 15:                                                             Below 1 metre.
#> 16:                                                            Around 1 metre.
#> 17:                 1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 18:                                                             Below 1 metre.
#> 19:                                                          Below 0.5 metres.
#> 20:         Below 0.5 metres, increasing to around 1 metre during the morning.
#> 21:              1 to 1.5 metres, decreasing below 1 metre during the morning.
#> 22:                                                             Below 1 metre.
#> 23:                                                             Below 1 metre.
#> 24:                                                             Below 1 metre.
#> 25:              1 to 1.5 metres, decreasing below 1 metre during the morning.
#> 26:                                                             Below 1 metre.
#> 27:                                                             Below 1 metre.
#> 28:                                                            Around 1 metre.
#> 29:                                                           1 to 1.5 metres.
#> 30:                                                             Below 1 metre.
#> 31:                                                          Below 0.5 metres.
#> 32:              Below 0.5 metres, increasing to around 1 metre around midday.
#> 33:                                                           1 to 1.5 metres.
#>                                                                  forecast_seas
#>                                                                                           forecast_weather
#>  1:                                                                                                   <NA>
#>  2:                                                                                         Partly cloudy.
#>  3:                        Partly cloudy. The chance of a thunderstorm in the afternoon and early evening.
#>  4:                              Partly cloudy. The chance of a thunderstorm in the morning and afternoon.
#>  5:                                                                                         Partly cloudy.
#>  6:                                                                                         Partly cloudy.
#>  7:                                          Partly cloudy. The chance of a thunderstorm in the afternoon.
#>  8:                        Partly cloudy. The chance of a thunderstorm in the afternoon and early evening.
#>  9:                                                                                         Partly cloudy.
#> 10:                                                                                         Partly cloudy.
#> 11:                                          Partly cloudy. The chance of a thunderstorm in the afternoon.
#> 12:                              Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 13:                                                                                         Partly cloudy.
#> 14:                                                                  Partly cloudy. 50% chance of showers.
#> 15:                                          Partly cloudy. The chance of a thunderstorm in the afternoon.
#> 16:                              Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 17:                                                                                         Partly cloudy.
#> 18:                                                                  Partly cloudy. 50% chance of showers.
#> 19:                                          Partly cloudy. The chance of a thunderstorm in the afternoon.
#> 20:                      Mostly sunny day. The chance of a thunderstorm in the late afternoon and evening.
#> 21:                                                                                         Partly cloudy.
#> 22:                                                                                         Partly cloudy.
#> 23:       Partly cloudy. 50% chance of showers. The chance of a thunderstorm in the afternoon and evening.
#> 24:                                                            Mostly sunny. The chance of a thunderstorm.
#> 25:                                                                                          Mostly sunny.
#> 26:                                                                                         Partly cloudy.
#> 27: Partly cloudy. 50% chance of showers. The chance of a thunderstorm in the afternoon and early evening.
#> 28:                                                           Partly cloudy. The chance of a thunderstorm.
#> 29:                                                                                          Mostly sunny.
#> 30:                                                                                         Partly cloudy.
#> 31:                       Partly cloudy. The chance of a thunderstorm mostly in the afternoon and evening.
#> 32:                                                           Partly cloudy. The chance of a thunderstorm.
#> 33:                                                                                          Mostly sunny.
#>                                                                                           forecast_weather
#>                                                                                                                                                       forecast_winds
#>  1:                                                                                                                                                             <NA>
#>  2:                                                                                                                                        Southerly 10 to 15 knots.
#>  3:                                                                        South to southeasterly about 10 knots becoming east to northeasterly in the late evening.
#>  4:                        North to northeasterly about 10 knots increasing to 15 to 20 knots in the morning. Winds offshore possibly reaching up to 25 knots later.
#>  5:  North to northwesterly 15 to 20 knots shifting southerly 10 to 15 knots in the morning, then turning northeasterly 15 to 20 knots in the afternoon and evening.
#>  6:                                                                                                                                        Southerly 10 to 15 knots.
#>  7:                                                                                     Southerly about 10 knots becoming east to northeasterly in the late evening.
#>  8:                                                                                            Northerly about 10 knots increasing to 15 to 20 knots in the morning.
#>  9: North to northwesterly 15 to 20 knots turning south to southeasterly during the morning then tending north to northeasterly 10 to 15 knots during the afternoon.
#> 10:                                                                                                                                        Southerly about 10 knots.
#> 11:                                                                             South to southwesterly about 10 knots becoming east to southeasterly in the evening.
#> 12:                                                                      Northeast to northwesterly about 10 knots becoming northerly 15 to 20 knots in the morning.
#> 13:                                                         North to northwesterly 15 to 20 knots shifting south to southeasterly 10 to 15 knots during the morning.
#> 14:                                                                                                                           South to southeasterly 10 to 15 knots.
#> 15:                                                                             South to southeasterly about 10 knots becoming east to northeasterly in the evening.
#> 16:                                                                     North to northeasterly about 10 knots increasing to 15 to 20 knots in the middle of the day.
#> 17:                         North to northwesterly 10 to 15 knots shifting south to southeasterly during the day, then tending east to northeasterly in the evening.
#> 18:                                                                                                                           South to southeasterly about 10 knots.
#> 19:                                                                        South to southeasterly below 10 knots becoming east to northeasterly in the late evening.
#> 20:                                                                               North to northeasterly about 10 knots increasing to 15 to 20 knots during the day.
#> 21:                 North to northwesterly 10 to 15 knots turning south to southeasterly during the afternoon then tending east to northeasterly during the evening.
#> 22:                                                                                                                           South to southeasterly about 10 knots.
#> 23:                                                                             South to southwesterly about 10 knots becoming east to southeasterly in the evening.
#> 24:                                                                                        Variable below 10 knots becoming northerly 10 to 15 knots in the morning.
#> 25:                                                                       North to northwesterly 15 to 20 knots tending northerly 10 to 15 knots during the evening.
#> 26:                                                                                                                           South to southeasterly about 10 knots.
#> 27:                                                                       Southeast to southwesterly below 10 knots becoming east to southeasterly in the afternoon.
#> 28:                                                                              Variable below 10 knots becoming northerly 15 to 20 knots in the middle of the day.
#> 29:                                                                                                                                        Northerly 15 to 20 knots.
#> 30:                                                                                                                                    Southeasterly about 10 knots.
#> 31:                                                                                          South to southeasterly below 10 knots becoming easterly in the evening.
#> 32:                                                                                        Variable below 10 knots becoming northerly 10 to 15 knots during the day.
#> 33:                                                                                                                           North to northwesterly 15 to 20 knots.
#>                                                                                                                                                       forecast_winds
#>              forecast_swell1 forecast_swell2 forecast_caution marine_forecast
#>  1:                     <NA>            <NA>             <NA>            <NA>
#>  2:  Easterly below 1 metre.            <NA>             <NA>            <NA>
#>  3: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  4: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  5: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  6: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  7: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  8: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>  9: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 10: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 11: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 12: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 13: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 14: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 15: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 16: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 17: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 18: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 19: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 20: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 21: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 22: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 23: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 24: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 25: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 26: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 27: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 28: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 29: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 30: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 31: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 32: Easterly around 1 metre.            <NA>             <NA>            <NA>
#> 33: Easterly around 1 metre.            <NA>             <NA>            <NA>
#>              forecast_swell1 forecast_swell2 forecast_caution marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#>     tropical_system_location forecast_waves