Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-01-18 16:05:25 2021-01-18 16:05:25      11:00 2021-01-18 16:05:25
#>  2: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#>  3: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#>  4: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#>  5: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#>  6: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#>  7: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#>  8: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#>  9: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 10: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 11: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 12: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 13: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 14: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 15: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 16: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 17: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 18: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 19: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 20: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 21: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 22: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 23: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 24: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 25: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 26: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 27: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 28: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 29: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 30: 2021-01-18 17:00:00 2021-01-18 17:00:00      11:00 2021-01-18 17:00:00
#> 31: 2021-01-19 00:00:00 2021-01-19 00:00:00      11:00 2021-01-19 00:00:00
#> 32: 2021-01-20 00:00:00 2021-01-20 00:00:00      11:00 2021-01-20 00:00:00
#> 33: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-01-18 16:05:25
#>  2: 2021-01-18 17:00:00
#>  3: 2021-01-19 00:00:00
#>  4: 2021-01-20 00:00:00
#>  5: 2021-01-21 00:00:00
#>  6: 2021-01-18 17:00:00
#>  7: 2021-01-19 00:00:00
#>  8: 2021-01-20 00:00:00
#>  9: 2021-01-21 00:00:00
#> 10: 2021-01-18 17:00:00
#> 11: 2021-01-19 00:00:00
#> 12: 2021-01-20 00:00:00
#> 13: 2021-01-21 00:00:00
#> 14: 2021-01-18 17:00:00
#> 15: 2021-01-19 00:00:00
#> 16: 2021-01-20 00:00:00
#> 17: 2021-01-21 00:00:00
#> 18: 2021-01-18 17:00:00
#> 19: 2021-01-19 00:00:00
#> 20: 2021-01-20 00:00:00
#> 21: 2021-01-21 00:00:00
#> 22: 2021-01-18 17:00:00
#> 23: 2021-01-19 00:00:00
#> 24: 2021-01-20 00:00:00
#> 25: 2021-01-21 00:00:00
#> 26: 2021-01-18 17:00:00
#> 27: 2021-01-19 00:00:00
#> 28: 2021-01-20 00:00:00
#> 29: 2021-01-21 00:00:00
#> 30: 2021-01-18 17:00:00
#> 31: 2021-01-19 00:00:00
#> 32: 2021-01-20 00:00:00
#> 33: 2021-01-21 00:00:00
#>            end_time_utc
#>                                                                                 forecast_seas
#>  1:                                                                                      <NA>
#>  2:                           Around 1 metre, increasing to 2 to 3 metres south of Merimbula.
#>  3:                   2 to 3 metres, increasing to 2 to 4 metres offshore during the morning.
#>  4:                               1 to 2 metres, decreasing below 1 metre during the morning.
#>  5:                           Around 1 metre, increasing to 1 to 2 metres during the morning.
#>  6:                                                                            Below 1 metre.
#>  7:                            Below 1 metre, increasing to 2 to 3 metres during the morning.
#>  8:                                  1 to 2 metres, decreasing to 1 metre during the morning.
#>  9:                        Below 1 metre, increasing to 1.5 to 2.5 metres during the morning.
#> 10:                                                                           Around 1 metre.
#> 11:                            Below 1 metre, increasing to 2 to 3 metres during the morning.
#> 12:                           1 to 1.5 metres, decreasing below 1 metre during the afternoon.
#> 13:                      Below 1 metre, increasing to 1.5 to 2.5 metres during the afternoon.
#> 14:                                                                           Around 1 metre.
#> 15:                           Around 1 metre, increasing to 2 to 3 metres during the morning.
#> 16:                              1 to 1.5 metres, decreasing to 1 metre during the afternoon.
#> 17:             Below 1 metre, increasing to 1 to 1.5 metres during the afternoon or evening.
#> 18:        Below 1 metre, increasing to 1 to 1.5 metres north of Norah Head by early evening.
#> 19:                            1 to 2 metres, increasing to 2 to 3 metres during the morning.
#> 20:                         1 to 1.5 metres, increasing to 1 to 2 metres north of Norah Head.
#> 21:               Below 1 metre, increasing to 1 to 2 metres during the afternoon or evening.
#> 22:                        1.5 metres, increasing to 1.5 to 2 metres north of Port Macquarie.
#> 23:                            1 to 2 metres, increasing to 2 to 3 metres during the morning.
#> 24:                                                                            1 to 2 metres.
#> 25: Around 1 metre, increasing to 1 to 1.5 metres south of Port Macquarie during the evening.
#> 26:                                                                            1 to 2 metres.
#> 27:                               1.5 to 2 metres, increasing to 2 to 3 metres around midday.
#> 28:                                                                        1.5 to 2.5 metres.
#> 29:                                1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 30:                                                                          1 to 1.5 metres.
#> 31:                              1 to 2 metres, increasing to 2 to 3 metres by early evening.
#> 32:                                                                          2 to 2.5 metres.
#> 33:                                  1 to 2 metres, decreasing to 1 metre during the morning.
#>                                                                                 forecast_seas
#>                                                                                                                  forecast_weather
#>  1:                                                                                                                          <NA>
#>  2:                                                                                                                        Clear.
#>  3:                                                                                                                Partly cloudy.
#>  4:                                                                                                                Partly cloudy.
#>  5:                                                                                                                        Sunny.
#>  6:                                                                                                                        Clear.
#>  7:                                                                                                                Partly cloudy.
#>  8:                                                                                                                Partly cloudy.
#>  9:                                                                                                                        Sunny.
#> 10:                                                                                                                        Clear.
#> 11:                                                                                                                Partly cloudy.
#> 12:                                                                                                                Partly cloudy.
#> 13:                                                                                                                        Sunny.
#> 14:                                                                                                                        Clear.
#> 15:                                                                                                                        Clear.
#> 16:                                                                                                                       Cloudy.
#> 17:                                                                                                               Cloud clearing.
#> 18:                                                   Mostly clear. The chance of a thunderstorm offshore north of Port Stephens.
#> 19:                                                      Partly cloudy. 50% chance of showers in the morning and early afternoon.
#> 20:                                                                                                                       Cloudy.
#> 21:                                                                                                                Partly cloudy.
#> 22:                                                                                  Partly cloudy. The chance of a thunderstorm.
#> 23:                      Partly cloudy. 50% chance of showers. The chance of a thunderstorm inshore in the afternoon and evening.
#> 24:                                                                                                                       Cloudy.
#> 25:                                                                                                                Partly cloudy.
#> 26:                                                                                  Partly cloudy. The chance of a thunderstorm.
#> 27: Partly cloudy. 60% chance of showers in the afternoon and evening. The chance of a thunderstorm in the afternoon and evening.
#> 28:                                                                                                                       Cloudy.
#> 29:                                                                                                                Partly cloudy.
#> 30:                                                                                  Partly cloudy. The chance of a thunderstorm.
#> 31:                              Partly cloudy. 80% chance of showers. The chance of a thunderstorm in the afternoon and evening.
#> 32:                                                                                                                Partly cloudy.
#> 33:                                                                                                                Partly cloudy.
#>                                                                                                                  forecast_weather
#>                                                                                                                                                          forecast_winds
#>  1:                                                                                                                                                                <NA>
#>  2:                                                      West to southwesterly 15 to 20 knots, reaching up to 35 knots offshore south of Merimbula in the late evening.
#>  3:                   Southwesterly 20 to 30 knots, reaching up to 35 knots offshore south of Merimbula early in the morning. Winds turning southerly in the afternoon.
#>  4:                    South to southwesterly 10 to 15 knots turning east to southeasterly below 10 knots in the early afternoon then becoming variable in the evening.
#>  5:                                                                                                                                           Northerly 15 to 25 knots.
#>  6:                                                                                                                                       Southeasterly 10 to 15 knots.
#>  7:                      South to southeasterly 15 to 20 knots becoming southerly 20 to 30 knots early in the morning then decreasing to 20 to 25 knots in the evening.
#>  8:                                                                                 Southerly 15 to 20 knots becoming variable about 10 knots in the middle of the day.
#>  9:                                                                          Variable about 10 knots becoming north to northeasterly 15 to 25 knots during the morning.
#> 10:                                                                                  Northeasterly 10 to 15 knots becoming variable about 10 knots in the late evening.
#> 11:                        Variable about 10 knots becoming southerly 15 to 25 knots before dawn. Winds reaching up to 30 knots during the morning and early afternoon.
#> 12:                                                                        Southerly 15 to 20 knots turning east to southeasterly below 10 knots in the late afternoon.
#> 13:                                                                          Variable about 10 knots becoming north to northeasterly 15 to 25 knots during the morning.
#> 14:                                                                                                                                       Northeasterly 10 to 15 knots.
#> 15: North to northeasterly 10 to 15 knots becoming southerly 15 to 25 knots early in the morning. Winds reaching up to 30 knots during the morning and early afternoon.
#> 16:                                                                                       Southerly 15 to 20 knots turning southeasterly 10 to 15 knots in the evening.
#> 17:                        Variable below 10 knots becoming northerly 10 to 15 knots during the morning then tending northeasterly 15 to 25 knots during the afternoon.
#> 18:                                                                                                                              North to northeasterly 15 to 20 knots.
#> 19:              Northerly 15 to 20 knots shifting southerly 15 to 25 knots early in the morning. Winds reaching up to 30 knots during the morning and early afternoon.
#> 20:                                                                                       Southerly 15 to 20 knots turning southeasterly 10 to 15 knots in the evening.
#> 21:                                                                                       Variable about 10 knots becoming northeasterly 15 to 25 knots during the day.
#> 22:                                                                                                                                           Northerly 15 to 20 knots.
#> 23:            North to northwesterly 15 to 20 knots shifting southerly 20 to 25 knots in the morning. Winds reaching up to 30 knots in the late morning and afternoon.
#> 24:                                                                                  Southerly 15 to 20 knots turning southeasterly 10 to 15 knots in the late evening.
#> 25:                                                                             Southeasterly 10 to 15 knots turning northeasterly 15 to 20 knots during the afternoon.
#> 26:                                                                                       Northeasterly 15 to 20 knots turning northerly 20 to 25 knots in the evening.
#> 27:                          Northerly 20 to 25 knots shifting southerly 15 to 25 knots in the morning. Winds reaching up to 30 knots during the afternoon and evening.
#> 28:                                                                                       Southerly 20 to 25 knots turning southeasterly 15 to 20 knots in the evening.
#> 29:                                                                                            Southeasterly 10 to 15 knots turning northeasterly during the afternoon.
#> 30:                               Northeasterly 10 to 15 knots turning northerly in the evening. Winds reaching up to 25 knots south of Cape Byron in the late evening.
#> 31:                               North to northwesterly 15 to 25 knots shifting south to southeasterly in the afternoon. Winds reaching up to 30 knots in the evening.
#> 32:                                                                                              Southeasterly 20 to 30 knots decreasing to 15 to 25 knots before dawn.
#> 33:                                                                                                 Southeasterly 10 to 15 knots turning easterly during the afternoon.
#>                                                                                                                                                          forecast_winds
#>                                                                                                                               forecast_swell1
#>  1:                                                                                                                                      <NA>
#>  2:                                                       South to southwesterly 1 to 2 metres inshore, increasing to 2 to 3 metres offshore.
#>  3: South to southwesterly 2 to 3 metres, increasing to 3 to 5 metres during the morning, then tending southerly 4 to 6 metres around midday.
#>  4:                                                                                                                  Southerly 4 to 5 metres.
#>  5:                                                                Southerly 3 to 4 metres, decreasing to 2.5 to 3 metres during the morning.
#>  6:                                                                Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#>  7:                                                                       Southerly 2 to 3 metres, increasing to 4 to 5 metres around midday.
#>  8:                                                                                                                  Southerly 4 to 5 metres.
#>  9:                                                                  Southerly 2 to 3 metres inshore, increasing to 2.5 to 4 metres offshore.
#> 10:                                                                     Southerly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 11:                                                                Southerly 2 to 3 metres, increasing to 4 to 5 metres during the afternoon.
#> 12:                                                                                                                  Southerly 4 to 5 metres.
#> 13:                                                                Southerly 3 to 4 metres, decreasing to 2.5 to 3 metres during the morning.
#> 14:                                                                     Southerly 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 15:                                                                Southerly 1.5 to 2 metres, increasing to 3 to 5 metres during the morning.
#> 16:                                                                                                                  Southerly 4 to 5 metres.
#> 17:                                                                       Southerly 3 to 4 metres, decreasing to 3 metres during the morning.
#> 18:                                                                                                                       Southerly 2 metres.
#> 19:                                                                   Southerly 1.5 to 2 metres, increasing to 2.5 to 4 metres around midday.
#> 20:                                                                                                                  Southerly 4 to 5 metres.
#> 21:                                                                     Southerly 3 to 4 metres, decreasing to 3 metres during the afternoon.
#> 22:                                                                Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 23:                                                              Southerly 1.5 to 2 metres, increasing to 2 to 3 metres during the afternoon.
#> 24:                                                                                                                  Southerly 3 to 5 metres.
#> 25:                                                          Southerly 3 to 4 metres, decreasing to 3 metres during the afternoon or evening.
#> 26:                                                                 Southerly around 1 metre inshore, increasing to 1.5 to 2 metres offshore.
#> 27:                                       Southerly 1 to 1.5 metres inshore, increasing to 1.5 to 2.5 metres offshore south of Coffs Harbour.
#> 28:                                                                Southerly 2.5 to 3 metres, increasing to 3 to 5 metres during the morning.
#> 29:                                                                  Southerly 2.5 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#> 30:                                                                      Southerly around 1 metre inshore, increasing to 1.5 metres offshore.
#> 31:                                                                 Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 32:                                                                Southerly 1.5 to 2 metres, increasing to 2 to 4 metres during the morning.
#> 33:                                                                    Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#>                                                                                                                               forecast_swell1
#>                   forecast_swell2
#>  1:                          <NA>
#>  2:                          <NA>
#>  3:                          <NA>
#>  4:                          <NA>
#>  5:                          <NA>
#>  6:                          <NA>
#>  7:                          <NA>
#>  8:                          <NA>
#>  9:                          <NA>
#> 10:                          <NA>
#> 11:                          <NA>
#> 12:                          <NA>
#> 13:                          <NA>
#> 14:                          <NA>
#> 15:                          <NA>
#> 16:                          <NA>
#> 17:                          <NA>
#> 18:                          <NA>
#> 19:                          <NA>
#> 20:                          <NA>
#> 21:                          <NA>
#> 22:                          <NA>
#> 23: Northeasterly around 1 metre.
#> 24:                          <NA>
#> 25:                          <NA>
#> 26:                          <NA>
#> 27:                          <NA>
#> 28:                          <NA>
#> 29:                          <NA>
#> 30:                          <NA>
#> 31:                          <NA>
#> 32:                          <NA>
#> 33:                          <NA>
#>                   forecast_swell2
#>                                                                                                                                                        forecast_caution
#>  1:                                                                                                                                                                <NA>
#>  2:                                                                                                                                                                <NA>
#>  3:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  4:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  5:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  6:                                                                                                                                                                <NA>
#>  7:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  8:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  9:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 10:                                                                                                                                                                <NA>
#> 11: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 12:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 13:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 14:                                                                                                                                                                <NA>
#> 15: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 16:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 17:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 18:                                                                                                                                                                <NA>
#> 19: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 20:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 21:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 22:                                                                                                                                                                <NA>
#> 23: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 24:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 25:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 26:                                                                                                                                                                <NA>
#> 27: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 28:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 29:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 30:                                                                                                                                                                <NA>
#> 31:                                                                                                                                                                <NA>
#> 32:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 33:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>                                                                                                                                                        forecast_caution
#>                                         marine_forecast
#>  1:                                                <NA>
#>  2:              Gale Warning for Monday for Eden Coast
#>  3:             Gale Warning for Tuesday for Eden Coast
#>  4:                                                <NA>
#>  5:                                                <NA>
#>  6:                                                <NA>
#>  7:  Strong Wind Warning for Tuesday for Batemans Coast
#>  8:                                                <NA>
#>  9:                                                <NA>
#> 10:                                                <NA>
#> 11: Strong Wind Warning for Tuesday for Illawarra Coast
#> 12:                                                <NA>
#> 13:                                                <NA>
#> 14:                                                <NA>
#> 15:    Strong Wind Warning for Tuesday for Sydney Coast
#> 16:                                                <NA>
#> 17:                                                <NA>
#> 18:                                                <NA>
#> 19:    Strong Wind Warning for Tuesday for Hunter Coast
#> 20:                                                <NA>
#> 21:                                                <NA>
#> 22:                                                <NA>
#> 23: Strong Wind Warning for Tuesday for Macquarie Coast
#> 24:                                                <NA>
#> 25:                                                <NA>
#> 26:                                                <NA>
#> 27:     Strong Wind Warning for Tuesday for Coffs Coast
#> 28:                                                <NA>
#> 29:                                                <NA>
#> 30:                                                <NA>
#> 31:     Strong Wind Warning for Tuesday for Byron Coast
#> 32:                                                <NA>
#> 33:                                                <NA>
#>                                         marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#>     tropical_system_location forecast_waves

# }