Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2020-12-13 16:05:11 2020-12-13 16:05:11      11:00 2020-12-13 16:05:11
#>  2: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#>  3: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#>  4: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#>  5: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#>  6: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#>  7: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#>  8: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#>  9: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 10: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 11: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 12: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 13: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 14: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 15: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 16: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 17: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 18: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 19: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 20: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 21: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 22: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 23: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 24: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 25: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 26: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 27: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 28: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 29: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#> 30: 2020-12-13 17:00:00 2020-12-13 17:00:00      11:00 2020-12-13 17:00:00
#> 31: 2020-12-14 00:00:00 2020-12-14 00:00:00      11:00 2020-12-14 00:00:00
#> 32: 2020-12-15 00:00:00 2020-12-15 00:00:00      11:00 2020-12-15 00:00:00
#> 33: 2020-12-16 00:00:00 2020-12-16 00:00:00      11:00 2020-12-16 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2020-12-13 16:05:11
#>  2: 2020-12-13 17:00:00
#>  3: 2020-12-14 00:00:00
#>  4: 2020-12-15 00:00:00
#>  5: 2020-12-16 00:00:00
#>  6: 2020-12-13 17:00:00
#>  7: 2020-12-14 00:00:00
#>  8: 2020-12-15 00:00:00
#>  9: 2020-12-16 00:00:00
#> 10: 2020-12-13 17:00:00
#> 11: 2020-12-14 00:00:00
#> 12: 2020-12-15 00:00:00
#> 13: 2020-12-16 00:00:00
#> 14: 2020-12-13 17:00:00
#> 15: 2020-12-14 00:00:00
#> 16: 2020-12-15 00:00:00
#> 17: 2020-12-16 00:00:00
#> 18: 2020-12-13 17:00:00
#> 19: 2020-12-14 00:00:00
#> 20: 2020-12-15 00:00:00
#> 21: 2020-12-16 00:00:00
#> 22: 2020-12-13 17:00:00
#> 23: 2020-12-14 00:00:00
#> 24: 2020-12-15 00:00:00
#> 25: 2020-12-16 00:00:00
#> 26: 2020-12-13 17:00:00
#> 27: 2020-12-14 00:00:00
#> 28: 2020-12-15 00:00:00
#> 29: 2020-12-16 00:00:00
#> 30: 2020-12-13 17:00:00
#> 31: 2020-12-14 00:00:00
#> 32: 2020-12-15 00:00:00
#> 33: 2020-12-16 00:00:00
#>            end_time_utc
#>                                                                                                                     forecast_seas
#>  1:                                                                                                                          <NA>
#>  2:                                                                                                              1.5 to 2 metres.
#>  3:                                                                                                              1.5 to 2 metres.
#>  4:                                                                                                                1 to 2 metres.
#>  5:                                                                        Around 1 metre, increasing to 1 to 1.5 metres inshore.
#>  6:                                                                                                              1 to 1.5 metres.
#>  7:                                                                                                              1 to 1.5 metres.
#>  8: 1 to 1.5 metres, decreasing to 1 metre during the morning, increasing to 1 to 1.5 metres north of Moruya Heads around midday.
#>  9:                                                 Around 1 metre, increasing to 1 to 1.5 metres offshore north of Moruya Heads.
#> 10:                                                                                                              1 to 1.5 metres.
#> 11:                                                                                                              1 to 1.5 metres.
#> 12:                                                 1 to 1.5 metres, increasing to 1.5 to 2 metres offshore north of Port Kembla.
#> 13:                                                                                                              1 to 1.5 metres.
#> 14:                                                                                                              1 to 1.5 metres.
#> 15:                                                                                                              1 to 1.5 metres.
#> 16:                                                                      1 to 1.5 metres, increasing to 1.5 to 2 metres offshore.
#> 17:                                                            1.5 to 2 metres, decreasing to 1 to 1.5 metres during the morning.
#> 18:                                                                                                                   1.5 metres.
#> 19:                                                                                                            1.5 to 2.5 metres.
#> 20:                                                    1.5 to 2 metres, increasing to 2 to 3 metres offshore north of Norah Head.
#> 21:                                                                                                            1.5 to 2.5 metres.
#> 22:                                                                                                              1.5 to 2 metres.
#> 23:                                                                                                                2 to 3 metres.
#> 24:                                                                                                                2 to 3 metres.
#> 25:                                                                                                            1.5 to 2.5 metres.
#> 26:                                                                                                                2 to 3 metres.
#> 27:                                                                                                              2.5 to 4 metres.
#> 28:                                                              2.5 to 4 metres, decreasing to 2 to 3 metres during the morning.
#> 29:                                                                                                            1.5 to 2.5 metres.
#> 30:                                                                                                                3 to 4 metres.
#> 31:                                                                                                                3 to 4 metres.
#> 32:                                                              2.5 to 4 metres, decreasing to 2 to 3 metres during the morning.
#> 33:                                                                                                            1.5 to 2.5 metres.
#>                                                                                                                     forecast_seas
#>                                                                              forecast_weather
#>  1:                                                                                      <NA>
#>  2:                                                                            Partly cloudy.
#>  3:                                                                            Partly cloudy.
#>  4:                                                     Partly cloudy. 50% chance of showers.
#>  5:                                                            Cloudy. 90% chance of showers.
#>  6:                                                                            Partly cloudy.
#>  7:                                                                                   Cloudy.
#>  8:                                                            Cloudy. 70% chance of showers.
#>  9:                              Cloudy. 90% chance of showers. The chance of a thunderstorm.
#> 10:                                                                            Partly cloudy.
#> 11:                                                            Cloudy. 50% chance of showers.
#> 12:                                                            Cloudy. 90% chance of showers.
#> 13:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 14:                                                                            Partly cloudy.
#> 15:                                                            Cloudy. 60% chance of showers.
#> 16:                                                            Cloudy. 90% chance of showers.
#> 17:                              Cloudy. 60% chance of showers. The chance of a thunderstorm.
#> 18:                                                            Cloudy. 50% chance of showers.
#> 19: Cloudy. 80% chance of showers. The chance of a thunderstorm in the afternoon and evening.
#> 20:          Cloudy. 90% chance of rain. The chance of a thunderstorm north of Port Stephens.
#> 21:                       Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#> 22:                                                               Cloudy. 80% chance of rain.
#> 23:    Cloudy. 95% chance of rain. The chance of a thunderstorm in the afternoon and evening.
#> 24:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 25:                       Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 26:                                                         Cloudy. Near 100% chance of rain.
#> 27:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 28:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 29:                       Partly cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 30:                           Cloudy. Near 100% chance of rain. The chance of a thunderstorm.
#> 31:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 32:                       Partly cloudy. 90% chance of showers. The chance of a thunderstorm.
#> 33:                       Partly cloudy. 80% chance of showers. The chance of a thunderstorm.
#>                                                                              forecast_weather
#>                                                                                            forecast_winds
#>  1:                                                                                                  <NA>
#>  2:                                 Northeasterly 15 to 20 knots, reaching up to 25 knots in the evening.
#>  3:                                                                         Northeasterly 15 to 25 knots.
#>  4:                                                                North to northeasterly 15 to 20 knots.
#>  5:           Northeasterly 10 to 15 knots shifting south to southwesterly below 10 knots during the day.
#>  6:                                 Easterly 10 to 15 knots, reaching up to 20 knots in the late evening.
#>  7:                                                                         Northeasterly 15 to 20 knots.
#>  8:                                                                         Northeasterly 15 to 20 knots.
#>  9:           North to northeasterly 10 to 15 knots shifting south to southwesterly during the afternoon.
#> 10:                                      Easterly 10 to 15 knots, reaching up to 20 knots in the evening.
#> 11:                                                                              Easterly 15 to 20 knots.
#> 12:                                                                 East to northeasterly 15 to 20 knots.
#> 13:                                    Northeasterly 15 to 20 knots turning northerly during the evening.
#> 14:                                                                              Easterly 15 to 20 knots.
#> 15:                                                                              Easterly 15 to 20 knots.
#> 16:                               Easterly 15 to 20 knots turning northeasterly in the middle of the day.
#> 17:                                                                         Northeasterly 15 to 20 knots.
#> 18:                                                                              Easterly 15 to 20 knots.
#> 19:                                                                              Easterly 20 to 30 knots.
#> 20: Easterly 20 to 30 knots turning northeasterly 20 to 25 knots in the late morning and early afternoon.
#> 21:                                                                         Northeasterly 20 to 30 knots.
#> 22:                                                                         Southeasterly 20 to 30 knots.
#> 23:                                                                 East to southeasterly 20 to 30 knots.
#> 24:                      Easterly 25 to 30 knots turning northeasterly 20 to 30 knots during the morning.
#> 25:                     Northeasterly 20 to 30 knots tending northerly 25 to 30 knots during the evening.
#> 26:                                                                         Southeasterly 25 to 30 knots.
#> 27:                                                                 East to southeasterly 25 to 35 knots.
#> 28:                    Easterly 25 to 30 knots turning northeasterly 20 to 30 knots early in the morning.
#> 29:                                                                North to northeasterly 20 to 30 knots.
#> 30:                                                                         Southeasterly 30 to 40 knots.
#> 31:      Southeasterly 30 to 40 knots turning east to northeasterly 25 to 35 knots in the late afternoon.
#> 32:                     Northeasterly 20 to 30 knots decreasing to 20 to 25 knots in the early afternoon.
#> 33:                                                                         Northeasterly 15 to 25 knots.
#>                                                                                            forecast_winds
#>                                                                                                                       forecast_swell1
#>  1:                                                                                                                              <NA>
#>  2:                                                                                                         Southerly around 1 metre.
#>  3:  Northeasterly below 1 metre, increasing to 1 to 1.5 metres during the morning, then increasing to 1.5 to 2 metres around midday.
#>  4:                                                    Northeasterly 1.5 to 2 metres, increasing to 2 to 3 metres during the morning.
#>  5:                                                                                                    Northeasterly 2.5 to 3 metres.
#>  6:                                                                                                         Southerly around 1 metre.
#>  7:                                                  Northeasterly 1 to 1.5 metres, increasing to 1.5 to 2 metres during the morning.
#>  8:                                                                                                      Northeasterly 2 to 3 metres.
#>  9:                                                                                                    Northeasterly 2.5 to 3 metres.
#> 10:                                                                                                         Northeasterly 1.5 metres.
#> 11:                                                       Northeasterly 1.5 metres, increasing to 2 to 2.5 metres during the morning.
#> 12:        Northeasterly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore north of Point Perpendicular during the morning.
#> 13:                                                           Northeasterly 2 to 3 metres, decreasing to 2 metres during the evening.
#> 14:                                                                                                    Northeasterly 1.5 to 2 metres.
#> 15:                                                                                                  Northeasterly 1.5 to 2.5 metres.
#> 16:                                     Northeasterly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore during the morning.
#> 17:                                              Northeasterly 2 to 3 metres, decreasing to 2 metres during the afternoon or evening.
#> 18:                                                                                                  Northeasterly 1.5 to 2.5 metres.
#> 19:                                                                                                      Northeasterly 2 to 3 metres.
#> 20:  Northeasterly 2.5 to 3 metres, increasing to 3 to 4 metres during the morning, then decreasing to 3 metres during the afternoon.
#> 21:                                                         Northeasterly 2 to 3 metres, decreasing to 2 metres during the afternoon.
#> 22:                                                                                                         Northeasterly 2.5 metres.
#> 23:                                                                                            East to northeasterly 2.5 to 3 metres.
#> 24:      Easterly 3 metres, tending northeasterly 3 to 4 metres during the morning, then decreasing to 3 metres during the afternoon.
#> 25:                                                                                                    Northeasterly 2 to 2.5 metres.
#> 26:                                                                                                                Easterly 3 metres.
#> 27:                                                                                                           Easterly 3 to 4 metres.
#> 28: Easterly 4 to 5 metres, tending northeasterly 3 to 4 metres during the morning, then decreasing to 2.5 to 3 metres around midday.
#> 29:                                                       Northeasterly 2 to 2.5 metres, decreasing to 1.5 metres during the evening.
#> 30:                                                                                                           Easterly 3 to 4 metres.
#> 31:                                                                                                           Easterly 3 to 5 metres.
#> 32:                                                         Easterly 3 to 4 metres, decreasing to 2.5 to 3 metres during the morning.
#> 33:                                                                                          East to northeasterly 1.5 to 2.5 metres.
#>                                                                                                                       forecast_swell1
#>                           forecast_swell2
#>  1:                                  <NA>
#>  2:                                  <NA>
#>  3:                                  <NA>
#>  4:                                  <NA>
#>  5:                                  <NA>
#>  6: East to northeasterly around 1 metre.
#>  7:             Southerly around 1 metre.
#>  8:                                  <NA>
#>  9:                                  <NA>
#> 10:             Southerly around 1 metre.
#> 11:             Southerly around 1 metre.
#> 12:                                  <NA>
#> 13:                                  <NA>
#> 14:             Southerly around 1 metre.
#> 15:             Southerly around 1 metre.
#> 16:                                  <NA>
#> 17:                                  <NA>
#> 18:             Southerly around 1 metre.
#> 19:             Southerly around 1 metre.
#> 20:                                  <NA>
#> 21:                                  <NA>
#> 22:             Southerly around 1 metre.
#> 23:             Southerly around 1 metre.
#> 24:                                  <NA>
#> 25:                                  <NA>
#> 26:             Southerly around 1 metre.
#> 27:             Southerly around 1 metre.
#> 28:                                  <NA>
#> 29:                                  <NA>
#> 30:             Southerly around 1 metre.
#> 31:                                  <NA>
#> 32:                                  <NA>
#> 33:                                  <NA>
#>                           forecast_swell2
#>                                                                                                                                                        forecast_caution
#>  1:                                                                                                                                                                <NA>
#>  2:                                                                                                                                                                <NA>
#>  3:                                                                                                                                                                <NA>
#>  4:                                                                                                                                                                <NA>
#>  5:                                                                                                                                                                <NA>
#>  6:                                                                                                                                                                <NA>
#>  7:                                                                                                                                                                <NA>
#>  8:                                                                                                                                                                <NA>
#>  9:                                                                                                                                                                <NA>
#> 10:                                                                                                                                                                <NA>
#> 11:                                                                                                                                                                <NA>
#> 12:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 13:                                                                                                                                                                <NA>
#> 14:                                                                                                                                                                <NA>
#> 15:                                                                                                                                                                <NA>
#> 16:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 17:                                                                                                                                                                <NA>
#> 18:                                                                                                                                                                <NA>
#> 19:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 20:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 21:                                                                                                                                                                <NA>
#> 22:                                                                                                                                                                <NA>
#> 23:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 24:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 25:                                                                                                                                                                <NA>
#> 26: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 27:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 28:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 29:                                                                                                                                                                <NA>
#> 30: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 31:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 32:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 33:                                                                                                                                                                <NA>
#>                                                                                                                                                        forecast_caution
#>                                        marine_forecast tropical_system_location
#>  1:                                               <NA>                       NA
#>  2:                                               <NA>                       NA
#>  3:                                               <NA>                       NA
#>  4:                                               <NA>                       NA
#>  5:                                               <NA>                       NA
#>  6:                                               <NA>                       NA
#>  7:                                               <NA>                       NA
#>  8:                                               <NA>                       NA
#>  9:                                               <NA>                       NA
#> 10:                                               <NA>                       NA
#> 11:                                               <NA>                       NA
#> 12:                                               <NA>                       NA
#> 13:                                               <NA>                       NA
#> 14:                                               <NA>                       NA
#> 15:                                               <NA>                       NA
#> 16:                                               <NA>                       NA
#> 17:                                               <NA>                       NA
#> 18:                                               <NA>                       NA
#> 19:    Strong Wind Warning for Monday for Hunter Coast                       NA
#> 20:                                               <NA>                       NA
#> 21:                                               <NA>                       NA
#> 22: Strong Wind Warning for Sunday for Macquarie Coast                       NA
#> 23: Strong Wind Warning for Monday for Macquarie Coast                       NA
#> 24:                                               <NA>                       NA
#> 25:                                               <NA>                       NA
#> 26:     Strong Wind Warning for Sunday for Coffs Coast                       NA
#> 27:            Gale Warning for Monday for Coffs Coast                       NA
#> 28:                                               <NA>                       NA
#> 29:                                               <NA>                       NA
#> 30:            Gale Warning for Sunday for Byron Coast                       NA
#> 31:            Gale Warning for Monday for Byron Coast                       NA
#> 32:                                               <NA>                       NA
#> 33:                                               <NA>                       NA
#>                                        marine_forecast tropical_system_location
#>     forecast_waves
#>  1:             NA
#>  2:             NA
#>  3:             NA
#>  4:             NA
#>  5:             NA
#>  6:             NA
#>  7:             NA
#>  8:             NA
#>  9:             NA
#> 10:             NA
#> 11:             NA
#> 12:             NA
#> 13:             NA
#> 14:             NA
#> 15:             NA
#> 16:             NA
#> 17:             NA
#> 18:             NA
#> 19:             NA
#> 20:             NA
#> 21:             NA
#> 22:             NA
#> 23:             NA
#> 24:             NA
#> 25:             NA
#> 26:             NA
#> 27:             NA
#> 28:             NA
#> 29:             NA
#> 30:             NA
#> 31:             NA
#> 32:             NA
#> 33:             NA
#>     forecast_waves

# }