Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-01-20 16:05:12 2021-01-20 16:05:12      11:00 2021-01-20 16:05:12
#>  2: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#>  3: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#>  4: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#>  5: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#>  6: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#>  7: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#>  8: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#>  9: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 10: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 11: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 12: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 13: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 14: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 15: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 16: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 17: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 18: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 19: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 20: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 21: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 22: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 23: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 24: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 25: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 26: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 27: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 28: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 29: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#> 30: 2021-01-20 17:00:00 2021-01-20 17:00:00      11:00 2021-01-20 17:00:00
#> 31: 2021-01-21 00:00:00 2021-01-21 00:00:00      11:00 2021-01-21 00:00:00
#> 32: 2021-01-22 00:00:00 2021-01-22 00:00:00      11:00 2021-01-22 00:00:00
#> 33: 2021-01-23 00:00:00 2021-01-23 00:00:00      11:00 2021-01-23 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-01-20 16:05:12
#>  2: 2021-01-20 17:00:00
#>  3: 2021-01-21 00:00:00
#>  4: 2021-01-22 00:00:00
#>  5: 2021-01-23 00:00:00
#>  6: 2021-01-20 17:00:00
#>  7: 2021-01-21 00:00:00
#>  8: 2021-01-22 00:00:00
#>  9: 2021-01-23 00:00:00
#> 10: 2021-01-20 17:00:00
#> 11: 2021-01-21 00:00:00
#> 12: 2021-01-22 00:00:00
#> 13: 2021-01-23 00:00:00
#> 14: 2021-01-20 17:00:00
#> 15: 2021-01-21 00:00:00
#> 16: 2021-01-22 00:00:00
#> 17: 2021-01-23 00:00:00
#> 18: 2021-01-20 17:00:00
#> 19: 2021-01-21 00:00:00
#> 20: 2021-01-22 00:00:00
#> 21: 2021-01-23 00:00:00
#> 22: 2021-01-20 17:00:00
#> 23: 2021-01-21 00:00:00
#> 24: 2021-01-22 00:00:00
#> 25: 2021-01-23 00:00:00
#> 26: 2021-01-20 17:00:00
#> 27: 2021-01-21 00:00:00
#> 28: 2021-01-22 00:00:00
#> 29: 2021-01-23 00:00:00
#> 30: 2021-01-20 17:00:00
#> 31: 2021-01-21 00:00:00
#> 32: 2021-01-22 00:00:00
#> 33: 2021-01-23 00:00:00
#>            end_time_utc
#>                                                                                                                   forecast_seas
#>  1:                                                                                                                        <NA>
#>  2:                                                                                                              Below 1 metre.
#>  3:                                                             Around 1 metre, increasing to 1 to 2 metres during the morning.
#>  4:                                                                 1 to 2 metres, decreasing to 1 to 1.5 metres around midday.
#>  5:                                                                  1 to 1.5 metres, decreasing to 1 metre during the morning.
#>  6:                                                                                                              Below 1 metre.
#>  7: Below 1 metre, increasing to 1 to 1.5 metres during the morning, then increasing to 1.5 to 2.5 metres during the afternoon.
#>  8:                                                                                                          1.5 to 2.5 metres.
#>  9:                                                                  1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 10:                                                                                                              Below 1 metre.
#> 11:                                                                   Below 1 metre, increasing to 1 to 2 metres around midday.
#> 12:                                                                                                          1.5 to 2.5 metres.
#> 13:                                                                                                              1 to 2 metres.
#> 14:                                                                                                              Below 1 metre.
#> 15:                                                                Below 1 metre, increasing to 1 to 2 metres in the afternoon.
#> 16:                                                                                                          1.5 to 2.5 metres.
#> 17:                                                        1.5 to 2.5 metres, decreasing to 1 to 1.5 metres during the morning.
#> 18:                                                                                                             Around 1 metre.
#> 19:                                                    Below 1 metre, increasing to 1 to 2 metres in the afternoon and evening.
#> 20:                                                                                                            1.5 to 2 metres.
#> 21:                                                                                                          1.5 to 2.5 metres.
#> 22:                                                                                                             Around 1 metre.
#> 23:                                                                                                             Around 1 metre.
#> 24:                                                 1 to 1.5 metres, increasing to 1.5 to 2 metres offshore during the morning.
#> 25:                                                                                                            1.5 to 2 metres.
#> 26:                                                                                                            1 to 1.5 metres.
#> 27:                                                                                                            1 to 1.5 metres.
#> 28:                                                                 Around 1 metre, increasing to 1 to 2 metres during the day.
#> 29:                                                      1 to 1.5 metres, increasing to 1.5 to 2 metres south of Coffs Harbour.
#> 30:                                                              1.5 metres, increasing to 1.5 to 2 metres north of Cape Byron.
#> 31:                                                                  1 to 1.5 metres, decreasing to 1 metre during the morning.
#> 32:                                                          Below 1 metre, increasing to 1 to 1.5 metres during the afternoon.
#> 33:                                                                                                            1 to 1.5 metres.
#>                                                                                                                   forecast_seas
#>                                                                            forecast_weather
#>  1:                                                                                    <NA>
#>  2:                                                                           Mostly clear.
#>  3:                                                                          Partly cloudy.
#>  4:                                                                           Mostly sunny.
#>  5:                                                                                  Sunny.
#>  6:                                                                           Mostly clear.
#>  7:                                                                           Mostly sunny.
#>  8:                                                                           Mostly sunny.
#>  9:                                                                                  Sunny.
#> 10:                                                                          Partly cloudy.
#> 11:                                                                           Mostly sunny.
#> 12:                                                                           Mostly sunny.
#> 13:                                                                                  Sunny.
#> 14:                                                                          Partly cloudy.
#> 15:                                                                          Partly cloudy.
#> 16:                                                                           Mostly sunny.
#> 17:                                                                                  Sunny.
#> 18:                                                                          Partly cloudy.
#> 19:                                                                          Partly cloudy.
#> 20:        Mostly sunny morning. The chance of a thunderstorm in the afternoon and evening.
#> 21:                                                                                  Sunny.
#> 22:                                                                                 Cloudy.
#> 23:                                                                          Partly cloudy.
#> 24:                    Mostly sunny. The chance of a thunderstorm inshore in the afternoon.
#> 25:                                                                                  Sunny.
#> 26:                                                                                 Cloudy.
#> 27:                                                                          Partly cloudy.
#> 28: Partly cloudy. The chance of a thunderstorm inshore in the afternoon and early evening.
#> 29:                                                                                  Sunny.
#> 30:                                                                                 Cloudy.
#> 31:                                                                          Partly cloudy.
#> 32:                                                                           Mostly sunny.
#> 33:                                                                           Mostly sunny.
#>                                                                            forecast_weather
#>                                                                                                                                                                                                                       forecast_winds
#>  1:                                                                                                                                                                                                                             <NA>
#>  2:                                                                                                                                                                                            East to southeasterly below 10 knots.
#>  3:                                                                                                   Variable below 10 knots becoming northerly 15 to 20 knots before dawn. Winds reaching up to 25 knots inshore in the afternoon.
#>  4:                                                                                                                                Northerly 15 to 25 knots tending southeast to southwesterly 15 to 20 knots in the late afternoon.
#>  5:                                                                             South to southwesterly 10 to 15 knots turning southeasterly in the afternoon, then tending north to northeasterly below 10 knots during the evening.
#>  6:                                                                                                                                                                                                         Variable below 10 knots.
#>  7:                                                                                           North to northeasterly about 10 knots increasing to 15 to 25 knots early in the morning. Winds reaching up to 30 knots in the evening.
#>  8:                                                                                                                                          Northerly 15 to 25 knots shifting south to southeasterly 15 to 20 knots in the evening.
#>  9:                                                                                                                                            Southerly 10 to 15 knots turning east to southeasterly below 10 knots during the day.
#> 10:                                                                                                                                                                                                         Variable about 10 knots.
#> 11:                                        Variable about 10 knots becoming north to northeasterly 15 to 25 knots in the middle of the day. Winds reaching up to 30 knots offshore south of Point Perpendicular in the late evening.
#> 12:                                                                                                                                                                Northerly 15 to 25 knots, shifting southerly in the late evening.
#> 13:                                                          North to northwesterly 15 to 20 knots shifting south to southwesterly 10 to 15 knots during the morning then tending northeasterly 15 to 20 knots during the afternoon.
#> 14:                                                                                                                                   South to southeasterly 10 to 15 knots turning southeasterly below 10 knots during the evening.
#> 15:                                                                                                                                            Variable about 10 knots becoming northeasterly 15 to 25 knots in the early afternoon.
#> 16:                                                                                                                                                                                                        Northerly 15 to 25 knots.
#> 17:                                                          North to northwesterly 15 to 20 knots turning south to southwesterly 10 to 15 knots in the early morning. Winds shifting northeasterly 15 to 25 knots in the afternoon.
#> 18:                                                                                                                                                    Southerly 10 to 15 knots turning southeasterly below 10 knots in the evening.
#> 19:                                                                                                                                                  Variable about 10 knots becoming northeasterly 15 to 25 knots in the afternoon.
#> 20:                                                                                                                                                                                                        Northerly 15 to 25 knots.
#> 21:                                                                                                                                                                                                        Northerly 15 to 25 knots.
#> 22:                                                                                                                                                       Southerly 10 to 15 knots decreasing to about 10 knots in the late evening.
#> 23: South to southeasterly about 10 knots tending east to northeasterly 10 to 15 knots in the early afternoon then tending north to northeasterly in the evening. Winds reaching up to 20 knots south of Crowdy Head in the evening.
#> 24:                                                                                                                                                                                                        Northerly 15 to 25 knots.
#> 25:                                                                                                                                                                                                        Northerly 15 to 25 knots.
#> 26:                                                                                                                                                                                           South to southeasterly 15 to 20 knots.
#> 27:                                                                                                                                             Southeasterly 10 to 15 knots turning northeasterly during the afternoon and evening.
#> 28:                                                                                                                                                                                                        Northerly 15 to 20 knots.
#> 29:                                                                                                                                                                                           North to northeasterly 15 to 20 knots.
#> 30:                                                                                                                                                                                                    Southeasterly 15 to 20 knots.
#> 31:                                                                                                                                 Southeasterly 10 to 15 knots turning east to northeasterly below 10 knots in the late afternoon.
#> 32:                                                                                                                                     North to northeasterly about 10 knots increasing to 15 to 20 knots in the middle of the day.
#> 33:                                                                                                                                                                   Northerly 15 to 20 knots turning northeasterly during the day.
#>                                                                                                                                                                                                                       forecast_winds
#>                                                                                      forecast_swell1
#>  1:                                                                                             <NA>
#>  2:                                   Southerly 2.5 to 3.5 metres, reaching up to 4 metres offshore.
#>  3:                  Southerly 2 to 3 metres, reaching up to 4 metres offshore in the early morning.
#>  4:                     Southerly 2 to 2.5 metres, decreasing to 1 to 1.5 metres during the morning.
#>  5:                                                                        Southerly around 1 metre.
#>  6:                           Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#>  7:                    Southerly 2 to 3 metres, reaching up to 4 metres offshore during the morning.
#>  8:                          Southerly 2 to 2.5 metres, decreasing to 1.5 metres during the morning.
#>  9:                     Northeasterly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 10:                                Southerly 2 to 3 metres inshore, increasing to 4 metres offshore.
#> 11:                    Southerly 2 to 3 metres, reaching up to 4 metres offshore during the morning.
#> 12:                               Southerly 2 to 2.5 metres, decreasing to 1.5 metres around midday.
#> 13: Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 14:                           Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#> 15:                  Southerly 2 to 3 metres, reaching up to 4 metres offshore in the early morning.
#> 16:                       Southerly 2 to 3 metres, decreasing to 1.5 to 2 metres during the morning.
#> 17:                                                                        Southerly around 1 metre.
#> 18:                                Southerly 2 to 3 metres inshore, increasing to 4 metres offshore.
#> 19:                        Southerly 2 to 3 metres, reaching up to 4 metres offshore in the morning.
#> 20:                       Southerly 2 to 3 metres, decreasing to 1.5 to 2 metres during the morning.
#> 21:                      Southerly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 22:                                                                              Southerly 4 metres.
#> 23:                              Southerly 3 to 4 metres, decreasing to 3 metres during the morning.
#> 24:                            Southerly 2 to 3 metres, decreasing to 2 metres during the afternoon.
#> 25:                        Southerly 1 to 2 metres, decreasing to around 1 metre during the morning.
#> 26:                                Southerly 2 to 3 metres inshore, increasing to 4 metres offshore.
#> 27:                           Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#> 28:                                                                         Southerly 2 to 3 metres.
#> 29:                            Southerly 2 metres, decreasing to 1 to 1.5 metres during the morning.
#> 30:                                Southerly 2 to 3 metres inshore, increasing to 4 metres offshore.
#> 31:                           Southerly 2 to 3 metres inshore, increasing to 3 to 4 metres offshore.
#> 32:                                                                         Southerly 2 to 3 metres.
#> 33:                   Southerly 1.5 to 2.5 metres, decreasing to 1 to 1.5 metres during the morning.
#>                                                                                      forecast_swell1
#>                                                       forecast_swell2
#>  1:                                                              <NA>
#>  2:                                                              <NA>
#>  3:                                                              <NA>
#>  4: North to northeasterly 1 to 1.5 metres developing during the day.
#>  5:                                     Northeasterly around 1 metre.
#>  6:                                                              <NA>
#>  7:                                                              <NA>
#>  8:                                                              <NA>
#>  9:                                         Southerly around 1 metre.
#> 10:                                                              <NA>
#> 11:                                                              <NA>
#> 12:                                                              <NA>
#> 13:                                         Southerly around 1 metre.
#> 14:                                                              <NA>
#> 15:                                                              <NA>
#> 16:                                                              <NA>
#> 17:                                     Northeasterly around 1 metre.
#> 18:                                                              <NA>
#> 19:                                                              <NA>
#> 20:                                                              <NA>
#> 21:                                                              <NA>
#> 22:                                                              <NA>
#> 23:                                                              <NA>
#> 24:                                                              <NA>
#> 25:                                                              <NA>
#> 26:                                                              <NA>
#> 27:                                                              <NA>
#> 28:                                                              <NA>
#> 29:                                                              <NA>
#> 30:                                                              <NA>
#> 31:                                                              <NA>
#> 32:                                                              <NA>
#> 33:                                                              <NA>
#>                                                       forecast_swell2
#>                                                                                                                                                        forecast_caution
#>  1:                                                                                                                                                                <NA>
#>  2: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  3:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  4:                                                                                                                                                                <NA>
#>  5:                                                                                                                                                                <NA>
#>  6: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  7:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#>  8:                                                                                                                                                                <NA>
#>  9:                                                                                                                                                                <NA>
#> 10: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 11:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 12:                                                                                                                                                                <NA>
#> 13:                                                                                                                                                                <NA>
#> 14: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 15:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 16:                                                                                                                                                                <NA>
#> 17:                                                                                                                                                                <NA>
#> 18: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 19:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 20:                                                                                                                                                                <NA>
#> 21:                                                                                                                                                                <NA>
#> 22: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 23:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 24:               Large and powerful surf conditions in the morning are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 25:                                                                                                                                                                <NA>
#> 26: Large and powerful surf conditions in the afternoon and evening are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 27:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 28:               Large and powerful surf conditions in the morning are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 29:                                                                                                                                                                <NA>
#> 30:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 31:                              Large and powerful surf conditions are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 32:               Large and powerful surf conditions in the morning are expected to be hazardous for coastal activities such as crossing bars by boat and rock fishing.
#> 33:                                                                                                                                                                <NA>
#>                                                                                                                                                        forecast_caution
#>                                          marine_forecast
#>  1:                                                 <NA>
#>  2:                                                 <NA>
#>  3:                                                 <NA>
#>  4:                                                 <NA>
#>  5:                                                 <NA>
#>  6:                                                 <NA>
#>  7:  Strong Wind Warning for Thursday for Batemans Coast
#>  8:                                                 <NA>
#>  9:                                                 <NA>
#> 10:                                                 <NA>
#> 11: Strong Wind Warning for Thursday for Illawarra Coast
#> 12:                                                 <NA>
#> 13:                                                 <NA>
#> 14:                                                 <NA>
#> 15:                                                 <NA>
#> 16:                                                 <NA>
#> 17:                                                 <NA>
#> 18:                                                 <NA>
#> 19:                                                 <NA>
#> 20:                                                 <NA>
#> 21:                                                 <NA>
#> 22:                                                 <NA>
#> 23:                                                 <NA>
#> 24:                                                 <NA>
#> 25:                                                 <NA>
#> 26:                                                 <NA>
#> 27:                                                 <NA>
#> 28:                                                 <NA>
#> 29:                                                 <NA>
#> 30:                                                 <NA>
#> 31:                                                 <NA>
#> 32:                                                 <NA>
#> 33:                                                 <NA>
#>                                          marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#>     tropical_system_location forecast_waves

# }