Get BOM coastal waters forecast

Fetch the BOM daily Coastal Waters Forecast and return a data frame of the forecast regions for a specified state or region.

get_coastal_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to “AUS” returning all state forecasts, see details for further information.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
coastal_forecast <- get_coastal_forecast(state = "NSW")
coastal_forecast

#>     index product_id type state_code
#>  1:  <NA>   IDN11001 <NA>        NSW
#>  2:     0   IDN11001 <NA>        NSW
#>  3:     1   IDN11001 <NA>        NSW
#>  4:     2   IDN11001 <NA>        NSW
#>  5:     3   IDN11001 <NA>        NSW
#>  6:     0   IDN11001 <NA>        NSW
#>  7:     1   IDN11001 <NA>        NSW
#>  8:     2   IDN11001 <NA>        NSW
#>  9:     3   IDN11001 <NA>        NSW
#> 10:     0   IDN11001 <NA>        NSW
#> 11:     1   IDN11001 <NA>        NSW
#> 12:     2   IDN11001 <NA>        NSW
#> 13:     3   IDN11001 <NA>        NSW
#> 14:     0   IDN11001 <NA>        NSW
#> 15:     1   IDN11001 <NA>        NSW
#> 16:     2   IDN11001 <NA>        NSW
#> 17:     3   IDN11001 <NA>        NSW
#> 18:     0   IDN11001 <NA>        NSW
#> 19:     1   IDN11001 <NA>        NSW
#> 20:     2   IDN11001 <NA>        NSW
#> 21:     3   IDN11001 <NA>        NSW
#> 22:     0   IDN11001 <NA>        NSW
#> 23:     1   IDN11001 <NA>        NSW
#> 24:     2   IDN11001 <NA>        NSW
#> 25:     3   IDN11001 <NA>        NSW
#> 26:     0   IDN11001 <NA>        NSW
#> 27:     1   IDN11001 <NA>        NSW
#> 28:     2   IDN11001 <NA>        NSW
#> 29:     3   IDN11001 <NA>        NSW
#> 30:     0   IDN11001 <NA>        NSW
#> 31:     1   IDN11001 <NA>        NSW
#> 32:     2   IDN11001 <NA>        NSW
#> 33:     3   IDN11001 <NA>        NSW
#>     index product_id type state_code
#>                                        dist_name pt_1_name pt_2_name       aac
#>  1:                              New South Wales      <NA>      <NA> NSW_FA001
#>  2:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  3:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  4:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  5:   Eden Coast: Montague Island to Gabo Island      <NA>      <NA> NSW_MW001
#>  6: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  7: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  8: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#>  9: Batemans Coast: Ulladulla to Montague Island      <NA>      <NA> NSW_MW002
#> 10:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 11:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 12:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 13:   Illawarra Coast: Port Hacking to Ulladulla      <NA>      <NA> NSW_MW003
#> 14:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 15:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 16:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 17:     Sydney Coast: Broken Bay to Port Hacking      <NA>      <NA> NSW_MW004
#> 18:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 19:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 20:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 21:       Hunter Coast: Seal Rocks to Broken Bay      <NA>      <NA> NSW_MW005
#> 22:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 23:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 24:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 25:    Macquarie Coast: Smoky Cape to Seal Rocks      <NA>      <NA> NSW_MW006
#> 26:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 27:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 28:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 29:             Coffs Coast: Wooli to Smoky Cape      <NA>      <NA> NSW_MW007
#> 30:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 31:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 32:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#> 33:           Byron Coast: Point Danger to Wooli      <NA>      <NA> NSW_MW008
#>                                        dist_name pt_1_name pt_2_name       aac
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>  1: 2021-01-13 16:05:14 2021-01-13 16:05:14      11:00 2021-01-13 16:05:14
#>  2: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#>  3: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#>  4: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#>  5: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#>  6: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#>  7: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#>  8: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#>  9: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 10: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 11: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 12: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 13: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 14: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 15: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 16: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 17: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 18: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 19: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 20: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 21: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 22: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 23: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 24: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 25: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 26: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 27: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 28: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 29: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#> 30: 2021-01-13 17:00:00 2021-01-13 17:00:00      11:00 2021-01-13 17:00:00
#> 31: 2021-01-14 00:00:00 2021-01-14 00:00:00      11:00 2021-01-14 00:00:00
#> 32: 2021-01-15 00:00:00 2021-01-15 00:00:00      11:00 2021-01-15 00:00:00
#> 33: 2021-01-16 00:00:00 2021-01-16 00:00:00      11:00 2021-01-16 00:00:00
#>        start_time_local      end_time_local utc_offset      start_time_utc
#>            end_time_utc
#>  1: 2021-01-13 16:05:14
#>  2: 2021-01-13 17:00:00
#>  3: 2021-01-14 00:00:00
#>  4: 2021-01-15 00:00:00
#>  5: 2021-01-16 00:00:00
#>  6: 2021-01-13 17:00:00
#>  7: 2021-01-14 00:00:00
#>  8: 2021-01-15 00:00:00
#>  9: 2021-01-16 00:00:00
#> 10: 2021-01-13 17:00:00
#> 11: 2021-01-14 00:00:00
#> 12: 2021-01-15 00:00:00
#> 13: 2021-01-16 00:00:00
#> 14: 2021-01-13 17:00:00
#> 15: 2021-01-14 00:00:00
#> 16: 2021-01-15 00:00:00
#> 17: 2021-01-16 00:00:00
#> 18: 2021-01-13 17:00:00
#> 19: 2021-01-14 00:00:00
#> 20: 2021-01-15 00:00:00
#> 21: 2021-01-16 00:00:00
#> 22: 2021-01-13 17:00:00
#> 23: 2021-01-14 00:00:00
#> 24: 2021-01-15 00:00:00
#> 25: 2021-01-16 00:00:00
#> 26: 2021-01-13 17:00:00
#> 27: 2021-01-14 00:00:00
#> 28: 2021-01-15 00:00:00
#> 29: 2021-01-16 00:00:00
#> 30: 2021-01-13 17:00:00
#> 31: 2021-01-14 00:00:00
#> 32: 2021-01-15 00:00:00
#> 33: 2021-01-16 00:00:00
#>            end_time_utc
#>                                                                                                   forecast_seas
#>  1:                                                                                                        <NA>
#>  2:                                                                                            2.5 to 3 metres.
#>  3:                                                                                              2 to 3 metres.
#>  4:                                          1 to 1.5 metres, increasing to 1.5 to 2.5 metres by early evening.
#>  5:                                    2 to 3 metres, increasing to 3 to 4 metres offshore south of Green Cape.
#>  6:                                                                                              2 to 3 metres.
#>  7:                                            2 to 3 metres, decreasing to 1 to 2 metres during the afternoon.
#>  8:                                                                                              1 to 2 metres.
#>  9:                                1.5 to 2.5 metres, decreasing below 1 metre during the afternoon or evening.
#> 10:                                                                                          1.5 to 2.5 metres.
#> 11:                                                                                          1.5 to 2.5 metres.
#> 12: 1 to 2 metres, decreasing below 1 metre around midday, then increasing to 1 to 1.5 metres by early evening.
#> 13:                                1.5 to 2.5 metres, decreasing below 1 metre during the afternoon or evening.
#> 14:                                                                                              1 to 2 metres.
#> 15:                                                                                          1.5 to 2.5 metres.
#> 16:                                                  1.5 to 2.5 metres, decreasing below 1 metre around midday.
#> 17:                                                  1 to 2 metres, decreasing to 1 metre during the afternoon.
#> 18:                                       Around 1 metre, increasing to 1 to 2 metres inshore by early evening.
#> 19:                                                                                          1.5 to 2.5 metres.
#> 20:                                                      2 to 3 metres, decreasing below 1 metre around midday.
#> 21:                                   1.5 to 2.5 metres, decreasing to 1 metre during the afternoon or evening.
#> 22:                                      Below 1 metre, increasing to 1 to 1.5 metres inshore by early evening.
#> 23:                  1 to 1.5 metres during the morning, then increasing to 1 to 2 metres during the afternoon.
#> 24:                                                                                              2 to 3 metres.
#> 25:                           1.5 to 2.5 metres, decreasing to 1 to 1.5 metres during the afternoon or evening.
#> 26:                                                                                              Below 1 metre.
#> 27:                                           Around 1 metre, increasing to 1 to 2 metres during the afternoon.
#> 28:                                                                                          1.5 to 2.5 metres.
#> 29:                                                1.5 to 2.5 metres, decreasing to 1 metre during the evening.
#> 30:                                                                                              Below 1 metre.
#> 31:                                                                                              Below 1 metre.
#> 32:                                      1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the afternoon.
#> 33:                                              2 to 3 metres, decreasing to 1 to 2 metres during the morning.
#>                                                                                                   forecast_seas
#>                                                                                               forecast_weather
#>  1:                                                                                                       <NA>
#>  2:                                                                                              Mostly clear.
#>  3: Partly cloudy. The chance of a thunderstorm offshore north of Merimbula in the late afternoon and evening.
#>  4:                                                                      Partly cloudy. 70% chance of showers.
#>  5:                                                                                              Mostly sunny.
#>  6:                                                                                              Mostly clear.
#>  7:                                  Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#>  8:      Partly cloudy. 60% chance of showers. The chance of a thunderstorm in the late afternoon and evening.
#>  9:                                                                                                     Sunny.
#> 10:                                                                                              Mostly clear.
#> 11:                                  Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 12:                             Partly cloudy. The chance of a thunderstorm in the late afternoon and evening.
#> 13:                                                                                                     Sunny.
#> 14:                                                                                                     Clear.
#> 15:                             Partly cloudy. The chance of a thunderstorm in the late afternoon and evening.
#> 16:                             Partly cloudy. The chance of a thunderstorm in the late afternoon and evening.
#> 17:                                                                                              Mostly sunny.
#> 18:                                                                                                     Clear.
#> 19:                                             Mostly sunny day. The chance of a thunderstorm in the evening.
#> 20:                                  Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#> 21:                                                                                              Mostly sunny.
#> 22:                                                                                                     Clear.
#> 23:                              Mostly sunny day. The chance of a thunderstorm south of Crowdy Head at night.
#> 24:                          Mostly sunny day. The chance of a thunderstorm in the late afternoon and evening.
#> 25:                                                                                              Mostly sunny.
#> 26:                                                                                             Partly cloudy.
#> 27:                                                                                              Mostly sunny.
#> 28:                                  Sunny morning. The chance of a thunderstorm in the afternoon and evening.
#> 29:                                                                                             Partly cloudy.
#> 30:                                                                                             Partly cloudy.
#> 31:                                                                                              Mostly sunny.
#> 32:             Sunny day. The chance of a thunderstorm south of Cape Byron in the late afternoon and evening.
#> 33:                                  Partly cloudy. The chance of a thunderstorm in the afternoon and evening.
#>                                                                                               forecast_weather
#>                                                                                                                                                                            forecast_winds
#>  1:                                                                                                                                                                                  <NA>
#>  2:                                                                                                                                                             Northerly 25 to 30 knots.
#>  3:                                                                                   Northerly 20 to 30 knots shifting southerly 15 to 25 knots in the late morning and early afternoon.
#>  4:                                                                        Southeast to southwesterly 10 to 15 knots tending south to southwesterly 25 to 35 knots in the late afternoon.
#>  5:                                                                                                                                                         Southwesterly 20 to 30 knots.
#>  6:                                                                                                                                                North to northeasterly 25 to 30 knots.
#>  7:                                               Northerly 15 to 25 knots, reaching up to 30 knots during the morning. Winds shifting southerly in the late morning and early afternoon.
#>  8:                                                Southerly 15 to 20 knots tending north to northeasterly in the middle of the day then tending southerly 20 to 30 knots in the evening.
#>  9:                               Southwesterly 15 to 25 knots tending southeast to southwesterly 10 to 15 knots during the day then becoming variable about 10 knots during the evening.
#> 10:                                                                                   North to northeasterly 15 to 20 knots, reaching up to 30 knots south of Port Kembla in the evening.
#> 11: Northerly 20 to 25 knots, reaching up to 30 knots offshore south of Point Perpendicular during the morning. Winds shifting southerly 15 to 25 knots during the afternoon and evening.
#> 12:            Southerly 15 to 25 knots shifting north to northeasterly 15 to 20 knots in the early afternoon then tending northwest to southwesterly 20 to 30 knots in the late evening.
#> 13:                                          Southerly 20 to 30 knots tending southwesterly 15 to 20 knots during the morning then becoming variable about 10 knots during the afternoon.
#> 14:                                                                                                                                                         Northeasterly 15 to 20 knots.
#> 15:                                                                                                        North to northeasterly 15 to 25 knots, reaching up to 30 knots in the evening.
#> 16:                              North to northeasterly 15 to 25 knots increasing to 20 to 30 knots before dawn then shifting south to southwesterly 15 to 25 knots early in the morning.
#> 17:                                 Southwesterly 20 to 30 knots decreasing to 15 to 25 knots during the morning then shifting east to southeasterly below 10 knots during the afternoon.
#> 18:                                                                                                         Northeasterly 10 to 15 knots, reaching up to 20 knots inshore in the evening.
#> 19:                                               North to northeasterly 15 to 20 knots increasing to 20 to 25 knots in the late afternoon. Winds reaching up to 30 knots in the evening.
#> 20:  North to northeasterly 20 to 30 knots tending southeast to southwesterly 15 to 25 knots in the morning then tending northwest to northeasterly 15 to 20 knots in the late afternoon.
#> 21:                                                                                      Southwesterly 20 to 30 knots shifting east to southeasterly 10 to 15 knots during the afternoon.
#> 22:                                                                                                                         Easterly 10 to 15 knots turning northeasterly in the evening.
#> 23:                        Northeasterly 10 to 15 knots turning northerly 15 to 20 knots early in the morning. Winds reaching up to 30 knots south of Port Macquarie in the late evening.
#> 24:                                                          Northerly 20 to 30 knots decreasing to 15 to 25 knots early in the morning then increasing to 20 to 30 knots in the evening.
#> 25:                                                                                  North to northwesterly 20 to 30 knots shifting south to southeasterly 10 to 15 knots during the day.
#> 26:                                                                                                                                                              Easterly about 10 knots.
#> 27:                               North to northeasterly about 10 knots increasing to 15 to 20 knots in the middle of the day. Winds reaching up to 25 knots inshore in the late evening.
#> 28:                                                                                                          Northerly 15 to 25 knots increasing to 20 to 30 knots in the late afternoon.
#> 29:                                                                                  North to northwesterly 20 to 30 knots shifting south to southeasterly 15 to 20 knots during the day.
#> 30:                                                                                                                                                 East to southeasterly about 10 knots.
#> 31:                                          Variable about 10 knots becoming northeast to southeasterly 10 to 15 knots in the early afternoon then tending northeasterly in the evening.
#> 32:                                                                                                          Northerly 15 to 25 knots increasing to 20 to 30 knots in the late afternoon.
#> 33:                                                                                           North to northwesterly 20 to 30 knots shifting southeasterly 15 to 20 knots during the day.
#>                                                                                                                                                                            forecast_winds
#>                                                                                               forecast_swell1
#>  1:                                                                                                      <NA>
#>  2:                                                                      East to northeasterly below 1 metre.
#>  3:                        Easterly around 1 metre, tending northeasterly 1 to 1.5 metres during the morning.
#>  4:             Southwesterly below 1 metre inshore, increasing to 1 to 1.5 metres offshore by early evening.
#>  5:                        Southwesterly 1 to 1.5 metres, increasing to 1.5 to 2.5 metres during the morning.
#>  6:                                                                                  Southerly below 1 metre.
#>  7:       East to northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore around midday.
#>  8:       Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore north of Moruya Heads.
#>  9:                 South to southwesterly 1 to 1.5 metres, increasing to 1 to 2 metres during the afternoon.
#> 10:                                                                                   Easterly below 1 metre.
#> 11: East to northeasterly below 1 metre, increasing to 1 to 1.5 metres south of Port Kembla by early evening.
#> 12:                             Northeasterly 1 to 1.5 metres, decreasing to around 1 metre by early evening.
#> 13:                                 Southerly around 1 metre, increasing to 1 to 2 metres during the morning.
#> 14:                                                                                   Easterly below 1 metre.
#> 15:                                                                      East to northeasterly below 1 metre.
#> 16:          Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 17:                             Southerly around 1 metre, increasing to 1.5 to 2 metres during the afternoon.
#> 18:                                                                                  Easterly around 1 metre.
#> 19:                                                                      East to northeasterly below 1 metre.
#> 20:          Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 21:                                  Southerly below 1 metre, increasing to 1 to 2 metres during the morning.
#> 22:                                                                                  Easterly around 1 metre.
#> 23:                                                                                  Easterly around 1 metre.
#> 24:    East to northeasterly below 1 metre, increasing to 1 to 1.5 metres south of Crowdy Head around midday.
#> 25:              Southerly around 1 metre inshore, increasing to 1 to 1.5 metres offshore during the morning.
#> 26:                                  Easterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 27:                                                                                 Easterly 1 to 1.5 metres.
#> 28:                                                                                  Easterly around 1 metre.
#> 29:  Southerly below 1 metre inshore, increasing to 1 to 1.5 metres offshore during the afternoon or evening.
#> 30:                                                                                      Easterly 1.5 metres.
#> 31:                                Easterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 32:                                                                                 Easterly 1 to 1.5 metres.
#> 33:                                                             Southerly around 1 metre south of Cape Byron.
#>                                                                                               forecast_swell1
#>                                                                                                             forecast_swell2
#>  1:                                                                                                                    <NA>
#>  2:                                                                                                                    <NA>
#>  3:                                                                                                                    <NA>
#>  4:                                                                                           Northeasterly around 1 metre.
#>  5:                                                                                                                    <NA>
#>  6:                                                                                                                    <NA>
#>  7:                                                                                                                    <NA>
#>  8:                                                                                                                    <NA>
#>  9:                                                                                                                    <NA>
#> 10:                                                                                                                    <NA>
#> 11:                                                                                                                    <NA>
#> 12:                                                                                                                    <NA>
#> 13:                                                                                           Northeasterly around 1 metre.
#> 14:                                                                                                                    <NA>
#> 15:                                                                                                                    <NA>
#> 16:                                                                                                                    <NA>
#> 17:                                                                                           Northeasterly around 1 metre.
#> 18:                                                                                                                    <NA>
#> 19:                                                                                                                    <NA>
#> 20:                                                                                                                    <NA>
#> 21:                                         Northeasterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 22:                                                                                                                    <NA>
#> 23:                                                                                                                    <NA>
#> 24:                                                                                                                    <NA>
#> 25:                                           Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 26:                                                                                                                    <NA>
#> 27:                                                                                                                    <NA>
#> 28:                                                                                                                    <NA>
#> 29: Northeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore south of Coffs Harbour during the morning.
#> 30:                                                                                                                    <NA>
#> 31:                                                                                                                    <NA>
#> 32:                                                                                                                    <NA>
#> 33:                                                                                   East to northeasterly around 1 metre.
#>                                                                                                             forecast_swell2
#>     forecast_caution                                       marine_forecast
#>  1:             <NA>                                                  <NA>
#>  2:             <NA>      Strong Wind Warning for Wednesday for Eden Coast
#>  3:             <NA>       Strong Wind Warning for Thursday for Eden Coast
#>  4:             <NA>                                                  <NA>
#>  5:             <NA>                                                  <NA>
#>  6:             <NA>  Strong Wind Warning for Wednesday for Batemans Coast
#>  7:             <NA>   Strong Wind Warning for Thursday for Batemans Coast
#>  8:             <NA>                                                  <NA>
#>  9:             <NA>                                                  <NA>
#> 10:             <NA> Strong Wind Warning for Wednesday for Illawarra Coast
#> 11:             <NA>  Strong Wind Warning for Thursday for Illawarra Coast
#> 12:             <NA>                                                  <NA>
#> 13:             <NA>                                                  <NA>
#> 14:             <NA>                                                  <NA>
#> 15:             <NA>     Strong Wind Warning for Thursday for Sydney Coast
#> 16:             <NA>                                                  <NA>
#> 17:             <NA>                                                  <NA>
#> 18:             <NA>                                                  <NA>
#> 19:             <NA>     Strong Wind Warning for Thursday for Hunter Coast
#> 20:             <NA>                                                  <NA>
#> 21:             <NA>                                                  <NA>
#> 22:             <NA>                                                  <NA>
#> 23:             <NA>  Strong Wind Warning for Thursday for Macquarie Coast
#> 24:             <NA>                                                  <NA>
#> 25:             <NA>                                                  <NA>
#> 26:             <NA>                                                  <NA>
#> 27:             <NA>                                                  <NA>
#> 28:             <NA>                                                  <NA>
#> 29:             <NA>                                                  <NA>
#> 30:             <NA>                                                  <NA>
#> 31:             <NA>                                                  <NA>
#> 32:             <NA>                                                  <NA>
#> 33:             <NA>                                                  <NA>
#>     forecast_caution                                       marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#>     tropical_system_location forecast_waves

# }