Get current weather observations of a BOM station

Get current weather observations of a BOM station

get_current_weather(
  station_name,
  strict = FALSE,
  latlon = NULL,
  emit_latlon_msg = TRUE
)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g., latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e., which station was found to be nearest to the given coordinate).

Value

A bomrang_tbl object (extension of a data.frame) of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see “Appendix 1 - Output from get_current_weather()” in the bomrang vignette
vignette("bomrang", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations: http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers: http://www.bom.gov.au/climate/cdo/about/site-num.shtml

See also

Author

Hugh Parsonage, hugh.parsonage@gmail.com

Examples

# \donttest{
  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>   --- Australian Bureau of Meteorology (BOM) Data Resource ---
#>   (Original Request Parameters)
#>   Station:		HOLSWORTHY AERODROME AWS [066161] 
#>   Location:		lat: -33.9925, lon: 150.9489
#>   Measurement / Origin:	All / Current
#>   Timespan:		2012 -- 2021 [9 years]
#>   ---------------------------------------------------------------  
#>      sort_order   wmo  full_name history_product local_date_time
#>   1:          0 95761 Holsworthy        IDN60801      09/02:30pm
#>   2:          1 95761 Holsworthy        IDN60801      09/02:00pm
#>   3:          2 95761 Holsworthy        IDN60801      09/01:30pm
#>   4:          3 95761 Holsworthy        IDN60801      09/01:00pm
#>   5:          4 95761 Holsworthy        IDN60801      09/12:30pm
#>  ---                                                            
#> 162:        161 95761 Holsworthy        IDN60801      06/05:00pm
#> 163:        162 95761 Holsworthy        IDN60801      06/04:30pm
#> 164:        163 95761 Holsworthy        IDN60801      06/04:00pm
#> 165:        164 95761 Holsworthy        IDN60801      06/03:30pm
#> 166:        165 95761 Holsworthy        IDN60801      06/03:00pm
#>      local_date_time_full        aifstime_utc      lat      lon apparent_t
#>   1:  2021-04-09 14:30:00 2021-04-09 04:30:00 -33.9925 150.9489       26.2
#>   2:  2021-04-09 14:00:00 2021-04-09 04:00:00 -33.9925 150.9489       25.9
#>   3:  2021-04-09 13:30:00 2021-04-09 03:30:00 -33.9925 150.9489       25.1
#>   4:  2021-04-09 13:00:00 2021-04-09 03:00:00 -33.9925 150.9489       24.3
#>   5:  2021-04-09 12:30:00 2021-04-09 02:30:00 -33.9925 150.9489       23.3
#>  ---                                                                      
#> 162:  2021-04-06 17:00:00 2021-04-06 07:00:00 -33.9925 150.9489       22.9
#> 163:  2021-04-06 16:30:00 2021-04-06 06:30:00 -33.9925 150.9489       23.1
#> 164:  2021-04-06 16:00:00 2021-04-06 06:00:00 -33.9925 150.9489       23.0
#> 165:  2021-04-06 15:30:00 2021-04-06 05:30:00 -33.9925 150.9489       22.3
#> 166:  2021-04-06 15:00:00 2021-04-06 05:00:00 -33.9925 150.9489       23.9
#>              cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#>   1:         Clear           NA           0          -            NA     6.7
#>   2:             -           NA          NA          -            NA     6.9
#>   3:             -           NA          NA          -            NA     6.2
#>   4:             -           NA          NA          -            NA     3.2
#>   5: Partly cloudy         3750           3          -            NA     2.4
#>  ---                                                                        
#> 162: Partly cloudy          540           3          -            NA     2.5
#> 163: Partly cloudy         2520           3          -            NA     3.1
#> 164: Partly cloudy         2580           5          -            NA     4.3
#> 165: Partly cloudy          630           3          -            NA     3.5
#> 166:        Cloudy          600           8          -            NA     4.1
#>      gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#>   1:       17       9     26.5  15.5 1001.8    1001.8    1002.0          -
#>   2:       11       6     26.1  14.6 1002.0    1002.0    1002.2          -
#>   3:       13       7     25.2  14.9 1002.6    1002.6    1002.8          -
#>   4:        7       4     22.8  17.7 1003.1    1003.1    1003.3          -
#>   5:       11       6     21.8  18.0 1003.5    1003.5    1003.7          -
#>  ---                                                                      
#> 162:       22      12     22.9  18.9 1020.0    1020.0    1020.2          -
#> 163:       26      14     23.3  18.4 1019.9    1019.9    1020.1          -
#> 164:       24      13     23.9  16.9 1019.7    1019.7    1019.9          -
#> 165:       33      18     23.4  17.8 1019.8    1019.8    1020.0          -
#> 166:       26      14     24.4  17.9 1019.6    1019.6    1019.8          -
#>      rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#>   1:          0      50         -                -           NA           NA
#>   2:          0      49         -                -           NA           NA
#>   3:          0      52         -                -           NA           NA
#>   4:          0      73         -                -           NA           NA
#>   5:          0      79         -                -           NA           NA
#>  ---                                                                        
#> 162:          0      78         -                -           NA           NA
#> 163:          0      74         -                -           NA           NA
#> 164:          0      65         -                -           NA           NA
#> 165:          0      71         -                -           NA           NA
#> 166:          0      67         -                -           NA           NA
#>      vis_km weather wind_dir wind_spd_kmh wind_spd_kt
#>   1:     10       -      NNW           11           6
#>   2:     10       -      WNW            9           5
#>   3:     10       -        W            9           5
#>   4:     10       -      WSW            6           3
#>   5:     10       -      SSW            7           4
#>  ---                                                 
#> 162:     10       -       SE           17           9
#> 163:     10       -      ESE           17           9
#> 164:     10       -      ESE           17           9
#> 165:     10       -      ESE           20          11
#> 166:     10       -      ESE           17           9

# }