Get current weather observations of a BOM station

Get current weather observations of a BOM station

get_current_weather(
  station_name,
  strict = FALSE,
  latlon = NULL,
  emit_latlon_msg = TRUE
)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g., latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e., which station was found to be nearest to the given coordinate).

Value

A bomrang_tbl object (extension of a data.frame) of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see “Appendix 1 - Output from get_current_weather()” in the bomrang vignette
vignette("bomrang", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations: http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers: http://www.bom.gov.au/climate/cdo/about/site-num.shtml

See also

Author

Hugh Parsonage, hugh.parsonage@gmail.com

Examples

# \donttest{
  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>   --- Australian Bureau of Meteorology (BOM) Data Resource ---
#>   (Original Request Parameters)
#>   Station:		HOLSWORTHY AERODROME AWS [066161] 
#>   Location:		lat: -33.9925, lon: 150.9489
#>   Measurement / Origin:	All / Current
#>   Timespan:		2012 -- 2021 [9 years]
#>   ---------------------------------------------------------------  
#>      sort_order   wmo  full_name history_product local_date_time
#>   1:          0 95761 Holsworthy        IDN60801      07/02:30pm
#>   2:          1 95761 Holsworthy        IDN60801      07/02:00pm
#>   3:          2 95761 Holsworthy        IDN60801      07/01:30pm
#>   4:          3 95761 Holsworthy        IDN60801      07/01:27pm
#>   5:          4 95761 Holsworthy        IDN60801      07/01:17pm
#>  ---                                                            
#> 147:        146 95761 Holsworthy        IDN60801      04/05:00pm
#> 148:        147 95761 Holsworthy        IDN60801      04/04:30pm
#> 149:        148 95761 Holsworthy        IDN60801      04/04:00pm
#> 150:        149 95761 Holsworthy        IDN60801      04/03:30pm
#> 151:        150 95761 Holsworthy        IDN60801      04/03:00pm
#>      local_date_time_full        aifstime_utc      lat      lon apparent_t
#>   1:  2021-04-07 14:30:00 2021-04-07 04:30:00 -33.9925 150.9489       22.9
#>   2:  2021-04-07 14:00:00 2021-04-07 04:00:00 -33.9925 150.9489       23.5
#>   3:  2021-04-07 13:30:00 2021-04-07 03:30:00 -33.9925 150.9489       22.3
#>   4:  2021-04-07 13:27:00 2021-04-07 03:27:00 -33.9925 150.9489       23.2
#>   5:  2021-04-07 13:17:00 2021-04-07 03:17:00 -33.9925 150.9489       22.9
#>  ---                                                                      
#> 147:  2021-04-04 17:00:00 2021-04-04 07:00:00 -33.9925 150.9489       27.0
#> 148:  2021-04-04 16:30:00 2021-04-04 06:30:00 -33.9925 150.9489       27.6
#> 149:  2021-04-04 16:00:00 2021-04-04 06:00:00 -33.9925 150.9489       28.1
#> 150:  2021-04-04 15:30:00 2021-04-04 05:30:00 -33.9925 150.9489       28.9
#> 151:  2021-04-04 15:00:00 2021-04-04 05:00:00 -33.9925 150.9489       28.8
#>              cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#>   1: Partly cloudy          540           3          -            NA     3.1
#>   2: Partly cloudy         1080           3          -            NA     2.4
#>   3: Partly cloudy          630           3          -            NA     2.1
#>   4: Partly cloudy          630           3          -            NA     2.1
#>   5: Partly cloudy         1290           5          -            NA     2.0
#>  ---                                                                        
#> 147:             -           NA          NA          -            NA     5.8
#> 148:             -           NA          NA          -            NA     6.8
#> 149:             -           NA          NA          -            NA     7.0
#> 150:             -           NA          NA          -            NA     9.6
#> 151:             -           NA          NA          -            NA     9.6
#>      gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#>   1:       20      11     22.9  18.0 1015.9    1015.9    1016.1          -
#>   2:       20      11     23.0  19.2 1016.1    1016.1    1016.3          -
#>   3:       24      13     21.8  18.4 1016.6    1016.6    1016.8          -
#>   4:       13       7     21.9  18.5 1016.6    1016.6    1016.8          -
#>   5:       17       9     21.9  18.7 1016.7    1016.7    1016.9          -
#>  ---                                                                      
#> 147:       19      10     26.8  17.6 1016.0    1016.0    1016.2          -
#> 148:       17       9     27.7  16.9 1015.8    1015.8    1016.0          -
#> 149:       20      11     28.4  17.3 1015.6    1015.6    1015.8          -
#> 150:       13       7     29.3  13.1 1015.6    1015.6    1015.8          -
#> 151:       13       7     29.5  13.3 1015.6    1015.6    1015.8          -
#>      rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#>   1:        0.2      74         -                -           NA           NA
#>   2:        0.2      79         -                -           NA           NA
#>   3:        0.2      81         -                -           NA           NA
#>   4:        0.2      81         -                -           NA           NA
#>   5:        0.2      82         -                -           NA           NA
#>  ---                                                                        
#> 147:        0.0      57         -                -           NA           NA
#> 148:        0.0      51         -                -           NA           NA
#> 149:        0.0      51         -                -           NA           NA
#> 150:        0.0      37         -                -           NA           NA
#> 151:        0.0      37         -                -           NA           NA
#>      vis_km weather wind_dir wind_spd_kmh wind_spd_kt
#>   1:     10       -       SE           15           8
#>   2:     10       -      SSE           15           8
#>   3:     10       -      SSE           13           7
#>   4:     10       -      SSE            9           5
#>   5:     10       -      SSE           11           6
#>  ---                                                 
#> 147:     10       -      ENE           13           7
#> 148:     10       -      ENE           13           7
#> 149:     10       -        E           15           8
#> 150:     10       -       NE            7           4
#> 151:     10       -        N            9           5

# }