Get Current Weather Observations of a BOM Station

Get Current Weather Observations of a BOM Station

get_current_weather(station_name, strict = FALSE, latlon = NULL,
  raw = FALSE, emit_latlon_msg = TRUE, as.data.table = FALSE)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g. latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

raw
Logical. Do not convert the columns data.table to the appropriate classes. (FALSE by default.)

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e. which station was found to be nearest to the given coordinate).

as.data.table
Logical. If TRUE, return result as a data.table.

Value

Tidy data frame of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see the vignette "Current Weather Fields":
vignette("Current Weather Fields", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations:
http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers:
http://www.bom.gov.au/climate/cdo/about/site-num.shtml

Examples

  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>     sort_order   wmo  full_name history_product local_date_time
#> 1            0 95761 Holsworthy        IDN60801      04/09:34pm
#> 2            1 95761 Holsworthy        IDN60801      04/09:32pm
#> 3            2 95761 Holsworthy        IDN60801      04/09:30pm
#> 4            3 95761 Holsworthy        IDN60801      04/09:00pm
#> 5            4 95761 Holsworthy        IDN60801      04/08:43pm
#> 6            5 95761 Holsworthy        IDN60801      04/08:40pm
#> 7            6 95761 Holsworthy        IDN60801      04/08:30pm
#> 8            7 95761 Holsworthy        IDN60801      04/08:00pm
#> 9            8 95761 Holsworthy        IDN60801      04/07:30pm
#> 10           9 95761 Holsworthy        IDN60801      04/07:00pm
#> 11          10 95761 Holsworthy        IDN60801      04/06:30pm
#> 12          11 95761 Holsworthy        IDN60801      04/06:00pm
#> 13          12 95761 Holsworthy        IDN60801      04/05:30pm
#> 14          13 95761 Holsworthy        IDN60801      04/05:00pm
#> 15          14 95761 Holsworthy        IDN60801      04/04:30pm
#> 16          15 95761 Holsworthy        IDN60801      04/04:00pm
#> 17          16 95761 Holsworthy        IDN60801      04/03:30pm
#> 18          17 95761 Holsworthy        IDN60801      04/03:07pm
#> 19          18 95761 Holsworthy        IDN60801      04/03:00pm
#> 20          19 95761 Holsworthy        IDN60801      04/02:30pm
#> 21          20 95761 Holsworthy        IDN60801      04/02:00pm
#> 22          21 95761 Holsworthy        IDN60801      04/01:30pm
#> 23          22 95761 Holsworthy        IDN60801      04/01:00pm
#> 24          23 95761 Holsworthy        IDN60801      04/12:30pm
#> 25          24 95761 Holsworthy        IDN60801      04/12:00pm
#> 26          25 95761 Holsworthy        IDN60801      04/11:30am
#> 27          26 95761 Holsworthy        IDN60801      04/11:23am
#> 28          27 95761 Holsworthy        IDN60801      04/11:05am
#>     local_date_time_full        aifstime_utc      lat      lon apparent_t
#> 1    2018-10-04 21:34:00 2018-10-04 11:34:00 -33.9925 150.9489       14.2
#> 2    2018-10-04 21:32:00 2018-10-04 11:32:00 -33.9925 150.9489       14.2
#> 3    2018-10-04 21:30:00 2018-10-04 11:30:00 -33.9925 150.9489       14.2
#> 4    2018-10-04 21:00:00 2018-10-04 11:00:00 -33.9925 150.9489       14.3
#> 5    2018-10-04 20:43:00 2018-10-04 10:43:00 -33.9925 150.9489       14.1
#> 6    2018-10-04 20:40:00 2018-10-04 10:40:00 -33.9925 150.9489       14.1
#> 7    2018-10-04 20:30:00 2018-10-04 10:30:00 -33.9925 150.9489       14.1
#> 8    2018-10-04 20:00:00 2018-10-04 10:00:00 -33.9925 150.9489       14.7
#> 9    2018-10-04 19:30:00 2018-10-04 09:30:00 -33.9925 150.9489       14.5
#> 10   2018-10-04 19:00:00 2018-10-04 09:00:00 -33.9925 150.9489       14.3
#> 11   2018-10-04 18:30:00 2018-10-04 08:30:00 -33.9925 150.9489       14.3
#> 12   2018-10-04 18:00:00 2018-10-04 08:00:00 -33.9925 150.9489       13.6
#> 13   2018-10-04 17:30:00 2018-10-04 07:30:00 -33.9925 150.9489       12.9
#> 14   2018-10-04 17:00:00 2018-10-04 07:00:00 -33.9925 150.9489       12.6
#> 15   2018-10-04 16:30:00 2018-10-04 06:30:00 -33.9925 150.9489       12.3
#> 16   2018-10-04 16:00:00 2018-10-04 06:00:00 -33.9925 150.9489       13.1
#> 17   2018-10-04 15:30:00 2018-10-04 05:30:00 -33.9925 150.9489       13.1
#> 18   2018-10-04 15:07:00 2018-10-04 05:07:00 -33.9925 150.9489       13.6
#> 19   2018-10-04 15:00:00 2018-10-04 05:00:00 -33.9925 150.9489       13.6
#> 20   2018-10-04 14:30:00 2018-10-04 04:30:00 -33.9925 150.9489       13.6
#> 21   2018-10-04 14:00:00 2018-10-04 04:00:00 -33.9925 150.9489       14.5
#> 22   2018-10-04 13:30:00 2018-10-04 03:30:00 -33.9925 150.9489       13.4
#> 23   2018-10-04 13:00:00 2018-10-04 03:00:00 -33.9925 150.9489       13.5
#> 24   2018-10-04 12:30:00 2018-10-04 02:30:00 -33.9925 150.9489       13.2
#> 25   2018-10-04 12:00:00 2018-10-04 02:00:00 -33.9925 150.9489       13.4
#> 26   2018-10-04 11:30:00 2018-10-04 01:30:00 -33.9925 150.9489       13.1
#> 27   2018-10-04 11:23:00 2018-10-04 01:23:00 -33.9925 150.9489       14.1
#> 28   2018-10-04 11:05:00 2018-10-04 01:05:00 -33.9925 150.9489       13.9
#>             cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#> 1   Partly cloudy          210           3          -            NA     0.3
#> 2   Partly cloudy          210           3          -            NA     0.3
#> 3   Partly cloudy          270           3          -            NA     0.3
#> 4          Cloudy          200           8          -            NA     0.3
#> 5   Partly cloudy          390           5          -            NA     0.3
#> 6   Partly cloudy          390           5          -            NA     0.3
#> 7          Cloudy          450           8          -            NA     0.3
#> 8   Partly cloudy          450           3          -            NA     0.3
#> 9   Partly cloudy          330           3          -            NA     0.4
#> 10         Cloudy         2790           8          -            NA     0.6
#> 11  Partly cloudy         2220           3          -            NA     0.6
#> 12         Cloudy         2000           8          -            NA     0.7
#> 13  Partly cloudy         2220           5          -            NA     0.6
#> 14         Cloudy         2220           8          -            NA     0.6
#> 15  Partly cloudy         1560           5          -            NA     0.6
#> 16  Partly cloudy         1500           5          -            NA     0.6
#> 17  Partly cloudy          330           3          -            NA     0.8
#> 18  Partly cloudy          330           3          -            NA     0.8
#> 19         Cloudy          300           8          -            NA     0.9
#> 20         Cloudy          330           8          -            NA     1.1
#> 21         Cloudy          330           8          -            NA     0.9
#> 22         Cloudy          330           8          -            NA     1.1
#> 23         Cloudy          330           8          -            NA     1.5
#> 24  Partly cloudy          330           5          -            NA     1.5
#> 25         Cloudy          300           8          -            NA     1.4
#> 26  Partly cloudy          270           5          -            NA     1.3
#> 27  Partly cloudy          270           5          -            NA     1.1
#> 28  Partly cloudy          270           3          -            NA     1.1
#>     gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#> 1         17       9     14.9  14.4 1017.7    1017.7    1017.9          -
#> 2         17       9     14.9  14.4 1017.7    1017.7    1017.9          -
#> 3         17       9     14.9  14.3 1017.7    1017.7    1017.9          -
#> 4         17       9     14.7  14.2 1017.7    1017.7    1017.9          -
#> 5         13       7     14.5  14.0 1017.6    1017.6    1017.8          -
#> 6         13       7     14.5  14.0 1017.7    1017.7    1017.9          -
#> 7         11       6     14.5  14.0 1017.7    1017.7    1017.9          -
#> 8          9       5     14.6  14.0 1017.4    1017.4    1017.6          -
#> 9          9       5     14.6  13.8 1017.3    1017.3    1017.5          -
#> 10        13       7     14.8  13.8 1017.0    1017.0    1017.2          -
#> 11        17       9     14.8  13.7 1016.7    1016.7    1016.9          -
#> 12        20      11     14.9  13.6 1016.5    1016.5    1016.7          -
#> 13        24      13     14.9  13.8 1016.1    1016.1    1016.3          -
#> 14        24      13     14.7  13.6 1015.9    1015.9    1016.1          -
#> 15        26      14     14.5  13.4 1015.9    1015.9    1016.1          -
#> 16        20      11     14.5  13.4 1016.0    1016.0    1016.2          -
#> 17        20      11     14.6  13.1 1016.0    1016.0    1016.2          -
#> 18        22      12     14.7  13.2 1016.0    1016.0    1016.2          -
#> 19        22      12     14.7  13.1 1016.1    1016.1    1016.3          -
#> 20        19      10     14.8  12.8 1016.3    1016.3    1016.5          -
#> 21        13       7     15.1  13.5 1016.4    1016.4    1016.6          -
#> 22        32      17     15.2  13.2 1016.5    1016.5    1016.7          -
#> 23        30      16     15.8  13.1 1016.7    1016.7    1016.9          -
#> 24        32      17     16.0  13.3 1016.8    1016.8    1017.0          -
#> 25        33      18     16.1  13.6 1016.9    1016.9    1017.1          -
#> 26        30      16     16.1  13.8 1017.0    1017.0    1017.2          -
#> 27        28      15     16.3  14.3 1017.1    1017.1    1017.3          -
#> 28        28      15     16.2  14.2 1017.2    1017.2    1017.4          -
#>     rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#> 1         18.6      97         -                -           NA           NA
#> 2         18.2      97         -                -           NA           NA
#> 3         17.4      96         -                -           NA           NA
#> 4         12.8      97         -                -           NA           NA
#> 5         12.0      97         -                -           NA           NA
#> 6         11.4      97         -                -           NA           NA
#> 7         11.4      97         -                -           NA           NA
#> 8         11.4      96         -                -           NA           NA
#> 9         11.4      95         -                -           NA           NA
#> 10        11.0      94         -                -           NA           NA
#> 11        10.8      93         -                -           NA           NA
#> 12        10.2      92         -                -           NA           NA
#> 13         9.6      93         -                -           NA           NA
#> 14         8.8      93         -                -           NA           NA
#> 15         7.6      93         -                -           NA           NA
#> 16         5.8      93         -                -           NA           NA
#> 17         4.4      91         -                -           NA           NA
#> 18         3.2      91         -                -           NA           NA
#> 19         3.0      90         -                -           NA           NA
#> 20         2.2      88         -                -           NA           NA
#> 21         2.0      90         -                -           NA           NA
#> 22         1.4      88         -                -           NA           NA
#> 23         1.4      84         -                -           NA           NA
#> 24         1.4      84         -                -           NA           NA
#> 25         1.4      85         -                -           NA           NA
#> 26         1.4      86         -                -           NA           NA
#> 27         1.4      88         -                -           NA           NA
#> 28         1.4      88         -                -           NA           NA
#>     vis_km        weather wind_dir wind_spd_kmh wind_spd_kt
#> 1        1              -       SW           11           6
#> 2        1              -       SW           11           6
#> 3        2              -       SW           11           6
#> 4        5           Rain      SSW            9           5
#> 5        2              -      WNW            9           5
#> 6       10              -       NW            9           5
#> 7       10              -      NNW            9           5
#> 8       10              -        W            6           3
#> 9       10              -        W            7           4
#> 10      10              -       SW            9           5
#> 11       8              -      SSW            9           5
#> 12      10           Rain        S           13           7
#> 13       9              -      SSE           17           9
#> 14       7              -        S           17           9
#> 15       5              -        S           17           9
#> 16       5              -        S           13           7
#> 17       5              -      SSW           13           7
#> 18       6              -      SSW           11           6
#> 19       7           Rain      SSW           11           6
#> 20      10              -      SSW           11           6
#> 21      10              -      SSW            9           5
#> 22      10              -        S           15           8
#> 23      10              -        S           17           9
#> 24      10              -        S           20          11
#> 25      10              -        S           20          11
#> 26      10              -        S           22          12
#> 27      10              -        S           19          10
#> 28      10              -        S           19          10
#>  [ reached getOption("max.print") -- omitted 130 rows ]