Get current weather observations of a BOM station

Get current weather observations of a BOM station

get_current_weather(
  station_name,
  strict = FALSE,
  latlon = NULL,
  emit_latlon_msg = TRUE
)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g., latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e., which station was found to be nearest to the given coordinate).

Value

A bomrang_tbl object (extension of a data.frame) of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see “Appendix 1 - Output from get_current_weather()” in the bomrang vignette
vignette("bomrang", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations: http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers: http://www.bom.gov.au/climate/cdo/about/site-num.shtml

See also

Author

Hugh Parsonage, hugh.parsonage@gmail.com

Examples

# \donttest{
  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>   --- Australian Bureau of Meteorology (BOM) Data Resource ---
#>   (Original Request Parameters)
#>   Station:		HOLSWORTHY AERODROME AWS [066161] 
#>   Location:		lat: -33.9925, lon: 150.9489
#>   Measurement / Origin:	All / Current
#>   Timespan:		2012 -- 2021 [9 years]
#>   ---------------------------------------------------------------  
#>      sort_order   wmo  full_name history_product local_date_time
#>   1:          0 95761 Holsworthy        IDN60801      02/04:00pm
#>   2:          1 95761 Holsworthy        IDN60801      02/03:30pm
#>   3:          2 95761 Holsworthy        IDN60801      02/03:00pm
#>   4:          3 95761 Holsworthy        IDN60801      02/02:30pm
#>   5:          4 95761 Holsworthy        IDN60801      02/02:00pm
#>  ---                                                            
#> 142:        141 95761 Holsworthy        IDN60801      30/06:30pm
#> 143:        142 95761 Holsworthy        IDN60801      30/06:00pm
#> 144:        143 95761 Holsworthy        IDN60801      30/05:30pm
#> 145:        144 95761 Holsworthy        IDN60801      30/05:00pm
#> 146:        145 95761 Holsworthy        IDN60801      30/04:30pm
#>      local_date_time_full        aifstime_utc      lat      lon apparent_t
#>   1:  2021-04-02 16:00:00 2021-04-02 05:00:00 -33.9925 150.9489       23.8
#>   2:  2021-04-02 15:30:00 2021-04-02 04:30:00 -33.9925 150.9489       24.5
#>   3:  2021-04-02 15:00:00 2021-04-02 04:00:00 -33.9925 150.9489       24.7
#>   4:  2021-04-02 14:30:00 2021-04-02 03:30:00 -33.9925 150.9489       25.8
#>   5:  2021-04-02 14:00:00 2021-04-02 03:00:00 -33.9925 150.9489       25.4
#>  ---                                                                      
#> 142:  2021-03-30 18:30:00 2021-03-30 07:30:00 -33.9925 150.9489       15.5
#> 143:  2021-03-30 18:00:00 2021-03-30 07:00:00 -33.9925 150.9489       15.7
#> 144:  2021-03-30 17:30:00 2021-03-30 06:30:00 -33.9925 150.9489       15.9
#> 145:  2021-03-30 17:00:00 2021-03-30 06:00:00 -33.9925 150.9489       16.2
#> 146:  2021-03-30 16:30:00 2021-03-30 05:30:00 -33.9925 150.9489       17.5
#>              cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#>   1:             -           NA          NA          -            NA     6.7
#>   2:             -           NA          NA          -            NA     6.7
#>   3:             -           NA          NA          -            NA     6.8
#>   4: Partly cloudy         1320           3          -            NA     6.5
#>   5: Partly cloudy         1260           3          -            NA     6.3
#>  ---                                                                        
#> 142:             -           NA          NA          -            NA     4.3
#> 143:             -           NA          NA          -            NA     5.0
#> 144:             -           NA          NA          -            NA     5.9
#> 145:             -           NA          NA          -            NA     6.3
#> 146:             -           NA          NA          -            NA     6.5
#>      gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#>   1:       19      10     25.1  13.6 1025.5    1025.5    1025.6          -
#>   2:       19      10     25.6  14.4 1025.4    1025.4    1025.5          -
#>   3:       19      10     25.8  14.3 1025.4    1025.4    1025.5          -
#>   4:       19      10     26.1  15.5 1025.6    1025.6    1025.7          -
#>   5:       15       8     25.5  14.9 1025.8    1025.8    1025.9          -
#>  ---                                                                      
#> 142:       24      13     18.5  10.6 1022.6    1022.6    1022.8          -
#> 143:       30      16     19.4  10.1 1022.5    1022.5    1022.7          -
#> 144:       30      16     20.0   8.7 1022.4    1022.4    1022.6          -
#> 145:       30      16     20.7   8.7 1022.2    1022.2    1022.4          -
#> 146:       28      15     21.4   9.3 1022.1    1022.1    1022.3          -
#>      rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#>   1:          0      49         -                -           NA           NA
#>   2:          0      50         -                -           NA           NA
#>   3:          0      49         -                -           NA           NA
#>   4:          0      52         -                -           NA           NA
#>   5:          0      52         -                -           NA           NA
#>  ---                                                                        
#> 142:          0      60         -                -           NA           NA
#> 143:          0      55         -                -           NA           NA
#> 144:          0      48         -                -           NA           NA
#> 145:          0      46         -                -           NA           NA
#> 146:          0      46         -                -           NA           NA
#>      vis_km weather wind_dir wind_spd_kmh wind_spd_kt
#>   1:     10       -      ENE           13           7
#>   2:     10       -        E           13           7
#>   3:     10       -      ENE           13           7
#>   4:     10       -      ENE           11           6
#>   5:     10       -      ENE            9           5
#>  ---                                                 
#> 142:     10       -      SSE           17           9
#> 143:     10       -      SSE           20          11
#> 144:     10       -      SSE           20          11
#> 145:     10       -       SE           22          12
#> 146:     10       -      SSE           20          11

# }