Get Current Weather Observations of a BOM Station

Get Current Weather Observations of a BOM Station

get_current_weather(station_name, strict = FALSE, latlon = NULL,
  emit_latlon_msg = TRUE)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g. latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e. which station was found to be nearest to the given coordinate).

Value

Tidy data frame of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see the vignette "Current Weather Fields":
vignette("Current Weather Fields", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations:
http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers:
http://www.bom.gov.au/climate/cdo/about/site-num.shtml

Examples

  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>   ---- Australian Bureau of Meteorology (BOM) Data Resource ----
#>   (Original Request Parameters)
#>   Station:		HOLSWORTHY AERODROME AWS [066161] 
#>   Location:		lat: -33.9925, lon: 150.9489
#>   Measurement / Origin:	All / Current
#>   Timespan:		2012 -- 2019 [7 years]
#>   ---------------------------------------------------------------  
#>      sort_order   wmo  full_name history_product local_date_time
#>   1:          0 95761 Holsworthy        IDN60801      21/04:00pm
#>   2:          1 95761 Holsworthy        IDN60801      21/03:30pm
#>   3:          2 95761 Holsworthy        IDN60801      21/03:00pm
#>   4:          3 95761 Holsworthy        IDN60801      21/02:30pm
#>   5:          4 95761 Holsworthy        IDN60801      21/02:00pm
#>  ---                                                            
#> 144:        143 95761 Holsworthy        IDN60801      18/06:30pm
#> 145:        144 95761 Holsworthy        IDN60801      18/06:00pm
#> 146:        145 95761 Holsworthy        IDN60801      18/05:30pm
#> 147:        146 95761 Holsworthy        IDN60801      18/05:00pm
#> 148:        147 95761 Holsworthy        IDN60801      18/04:30pm
#>      local_date_time_full        aifstime_utc      lat      lon apparent_t
#>   1:  2019-03-21 16:00:00 2019-03-21 05:00:00 -33.9925 150.9489       25.3
#>   2:  2019-03-21 15:30:00 2019-03-21 04:30:00 -33.9925 150.9489       24.7
#>   3:  2019-03-21 15:00:00 2019-03-21 04:00:00 -33.9925 150.9489       25.8
#>   4:  2019-03-21 14:30:00 2019-03-21 03:30:00 -33.9925 150.9489       25.5
#>   5:  2019-03-21 14:00:00 2019-03-21 03:00:00 -33.9925 150.9489       24.7
#>  ---                                                                      
#> 144:  2019-03-18 18:30:00 2019-03-18 07:30:00 -33.9925 150.9489       20.8
#> 145:  2019-03-18 18:00:00 2019-03-18 07:00:00 -33.9925 150.9489       21.1
#> 146:  2019-03-18 17:30:00 2019-03-18 06:30:00 -33.9925 150.9489       21.5
#> 147:  2019-03-18 17:00:00 2019-03-18 06:00:00 -33.9925 150.9489       20.5
#> 148:  2019-03-18 16:30:00 2019-03-18 05:30:00 -33.9925 150.9489       20.3
#>              cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#>   1: Partly cloudy         1860           3          -            NA     4.7
#>   2: Partly cloudy         1800           5          -            NA     5.4
#>   3: Partly cloudy          600           3          -            NA     5.1
#>   4:             -           NA          NA          -            NA     5.3
#>   5: Partly cloudy          870           3          -            NA     5.3
#>  ---                                                                        
#> 144:        Cloudy         1350           8          -            NA     1.2
#> 145:        Cloudy          600           8          -            NA     1.4
#> 146: Partly cloudy         1230           3          -            NA     1.1
#> 147: Partly cloudy          480           3          -            NA     1.0
#> 148: Partly cloudy          420           3          -            NA     0.8
#>      gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#>   1:       32      17     26.0  18.7 1014.9    1014.9    1015.2          -
#>   2:       30      16     26.2  17.6 1015.0    1015.0    1015.3          -
#>   3:       26      14     26.2  18.1 1015.1    1015.1    1015.4          -
#>   4:       28      15     26.3  17.9 1015.2    1015.2    1015.5          -
#>   5:       32      17     26.2  17.8 1015.3    1015.3    1015.6          -
#>  ---                                                                      
#> 144:       11       6     19.4  17.4 1011.3    1011.3    1011.5          -
#> 145:       11       6     19.8  17.6 1010.9    1010.9    1011.1          -
#> 146:        7       4     19.8  18.1 1010.9    1010.9    1011.1          -
#> 147:       15       8     19.5  17.8 1010.6    1010.6    1010.8          -
#> 148:       13       7     19.2  17.9 1010.7    1010.7    1010.9          -
#>      rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#>   1:        0.4      NA         -                -           NA           NA
#>   2:        0.4      NA         -                -           NA           NA
#>   3:        0.4      NA         -                -           NA           NA
#>   4:        0.4      NA         -                -           NA           NA
#>   5:        0.4      NA         -                -           NA           NA
#>  ---                                                                        
#> 144:        1.2      NA         -                -           NA           NA
#> 145:        1.2      NA         -                -           NA           NA
#> 146:        1.2      NA         -                -           NA           NA
#> 147:        1.2      NA         -                -           NA           NA
#> 148:        1.2      NA         -                -           NA           NA
#>      vis_km weather wind_dir wind_spd_kmh wind_spd_kt
#>   1:     10       -       SE           20          11
#>   2:     10       -       SE           22          12
#>   3:     10       -       SE           17           9
#>   4:     10       -       SE           19          10
#>   5:     10       -       SE           22          12
#>  ---                                                 
#> 144:     10       -      SSW            6           3
#> 145:     10       -      SSW            7           4
#> 146:     10       -      SSE            6           3
#> 147:     10       -      ESE            9           5
#> 148:     10       -      ESE            9           5