Get current weather observations of a BOM station

Get current weather observations of a BOM station

get_current_weather(
  station_name,
  strict = FALSE,
  latlon = NULL,
  emit_latlon_msg = TRUE
)

Arguments

station_name
The name of the weather station. Fuzzy string matching via agrep is done.

strict
(logical) If TRUE, station_name must match the station name exactly, except that station_name need not be upper case. Note this may be different to full_name in the response. See Details.

latlon
A length-2 numeric vector giving the decimal degree latitude and longitude (in that order), e.g., latlon = c(-34, 151) for Sydney. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e., which station was found to be nearest to the given coordinate).

Value

A bomrang_tbl object (extension of a data.frame) of requested BOM station's current and prior 72hr data. For full details of fields and units returned, see Appendix 1 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Station names are not consistently named within the Bureau, so the response may contain a different full_name to the one matched, even if strict = TRUE. For example,
get_current_weather("CASTLEMAINE PRISON")[["full_name"]][1]
is Castlemaine, not Castlemaine Prison.

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user. For more details see “Appendix 1 - Output from get_current_weather()” in the bomrang vignette
vignette("bomrang", package = "bomrang")
for a complete list of fields and units.

References

Weather data observations are retrieved from: Australian Bureau of Meteorology (BOM) Weather Data Services, Observations - individual stations: http://www.bom.gov.au/catalogue/data-feeds.shtml

Station location and other metadata are sourced from the Australian Bureau of Meteorology (BOM) webpage, Bureau of Meteorology Site Numbers: http://www.bom.gov.au/climate/cdo/about/site-num.shtml

See also

Author

Hugh Parsonage, hugh.parsonage@gmail.com

Examples

# \donttest{
  # warning
  Melbourne_weather <- get_current_weather("Melbourne")

#> Warning: Multiple stations match station_name. Using
#> 	station_name = 'MELBOURNE AIRPORT'
#> Did you mean any of the following?
#> 	station_name = 'MELBOURNE (OLYMPIC PARK)'

  # no warning
  Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

  # Get weather by latitude and longitude:
  get_current_weather(latlon = c(-34, 151))

#> Using station_name = 'HOLSWORTHY AERODROME AWS', at latitude = -33.9925, longitude = 150.9489 (4.78 km away).

#>   --- Australian Bureau of Meteorology (BOM) Data Resource ---
#>   (Original Request Parameters)
#>   Station:		HOLSWORTHY AERODROME AWS [066161] 
#>   Location:		lat: -33.9925, lon: 150.9489
#>   Measurement / Origin:	All / Current
#>   Timespan:		2012 -- 2021 [9 years]
#>   ---------------------------------------------------------------  
#>      sort_order   wmo  full_name history_product local_date_time
#>   1:          0 95761 Holsworthy        IDN60801      04/02:30pm
#>   2:          1 95761 Holsworthy        IDN60801      04/02:00pm
#>   3:          2 95761 Holsworthy        IDN60801      04/01:30pm
#>   4:          3 95761 Holsworthy        IDN60801      04/01:00pm
#>   5:          4 95761 Holsworthy        IDN60801      04/12:30pm
#>  ---                                                            
#> 140:        139 95761 Holsworthy        IDN60801      01/06:00pm
#> 141:        140 95761 Holsworthy        IDN60801      01/05:30pm
#> 142:        141 95761 Holsworthy        IDN60801      01/05:00pm
#> 143:        142 95761 Holsworthy        IDN60801      01/04:30pm
#> 144:        143 95761 Holsworthy        IDN60801      01/04:00pm
#>      local_date_time_full        aifstime_utc      lat      lon apparent_t
#>   1:  2021-04-04 14:30:00 2021-04-04 04:30:00 -33.9925 150.9489       29.2
#>   2:  2021-04-04 14:00:00 2021-04-04 04:00:00 -33.9925 150.9489       29.5
#>   3:  2021-04-04 13:30:00 2021-04-04 03:30:00 -33.9925 150.9489       28.8
#>   4:  2021-04-04 13:00:00 2021-04-04 03:00:00 -33.9925 150.9489       28.4
#>   5:  2021-04-04 12:30:00 2021-04-04 02:30:00 -33.9925 150.9489       28.0
#>  ---                                                                      
#> 140:  2021-04-01 18:00:00 2021-04-01 07:00:00 -33.9925 150.9489       20.7
#> 141:  2021-04-01 17:30:00 2021-04-01 06:30:00 -33.9925 150.9489       21.7
#> 142:  2021-04-01 17:00:00 2021-04-01 06:00:00 -33.9925 150.9489       22.2
#> 143:  2021-04-01 16:30:00 2021-04-01 05:30:00 -33.9925 150.9489       22.7
#> 144:  2021-04-01 16:00:00 2021-04-01 05:00:00 -33.9925 150.9489       23.8
#>              cloud cloud_base_m cloud_oktas cloud_type cloud_type_id delta_t
#>   1:         Clear           NA           0          -            NA     8.6
#>   2:             -           NA          NA          -            NA     7.9
#>   3:             -           NA          NA          -            NA     7.5
#>   4:             -           NA          NA          -            NA     7.1
#>   5:             -           NA          NA          -            NA     7.3
#>  ---                                                                        
#> 140:             -           NA          NA          -            NA     4.1
#> 141:             -           NA          NA          -            NA     4.4
#> 142:             -           NA          NA          -            NA     4.7
#> 143: Partly cloudy          960           3          -            NA     5.1
#> 144: Partly cloudy         1020           5          -            NA     5.1
#>      gust_kmh gust_kt air_temp dewpt  press press_msl press_qnh press_tend
#>   1:       11       6     29.0  14.8 1015.9    1015.9    1016.1          -
#>   2:       13       7     28.8  16.0 1016.2    1016.2    1016.4          -
#>   3:       15       8     28.4  16.3 1016.8    1016.8    1017.0          -
#>   4:       15       8     27.9  16.5 1017.3    1017.3    1017.5          -
#>   5:       13       7     27.4  15.4 1017.5    1017.5    1017.7          -
#>  ---                                                                      
#> 140:       20      11     21.9  15.0 1023.3    1023.3    1023.4          -
#> 141:       20      11     22.8  15.4 1023.0    1023.0    1023.1          -
#> 142:       24      13     23.3  15.4 1023.0    1023.0    1023.1          -
#> 143:       22      12     23.8  15.3 1022.9    1022.9    1023.0          -
#> 144:       20      11     23.5  15.0 1023.0    1023.0    1023.1          -
#>      rain_trace rel_hum sea_state swell_dir_worded swell_height swell_period
#>   1:          0      42         -                -           NA           NA
#>   2:          0      45         -                -           NA           NA
#>   3:          0      47         -                -           NA           NA
#>   4:          0      50         -                -           NA           NA
#>   5:          0      48         -                -           NA           NA
#>  ---                                                                        
#> 140:          0      65         -                -           NA           NA
#> 141:          0      63         -                -           NA           NA
#> 142:          0      61         -                -           NA           NA
#> 143:          0      59         -                -           NA           NA
#> 144:          0      59         -                -           NA           NA
#>      vis_km weather wind_dir wind_spd_kmh wind_spd_kt
#>   1:     10       -       NE            7           4
#>   2:     10       -      ENE            7           4
#>   3:     10       -        N            9           5
#>   4:     10       -       NW            9           5
#>   5:     10       -        N            6           3
#>  ---                                                 
#> 140:     10       -        E           15           8
#> 141:     10       -        E           15           8
#> 142:     10       -        E           15           8
#> 143:     10       -      ESE           15           8
#> 144:     10       -        S            7           4

# }