Current weather observations of a station

Current weather observations of a station

get_current_weather(station_name, latlon = NULL, raw = FALSE,
  emit_latlon_msg = TRUE)

Arguments

station_name
The name of the weather station. Fuzzy string matching via base::agrep is done.

latlon
A length-2 numeric vector. When given instead of station_name, the nearest station (in this package) is used, with a message indicating the nearest such station. (See also sweep_for_stations.) Ignored if used in combination with station_name, with a warning.

raw
Logical. Do not convert the columns data.table to the appropriate classes. (FALSE by default.)

emit_latlon_msg
Logical. If TRUE (the default), and latlon is selected, a message is emitted before the table is returned indicating which station was actually used (i.e. which station was found to be nearest to the given coordinate).

Details

Note that the column local_date_time_full is set to a POSIXct object in the local time of the user.

Examples

not_run({
    # warning
    Melbourne_weather <- get_current_weather("Melbourne")

    # no warning
    Melbourne_weather <- get_current_weather("Melbourne (Olympic Park)")

    # Get weather by latitude and longitude:
    get_current_weather(latlon = c(-34, 151))
})