Obtain Historical BOM Data

Retrieves daily observations for a given station.

get_historical(stationid = NULL, latlon = NULL, type = c("rain",
  "min", "max", "solar"))

Arguments

stationid
BOM station 'ID'. See Details.

latlon
Length-2 numeric vector of Latitude/Longitude. See Details.

type
Measurement type, either daily "rain", "min" (temp), "max" (temp), or "solar" (exposure). Partial matching is performed. If not specified returns the first matching type in the order listed.

Value

A bomrang_tbl object (extension of a data.frame) of historical observations for the chosen station/product type, with some subset of the following columns

Product_code

:	BOM internal code.

Station_number

:	BOM station ID.

Year

:	Year of observation (YYYY).

Month

:	Month of observation (1-12).

Day

:	Day of observation (1-31).

Min_temperature

:	Minimum daily recorded temperature (degrees C).

Max_temperature

:	Maximum daily recorded temperature (degrees C).

Accum_days_min

:	Accumulated number of days of minimum temperature.

Accum_days_max

:	Accumulated number of days of maximum temperature.

Rainfall

:	Daily recorded rainfall in mm.

Period

:	Period over which rainfall was measured.

Solar_exposure

:	Daily global solar exposure in MJ/m^2.

Quality

:	Y, N, or missing. Data which have not yet completed the

routine quality control process are marked accordingly.

The following attributes are set on the data, and these are used to generate the header

site

:	BOM station ID.

name

:	BOM station name.

lat

:	Latitude in decimal degrees.

lon

:	Longitude in decimal degrees.

start

:	Date observations start.

end

:	Date observations end.

years

:	Available number of years data.

percent

:	Percent complete.

AWS

:	Automated weather station?

type

:	Measurement types available for the station.

Caution

Temperature data prior to 1910 should be used with extreme caution as many stations, prior to that date, were exposed in non-standard shelters, some of which give readings which are several degrees warmer or cooler than those measured according to post-1910 standards.

Daily maximum temperatures usually occur in the afternoon and daily minimum temperatures overnight or near dawn. Occasionally, however, the lowest temperature in the 24 hours to prior to 9 AM can occur around 9 AM the previous day if the night was particularly warm.

Either stationid or latlon must be provided, but if both are, then stationid will be used as it is more reliable.

In some cases data is available back to the 1800s, so tens of thousands of daily records will be returned. Other stations will be newer and will return fewer observations.

dplyr Compatibility

The bomrang_tbl class is compatible with dplyr as long as the bomrang package is on the search path. Common functions (filter, select, arrange, mutate, rename, arrange, slice, group_by) are provided which mask the dplyr versions (but use those internally, maintaining attributes).

Examples

get_historical(stationid = "023000", type = "max") ## ~48,000+ daily records

#> Data saved as /var/folders/47/rh3mqv4j3w31bch7f_tcj3jr0000gn/T//Rtmp18mLyy/IDCJAC0010_023000_1800_Data.csv

#>   ---- Australian Bureau of Meteorology (BOM) Data Resource ----
#>   (Original Request Parameters)
#>   Station:		ADELAIDE (WEST TERRACE / NGAYIRDAPIRA) [023000] 
#>   Location:		lat: -34.9257, lon: 138.5832
#>   Measurement / Origin:	Max / Historical
#>   Timespan:		1887-01-01 -- 2019-03-01 [94 years]
#>   ---------------------------------------------------------------  
#>        product_code station_number year month day max_temperature
#>     1:   IDCJAC0010          23000 1887     1   1              NA
#>     2:   IDCJAC0010          23000 1887     1   2              NA
#>     3:   IDCJAC0010          23000 1887     1   3              NA
#>     4:   IDCJAC0010          23000 1887     1   4              NA
#>     5:   IDCJAC0010          23000 1887     1   5              NA
#>    ---                                                           
#> 48288:   IDCJAC0010          23000 2019     3  17            31.3
#> 48289:   IDCJAC0010          23000 2019     3  18            30.5
#> 48290:   IDCJAC0010          23000 2019     3  19            28.6
#> 48291:   IDCJAC0010          23000 2019     3  20            27.1
#> 48292:   IDCJAC0010          23000 2019     3  21            28.8
#>        accum_days_max quality
#>     1:             NA        
#>     2:             NA        
#>     3:             NA        
#>     4:             NA        
#>     5:             NA        
#>    ---                       
#> 48288:              1       N
#> 48289:              1       N
#> 48290:              1       N
#> 48291:              1       N
#> 48292:              1       N

get_historical(latlon = c(-35.2809, 149.1300),
               type = "min") ## 3,500+ daily records

#> Closest station: 070351 (CANBERRA AIRPORT)

#> Data saved as /var/folders/47/rh3mqv4j3w31bch7f_tcj3jr0000gn/T//Rtmp18mLyy/IDCJAC0011_070351_1800_Data.csv

#>   ---- Australian Bureau of Meteorology (BOM) Data Resource ----
#>   (Original Request Parameters)
#>   Station:		CANBERRA AIRPORT [070351] 
#>   Location:		lat: -35.3088, lon: 149.2004
#>   Measurement / Origin:	Min / Historical
#>   Timespan:		2008-09-01 -- 2019-03-01 [10.6 years]
#>   ---------------------------------------------------------------  
#>       product_code station_number year month day min_temperature accum_days_min
#>    1:   IDCJAC0011          70351 2008     1   1              NA             NA
#>    2:   IDCJAC0011          70351 2008     1   2              NA             NA
#>    3:   IDCJAC0011          70351 2008     1   3              NA             NA
#>    4:   IDCJAC0011          70351 2008     1   4              NA             NA
#>    5:   IDCJAC0011          70351 2008     1   5              NA             NA
#>   ---                                                                          
#> 4095:   IDCJAC0011          70351 2019     3  18            12.7              1
#> 4096:   IDCJAC0011          70351 2019     3  19            15.7              1
#> 4097:   IDCJAC0011          70351 2019     3  20            12.2              1
#> 4098:   IDCJAC0011          70351 2019     3  21            14.7              1
#> 4099:   IDCJAC0011          70351 2019     3  22            16.7              1
#>       quality
#>    1:        
#>    2:        
#>    3:        
#>    4:        
#>    5:        
#>   ---        
#> 4095:       N
#> 4096:       N
#> 4097:       N
#> 4098:       N
#> 4099:       N