Get BoM Daily Précis Forecast for Select Towns

Fetch the BoM daily précis forecast and return a tidy data frame of the seven day town forecast for a specified state or territory.

get_precis_forecast(state = "AUS")

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done. Defaults to "AUS" returning all state bulletins, see details for further information.

Value

Tidy data frame of a Australia BoM précis seven day forecasts for select towns. For full details of fields and units returned see Appendix 2 in the bomrang vignette, use vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

References

Forecast data come from Australian Bureau of Meteorology (BoM) Weather Data Services http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata for towns come from the BoM anonymous FTP server with spatial data ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile, ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00013.dbf

Examples

# NOT RUN {
BoM_forecast <- get_precis_forecast(state = "QLD")
# }