Get BOM radar imagery

Fetch BOM radar imagery from ftp://ftp.bom.gov.au/anon/gen/radar/ and return a SpatRaster layer object. Files available are the most recent radar snapshot which are updated approximately every 6 to 10 minutes. Suggested to check file availability first by using get_available_radar.

get_radar_imagery(product_id, path = NULL, download_only = FALSE)

Arguments

product_id
Character. BOM product ID to download and import as a magick object. Value is required.

path
Character. A character string with the name where the downloaded file is saved. If not provided, the default value NULL is used which saves the file in an R session temp directory.

download_only
Logical. Whether the radar image is loaded into the environment as a magick object or just downloaded.

Value

A magick object of the most recent radar image snapshot published by the BOM. If download_only = TRUE there will be a NULL return value with the download path printed in the console as a message.

Details

Valid BOM Radar Product IDs for radar imagery can be obtained from get_available_radar.

References

Australian Bureau of Meteorology (BOM) radar images
http://www.bom.gov.au/australia/radar/

See also

Author

Dean Marchiori, deanmarchiori@gmail.com

Examples

# \donttest{
# Fetch most recent radar image for Wollongong 256km radar
imagery <- get_radar_imagery(product_id = "IDR032")

#> file downloaded to:/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2LAXaf/file11ee7e5fead1.gif

imagery

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 GIF      524    564 sRGB       TRUE     41653 72x72  

# }