Get BOM Satellite GeoTIFF Imagery

Fetch BOM satellite GeoTIFF imagery from ftp://ftp.bom.gov.au/anon/gen/gms/ and return a raster SpatRaster object of 'GeoTIFF' files. Files are available at ten minutes update frequency with a 24 hour delete time. It is suggested to check file availability first by using get_available_imagery.

get_satellite_imagery(product_id, scans = 1, cache = FALSE)

Arguments

product_id
Character. BOM product ID to download in 'GeoTIFF' format and import as a SpatRaster object. A vector of values from get_available_imagery may be used here. Value is required.

scans
Numeric. Number of scans to download, starting with most recent and progressing backwards, e.g., 1 - the most recent single scan available , 6 - the most recent hour available, 12 - the most recent 2 hours available, etc. Negating will return the oldest files first. Defaults to 1. Value is optional.

cache
Logical. Store image files locally for later use? If FALSE, the downloaded files are removed when R session is closed. To take advantage of cached files in future sessions, use cache = TRUE. Defaults to FALSE. Value is optional.

Value

A SpatRaster object of GeoTIFF images with layers named by BOM Product ID, timestamp and band.

Details

Valid BOM satellite Product IDs for use with product_id include:

IDE00420

AHI cloud cover only 2km FD GEOS GIS

IDE00421

AHI IR (Ch13) greyscale 2km FD GEOS GIS

IDE00422

AHI VIS (Ch3) greyscale 2km FD GEOS GIS

IDE00423

AHI IR (Ch13) Zehr 2km FD GEOS GIS

IDE00425

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD GEOS GIS

IDE00426

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km FD GEOS GIS

IDE00427

AHI WV (Ch8) 2km FD GEOS GIS

IDE00430

AHI cloud cover only 2km AUS equirect. GIS

IDE00431

AHI IR (Ch13) greyscale 2km AUS equirect. GIS

IDE00432

AHI VIS (Ch3) greyscale 2km AUS equirect. GIS

IDE00433

AHI IR (Ch13) Zehr 2km AUS equirect. GIS

IDE00435

AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km AUS equirect. GIS

IDE00436

AHI VIS (true colour) / IR (Ch13 greyscale) composite 2km AUS equirect. GIS

IDE00437

AHI WV (Ch8) 2km AUS equirect. GIS

IDE00439

AHI VIS (Ch3) greyscale 0.5km AUS equirect. GIS

We cache using hoardr, find your cache folder by executing manage_cache$cache_path_get.

References

Australian Bureau of Meteorology (BOM) high-definition satellite images
http://www.bom.gov.au/australia/satellite/index.shtml

See also

Author

Adam H. Sparks, adamhsparks@gmail.com

Examples

# \donttest{
# Fetch AHI VIS (true colour) / IR (Ch13 greyscale) composite 1km FD
# GEOS GIS \code{SpatRaster} object for most recent single scan available

imagery <- get_satellite_imagery(product_id = "IDE00425", scans = 1)

# Get a list of available image files and use that to specify files for
# download, downloading the two most recent files available

avail <- get_available_imagery(product_id = "IDE00425")

#> 
#> The following files are currently available for download:

#>   [1] "IDE00425.202104030410.tif" "IDE00425.202104030420.tif"
#>   [3] "IDE00425.202104030430.tif" "IDE00425.202104030440.tif"
#>   [5] "IDE00425.202104030450.tif" "IDE00425.202104030500.tif"
#>   [7] "IDE00425.202104030510.tif" "IDE00425.202104030520.tif"
#>   [9] "IDE00425.202104030530.tif" "IDE00425.202104030540.tif"
#>  [11] "IDE00425.202104030550.tif" "IDE00425.202104030600.tif"
#>  [13] "IDE00425.202104030610.tif" "IDE00425.202104030620.tif"
#>  [15] "IDE00425.202104030630.tif" "IDE00425.202104030640.tif"
#>  [17] "IDE00425.202104030650.tif" "IDE00425.202104030700.tif"
#>  [19] "IDE00425.202104030710.tif" "IDE00425.202104030720.tif"
#>  [21] "IDE00425.202104030730.tif" "IDE00425.202104030740.tif"
#>  [23] "IDE00425.202104030750.tif" "IDE00425.202104030800.tif"
#>  [25] "IDE00425.202104030810.tif" "IDE00425.202104030820.tif"
#>  [27] "IDE00425.202104030830.tif" "IDE00425.202104030840.tif"
#>  [29] "IDE00425.202104030850.tif" "IDE00425.202104030900.tif"
#>  [31] "IDE00425.202104030910.tif" "IDE00425.202104030920.tif"
#>  [33] "IDE00425.202104030930.tif" "IDE00425.202104030940.tif"
#>  [35] "IDE00425.202104030950.tif" "IDE00425.202104031000.tif"
#>  [37] "IDE00425.202104031010.tif" "IDE00425.202104031020.tif"
#>  [39] "IDE00425.202104031030.tif" "IDE00425.202104031040.tif"
#>  [41] "IDE00425.202104031050.tif" "IDE00425.202104031100.tif"
#>  [43] "IDE00425.202104031110.tif" "IDE00425.202104031120.tif"
#>  [45] "IDE00425.202104031130.tif" "IDE00425.202104031140.tif"
#>  [47] "IDE00425.202104031150.tif" "IDE00425.202104031200.tif"
#>  [49] "IDE00425.202104031210.tif" "IDE00425.202104031220.tif"
#>  [51] "IDE00425.202104031230.tif" "IDE00425.202104031240.tif"
#>  [53] "IDE00425.202104031250.tif" "IDE00425.202104031300.tif"
#>  [55] "IDE00425.202104031310.tif" "IDE00425.202104031320.tif"
#>  [57] "IDE00425.202104031330.tif" "IDE00425.202104031340.tif"
#>  [59] "IDE00425.202104031350.tif" "IDE00425.202104031400.tif"
#>  [61] "IDE00425.202104031410.tif" "IDE00425.202104031420.tif"
#>  [63] "IDE00425.202104031430.tif" "IDE00425.202104031450.tif"
#>  [65] "IDE00425.202104031500.tif" "IDE00425.202104031510.tif"
#>  [67] "IDE00425.202104031520.tif" "IDE00425.202104031530.tif"
#>  [69] "IDE00425.202104031540.tif" "IDE00425.202104031550.tif"
#>  [71] "IDE00425.202104031600.tif" "IDE00425.202104031610.tif"
#>  [73] "IDE00425.202104031620.tif" "IDE00425.202104031630.tif"
#>  [75] "IDE00425.202104031640.tif" "IDE00425.202104031650.tif"
#>  [77] "IDE00425.202104031700.tif" "IDE00425.202104031710.tif"
#>  [79] "IDE00425.202104031720.tif" "IDE00425.202104031730.tif"
#>  [81] "IDE00425.202104031740.tif" "IDE00425.202104031750.tif"
#>  [83] "IDE00425.202104031800.tif" "IDE00425.202104031810.tif"
#>  [85] "IDE00425.202104031820.tif" "IDE00425.202104031830.tif"
#>  [87] "IDE00425.202104031840.tif" "IDE00425.202104031850.tif"
#>  [89] "IDE00425.202104031900.tif" "IDE00425.202104031910.tif"
#>  [91] "IDE00425.202104031920.tif" "IDE00425.202104031930.tif"
#>  [93] "IDE00425.202104031940.tif" "IDE00425.202104031950.tif"
#>  [95] "IDE00425.202104032000.tif" "IDE00425.202104032010.tif"
#>  [97] "IDE00425.202104032020.tif" "IDE00425.202104032030.tif"
#>  [99] "IDE00425.202104032040.tif" "IDE00425.202104032050.tif"
#> [101] "IDE00425.202104032100.tif" "IDE00425.202104032110.tif"
#> [103] "IDE00425.202104032120.tif" "IDE00425.202104032130.tif"
#> [105] "IDE00425.202104032140.tif" "IDE00425.202104032150.tif"
#> [107] "IDE00425.202104032200.tif" "IDE00425.202104032210.tif"
#> [109] "IDE00425.202104032220.tif" "IDE00425.202104032230.tif"
#> [111] "IDE00425.202104032240.tif" "IDE00425.202104032250.tif"
#> [113] "IDE00425.202104032300.tif" "IDE00425.202104032310.tif"
#> [115] "IDE00425.202104032320.tif" "IDE00425.202104032330.tif"
#> [117] "IDE00425.202104032340.tif" "IDE00425.202104032350.tif"
#> [119] "IDE00425.202104040000.tif" "IDE00425.202104040010.tif"
#> [121] "IDE00425.202104040020.tif" "IDE00425.202104040030.tif"
#> [123] "IDE00425.202104040040.tif" "IDE00425.202104040050.tif"
#> [125] "IDE00425.202104040100.tif" "IDE00425.202104040110.tif"
#> [127] "IDE00425.202104040120.tif" "IDE00425.202104040130.tif"
#> [129] "IDE00425.202104040140.tif" "IDE00425.202104040150.tif"
#> [131] "IDE00425.202104040200.tif" "IDE00425.202104040210.tif"
#> [133] "IDE00425.202104040220.tif" "IDE00425.202104040230.tif"
#> [135] "IDE00425.202104040250.tif" "IDE00425.202104040300.tif"
#> [137] "IDE00425.202104040310.tif" "IDE00425.202104040320.tif"
#> [139] "IDE00425.202104040330.tif" "IDE00425.202104040340.tif"
#> [141] "IDE00425.202104040350.tif" "IDE00425.202104040400.tif"
#> [143] "IDE00425.202104040410.tif" "IDE00425.202104040420.tif"
#> [145] "IDE00425.202104040430.tif"

imagery <- get_satellite_imagery(product_id = avail, scans = 2)
# }