Get BOM 0900 or 1500 Weather Bulletin

Fetch the daily BOM 0900 or 1500 weather bulletins and return a tidy data frame for a specified state or territory.

get_weather_bulletin(state = "qld", morning = TRUE)

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

morning
If TRUE, return the 9am bulletin for the nominated state; otherwise return the 3pm bulletin.

Value

Tidy data frame of Australian 9am or 3pm weather observations for a state. For full details of fields and units returned see Appendix 4 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes:

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

It is not possible to return weather bulletins for the entire country in a single call. Rainfall figures for the 9am bulletin are generally for the preceding 24 hours, while those for the 3pm bulletin are for the preceding 6 hours since 9am. Note that values are manually entered into the bulletins and sometimes contain typographical errors which may lead to warnings about "NAs introduced by coercion".

References

Daily observation data come from Australian Bureau of Meteorology (BOM) website. The 3pm bulletin for Queensland is, for example,
http://www.bom.gov.au/qld/observations/3pm_bulletin.shtml

Examples

qld_weather <- get_weather_bulletin (state = "QLD", morning = FALSE)
qld_weather

#>              stations cld8ths wind_dir wind_speed_kmh temp_c_dry temp_c_dew
#> 1          Coconut Is      NA      WNW             13         30         NA
#> 2             Coen Ap      NA      SSW             13         34         21
#> 3             Horn Is      NA      WNW             19         29         23
#> 4      Lockhart River       7        E             13         33         27
#> 5         Palmerville      NA      SSW             11         34         24
#> 6            Scherger       3        W             15         33         24
#> 7            Weipa Ap      NA        W             22         31         25
#> 8        Burketown AP       2      ENE             13         37         24
#> 9        Century Mine      NA        S              9         37         13
#> 10       Kowanyama AP       1        W             20         33         25
#> 11      Mornington Is      NA      NNW             19         33         27
#> 12       Normanton AP      NA        N             13         34         25
#> 13      Charters Twrs       0       SW              2         37         13
#> 14      Georgetown Ap      NA      WNW             13         34         23
#> 15       Hughenden Ap       3       SW             19         36         10
#> 16        Richmond Ap       3        S             19         35         11
#> 17      Blackwater Ap      NA      ENE             15         39          8
#> 18           Clermont      NA       SE             11         39         10
#> 19         Emerald Ap      NA       SW              9         39          7
#> 20         Lochington      NA      WSW              9         39          2
#> 21           Moranbah      NA      SSE             11         39         14
#> 22       Rolleston Ap      NA      ENE              9         39          0
#> 23         Springsure      NA                      NA         NA         NA
#> 24         Barcaldine       0     CALM             NA         37          0
#> 25        Blackall Ap      NA       SW             15         36         -5
#> 26           Isisford      NA                      NA         NA         NA
#> 27          Longreach      NA      SSW             19         36          7
#> 28              Tambo      NA                      NA         NA         NA
#> 29             Winton      NA       SW             17         35          7
#> 30          Camooweal      NA        S             19         35          9
#> 31       Carters Bore      NA        S             13         34          6
#> 32       Cloncurry AP       1       SW             13         35          7
#> 33     Julia Creek Ap       1      SSE             17         35         12
#> 34    Lake Julius AWS      NA      SSW              9         36          9
#> 35             Mt Isa      NA        S             15         35          8
#> 36      New May Downs      NA        S             13         35          6
#> 37    The Monument AP      NA      SSW             19         33          5
#> 38         Trepell AP      NA       SW             20         34          6
#> 39         Urandangie      NA      SSE             15         35          1
#> 40           Bedourie      NA                      NA         NA         NA
#> 41      Birdsville Ap      NA      SSE             22         31         -2
#> 42             Boulia       0      SSW             15         32          1
#> 43    Thargomindah Ap      NA      SSE             17         30         -4
#> 44   Windorah Airport      NA       SE             19         32         -3
#> 45             Bollon      NA                      NA         NA         NA
#> 46        Charleville      NA        S             22         33         -3
#> 47         Cunnamulla      NA                      NA         NA         NA
#> 48             Injune      NA                      NA         NA         NA
#> 49           Mitchell      NA                      NA         NA         NA
#> 50               Roma      NA       SE             11         34         -1
#> 51       St George Ap      NA      WNW              9         33        -10
#> 52              Surat      NA                      NA         NA         NA
#> 53              Dalby      NA      ENE             19         35         10
#> 54              Oakey      NA      ESE             35         32         15
#> 55         Stanthorpe      NA                      NA         NA         NA
#> 56        Applethorpe      NA        E             13         21         14
#> 57              Texas      NA                      NA         NA         NA
#> 58       Toowoomba Ap      NA        E             39         29         16
#> 59        Warwick AWS      NA       SE             28         26         15
#> 60          Cairns Ap      NA        N              9         37         21
#> 61  Cairns Racecourse      NA       NE              9         37         19
#> 62      Cape Flattery      NA        N             20         32         26
#> 63           Cardwell      NA                      NA         NA         NA
#> 64        Cooktown Ap       1       SE             11         37         17
#> 65     Arlington Reef      NA        N              6         NA         NA
#> 66          Innisfail      NA                      NA         NA         NA
#> 67          Low Isles      NA        N             15         34         27
#> 68            Lucinda      NA      ESE             11         35         22
#> 69         Mareeba Ap      NA        N              9         33         19
#> 70    South Johnstone      NA      ESE             11         38         17
#> 71          Willis Is       4        W             15         31         26
#> 72         Alva Beach      NA        E             24         31         24
#> 73            Ayr DPI      NA      ENE             19         33         21
#> 74       Bowen Ap AWS      NA        E             17         33         23
#> 75             Ingham      NA                      NA         NA         NA
#> 76         Townsville       8        S             15         39         13
#> 77           Woolshed      NA      SSW              7         35         16
#> 78        Hamilton Is      NA      ESE             22         30         25
#> 79             Mackay      NA      ESE             11         30         24
#> 80          Mackay Ap       1        E             19         30         24
#> 81  Mackay Racecourse      NA        E             15         31         24
#> 82    Middle Percy Is      NA      ESE             17         29         24
#> 83      Proserpine Ap      NA      ESE             20         31         25
#> 84          Gladstone      NA      ENE             24         32         23
#> 85       Gladstone Ap      NA      ENE             26         31         22
#> 86           Heron Is      NA        E             19         NA         NA
#> 87        Rockhampton      NA        E             19         35         22
#> 88          Rundle Is      NA        E             24         29         NA
#> 89        Samuel Hill       1        E             19         31         24
#> 90        St Lawrence      NA      NNE             17         33         23
#> 91           Thangool       8       SE              9         39          6
#> 92            Yeppoon      NA       NE             19         30         25
#> 93          Bundaberg       8      ESE             26         31         23
#> 94       Double Is Pt      NA        S             46         27         25
#> 95         Gayndah Ap      NA        E             15         37         16
#> 96             Gympie      NA       SE             17         31         22
#> 97         Hervey Bay       2       SE             22         31         22
#> 98     Jimna Forestry      NA                      NA         NA         NA
#> 99           Kingaroy       8        E             19         33         16
#> 100    Lady Elliot Is      NA      ESE             22         30         24
#> 101       Maryborough      NA      ESE             30         32         23
#> 102    Monto Township      NA                      NA         NA         NA
#> 103     Rainbow Beach      NA                      NA         NA         NA
#> 104        Sandy Cape      NA                      NA         NA         NA
#> 105 Seventeen Seventy      NA                      NA         NA         NA
#> 106       Tin Can Bay      NA       SE             19         29         23
#> 107       Toolara AWS      NA                      NA         NA         NA
#> 108          Amberley       8       SE             28         31         19
#> 109       Archerfield       8      SSE             33         29         19
#> 110    Beaudesert AWS      NA      SSW             22         27         18
#> 111        Beerburrum      NA        S             13         27         20
#> 112       Brisbane Ap       8       SE             26         28         20
#> 113          Brisbane      NA       SE             17         28         20
#> 114      Canungra Def      NA        S              9         26         19
#> 115      Cape Moreton      NA      SSE             65         25         25
#> 116       Coolangatta       8        S             32         25         20
#> 117    G Coast Seaway      NA        S             50         26         23
#> 118         Gatton UQ      NA       SE             32         32         18
#> 119     Greenbank Def      NA      SSE             13         29         20
#> 120        Logan City       8        S             13         28         20
#> 121            Maleny      NA                      NA         NA         NA
#> 122           Nambour      NA        S              7         28         22
#> 123     Point Lookout      NA                      NA         NA         NA
#> 124         Redcliffe      NA      SSE             54         27         20
#> 125           Redland      NA      SSE             24         27         20
#> 126 Sunshine Coast Ap       8      SSE             35         29         23
#> 127          Tewantin      NA      SSE             22         29         23
#> 128            Sydney      NA                      NA         25         11
#> 129       Canberra Ap      NA        N             13         26          4
#> 130         Melbourne      NA      SSW             11         22          9
#> 131            Hobart      NA      NNW             26         19         10
#> 132          Adelaide      NA      WSW             13         23          9
#> 133          Perth Ap      NA        W             24         30         11
#> 134            Darwin       2       NW             24         33         25
#> 135     Alice Springs      NA        E             19         30         -1
#>     temp_c_max temp_c_min temp_c_gr barhpa rain_mm        weather seastate
#> 1           31         NA        NA   1008      NA                        
#> 2           35         NA        NA   1007     0.2                        
#> 3           30         NA        NA   1009      NA                        
#> 4           35         NA        NA   1007      NA                        
#> 5           35         NA        NA   1006      NA                        
#> 6           34         NA        NA   1008      NA                        
#> 7           32         NA        NA   1008      NA                        
#> 8           38         NA        NA   1006      NA                        
#> 9           38         NA        NA   1008      NA                        
#> 10          34         NA        NA   1008      NA                        
#> 11          33         NA        NA   1008      NA                        
#> 12          35         NA        NA   1008      NA                        
#> 13          38         NA        NA   1006      NA           Dust         
#> 14          34         NA        NA   1007      NA                        
#> 15          36         NA        NA   1007      NA                        
#> 16          36         NA        NA   1008      NA                        
#> 17          40         NA        NA   1006      NA                        
#> 18          41         NA        NA   1005      NA                        
#> 19          40         NA        NA   1006      NA                        
#> 20          40         NA        NA   1006      NA                        
#> 21          40         NA        NA   1006      NA                        
#> 22          39         NA        NA   1007      NA                        
#> 23          NA         NA        NA     NA      NA                        
#> 24          38         NA        NA     NA      NA                        
#> 25          37         NA        NA   1008      NA                        
#> 26          NA         NA        NA     NA      NA                        
#> 27          36         NA        NA   1008      NA                        
#> 28          NA         NA        NA     NA      NA                        
#> 29          35         NA        NA   1008      NA                        
#> 30          36         NA        NA   1009      NA                        
#> 31          34         NA        NA     NA      NA                        
#> 32          36         NA        NA   1008      NA                        
#> 33          36         NA        NA   1008      NA                        
#> 34          36         NA        NA     NA      NA                        
#> 35          35         NA        NA   1009      NA                        
#> 36          35         NA        NA     NA      NA                        
#> 37          33         NA        NA   1010      NA                        
#> 38          34         NA        NA   1009      NA                        
#> 39          35         NA        NA   1010      NA                        
#> 40          NA         NA        NA     NA      NA                        
#> 41          32         NA        NA   1013      NA                        
#> 42          33         NA        NA   1011      NA                        
#> 43          31         NA        NA   1013      NA                        
#> 44          33         NA        NA   1011      NA                        
#> 45          NA         NA        NA     NA      NA                        
#> 46          34         NA        NA   1010      NA                        
#> 47          NA         NA        NA     NA      NA                        
#> 48          NA         NA        NA     NA      NA                        
#> 49          NA         NA        NA     NA      NA                        
#> 50          35         NA        NA   1010      NA                        
#> 51          33         NA        NA   1011      NA                        
#> 52          NA         NA        NA     NA      NA                        
#> 53          36         NA        NA   1008      NA                        
#> 54          34         NA        NA   1010      NA                        
#> 55          NA         NA        NA     NA      NA                        
#> 56          23         NA        NA   1014      NA                        
#> 57          NA         NA        NA     NA      NA                        
#> 58          30         NA        NA   1011      NA                        
#> 59          28         NA        NA   1012      NA                        
#> 60          NA         NA        NA   1006      NA                        
#> 61          39         NA        NA     NA      NA                        
#> 62          34         NA        NA   1006      NA                        
#> 63          NA         NA        NA     NA      NA                        
#> 64          38         NA        NA   1005      NA                        
#> 65          NA         NA        NA   1006      NA                        
#> 66          NA         NA        NA     NA      NA                        
#> 67          37         NA        NA   1006      NA                        
#> 68          37         NA        NA   1005      NA                        
#> 69          35         NA        NA   1006      NA                        
#> 70          38         NA        NA   1006      NA                        
#> 71          31         NA        NA   1006      NA                SL LS NW
#> 72          36         NA        NA   1006      NA                        
#> 73          38         NA        NA     NA      NA                        
#> 74          33         NA        NA   1006      NA                        
#> 75          NA         NA        NA     NA      NA                        
#> 76          39         NA        NA   1005      NA                        
#> 77          NA         NA        NA   1006      NA                        
#> 78          30         NA        NA   1007      NA                        
#> 79          31         NA        NA   1007      NA                        
#> 80          31         NA        NA   1008      NA                        
#> 81          32         NA        NA     NA      NA                        
#> 82          30         NA        NA   1008      NA                        
#> 83          32         NA        NA   1007      NA                        
#> 84          35         NA        NA   1008      NA                        
#> 85          33         NA        NA   1008      NA                        
#> 86          NA         NA        NA     NA      NA                        
#> 87          36         NA        NA   1007      NA                        
#> 88          31         NA        NA   1008      NA                        
#> 89          34         NA        NA   1008      NA                        
#> 90          34         NA        NA   1007      NA                        
#> 91          40         NA        NA   1006      NA                        
#> 92          30         NA        NA   1008      NA                        
#> 93          33         NA        NA   1009      NA                        
#> 94          27         NA        NA   1010     0.4                        
#> 95          37         NA        NA   1008      NA                        
#> 96          33         NA        NA   1010      NA                        
#> 97          34         NA        NA   1010      NA                        
#> 98          NA         NA        NA     NA      NA                        
#> 99          NA         NA        NA   1010      NA                        
#> 100         32         NA        NA   1010      NA                        
#> 101         34         NA        NA   1010      NA                        
#> 102         NA         NA        NA     NA      NA                        
#> 103         NA         NA        NA     NA      NA                        
#> 104         NA         NA        NA     NA      NA                        
#> 105         NA         NA        NA     NA      NA                        
#> 106         31         NA        NA     NA      NA                        
#> 107         NA         NA        NA     NA      NA                        
#> 108         33         NA        NA   1012      NA                        
#> 109         33         NA        NA   1013      NA                        
#> 110         31         NA        NA   1014      NA                        
#> 111         31         NA        NA     NA      NA                        
#> 112         31         NA        NA   1013      NA           Haze         
#> 113         32         NA        NA   1013      NA                        
#> 114         27         NA        NA     NA      NA                        
#> 115         27         NA        NA   1011     0.6 Recent precip.         
#> 116         26         NA        NA   1014     1.0                        
#> 117         28         NA        NA   1014      NA                        
#> 118         34         NA        NA     NA      NA                        
#> 119         31         NA        NA     NA      NA                        
#> 120         29         NA        NA     NA      NA                        
#> 121         NA         NA        NA     NA      NA                        
#> 122         30         NA        NA   1012      NA                        
#> 123         NA         NA        NA     NA      NA                        
#> 124         31         NA        NA   1012      NA                        
#> 125         30         NA        NA     NA      NA                        
#> 126         31         NA        NA   1012      NA                        
#> 127         31         NA        NA   1012      NA                        
#> 128         NA         NA        NA   1018      NA                        
#> 129         NA         NA        NA   1016      NA                        
#> 130         NA         NA        NA   1020      NA                        
#> 131         NA         NA        NA   1011      NA                        
#> 132         NA         NA        NA   1021      NA                        
#> 133         NA         NA        NA   1012      NA                        
#> 134         33         NA        NA   1008      NA                        
#> 135         31         NA        NA   1012      NA