Get BOM 0900 or 1500 Weather Bulletin

Fetch the daily BOM 0900 or 1500 weather bulletins and return a tidy data frame for a specified state or territory.

get_weather_bulletin(state = "qld", morning = TRUE)

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

morning
If TRUE, return the 9am bulletin for the nominated state; otherwise return the 3pm bulletin.

Value

Tidy data frame of Australian 9am or 3pm weather observations for a state. For full details of fields and units returned see Appendix 4 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes:

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

It is not possible to return weather bulletins for the entire country in a single call. Rainfall figures for the 9am bulletin are generally for the preceding 24 hours, while those for the 3pm bulletin are for the preceding 6 hours since 9am. Note that values are manually entered into the bulletins and sometimes contain typographical errors which may lead to warnings about "NAs introduced by coercion".

References

Daily observation data come from Australian Bureau of Meteorology (BOM) website. The 3pm bulletin for Queensland is, for example,
http://www.bom.gov.au/qld/observations/3pm_bulletin.shtml

Examples

qld_weather <- get_weather_bulletin (state = "QLD", morning = FALSE)
qld_weather

#>            stations cld8ths wind_dir wind_speed_kmh temp_c_dry temp_c_dew
#>   1:     Coconut Is      NA      NNW             30         29         NA
#>   2:        Coen Ap      NA      ENE             20         25         25
#>   3:        Horn Is       8       NW             33         27         25
#>   4: Lockhart River       8      NNE             24         27         25
#>   5:    Palmerville      NA        E             13         26         23
#>  ---                                                                     
#> 131:         Hobart      NA       SE             15         21         16
#> 132:       Adelaide      NA        W             13         29         15
#> 133:       Perth Ap      NA      WSW             22         31         19
#> 134:         Darwin       7        W             24         33         24
#> 135:  Alice Springs      NA        E             17         36          2
#>      temp_c_max temp_c_min temp_c_gr barhpa rain_mm weather seastate
#>   1:         29         NA        NA   1004     0.4                 
#>   2:         27         NA        NA   1000    29.0    Rain         
#>   3:         28         NA        NA   1003     2.0                 
#>   4:         29         NA        NA   1000     9.0                 
#>   5:         27         NA        NA   1005     0.2                 
#>  ---                                                                
#> 131:         NA         NA        NA   1017      NA                 
#> 132:         NA         NA        NA   1014      NA                 
#> 133:         NA         NA        NA   1010      NA                 
#> 134:         34         NA        NA   1006      NA                 
#> 135:         38         NA        NA   1009      NA