Get BOM 0900 or 1500 Weather Bulletin

Fetch the daily BOM 0900 or 1500 weather bulletins and return a tidy data frame for a specified state or territory.

get_weather_bulletin(state = "qld", morning = TRUE)

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

morning
If TRUE, return the 9am bulletin for the nominated state; otherwise return the 3pm bulletin.

Value

Tidy data frame of Australian 9am or 3pm weather observations for a state. For full details of fields and units returned see Appendix 4 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes:

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

It is not possible to return weather bulletins for the entire country in a single call. Rainfall figures for the 9am bulletin are generally for the preceding 24 hours, while those for the 3pm bulletin are for the preceding 6 hours since 9am. Note that values are manually entered into the bulletins and sometimes contain typographical errors which may lead to warnings about "NAs introduced by coercion".

References

Daily observation data come from Australian Bureau of Meteorology (BOM) website. The 3pm bulletin for Queensland is, for example,
http://www.bom.gov.au/qld/observations/3pm_bulletin.shtml

Examples

qld_weather <- get_weather_bulletin (state = "QLD", morning = FALSE)
qld_weather

#>              stations cld8ths wind_dir wind_speed_kmh temp_c_dry temp_c_dew
#> 1          Coconut Is      NA      NNW             26         31         NA
#> 2             Coen Ap      NA       NE             26         25         25
#> 3             Horn Is       7       NW             33         29         25
#> 4      Lockhart River       8        N             20         26         25
#> 5         Palmerville      NA        E              9         23         23
#> 6            Scherger       8        N             20         25         24
#> 7            Weipa Ap       7      NNW             32         25         24
#> 8        Burketown AP      NA       NE             32         35         21
#> 9        Century Mine      NA      ESE             17         37         14
#> 10       Kowanyama AP       8      ESE             30         26         23
#> 11      Mornington Is       7        E             30         31         23
#> 12       Normanton AP       8      ENE             17         31         21
#> 13      Charters Twrs       8      NNE              6         28         21
#> 14      Georgetown Ap       6      ENE             22         29         20
#> 15       Hughenden Ap       4        E             17         35         15
#> 16        Richmond Ap      NA       SE             20         36         17
#> 17      Blackwater Ap      NA        N             15         33         17
#> 18           Clermont      NA      NNE              7         33         15
#> 19         Emerald Ap       2      ENE             17         33         17
#> 20         Lochington      NA      ENE              9         33         15
#> 21           Moranbah       1      ENE             13         33         17
#> 22       Rolleston Ap      NA      NNE             11         34         16
#> 23         Springsure      NA                      NA         NA         NA
#> 24         Barcaldine       5       SE             24         35         17
#> 25        Blackall Ap      NA       NE              6         37         12
#> 26           Isisford      NA                      NA         NA         NA
#> 27          Longreach       1      ESE             13         36         17
#> 28              Tambo      NA                      NA         NA         NA
#> 29             Winton      NA        E             11         37         17
#> 30          Camooweal      NA       SE             22         38         13
#> 31       Carters Bore      NA        E             15         37         13
#> 32       Cloncurry AP      NA        E             19         37         14
#> 33     Julia Creek Ap      NA        E             19         37         14
#> 34    Lake Julius AWS      NA      ESE             11         36         15
#> 35             Mt Isa       2       SE             17         36         14
#> 36      New May Downs      NA      ESE             15         36         15
#> 37    The Monument AP       1        E              9         37         14
#> 38         Trepell AP      NA        E             11         37         15
#> 39         Urandangie      NA      SSW              6         39         12
#> 40           Bedourie      NA                      NA         NA         NA
#> 41      Birdsville Ap      NA        S             19         39          2
#> 42             Boulia       6      SSE              9         38         13
#> 43    Thargomindah Ap       1        E             13         38          3
#> 44   Windorah Airport       3      ENE             13         37         12
#> 45             Bollon      NA                      NA         NA         NA
#> 46        Charleville      NA        E             13         36          4
#> 47         Cunnamulla      NA                      NA         NA         NA
#> 48             Injune      NA                      NA         NA         NA
#> 49           Mitchell      NA                      NA         NA         NA
#> 50               Roma      NA        S              9         35          5
#> 51       St George Ap      NA       NE              7         36          6
#> 52              Surat      NA                      NA         NA         NA
#> 53              Dalby      NA        W             11         33          7
#> 54              Oakey       1      ENE             15         31         14
#> 55         Stanthorpe      NA                      NA         27         15
#> 56        Applethorpe      NA        E              7         27         14
#> 57              Texas      NA                      NA         NA         NA
#> 58       Toowoomba Ap      NA        E              9         27         15
#> 59        Warwick AWS      NA      SSE             11         30         16
#> 60          Cairns Ap       8        S             19         23         21
#> 61  Cairns Racecourse       7        S             13         23         22
#> 62      Cape Flattery      NA        E             19         27         25
#> 63           Cardwell      NA                      NA         NA         NA
#> 64        Cooktown Ap       8      ESE             24         26         25
#> 65     Arlington Reef      NA        S             20         NA         NA
#> 66          Innisfail      NA                      NA         NA         NA
#> 67          Low Isles      NA      ESE             33         24         24
#> 68            Lucinda      NA       NE             20         25         21
#> 69         Mareeba Ap      NA       SE             11         22         21
#> 70    South Johnstone      NA      ENE             13         22         21
#> 71          Willis Is       7       NE             13         29         25
#> 72         Alva Beach      NA        E             17         25         23
#> 73            Ayr DPI      NA      ENE             11         24         21
#> 74       Bowen Ap AWS      NA      NNE             20         25         23
#> 75             Ingham      NA                      NA         NA         NA
#> 76         Townsville       8      WSW             11         24         23
#> 77           Woolshed      NA        N              6         22         21
#> 78        Hamilton Is       1      ESE             26         28         24
#> 79             Mackay      NA        E             13         28         23
#> 80          Mackay Ap      NA        E             17         28         23
#> 81  Mackay Racecourse      NA        E             13         29         23
#> 82    Middle Percy Is      NA      ESE             13         27         22
#> 83      Proserpine Ap       1      ESE             19         29         24
#> 84          Gladstone      NA       NE             13         29         23
#> 85       Gladstone Ap       4      ENE             17         29         22
#> 86           Heron Is      NA      ENE             11         NA         NA
#> 87        Rockhampton       1       NE             17         32         21
#> 88          Rundle Is      NA                      NA         NA         NA
#> 89        Samuel Hill      NA        E             17         31         22
#> 90        St Lawrence      NA      NNE             19         31         22
#> 91           Thangool      NA      WNW              9         34         15
#> 92            Yeppoon      NA       NE             19         29         24
#> 93          Bundaberg       1        N             19         31         21
#> 94       Double Is Pt      NA        N              7         31         24
#> 95         Gayndah Ap      NA       NE              7         34         16
#> 96             Gympie      NA      NNW              6         33         19
#> 97         Hervey Bay      NA        N             17         29         21
#> 98     Jimna Forestry      NA                      NA         NA         NA
#> 99           Kingaroy      NA        E              7         30         15
#> 100    Lady Elliot Is      NA       NE              9         30         22
#> 101       Maryborough       2      NNE             11         31         21
#> 102    Monto Township      NA                      NA         NA         NA
#> 103     Rainbow Beach      NA                      NA         NA         NA
#> 104        Sandy Cape      NA                      NA         NA         NA
#> 105 Seventeen Seventy      NA                      NA         NA         NA
#> 106       Tin Can Bay      NA      ENE             11         30         23
#> 107       Toolara AWS      NA                      NA         NA         NA
#> 108          Amberley      NA       NE             17         33         18
#> 109       Archerfield       3        E             22         29         20
#> 110    Beaudesert AWS      NA      NNE              6         31         19
#> 111        Beerburrum      NA      ENE              9         31         20
#> 112       Brisbane Ap       3        E             22         29         22
#> 113          Brisbane      NA      ENE             13         30         22
#> 114      Canungra Def      NA      ENE              9         28         22
#> 115      Cape Moreton      NA        E             17         28         22
#> 116       Coolangatta       8      ESE             26         27         21
#> 117    G Coast Seaway      NA      ESE             28         28         25
#> 118         Gatton UQ      NA      NNE             13         32         17
#> 119     Greenbank Def      NA        E             11         30         22
#> 120        Logan City       3      ESE             19         29         22
#> 121            Maleny      NA                      NA         NA         NA
#> 122           Nambour      NA      ESE              6         28         21
#> 123     Point Lookout      NA                      NA         NA         NA
#> 124         Redcliffe      NA        E             24         28         21
#> 125           Redland      NA      ESE             13         28         22
#> 126 Sunshine Coast Ap      NA       NE             17         27         21
#> 127          Tewantin      NA      NNE             15         28         23
#> 128            Sydney      NA                      NA         25         18
#> 129       Canberra Ap      NA      ENE             13         26         13
#> 130         Melbourne      NA        S              9         24         17
#> 131            Hobart      NA       SE             17         17         14
#> 132          Adelaide      NA       SE              9         27         11
#> 133          Perth Ap      NA      WSW             24         30         18
#> 134            Darwin       7        W             26         33         24
#> 135     Alice Springs      NA        E             24         37          3
#>     temp_c_max temp_c_min temp_c_gr barhpa rain_mm         weather seastate
#> 1           32         NA        NA   1003      NA                         
#> 2           27         NA        NA   1000     6.0  Recent precip.         
#> 3           29         NA        NA   1003     1.0                         
#> 4           28         NA        NA   1001     6.0                         
#> 5           24         NA        NA   1006    20.0            Rain         
#> 6           27         NA        NA    996    23.0            Rain         
#> 7           27         NA        NA    995      NA            Rain         
#> 8           36         NA        NA   1005      NA                         
#> 9           39         NA        NA   1005      NA                         
#> 10          27         NA        NA   1002     2.0                         
#> 11          33         NA        NA   1005      NA                         
#> 12          34         NA        NA   1006      NA                         
#> 13          32         NA        NA   1009     0.1                         
#> 14          30         NA        NA   1007      NA                         
#> 15          36         NA        NA   1006      NA                         
#> 16          37         NA        NA   1006      NA                         
#> 17          34         NA        NA   1010      NA                         
#> 18          34         NA        NA   1009      NA                         
#> 19          35         NA        NA   1009      NA                         
#> 20          35         NA        NA   1009      NA                         
#> 21          34         NA        NA   1009      NA                         
#> 22          35         NA        NA   1009      NA                         
#> 23          NA         NA        NA     NA      NA                         
#> 24          35         NA        NA     NA      NA                         
#> 25          37         NA        NA   1008      NA                         
#> 26          NA         NA        NA     NA      NA                         
#> 27          37         NA        NA   1007      NA                         
#> 28          NA         NA        NA     NA      NA                         
#> 29          37         NA        NA   1007      NA                         
#> 30          40         NA        NA   1004      NA                         
#> 31          38         NA        NA     NA      NA                         
#> 32          38         NA        NA   1006      NA                         
#> 33          38         NA        NA   1006      NA                         
#> 34          37         NA        NA     NA      NA                         
#> 35          38         NA        NA   1005      NA                         
#> 36          37         NA        NA     NA      NA                         
#> 37          38         NA        NA   1006      NA                         
#> 38          37         NA        NA   1006      NA                         
#> 39          NA         NA        NA   1005      NA                         
#> 40          NA         NA        NA     NA      NA                         
#> 41          41         NA        NA   1007      NA                         
#> 42          40         NA        NA   1005      NA                         
#> 43          39         NA        NA   1008      NA                         
#> 44          38         NA        NA   1008      NA                         
#> 45          NA         NA        NA     NA      NA                         
#> 46          37         NA        NA   1009      NA                         
#> 47          NA         NA        NA     NA      NA                         
#> 48          NA         NA        NA     NA      NA                         
#> 49          NA         NA        NA     NA      NA                         
#> 50          37         NA        NA   1009      NA                         
#> 51          36         NA        NA   1009      NA                         
#> 52          NA         NA        NA     NA      NA                         
#> 53          34         NA        NA   1010      NA                         
#> 54          33         NA        NA   1010      NA                         
#> 55          NA         NA        NA     NA     0.2                         
#> 56          29         NA        NA   1012      NA                         
#> 57          NA         NA        NA     NA      NA                         
#> 58          29         NA        NA   1012    26.0                         
#> 59          31         NA        NA   1011      NA                         
#> 60          27         NA        NA   1009    10.0            Rain         
#> 61          25         NA        NA     NA    10.0            Rain         
#> 62          29         NA        NA   1006     0.8  Recent precip.         
#> 63          NA         NA        NA     NA      NA                         
#> 64          27         NA        NA   1006    26.0  Recent precip.         
#> 65          NA         NA        NA   1009      NA                         
#> 66          NA         NA        NA     NA      NA                         
#> 67          27         NA        NA   1009    63.0            Rain         
#> 68          27         NA        NA   1010    11.0  Recent precip.         
#> 69          23         NA        NA   1009     6.0                         
#> 70          25         NA        NA   1011    34.0                         
#> 71          30         NA        NA   1008      NA Distant precip. SL LS NW
#> 72          30         NA        NA   1011     2.0            Rain         
#> 73          31         NA        NA     NA     3.0                         
#> 74          33         NA        NA   1011     1.0            Rain         
#> 75          NA         NA        NA     NA      NA                         
#> 76          30         NA        NA   1010    13.0            Rain         
#> 77          26         NA        NA   1010    21.0  Recent precip.         
#> 78          30         NA        NA   1010      NA                         
#> 79          29         NA        NA   1011      NA                         
#> 80          30         NA        NA   1011      NA                         
#> 81          30         NA        NA     NA      NA                         
#> 82          29         NA        NA   1012      NA                         
#> 83          31         NA        NA   1010      NA                         
#> 84          33         NA        NA   1011      NA                         
#> 85          31         NA        NA   1011      NA                         
#> 86          NA         NA        NA     NA      NA                         
#> 87          35         NA        NA   1010     0.2                         
#> 88          NA         NA        NA     NA      NA                         
#> 89          32         NA        NA   1011      NA                         
#> 90          32         NA        NA   1010      NA                         
#> 91          35         NA        NA   1010      NA                         
#> 92          30         NA        NA   1011      NA                         
#> 93          32         NA        NA   1011     1.0                         
#> 94          32         NA        NA   1011      NA                         
#> 95          NA         NA        NA   1010      NA                         
#> 96          NA         NA        NA   1010      NA                         
#> 97          NA         NA        NA   1012      NA                         
#> 98          NA         NA        NA     NA      NA                         
#> 99          NA         NA        NA   1011      NA                         
#> 100         31         NA        NA   1012      NA                         
#> 101         32         NA        NA   1011      NA                         
#> 102         NA         NA        NA     NA      NA                         
#> 103         NA         NA        NA     NA      NA                         
#> 104         NA         NA        NA     NA      NA                         
#> 105         NA         NA        NA     NA      NA                         
#> 106         31         NA        NA     NA     0.2                         
#> 107         NA         NA        NA     NA      NA                         
#> 108         33         NA        NA   1010      NA                         
#> 109         33         NA        NA   1012      NA                         
#> 110         32         NA        NA   1011      NA                         
#> 111         32         NA        NA     NA      NA                         
#> 112         30         NA        NA   1012      NA                         
#> 113         32         NA        NA   1012      NA                         
#> 114         31         NA        NA     NA      NA                         
#> 115         30         NA        NA   1012      NA                         
#> 116         30         NA        NA   1013      NA                         
#> 117         29         NA        NA   1012      NA                         
#> 118         34         NA        NA     NA      NA                         
#> 119         33         NA        NA     NA      NA                         
#> 120         31         NA        NA     NA      NA                         
#> 121         NA         NA        NA     NA      NA                         
#> 122         31         NA        NA   1012      NA                         
#> 123         NA         NA        NA     NA      NA                         
#> 124         30         NA        NA   1012      NA                         
#> 125         29         NA        NA     NA      NA                         
#> 126         30         NA        NA   1012      NA                         
#> 127         30         NA        NA   1012      NA                         
#> 128         NA         NA        NA   1015      NA                         
#> 129         NA         NA        NA   1014      NA                         
#> 130         NA         NA        NA   1015      NA                         
#> 131         NA         NA        NA   1020      NA                         
#> 132         NA         NA        NA   1014      NA                         
#> 133         NA         NA        NA   1009      NA                         
#> 134         35         NA        NA   1005      NA                         
#> 135         38         NA        NA   1008      NA