get_weather_bulletin

Fetch the daily BoM 0900 or 1500 weather bulletins and return a tidy data frame for a specified state or territory.

get_weather_bulletin(state = "qld", morning = TRUE)

Arguments

state
Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

morning
If TRUE, return the 9am bulletin for the nominated state; otherwise return the 3pm bulletin.

Value

Tidy data frame of Australian 9am or 3pm weather observations for a state.

Details

Allowed state and territory postal codes:

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

It is not possible to return weather bulletins for the entire country in a single call. Rainfall figures for the 9am bulletin are for the preceding 24 hours, while those for the 3pm bulletin are for the preceding 6 hours since 9am.

References

Daily observation data come from Australian Bureau of Meteorology (BoM) website. The 3pm bulletin for Queensland is, for example, http://www.bom.gov.au/qld/observations/3pm_bulletin.shtml

Examples

# NOT RUN {
qld_weather <- get_weather_bulletin (state = "QLD", morning = FALSE)
# }