Parse local BOM coastal waters forecast XML files

Parse local BOM daily coastal waters forecast XML file(s) and return a data frame for a specified state or territory or all Australia.

parse_coastal_forecast(state = "AUS", filepath)

Arguments

state
Required value of an Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

filepath
A string providing the directory location of the coastal forecast file(s) to parse. See Details for more.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

The filepath argument will only accept a directory where files are located for parsing. DO NOT supply the full path including the file name. This function will only parse the requested state or all of Australia in the same fashion as get_coastal_forecast, provided that the files are all present in the directory.

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
# parse the coastal forecast for Queensland

#download to tempfile() using basename() to keep original name
download.file(url = "ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml",
              destfile = file.path(tempdir(),
              basename("ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml")),
              mode = "wb")

coastal_forecast <- parse_coastal_forecast(state = "QLD",
                                           filepath = tempdir())
coastal_forecast

#>     index product_id    type state_code
#>  1:  <NA>   IDQ11290    <NA>        QLD
#>  2:  <NA>   IDQ11290    <NA>        QLD
#>  3:  <NA>   IDQ11290    <NA>        QLD
#>  4:  <NA>   IDQ11290    <NA>        QLD
#>  5:     0   IDQ11290    <NA>        QLD
#>  6:     1   IDQ11290    <NA>        QLD
#>  7:     2   IDQ11290    <NA>        QLD
#>  8:     0   IDQ11290    <NA>        QLD
#>  9:     1   IDQ11290    <NA>        QLD
#> 10:     2   IDQ11290    <NA>        QLD
#> 11:     0   IDQ11290 Coastal        QLD
#> 12:     1   IDQ11290 Coastal        QLD
#> 13:     2   IDQ11290 Coastal        QLD
#> 14:     0   IDQ11290    <NA>        QLD
#> 15:     1   IDQ11290    <NA>        QLD
#> 16:     2   IDQ11290    <NA>        QLD
#> 17:     0   IDQ11290    <NA>        QLD
#> 18:     1   IDQ11290    <NA>        QLD
#> 19:     2   IDQ11290    <NA>        QLD
#> 20:     0   IDQ11290    <NA>        QLD
#> 21:     1   IDQ11290    <NA>        QLD
#> 22:     2   IDQ11290    <NA>        QLD
#> 23:     0   IDQ11290    <NA>        QLD
#> 24:     1   IDQ11290    <NA>        QLD
#> 25:     2   IDQ11290    <NA>        QLD
#> 26:     0   IDQ11290    <NA>        QLD
#> 27:     1   IDQ11290    <NA>        QLD
#> 28:     2   IDQ11290    <NA>        QLD
#> 29:     0   IDQ11290    <NA>        QLD
#> 30:     1   IDQ11290    <NA>        QLD
#> 31:     2   IDQ11290    <NA>        QLD
#> 32:     0   IDQ11290   Local        QLD
#> 33:     1   IDQ11290   Local        QLD
#> 34:     2   IDQ11290   Local        QLD
#> 35:     0   IDQ11290    <NA>        QLD
#> 36:     1   IDQ11290    <NA>        QLD
#> 37:     2   IDQ11290    <NA>        QLD
#> 38:     0   IDQ11290    <NA>        QLD
#> 39:     1   IDQ11290    <NA>        QLD
#> 40:     2   IDQ11290    <NA>        QLD
#> 41:     0   IDQ11290   Local        QLD
#> 42:     1   IDQ11290   Local        QLD
#> 43:     2   IDQ11290   Local        QLD
#> 44:     0   IDQ11290    <NA>        QLD
#> 45:     1   IDQ11290    <NA>        QLD
#> 46:     2   IDQ11290    <NA>        QLD
#> 47:     0   IDQ11290    <NA>        QLD
#> 48:     1   IDQ11290    <NA>        QLD
#> 49:     2   IDQ11290    <NA>        QLD
#>     index product_id    type state_code
#>                                                          dist_name pt_1_name
#>  1:                                                     Queensland      <NA>
#>  2:                                 Queensland Gulf of Carpentaria      <NA>
#>  3:                                               North Queensland      <NA>
#>  4:                                               South Queensland      <NA>
#>  5: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  6: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  7: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  8:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#>  9:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 10:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 11:                                                  Torres Strait      <NA>
#> 12:                                                  Torres Strait      <NA>
#> 13:                                                  Torres Strait      <NA>
#> 14:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 15:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 16:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 17:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 18:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 19:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 20:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 21:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 22:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 23:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 24:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 25:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 26:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 27:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 28:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 29:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 30:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 31:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 32:                                                     Hervey Bay      <NA>
#> 33:                                                     Hervey Bay      <NA>
#> 34:                                                     Hervey Bay      <NA>
#> 35:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 36:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 37:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 38:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 39:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 40:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 41:                                                    Moreton Bay      <NA>
#> 42:                                                    Moreton Bay      <NA>
#> 43:                                                    Moreton Bay      <NA>
#> 44:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 45:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 46:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 47:                             Great Barrier Reef Offshore Waters      <NA>
#> 48:                             Great Barrier Reef Offshore Waters      <NA>
#> 49:                             Great Barrier Reef Offshore Waters      <NA>
#>                                                          dist_name pt_1_name
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>  1:      <NA> QLD_FA001 2021-04-02 09:59:04 2021-04-02 09:59:04      10:00
#>  2:      <NA> QLD_FA002 2021-04-02 05:00:00 2021-04-02 05:00:00      10:00
#>  3:      <NA> QLD_FA003 2021-04-02 05:00:00 2021-04-02 05:00:00      10:00
#>  4:      <NA> QLD_FA004 2021-04-02 05:00:00 2021-04-02 05:00:00      10:00
#>  5:      <NA> QLD_MW001 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#>  6:      <NA> QLD_MW001 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#>  7:      <NA> QLD_MW001 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#>  8:      <NA> QLD_MW002 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#>  9:      <NA> QLD_MW002 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 10:      <NA> QLD_MW002 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 11:      <NA> QLD_MW003 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 12:      <NA> QLD_MW003 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 13:      <NA> QLD_MW003 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 14:      <NA> QLD_MW004 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 15:      <NA> QLD_MW004 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 16:      <NA> QLD_MW004 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 17:      <NA> QLD_MW005 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 18:      <NA> QLD_MW005 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 19:      <NA> QLD_MW005 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 20:      <NA> QLD_MW006 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 21:      <NA> QLD_MW006 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 22:      <NA> QLD_MW006 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 23:      <NA> QLD_MW007 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 24:      <NA> QLD_MW007 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 25:      <NA> QLD_MW007 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 26:      <NA> QLD_MW008 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 27:      <NA> QLD_MW008 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 28:      <NA> QLD_MW008 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 29:      <NA> QLD_MW009 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 30:      <NA> QLD_MW009 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 31:      <NA> QLD_MW009 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 32:      <NA> QLD_MW010 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 33:      <NA> QLD_MW010 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 34:      <NA> QLD_MW010 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 35:      <NA> QLD_MW011 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 36:      <NA> QLD_MW011 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 37:      <NA> QLD_MW011 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 38:      <NA> QLD_MW012 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 39:      <NA> QLD_MW012 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 40:      <NA> QLD_MW012 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 41:      <NA> QLD_MW013 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 42:      <NA> QLD_MW013 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 43:      <NA> QLD_MW013 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 44:      <NA> QLD_MW014 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 45:      <NA> QLD_MW014 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 46:      <NA> QLD_MW014 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#> 47:      <NA> QLD_MW015 2021-04-02 10:00:00 2021-04-02 10:00:00      10:00
#> 48:      <NA> QLD_MW015 2021-04-03 00:00:00 2021-04-03 00:00:00      10:00
#> 49:      <NA> QLD_MW015 2021-04-04 00:00:00 2021-04-04 00:00:00      10:00
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>          start_time_utc        end_time_utc
#>  1: 2021-04-02 09:59:04 2021-04-02 09:59:04
#>  2: 2021-04-02 05:00:00 2021-04-02 05:00:00
#>  3: 2021-04-02 05:00:00 2021-04-02 05:00:00
#>  4: 2021-04-02 05:00:00 2021-04-02 05:00:00
#>  5: 2021-04-02 10:00:00 2021-04-02 10:00:00
#>  6: 2021-04-03 00:00:00 2021-04-03 00:00:00
#>  7: 2021-04-04 00:00:00 2021-04-04 00:00:00
#>  8: 2021-04-02 10:00:00 2021-04-02 10:00:00
#>  9: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 10: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 11: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 12: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 13: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 14: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 15: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 16: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 17: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 18: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 19: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 20: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 21: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 22: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 23: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 24: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 25: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 26: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 27: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 28: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 29: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 30: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 31: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 32: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 33: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 34: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 35: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 36: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 37: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 38: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 39: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 40: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 41: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 42: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 43: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 44: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 45: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 46: 2021-04-04 00:00:00 2021-04-04 00:00:00
#> 47: 2021-04-02 10:00:00 2021-04-02 10:00:00
#> 48: 2021-04-03 00:00:00 2021-04-03 00:00:00
#> 49: 2021-04-04 00:00:00 2021-04-04 00:00:00
#>          start_time_utc        end_time_utc
#>                                                                 forecast_seas
#>  1:                                                                      <NA>
#>  2:                                                                      <NA>
#>  3:                                                                      <NA>
#>  4:                                                                      <NA>
#>  5:        1.5 to 2.5 metres, decreasing to 1 to 1.5 metres by early evening.
#>  6:                                                          1 to 1.5 metres.
#>  7: Around 1 metre, increasing to 1 to 1.5 metres inshore during the morning.
#>  8:            1 to 2 metres, decreasing to 1 to 1.5 metres by early evening.
#>  9:        1 to 2 metres, decreasing to 1 to 1.5 metres during the afternoon.
#> 10:                                                           Around 1 metre.
#> 11:                                                          1 to 1.5 metres.
#> 12:           1 to 1.5 metres, decreasing below 1 metre during the afternoon.
#> 13:                                                            Below 1 metre.
#> 14:          1 to 1.5 metres, increasing to 2 to 2.5 metres outside the reef.
#> 15:        1 to 1.5 metres, increasing to 1.5 to 2.5 metres outside the reef.
#> 16:           Around 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 17:              1 to 2 metres, increasing to 2 to 3 metres outside the reef.
#> 18:        1.5 to 2.5 metres, decreasing to 1 to 1.5 metres by early evening.
#> 19:           Around 1 metre, increasing to 1 to 1.5 metres outside the reef.
#> 20:            1 to 2 metres, increasing to 2.5 to 3 metres outside the reef.
#> 21:          2 to 3 metres, decreasing to 1 to 2 metres during the afternoon.
#> 22:             Around 1 metre, increasing to 1 to 2 metres outside the reef.
#> 23:                                                            2 to 3 metres.
#> 24:            2 to 3 metres, decreasing below 2 metres during the afternoon.
#> 25:                                                          1 to 1.5 metres.
#> 26:                                                        1.5 to 2.5 metres.
#> 27:                                                        1.5 to 2.5 metres.
#> 28:                                                        1.5 to 2.5 metres.
#> 29:                                                          2 to 2.5 metres.
#> 30:                                                            2 to 3 metres.
#> 31:                                                            2 to 3 metres.
#> 32:                                                        1.5 to 2.5 metres.
#> 33:                                                        1.5 to 2.5 metres.
#> 34:                                                        1.5 to 2.5 metres.
#> 35:                 2 metres, increasing to 2.5 to 3 metres by early evening.
#> 36:                                                            2 to 3 metres.
#> 37:                                                            2 to 3 metres.
#> 38:                                                        1.5 to 2.5 metres.
#> 39:                                                          2 to 2.5 metres.
#> 40:                                                            2 to 3 metres.
#> 41:            1 to 1.5 metres, increasing to 1 to 2 metres in the far north.
#> 42:          1 to 1.5 metres, increasing to 1 to 2 metres in the western bay.
#> 43:          Around 1 metre, increasing to 1 to 2 metres in the northern bay.
#> 44:                                                          1.5 to 2 metres.
#> 45:                                                        1.5 to 2.5 metres.
#> 46:                  1 to 1.5 metres, increasing to 1.5 to 2 metres offshore.
#> 47:                                                            2 to 3 metres.
#> 48:                                                            2 to 3 metres.
#> 49:            2 to 3 metres, decreasing to 1 to 2 metres during the evening.
#>                                                                 forecast_seas
#>                                                                              forecast_weather
#>  1:                                                                                      <NA>
#>  2:                                                                                      <NA>
#>  3:                                                                                      <NA>
#>  4:                                                                                      <NA>
#>  5:       Sunny day. The chance of a thunderstorm offshore north of Kowanyama in the evening.
#>  6: The chance of a thunderstorm offshore north of Kowanyama in the morning. Sunny afternoon.
#>  7:                                                                                    Sunny.
#>  8:                                              Partly cloudy. The chance of a thunderstorm.
#>  9:                       Partly cloudy. 50% chance of showers. The chance of a thunderstorm.
#> 10:                                               Mostly sunny. The chance of a thunderstorm.
#> 11:                              Cloudy. 95% chance of showers. The chance of a thunderstorm.
#> 12:                              Cloudy. 90% chance of showers. The chance of a thunderstorm.
#> 13:                       Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 14:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 15:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 16:                       Partly cloudy. 50% chance of showers. The chance of a thunderstorm.
#> 17:                                                     Partly cloudy. 70% chance of showers.
#> 18:                       Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 19:                       Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 20:                                                     Partly cloudy. 80% chance of showers.
#> 21:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 22:                              Cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 23:                                                     Partly cloudy. 50% chance of showers.
#> 24:                       Partly cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 25:                              Cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 26:   Partly cloudy. 70% chance of showers. The chance of a thunderstorm in the late morning.
#> 27:                              Cloudy. 90% chance of showers. The chance of a thunderstorm.
#> 28:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 29:                                                     Partly cloudy. 80% chance of showers.
#> 30:               Partly cloudy. 80% chance of showers. The chance of a thunderstorm inshore.
#> 31:                                 Cloudy. 90% chance of rain. The chance of a thunderstorm.
#> 32:                                                            Cloudy. 90% chance of showers.
#> 33:                                                     Partly cloudy. 70% chance of showers.
#> 34:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 35:                                                            Cloudy. 90% chance of showers.
#> 36:                                                     Partly cloudy. 60% chance of showers.
#> 37:                                 Cloudy. 95% chance of rain. The chance of a thunderstorm.
#> 38:                                                     Partly cloudy. 70% chance of showers.
#> 39:                                                     Partly cloudy. 60% chance of showers.
#> 40:                                                               Cloudy. 95% chance of rain.
#> 41:                                                                            Partly cloudy.
#> 42:                                                            Cloudy. 60% chance of showers.
#> 43:                                                            Cloudy. 95% chance of showers.
#> 44:                                                     Partly cloudy. 50% chance of showers.
#> 45:                                                     Partly cloudy. 50% chance of showers.
#> 46:                                                            Cloudy. 90% chance of showers.
#> 47:                                                     Partly cloudy. 60% chance of showers.
#> 48:                              Cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 49:                              Cloudy. 90% chance of showers. The chance of a thunderstorm.
#>                                                                              forecast_weather
#>                                                                                                                                                                                                                                                                                                                          forecast_winds
#>  1:                                                                                                                                                                                                                                                                                                                                <NA>
#>  2:                                                                                                                                                                                                                                                                                                                                <NA>
#>  3:                                                                                                                                                                                                                                                                                                                                <NA>
#>  4:                                                                                                                                                                                                                                                                                                                                <NA>
#>  5:                                                                                                                                                                                                                   East to southeasterly 15 to 20 knots, reaching up to 25 knots offshore north of Kowanyama in the early afternoon.
#>  6:                                                                                                                                                                                                                              East to southeasterly 15 to 20 knots tending southeast to southwesterly 10 to 15 knots in the evening.
#>  7:                                                                                                                                                                                      South to southeasterly 10 to 15 knots, reaching up to 20 knots west of Mornington Island. Becoming variable about 10 knots during the evening.
#>  8:                                                                                                                                                                                                                                                                                               East to southeasterly 15 to 20 knots.
#>  9:                                                                                                                                                                                                                                          Southeasterly 10 to 15 knots, reaching up to 20 knots offshore during the early afternoon.
#> 10:                                                                                                                                                                                                                                                 Southeasterly 10 to 15 knots becoming variable about 10 knots during the afternoon.
#> 11:                                                                                                                                                                                                                                                                                               East to southeasterly 15 to 20 knots.
#> 12:                                                                                                                                                                                                                                                           Southeasterly 15 to 20 knots decreasing to about 10 knots in the evening.
#> 13:                                                                                                                                                                                                                                                                                                            Variable about 10 knots.
#> 14:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 20 knots.
#> 15:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 25 knots.
#> 16:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 20 knots.
#> 17:                                                                                                                                                                                                                                                      Southeasterly 15 to 25 knots, reaching up to 30 knots at times during the day.
#> 18:                                                                                                                                                                                                                     Southeasterly 15 to 25 knots, reaching up to 30 knots south of Cooktown during the morning and early afternoon.
#> 19:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 20 knots.
#> 20:                                                                                                                                                                                                                                                                     Southeasterly 20 to 25 knots, reaching 25 to 30 knots at times.
#> 21:                                                                                                                                                                                                                                       Southeasterly 20 to 25 knots, reaching up to 30 knots during the morning and early afternoon.
#> 22:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 20 knots.
#> 23:                                                                                                                                                                                                                                                                     Southeasterly 20 to 25 knots, reaching 25 to 30 knots at times.
#> 24:                                                                                                                                                                                                                                                                     Southeasterly 20 to 25 knots, reaching 25 to 30 knots at times.
#> 25:                                                                                                                                                                                                                                                                                              South to southeasterly 15 to 20 knots.
#> 26:                                                                                                                                                                                                                                                                                                       Southeasterly 20 to 30 knots.
#> 27:                                                                                                                                                                                                                                                                                                       Southeasterly 20 to 30 knots.
#> 28: Southeasterly 25 to 30 knots turning southerly 20 to 30 knots during the day, as a north-south oriented trough develops east of the coastal waters area. There is a chance of this trough skirting the eastern edge of the coastal waters area, in which case the winds on its eastern side will tend northeasterly 15 to 20 knots.
#> 29:                                                                                                                                                                                                                                                                     Southeasterly 20 to 25 knots, reaching 25 to 30 knots offshore.
#> 30:                                                                                                                                                                                                                                                                     Southeasterly 20 to 25 knots, reaching 25 to 30 knots offshore.
#> 31:                                                                                                                                       Generally southeasterly 25 to 30 knots. A trough will move through the coastal waters area late Sunday or early Monday, and specific wind directions will depend on its placement and timing.
#> 32:                                                                                                                                                                                                                         Southeasterly 15 to 20 knots, reaching 20 to 30 knots in the northern bay during the afternoon and evening.
#> 33:                                                                                                                                                                                                                                                           Southeasterly 15 to 20 knots, reaching 20 to 30 knots in the northern bay
#> 34:                                                                                                                                                                                                                                                 Southeasterly 20 to 25 knots, reaching up to 30 knots in northern parts of the bay.
#> 35:                                                                                                                                                                                                                                                            Southeasterly 15 to 25 knots, reaching up 25 to 30 knots in the evening.
#> 36:                                                                                                                                                                                                                                                            Southeasterly 15 to 25 knots, reaching up 25 to 30 knots in the morning.
#> 37:                                                                                                                                                                                                                                                                                                            Easterly 25 to 30 knots.
#> 38:                                                                                                                                                                                                                                                      Southeasterly 20 to 25 knots, reaching up to 30 knots offshore in the evening.
#> 39:                                                                                                                                                                                                                                                                                                       Southeasterly 20 to 25 knots.
#> 40:                                                                                                                                                                                                                                                                                               East to southeasterly 20 to 30 knots.
#> 41:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 25 knots.
#> 42:                                                                                                                                                                                                                                                              Southeasterly 15 to 25 knots, easing to 15 to 20 knots in the evening.
#> 43:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 20 knots.
#> 44:                                                                                                                                                                                                                                                                                                       Southeasterly 15 to 25 knots.
#> 45:                                                                                                                                                                                                                                                Southeasterly 20 to 25 knots, turning easterly 15 to 20 knots in the late afternoon.
#> 46:                                                                                                                                                                                                                                                                                               East to southeasterly 20 to 25 knots.
#> 47:                                                                                                                                                                                                                                                                                                       Southeasterly 25 to 30 knots.
#> 48:                                                                                                                                                                                                                                                                                                       Southeasterly 20 to 30 knots.
#> 49:                                                                                                                                                                                                                                              South to southeasterly 20 to 30 knots decreasing to 15 to 20 knots during the evening.
#>                                                                                                                                                                                                                                                                                                                          forecast_winds
#>                                                                                                forecast_swell1
#>  1:                                                                                                       <NA>
#>  2:                                                                                                       <NA>
#>  3:                                                                                                       <NA>
#>  4:                                                                                                       <NA>
#>  5:                                                                                                       <NA>
#>  6:                                                                                          Below 0.5 metres.
#>  7:                                                                                          Below 0.5 metres.
#>  8:                                                                                                       <NA>
#>  9:                                                                                          Below 0.5 metres.
#> 10:                                                                                          Below 0.5 metres.
#> 11:                       East to southeasterly below 1 metre, increasing to 1.5 to 2 metres outside the reef.
#> 12:                         East to southeasterly below 1 metre, increasing to 1 to 2 metres outside the reef.
#> 13:                    East to southeasterly 1 to 1.5 metres, decreasing to around 1 metre during the evening.
#> 14:                             Southeasterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 15:                             Southeasterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 16:                             Southeasterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 17:                         East to southeasterly below 1 metre, increasing to 2 to 3 metres outside the reef.
#> 18:                         East to southeasterly below 1 metre, increasing to 2 to 3 metres outside the reef.
#> 19:                             Southeasterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 20:                                 Southeasterly below 1 metre, increasing to 2 to 3 metres outside the reef.
#> 21:                         East to southeasterly below 1 metre, increasing to 2 to 3 metres outside the reef.
#> 22:                            Southeasterly around 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 23:                                                               Easterly 1.5 to 2.5 metres outside the reef.
#> 24:                     East to northeasterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 25:                                  Easterly below 1 metre, increasing to 1.5 to 2.5 metres outside the reef.
#> 26:                       East to northeasterly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 27:                         East to northeasterly below 1 metre inshore, increasing to 1 to 2 metres offshore.
#> 28:                         East to northeasterly below 1 metre inshore, increasing to 1 to 2 metres offshore.
#> 29:                         East to southeasterly below 1 metre inshore, increasing to 1 to 2 metres offshore.
#> 30:                                Easterly 1 to 1.5 metres inshore, increasing to 1.5 to 2.5 metres offshore.
#> 31: Easterly 2 to 3 metres inshore, increasing to 2 to 4 metres offshore southern waters during the afternoon.
#> 32:                                                                               Northeasterly below 1 metre.
#> 33:       Northeasterly below 1 metre, increasing to 1 to 1.5 metres in the northern bay during the afternoon.
#> 34:                                                                      North to northeasterly 1 to 2 metres.
#> 35:                             Southeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 36:                             Southeasterly 1.5 to 2 metres, tending easterly 2 to 2.5 metres around midday.
#> 37:                                                                                    Easterly 2 to 3 metres.
#> 38:                                   Southeasterly around 1 metre inshore, increasing to 1.5 metres offshore.
#> 39:                     East to southeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 40:                                                                                  Easterly 2 to 2.5 metres.
#> 41:                                                                                                       <NA>
#> 42:                                                                                                       <NA>
#> 43:                                                                                                       <NA>
#> 44:                              Southeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 45:                     East to southeasterly 1 to 1.5 metres inshore, increasing to 1.5 to 2 metres offshore.
#> 46:                                                                                  Easterly 1.5 to 2 metres.
#> 47:        East to southeasterly 1 to 1.5 metres, increasing to 1.5 to 2 metres through the Capricorn Channel.
#> 48:                         Easterly 1 to 2 metres, increasing to 2 to 3 metres through the Capricorn Channel.
#> 49:                       Easterly 1 to 2 metres, increasing to 2.5 to 4 metres through the Capricorn Channel.
#>                                                                                                forecast_swell1
#>     forecast_swell2 forecast_caution
#>  1:            <NA>             <NA>
#>  2:            <NA>             <NA>
#>  3:            <NA>             <NA>
#>  4:            <NA>             <NA>
#>  5:            <NA>             <NA>
#>  6:            <NA>             <NA>
#>  7:            <NA>             <NA>
#>  8:            <NA>             <NA>
#>  9:            <NA>             <NA>
#> 10:            <NA>             <NA>
#> 11:            <NA>             <NA>
#> 12:            <NA>             <NA>
#> 13:            <NA>             <NA>
#> 14:            <NA>             <NA>
#> 15:            <NA>             <NA>
#> 16:            <NA>             <NA>
#> 17:            <NA>             <NA>
#> 18:            <NA>             <NA>
#> 19:            <NA>             <NA>
#> 20:            <NA>             <NA>
#> 21:            <NA>             <NA>
#> 22:            <NA>             <NA>
#> 23:            <NA>             <NA>
#> 24:            <NA>             <NA>
#> 25:            <NA>             <NA>
#> 26:            <NA>             <NA>
#> 27:            <NA>             <NA>
#> 28:            <NA>             <NA>
#> 29:            <NA>             <NA>
#> 30:            <NA>             <NA>
#> 31:            <NA>             <NA>
#> 32:            <NA>             <NA>
#> 33:            <NA>             <NA>
#> 34:            <NA>             <NA>
#> 35:            <NA>             <NA>
#> 36:            <NA>             <NA>
#> 37:            <NA>             <NA>
#> 38:            <NA>             <NA>
#> 39:            <NA>             <NA>
#> 40:            <NA>             <NA>
#> 41:            <NA>             <NA>
#> 42:            <NA>             <NA>
#> 43:            <NA>             <NA>
#> 44:            <NA>             <NA>
#> 45:            <NA>             <NA>
#> 46:            <NA>             <NA>
#> 47:            <NA>             <NA>
#> 48:            <NA>             <NA>
#> 49:            <NA>             <NA>
#>     forecast_swell2 forecast_caution
#>                                              marine_forecast
#>  1:                                                     <NA>
#>  2:                                                     <NA>
#>  3:                                                     <NA>
#>  4:                                                     <NA>
#>  5:                                                     <NA>
#>  6:                                                     <NA>
#>  7:                                                     <NA>
#>  8:                                                     <NA>
#>  9:                                                     <NA>
#> 10:                                                     <NA>
#> 11:                                                     <NA>
#> 12:                                                     <NA>
#> 13:                                                     <NA>
#> 14:                                                     <NA>
#> 15:                                                     <NA>
#> 16:                                                     <NA>
#> 17:        Strong Wind Warning for Friday for Cooktown Coast
#> 18:      Strong Wind Warning for Saturday for Cooktown Coast
#> 19:                                                     <NA>
#> 20:          Strong Wind Warning for Friday for Cairns Coast
#> 21:        Strong Wind Warning for Saturday for Cairns Coast
#> 22:                                                     <NA>
#> 23:      Strong Wind Warning for Friday for Townsville Coast
#> 24:    Strong Wind Warning for Saturday for Townsville Coast
#> 25:                                                     <NA>
#> 26:          Strong Wind Warning for Friday for Mackay Coast
#> 27:        Strong Wind Warning for Saturday for Mackay Coast
#> 28:                                                     <NA>
#> 29:     Strong Wind Warning for Friday for Capricornia Coast
#> 30:   Strong Wind Warning for Saturday for Capricornia Coast
#> 31:                                                     <NA>
#> 32:            Strong Wind Warning for Friday for Hervey Bay
#> 33:          Strong Wind Warning for Saturday for Hervey Bay
#> 34:                                                     <NA>
#> 35:   Strong Wind Warning for Friday for Fraser Island Coast
#> 36: Strong Wind Warning for Saturday for Fraser Island Coast
#> 37:                                                     <NA>
#> 38: Strong Wind Warning for Friday for Sunshine Coast Waters
#> 39:                                                     <NA>
#> 40:                                                     <NA>
#> 41:                                                     <NA>
#> 42:                                                     <NA>
#> 43:                                                     <NA>
#> 44:                                                     <NA>
#> 45:                                                     <NA>
#> 46:                                                     <NA>
#> 47:                                                     <NA>
#> 48:                                                     <NA>
#> 49:                                                     <NA>
#>                                              marine_forecast
#>     tropical_system_location forecast_waves
#>  1:                       NA             NA
#>  2:                       NA             NA
#>  3:                       NA             NA
#>  4:                       NA             NA
#>  5:                       NA             NA
#>  6:                       NA             NA
#>  7:                       NA             NA
#>  8:                       NA             NA
#>  9:                       NA             NA
#> 10:                       NA             NA
#> 11:                       NA             NA
#> 12:                       NA             NA
#> 13:                       NA             NA
#> 14:                       NA             NA
#> 15:                       NA             NA
#> 16:                       NA             NA
#> 17:                       NA             NA
#> 18:                       NA             NA
#> 19:                       NA             NA
#> 20:                       NA             NA
#> 21:                       NA             NA
#> 22:                       NA             NA
#> 23:                       NA             NA
#> 24:                       NA             NA
#> 25:                       NA             NA
#> 26:                       NA             NA
#> 27:                       NA             NA
#> 28:                       NA             NA
#> 29:                       NA             NA
#> 30:                       NA             NA
#> 31:                       NA             NA
#> 32:                       NA             NA
#> 33:                       NA             NA
#> 34:                       NA             NA
#> 35:                       NA             NA
#> 36:                       NA             NA
#> 37:                       NA             NA
#> 38:                       NA             NA
#> 39:                       NA             NA
#> 40:                       NA             NA
#> 41:                       NA             NA
#> 42:                       NA             NA
#> 43:                       NA             NA
#> 44:                       NA             NA
#> 45:                       NA             NA
#> 46:                       NA             NA
#> 47:                       NA             NA
#> 48:                       NA             NA
#> 49:                       NA             NA
#>     tropical_system_location forecast_waves

# }