Parse local BOM coastal waters forecast XML files

Parse local BOM daily coastal waters forecast XML file(s) and return a data frame for a specified state or territory or all Australia.

parse_coastal_forecast(state = "AUS", filepath)

Arguments

state
Required value of an Australian state or territory as full name or postal code. Fuzzy string matching via agrep is done.

filepath
A string providing the directory location of the coastal forecast file(s) to parse. See Details for more.

Value

A data.table of an Australia BOM Coastal Waters Forecast. For full details of fields and units returned see Appendix 5 in the bomrang vignette, use
vignette("bomrang", package = "bomrang") to view.

Details

Allowed state and territory postal codes, only one state per request or all using AUS.

ACT

Australian Capital Territory (will return NSW)

NSW

New South Wales

NT

Northern Territory

QLD

Queensland

SA

South Australia

TAS

Tasmania

VIC

Victoria

WA

Western Australia

AUS

Australia, returns forecast for all states, NT and ACT

The filepath argument will only accept a directory where files are located for parsing. DO NOT supply the full path including the file name. This function will only parse the requested state or all of Australia in the same fashion as get_coastal_forecast, provided that the files are all present in the directory.

References

Forecast data come from Australian Bureau of Meteorology (BOM) Weather Data Services
http://www.bom.gov.au/catalogue/data-feeds.shtml

Location data and other metadata come from the BOM anonymous FTP server with spatial data
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/, specifically the DBF file portion of a shapefile,
ftp://ftp.bom.gov.au/anon/home/adfd/spatial/IDM00003.dbf

See also

Author

Dean Marchiori, deanmarchiori@gmail.com and Paul Melloy paul@melloy.com.au

Examples

# \donttest{
# parse the coastal forecast for Queensland

#download to tempfile() using basename() to keep original name
download.file(url = "ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml",
              destfile = file.path(tempdir(),
              basename("ftp://ftp.bom.gov.au/anon/gen/fwo/IDQ11290.xml")),
              mode = "wb")

coastal_forecast <- parse_coastal_forecast(state = "QLD",
                                           filepath = tempdir())
coastal_forecast

#>     index product_id    type state_code
#>  1:  <NA>   IDQ11290    <NA>        QLD
#>  2:  <NA>   IDQ11290    <NA>        QLD
#>  3:  <NA>   IDQ11290    <NA>        QLD
#>  4:  <NA>   IDQ11290    <NA>        QLD
#>  5:     0   IDQ11290    <NA>        QLD
#>  6:     1   IDQ11290    <NA>        QLD
#>  7:     2   IDQ11290    <NA>        QLD
#>  8:     0   IDQ11290    <NA>        QLD
#>  9:     1   IDQ11290    <NA>        QLD
#> 10:     2   IDQ11290    <NA>        QLD
#> 11:     0   IDQ11290 Coastal        QLD
#> 12:     1   IDQ11290 Coastal        QLD
#> 13:     2   IDQ11290 Coastal        QLD
#> 14:     0   IDQ11290    <NA>        QLD
#> 15:     1   IDQ11290    <NA>        QLD
#> 16:     2   IDQ11290    <NA>        QLD
#> 17:     0   IDQ11290    <NA>        QLD
#> 18:     1   IDQ11290    <NA>        QLD
#> 19:     2   IDQ11290    <NA>        QLD
#> 20:     0   IDQ11290    <NA>        QLD
#> 21:     1   IDQ11290    <NA>        QLD
#> 22:     2   IDQ11290    <NA>        QLD
#> 23:     0   IDQ11290    <NA>        QLD
#> 24:     1   IDQ11290    <NA>        QLD
#> 25:     2   IDQ11290    <NA>        QLD
#> 26:     0   IDQ11290    <NA>        QLD
#> 27:     1   IDQ11290    <NA>        QLD
#> 28:     2   IDQ11290    <NA>        QLD
#> 29:     0   IDQ11290    <NA>        QLD
#> 30:     1   IDQ11290    <NA>        QLD
#> 31:     2   IDQ11290    <NA>        QLD
#> 32:     0   IDQ11290   Local        QLD
#> 33:     1   IDQ11290   Local        QLD
#> 34:     2   IDQ11290   Local        QLD
#> 35:     0   IDQ11290    <NA>        QLD
#> 36:     1   IDQ11290    <NA>        QLD
#> 37:     2   IDQ11290    <NA>        QLD
#> 38:     0   IDQ11290    <NA>        QLD
#> 39:     1   IDQ11290    <NA>        QLD
#> 40:     2   IDQ11290    <NA>        QLD
#> 41:     0   IDQ11290   Local        QLD
#> 42:     1   IDQ11290   Local        QLD
#> 43:     2   IDQ11290   Local        QLD
#> 44:     0   IDQ11290    <NA>        QLD
#> 45:     1   IDQ11290    <NA>        QLD
#> 46:     2   IDQ11290    <NA>        QLD
#> 47:     0   IDQ11290    <NA>        QLD
#> 48:     1   IDQ11290    <NA>        QLD
#> 49:     2   IDQ11290    <NA>        QLD
#>     index product_id    type state_code
#>                                                          dist_name pt_1_name
#>  1:                                                     Queensland      <NA>
#>  2:                                 Queensland Gulf of Carpentaria      <NA>
#>  3:                                               North Queensland      <NA>
#>  4:                                               South Queensland      <NA>
#>  5: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  6: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  7: South East Gulf of Carpentaria: QLD-NT Border to Cape Keerweer      <NA>
#>  8:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#>  9:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 10:   North East Gulf of Carpentaria: Cape Keerweer to Crab Island      <NA>
#> 11:                                                  Torres Strait      <NA>
#> 12:                                                  Torres Strait      <NA>
#> 13:                                                  Torres Strait      <NA>
#> 14:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 15:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 16:                  Peninsula Coast: Sharp Point to Cape Melville      <NA>
#> 17:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 18:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 19:              Cooktown Coast: Cape Melville to Cape Tribulation      <NA>
#> 20:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 21:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 22:                     Cairns Coast: Cape Tribulation to Cardwell      <NA>
#> 23:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 24:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 25:                            Townsville Coast: Cardwell to Bowen      <NA>
#> 26:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 27:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 28:                             Mackay Coast: Bowen to St Lawrence      <NA>
#> 29:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 30:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 31:                Capricornia Coast: St Lawrence to Burnett Heads      <NA>
#> 32:                                                     Hervey Bay      <NA>
#> 33:                                                     Hervey Bay      <NA>
#> 34:                                                     Hervey Bay      <NA>
#> 35:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 36:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 37:         Fraser Island Coast: Sandy Cape to Double Island Point      <NA>
#> 38:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 39:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 40:     Sunshine Coast Waters: Double Island Point to Cape Moreton      <NA>
#> 41:                                                    Moreton Bay      <NA>
#> 42:                                                    Moreton Bay      <NA>
#> 43:                                                    Moreton Bay      <NA>
#> 44:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 45:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 46:                Gold Coast Waters: Cape Moreton to Point Danger      <NA>
#> 47:                             Great Barrier Reef Offshore Waters      <NA>
#> 48:                             Great Barrier Reef Offshore Waters      <NA>
#> 49:                             Great Barrier Reef Offshore Waters      <NA>
#>                                                          dist_name pt_1_name
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>  1:      <NA> QLD_FA001 2021-04-08 10:00:18 2021-04-08 10:00:18      10:00
#>  2:      <NA> QLD_FA002 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#>  3:      <NA> QLD_FA003 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#>  4:      <NA> QLD_FA004 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#>  5:      <NA> QLD_MW001 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#>  6:      <NA> QLD_MW001 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#>  7:      <NA> QLD_MW001 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#>  8:      <NA> QLD_MW002 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#>  9:      <NA> QLD_MW002 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 10:      <NA> QLD_MW002 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 11:      <NA> QLD_MW003 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 12:      <NA> QLD_MW003 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 13:      <NA> QLD_MW003 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 14:      <NA> QLD_MW004 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 15:      <NA> QLD_MW004 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 16:      <NA> QLD_MW004 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 17:      <NA> QLD_MW005 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 18:      <NA> QLD_MW005 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 19:      <NA> QLD_MW005 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 20:      <NA> QLD_MW006 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 21:      <NA> QLD_MW006 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 22:      <NA> QLD_MW006 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 23:      <NA> QLD_MW007 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 24:      <NA> QLD_MW007 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 25:      <NA> QLD_MW007 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 26:      <NA> QLD_MW008 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 27:      <NA> QLD_MW008 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 28:      <NA> QLD_MW008 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 29:      <NA> QLD_MW009 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 30:      <NA> QLD_MW009 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 31:      <NA> QLD_MW009 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 32:      <NA> QLD_MW010 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 33:      <NA> QLD_MW010 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 34:      <NA> QLD_MW010 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 35:      <NA> QLD_MW011 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 36:      <NA> QLD_MW011 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 37:      <NA> QLD_MW011 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 38:      <NA> QLD_MW012 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 39:      <NA> QLD_MW012 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 40:      <NA> QLD_MW012 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 41:      <NA> QLD_MW013 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 42:      <NA> QLD_MW013 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 43:      <NA> QLD_MW013 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 44:      <NA> QLD_MW014 2021-04-08 10:00:00 2021-04-08 10:00:00      10:00
#> 45:      <NA> QLD_MW014 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 46:      <NA> QLD_MW014 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#> 47:      <NA> QLD_MW015 2021-04-08 05:00:00 2021-04-08 05:00:00      10:00
#> 48:      <NA> QLD_MW015 2021-04-09 00:00:00 2021-04-09 00:00:00      10:00
#> 49:      <NA> QLD_MW015 2021-04-10 00:00:00 2021-04-10 00:00:00      10:00
#>     pt_2_name       aac    start_time_local      end_time_local utc_offset
#>          start_time_utc        end_time_utc
#>  1: 2021-04-08 10:00:18 2021-04-08 10:00:18
#>  2: 2021-04-08 05:00:00 2021-04-08 05:00:00
#>  3: 2021-04-08 05:00:00 2021-04-08 05:00:00
#>  4: 2021-04-08 05:00:00 2021-04-08 05:00:00
#>  5: 2021-04-08 05:00:00 2021-04-08 05:00:00
#>  6: 2021-04-09 00:00:00 2021-04-09 00:00:00
#>  7: 2021-04-10 00:00:00 2021-04-10 00:00:00
#>  8: 2021-04-08 05:00:00 2021-04-08 05:00:00
#>  9: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 10: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 11: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 12: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 13: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 14: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 15: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 16: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 17: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 18: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 19: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 20: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 21: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 22: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 23: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 24: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 25: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 26: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 27: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 28: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 29: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 30: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 31: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 32: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 33: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 34: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 35: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 36: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 37: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 38: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 39: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 40: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 41: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 42: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 43: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 44: 2021-04-08 10:00:00 2021-04-08 10:00:00
#> 45: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 46: 2021-04-10 00:00:00 2021-04-10 00:00:00
#> 47: 2021-04-08 05:00:00 2021-04-08 05:00:00
#> 48: 2021-04-09 00:00:00 2021-04-09 00:00:00
#> 49: 2021-04-10 00:00:00 2021-04-10 00:00:00
#>          start_time_utc        end_time_utc
#>                                                            forecast_seas
#>  1:                                                                 <NA>
#>  2:                                                                 <NA>
#>  3:                                                                 <NA>
#>  4:                                                                 <NA>
#>  5:                                                       Below 1 metre.
#>  6:                                                       Below 1 metre.
#>  7:                                                       Below 1 metre.
#>  8:                                                    Below 0.5 metres.
#>  9:                                                    Below 0.5 metres.
#> 10:                                                       Below 1 metre.
#> 11:                                                       Below 1 metre.
#> 12:                                                    Below 0.5 metres.
#> 13:                                                    Below 0.5 metres.
#> 14:                                                       Below 1 metre.
#> 15:                                                       Below 1 metre.
#> 16:                                                    Below 0.5 metres.
#> 17:                                                       Below 1 metre.
#> 18:                                                       Below 1 metre.
#> 19:                                                    Below 0.5 metres.
#> 20:                                                       Below 1 metre.
#> 21:                                                       Below 1 metre.
#> 22:                                                    Below 0.5 metres.
#> 23:                                                       Below 1 metre.
#> 24:                                                       Below 1 metre.
#> 25:                                                    Below 0.5 metres.
#> 26:                                                       Below 1 metre.
#> 27:                                                    Below 0.5 metres.
#> 28:                                                       Below 1 metre.
#> 29:                                                      Around 1 metre.
#> 30:                                                       Below 1 metre.
#> 31:                                                       Below 1 metre.
#> 32:                                                      Around 1 metre.
#> 33:                                                       Below 1 metre.
#> 34:                                                       Below 1 metre.
#> 35:                1 to 1.5 metres, decreasing to 1 metre around midday.
#> 36:                                                       Below 1 metre.
#> 37:                                                       Below 1 metre.
#> 38: 1.5 to 2 metres, decreasing to 1 to 1.5 metres during the afternoon.
#> 39:                                                      Around 1 metre.
#> 40:                                                      Around 1 metre.
#> 41:                                                      Around 1 metre.
#> 42:                                                       Below 1 metre.
#> 43:                                                       Below 1 metre.
#> 44:          1.5 to 2 metres, decreasing to 1.5 metres by early evening.
#> 45:              Around 1 metre, decreasing below 1 metre around midday.
#> 46:                                                      Around 1 metre.
#> 47:                                                      Around 1 metre.
#> 48:                                                       Below 1 metre.
#> 49:                                                       Below 1 metre.
#>                                                            forecast_seas
#>                                                                                  forecast_weather
#>  1:                                                                                          <NA>
#>  2:                                                                                          <NA>
#>  3:                                                                                          <NA>
#>  4:                                                                                          <NA>
#>  5: Mostly sunny morning. The chance of a thunderstorm inshore north of Kowanyama this afternoon.
#>  6:       Mostly sunny. The chance of a thunderstorm inshore north of Kowanyama in the afternoon.
#>  7:                                                   Mostly sunny. The chance of a thunderstorm.
#>  8:                           Partly cloudy. 50% chance of showers. The chance of a thunderstorm.
#>  9:                           Partly cloudy. 50% chance of showers. The chance of a thunderstorm.
#> 10:                                                  Partly cloudy. The chance of a thunderstorm.
#> 11:                           Partly cloudy. 80% chance of showers. The chance of a thunderstorm.
#> 12:                           Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#> 13:                                                   Mostly sunny. The chance of a thunderstorm.
#> 14:                           Partly cloudy. 70% chance of showers. The chance of a thunderstorm.
#> 15:                           Partly cloudy. 60% chance of showers. The chance of a thunderstorm.
#> 16:                                                  Partly cloudy. The chance of a thunderstorm.
#> 17:                 Partly cloudy. The chance of a thunderstorm during the morning and afternoon.
#> 18:             Partly cloudy. The chance of a thunderstorm inshore in the morning and afternoon.
#> 19:                                The chance of a thunderstorm in the morning. Mostly sunny day.
#> 20:                                  Partly cloudy. The chance of a thunderstorm north of Cairns.
#> 21:                                                                                 Mostly sunny.
#> 22:                                                                                 Mostly sunny.
#> 23:                                                                                 Mostly sunny.
#> 24:                                                                                 Mostly sunny.
#> 25:                                                                                        Sunny.
#> 26:                                                                                 Mostly sunny.
#> 27:                                                                                        Sunny.
#> 28:                                                                                        Sunny.
#> 29:                                                                                 Mostly sunny.
#> 30:                                                                                        Sunny.
#> 31:                                                                                        Sunny.
#> 32:                                                                                        Sunny.
#> 33:                                                                                        Sunny.
#> 34:                                                                                        Sunny.
#> 35:                                                                                Partly cloudy.
#> 36:                                                                                        Sunny.
#> 37:                                                                                        Sunny.
#> 38:                                                                                Partly cloudy.
#> 39:                                                                                        Sunny.
#> 40:                                                                                        Sunny.
#> 41:                                                                                 Mostly sunny.
#> 42:                                                                                        Sunny.
#> 43:                                                                                        Sunny.
#> 44:                                                                                Partly cloudy.
#> 45:                                                                                        Sunny.
#> 46:                                                                                        Sunny.
#> 47:                                                                                 Mostly sunny.
#> 48:                                                                                        Sunny.
#> 49:                                                                                        Sunny.
#>                                                                                  forecast_weather
#>                                                                                                                                                                     forecast_winds
#>  1:                                                                                                                                                                           <NA>
#>  2:                                                                                                                                                                           <NA>
#>  3:                                                                                                                                                                           <NA>
#>  4:                                                                                                                                                                           <NA>
#>  5:                                                                                        West to southwesterly 10 to 15 knots becoming variable about 10 knots in the afternoon.
#>  6:                      West to northwesterly 10 to 15 knots becoming variable about 10 knots in the morning. Winds tending north to northwesterly 10 to 15 knots in the evening.
#>  7:                                                                                                                                                       Variable about 10 knots.
#>  8:                                                                                                        Variable about 10 knots tending west to southwesterly in the afternoon.
#>  9:                                                                                                                                                       Variable below 10 knots.
#> 10:                                                                                                                                                       Variable about 10 knots.
#> 11:                                                                                           Variable about 10 knots becoming west to southwesterly in the afternoon and evening.
#> 12:                                                                                                                                          West to southwesterly about 10 knots.
#> 13:                                                                                                                                                       Variable below 10 knots.
#> 14:                                                                                                                                                       Variable about 10 knots.
#> 15:                                                                                                                                                       Variable about 10 knots.
#> 16:                                                                                                                                          East to southeasterly about 10 knots.
#> 17:                                                                                                                                                  Southeasterly 10 to 15 knots.
#> 18:                                                                                                                                                  Southeasterly 10 to 15 knots.
#> 19:                                                                                                                                          East to southeasterly below 10 knots.
#> 20:                                                                                                    Southeasterly 10 to 15 knots decreasing to about 10 knots in the afternoon.
#> 21:                                                                                Southeasterly about 10 knots. Winds south of Cape Grafton turning northeasterly in the evening.
#> 22:                                                                                                                                                       Variable below 10 knots.
#> 23:                                                                                Southeasterly 10 to 15 knots, tending northeasterly inshore in the afternoon and early evening.
#> 24:                                                                             East to southeasterly about 10 knots becoming north to northeasterly in the afternoon and evening.
#> 25:                                                                                            Variable about 10 knots, turning northeasterly inshore in the afternoon seabreezes.
#> 26:                                                                           South to southeasterly 10 to 15 knots tending east to southeasterly about 10 knots in the afternoon.
#> 27:                                                                                        Variable below 10 knots becoming northerly about 10 knots in the afternoon and evening.
#> 28:                                                                                                                     Variable about 10 knots with inshore afternoon seabreezes.
#> 29:                                                Southerly 10 to 15 knots turning east to southeasterly about 10 knots in the afternoon, with northeasterly sea breezes inshore.
#> 30:                                                                                         Variable below 10 knots tending northerly about 10 knots in the afternoon and evening.
#> 31:                                                                                                                     Variable about 10 knots with inshore afternoon seabreezes.
#> 32:                                                                                Southerly 10 to 15 knots turning east to southeasterly about 10 knots in the middle of the day.
#> 33:                     Southeasterly below 10 knots becoming south to southwesterly before dawn then becoming north to northwesterly about 10 knots in the afternoon and evening.
#> 34:                                                                                                                                                       Variable about 10 knots.
#> 35:                     South to southwesterly 15 to 20 knots, easing to 10 to 15 knots in the middle of the day then turning south to southeasterly in the afternoon and evening.
#> 36:                          Southerly 10 to 15 knots becoming variable about 10 knots in the middle of the day then tending north to northwesterly 10 to 15 knots in the evening.
#> 37:                                                                                      West to northwesterly 10 to 15 knots tending north to northwesterly during the afternoon.
#> 38: South to southwesterly 15 to 20 knots, reaching up to 25 knots offshore in the morning. Winds turning south to southeasterly 10 to 15 knots in the late afternoon and evening.
#> 39:                         South to southwesterly 10 to 15 knots becoming variable about 10 knots in the middle of the day then becoming northerly 10 to 15 knots in the evening.
#> 40:                                                                                                                                         North to northwesterly 10 to 15 knots.
#> 41:                            Southwesterly 10 to 15 knots, reaching up to 20 knots in the eastern bay early this morning. Winds tending south to southeasterly in the afternoon.
#> 42:                                                                                        South to southwesterly about 10 knots becoming north to northeasterly in the afternoon.
#> 43:                                                                                                                                          West to northwesterly 10 to 15 knots.
#> 44:                                        South to southwesterly 20 to 25 knots. Winds decreasing to 15 to 20 knots in the afternoon, then to 10 to 15 knots in the late evening.
#> 45:                       South to southwesterly 10 to 15 knots becoming variable about 10 knots in the morning then tending north to northwesterly 10 to 15 knots in the evening.
#> 46:                                                                                    West to northwesterly 10 to 15 knots, tending north to northeasterly in the late afternoon.
#> 47:                                                                                                      Southeasterly 10 to 15 knots decreasing to about 10 knots in the evening.
#> 48:                                                                                                   Southeasterly about 10 knots becoming north to northeasterly in the evening.
#> 49:                                                                                                                                                       Variable about 10 knots.
#>                                                                                                                                                                     forecast_winds
#>                                                                                                   forecast_swell1
#>  1:                                                                                                          <NA>
#>  2:                                                                                                          <NA>
#>  3:                                                                                                          <NA>
#>  4:                                                                                                          <NA>
#>  5:                                                                                                          <NA>
#>  6:                                                                                                          <NA>
#>  7:                                                                                                          <NA>
#>  8:                                                                                                          <NA>
#>  9:                                                                                                          <NA>
#> 10:                                                                                                          <NA>
#> 11:                                                                          East to southeasterly below 1 metre.
#> 12:                                                                          East to southeasterly below 1 metre.
#> 13:                                                                          East to southeasterly below 1 metre.
#> 14:                                                                          East to southeasterly below 1 metre.
#> 15:                                                                          East to southeasterly below 1 metre.
#> 16:                                                                          East to southeasterly below 1 metre.
#> 17:                                                                                       Easterly below 1 metre.
#> 18:                                                                                       Easterly below 1 metre.
#> 19:                                                                          East to southeasterly below 1 metre.
#> 20:                                                                          East to southeasterly below 1 metre.
#> 21:                                                                                       Easterly below 1 metre.
#> 22:                                                                                       Easterly below 1 metre.
#> 23:                                                                          East to northeasterly below 1 metre.
#> 24:                                                                          East to northeasterly below 1 metre.
#> 25:                                                                          East to northeasterly below 1 metre.
#> 26:                                                                                  Southeasterly below 1 metre.
#> 27:                                                                                  Southeasterly below 1 metre.
#> 28:                                                                          East to northeasterly below 1 metre.
#> 29:                            East to southeasterly below 1 metre inshore, increasing to 1 to 2 metres offshore.
#> 30:                          East to southeasterly below 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 31: East to southeasterly 1 to 1.5 metres offshore, decreasing to around 1 metre during the afternoon or evening.
#> 32:                                                     North to northeasterly below 1 metre in the northern bay.
#> 33:                                                                                             Below 0.5 metres.
#> 34:                                                                                             Below 0.5 metres.
#> 35:                                        Easterly 1.5 to 2 metres, decreasing to 1 to 1.5 metres around midday.
#> 36:                                    Easterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 37:                  Southeasterly 1 to 1.5 metres, decreasing to around 1 metre during the afternoon or evening.
#> 38:                                       Easterly 1.5 to 2 metres, decreasing to 1 to 1.5 metres in the evening.
#> 39:                                    Easterly 1 to 1.5 metres, decreasing to around 1 metre during the morning.
#> 40:                             Southeasterly 1 to 1.5 metres, decreasing to around 1 metre during the afternoon.
#> 41:                                                                                                          <NA>
#> 42:                                                                                                          <NA>
#> 43:                                                                                                          <NA>
#> 44:                                                                                   Easterly 1.5 to 2.5 metres.
#> 45:                                    Easterly 1.5 to 2 metres, decreasing to around 1 metre during the morning.
#> 46:                               Southerly 1.5 metres, tending southeasterly 1 to 1.5 metres during the morning.
#> 47:             East to southeasterly below 1 metre, increasing to 1 to 1.5 metres through the Capricorn Channel.
#> 48:             East to southeasterly below 1 metre, increasing to 1 to 1.5 metres through the Capricorn Channel.
#> 49:             East to southeasterly below 1 metre, increasing to 1 to 1.5 metres through the Capricorn Channel.
#>                                                                                                   forecast_swell1
#>                                                                                            forecast_swell2
#>  1:                                                                                                   <NA>
#>  2:                                                                                                   <NA>
#>  3:                                                                                                   <NA>
#>  4:                                                                                                   <NA>
#>  5:                                                                                                   <NA>
#>  6:                                                                                                   <NA>
#>  7:                                                                                                   <NA>
#>  8:                                                                                                   <NA>
#>  9:                                                                                                   <NA>
#> 10:                                                                                                   <NA>
#> 11:                                                                                                   <NA>
#> 12:                                                                                                   <NA>
#> 13:                                                                                                   <NA>
#> 14:                                                                                                   <NA>
#> 15:                                                                                                   <NA>
#> 16:                                                                                                   <NA>
#> 17:                                                                                                   <NA>
#> 18:                                                                                                   <NA>
#> 19:                                                                                                   <NA>
#> 20:                                                                                                   <NA>
#> 21:                                                                                                   <NA>
#> 22:                                                                                                   <NA>
#> 23:                                                                                                   <NA>
#> 24:                                                                                                   <NA>
#> 25:                                                                                                   <NA>
#> 26:                                                                                                   <NA>
#> 27:                                                                                                   <NA>
#> 28:                                                                                                   <NA>
#> 29: Southerly below 1 metre offshore east of Sandy Cape, increasing to 1 to 1.5 metres during the morning.
#> 30:                                                 Southerly 1 to 1.5 metres offshore east of Sandy Cape.
#> 31:                                                                                                   <NA>
#> 32:                                                                                                   <NA>
#> 33:                                                                                                   <NA>
#> 34:                                                                                                   <NA>
#> 35:                    Southerly below 1 metre, increasing to 1 to 1.5 metres offshore during the morning.
#> 36:                                                                South to southeasterly 1 to 1.5 metres.
#> 37:                                                                                                   <NA>
#> 38:                       Southerly below 1 metre offshore, increasing to 1.5 metres later in the evening.
#> 39:                                                       South to southeasterly 1 to 1.5 metres offshore.
#> 40:                                                                                                   <NA>
#> 41:                                                                                                   <NA>
#> 42:                                                                                                   <NA>
#> 43:                                                                                                   <NA>
#> 44:                       Southeasterly below 1 metre inshore, tending southerly 1 to 1.5 metres offshore.
#> 45:                 South to southeasterly around 1 metre inshore, increasing to 1 to 1.5 metres offshore.
#> 46:                                                                                                   <NA>
#> 47:                                                                                                   <NA>
#> 48:                                                                                                   <NA>
#> 49:                                                                                                   <NA>
#>                                                                                            forecast_swell2
#>     forecast_caution marine_forecast tropical_system_location forecast_waves
#>  1:             <NA>            <NA>                       NA             NA
#>  2:             <NA>            <NA>                       NA             NA
#>  3:             <NA>            <NA>                       NA             NA
#>  4:             <NA>            <NA>                       NA             NA
#>  5:             <NA>            <NA>                       NA             NA
#>  6:             <NA>            <NA>                       NA             NA
#>  7:             <NA>            <NA>                       NA             NA
#>  8:             <NA>            <NA>                       NA             NA
#>  9:             <NA>            <NA>                       NA             NA
#> 10:             <NA>            <NA>                       NA             NA
#> 11:             <NA>            <NA>                       NA             NA
#> 12:             <NA>            <NA>                       NA             NA
#> 13:             <NA>            <NA>                       NA             NA
#> 14:             <NA>            <NA>                       NA             NA
#> 15:             <NA>            <NA>                       NA             NA
#> 16:             <NA>            <NA>                       NA             NA
#> 17:             <NA>            <NA>                       NA             NA
#> 18:             <NA>            <NA>                       NA             NA
#> 19:             <NA>            <NA>                       NA             NA
#> 20:             <NA>            <NA>                       NA             NA
#> 21:             <NA>            <NA>                       NA             NA
#> 22:             <NA>            <NA>                       NA             NA
#> 23:             <NA>            <NA>                       NA             NA
#> 24:             <NA>            <NA>                       NA             NA
#> 25:             <NA>            <NA>                       NA             NA
#> 26:             <NA>            <NA>                       NA             NA
#> 27:             <NA>            <NA>                       NA             NA
#> 28:             <NA>            <NA>                       NA             NA
#> 29:             <NA>            <NA>                       NA             NA
#> 30:             <NA>            <NA>                       NA             NA
#> 31:             <NA>            <NA>                       NA             NA
#> 32:             <NA>            <NA>                       NA             NA
#> 33:             <NA>            <NA>                       NA             NA
#> 34:             <NA>            <NA>                       NA             NA
#> 35:             <NA>            <NA>                       NA             NA
#> 36:             <NA>            <NA>                       NA             NA
#> 37:             <NA>            <NA>                       NA             NA
#> 38:             <NA>            <NA>                       NA             NA
#> 39:             <NA>            <NA>                       NA             NA
#> 40:             <NA>            <NA>                       NA             NA
#> 41:             <NA>            <NA>                       NA             NA
#> 42:             <NA>            <NA>                       NA             NA
#> 43:             <NA>            <NA>                       NA             NA
#> 44:             <NA>            <NA>                       NA             NA
#> 45:             <NA>            <NA>                       NA             NA
#> 46:             <NA>            <NA>                       NA             NA
#> 47:             <NA>            <NA>                       NA             NA
#> 48:             <NA>            <NA>                       NA             NA
#> 49:             <NA>            <NA>                       NA             NA
#>     forecast_caution marine_forecast tropical_system_location forecast_waves

# }